### 2024-06-03 #6829 @editorialbot
[PRE REVIEW]: GEOS-2024: A portable multi-physics simulation framework
**Submitting author:** @rrsettgast (<a href="http://orcid.org/0000-0002-2536-7867">Randolph Settgast</a>)
**Repository:** https://github.com/GEOS-DEV/GEOS
**Branch with paper.md** (empty if default branch): docs/JOSS
**Version:** v1.0.1
