+++ src/coreComponents/mesh/generators/BuildPods.cpp
@@ -24,6 +24,7 @@
 namespace geos::ghosting
 {
 
+// TODO Do we want 2 separated `g2l` and `l2g` structs?
 struct GlobalNumberings
 {
   std::map< NodeGlbIdx, NodeLocIdx > ng2l;
@@ -94,6 +95,22 @@ std::vector< integer > buildGhostRank( std::map< GI, LI > const & g2l,
   }
 
   return ghostRank;
+}
+
+
+MeshGraph mergeMeshGraph( MeshGraph const & owned,
+                          MeshGraph const & present,
+                          MeshGraph const & ghosts )
+{
+  MeshGraph result{ owned };
+  for( MeshGraph const & graph: { present, ghosts } )
+  {
+    result.c2f.insert( std::cbegin( graph.c2f ), std::cend( graph.c2f ) );
+    result.f2e.insert( std::cbegin( graph.f2e ), std::cend( graph.f2e ) );
+    result.e2n.insert( std::cbegin( graph.e2n ), std::cend( graph.e2n ) );
+    result.n.insert( std::cbegin( graph.n ), std::cend( graph.n ) );
+  }
+  return result;
 }
 
 EdgeMgrImpl makeFlavorlessEdgeMgrImpl( std::size_t const & numEdges,
@@ -212,20 +229,78 @@ void buildFaceMgr( GlobalNumberings const & numberings,
   }
 }
 
-
-MeshGraph mergeMeshGraph( MeshGraph const & owned,
-                          MeshGraph const & present,
-                          MeshGraph const & ghosts )
+std::vector< NodeLocIdx > resetFaceNodes( std::vector< NodeLocIdx > const & nodes,
+                                          bool const & isFlipped,
+                                          std::uint8_t const & start )
 {
-  MeshGraph result{ owned };
-  for( MeshGraph const & graph: { present, ghosts } )
+  std::vector< NodeLocIdx > result( nodes );
+  std::rotate( std::begin( result ), std::begin( result ) + start, std::end( result ) );
+  if( isFlipped )  // TODO before or after?
   {
-    result.c2f.insert( std::cbegin( graph.c2f ), std::cend( graph.c2f ) );
-    result.f2e.insert( std::cbegin( graph.f2e ), std::cend( graph.f2e ) );
-    result.e2n.insert( std::cbegin( graph.e2n ), std::cend( graph.e2n ) );
-    result.n.insert( std::cbegin( graph.n ), std::cend( graph.n ) );
+    std::reverse( std::begin( result ), std::end( result ) );
   }
   return result;
+}
+
+void buildCellBlock( GlobalNumberings const & numberings,
+                     MeshGraph const & graph,
+                     GhostRecv const & recv,
+                     GhostSend const & send,
+                     std::map< FaceLocIdx, std::vector< EdgeLocIdx > > const & f2e,
+                     std::map< FaceLocIdx, std::vector< NodeLocIdx > > const & f2n )
+{
+  // TODO MISSING cell type. Should be OK, the information is conveyed.
+  // TODO MISSING get the cell -> numNodesPerElement... from the original CellBlock
+
+  // Total number of faces available in the rank (including the ghosted edges).
+  std::size_t const numElements = std::size( numberings.cg2l );
+
+  std::map< CellLocIdx, std::vector< FaceLocIdx > > c2f;
+  std::map< CellLocIdx, std::vector< EdgeLocIdx > > c2e;
[…]
