### 2024-06-03 3c1ccc5
Add generic parameters
