+++ src/coreComponents/constitutive/fluid/multifluid/compositional/models/ConstantViscosity.hpp
@@ -33,7 +33,7 @@ namespace compositional
 class ConstantViscosityUpdate final : public FunctionBaseUpdate
 {
 public:
-  ConstantViscosityUpdate() = default;
+  explicit ConstantViscosityUpdate(real64 const constantPhaseViscosity);
 
   template< integer USD1, integer USD2 >
   GEOS_HOST_DEVICE
@@ -46,13 +46,20 @@ public:
                 real64 & viscosity,
                 arraySlice1d< real64, USD2 > const & dViscosity,
                 bool useMass ) const;
+
+private:
+real64 const m_constantPhaseViscosity;
 };
 
 class ConstantViscosity : public FunctionBase
 {
+    public:
+    static constexpr real64 defaultViscosity = 0.001;
 public:
   ConstantViscosity( string const & name,
-                     ComponentProperties const & componentProperties );
+                     ComponentProperties const & componentProperties,
+                     integer const phaseIndex,
+                             ModelParameters const & modelParameters );
 
   static string catalogName() { return ""; }
 
@@ -69,6 +76,26 @@ public:
    * @return the wrapper
    */
   KernelWrapper createKernelWrapper() const;
+
+  // Parameters for constant viscosity model
+  class Parameters : public ModelParameters
+{
+public:
+  Parameters() = default;
+~Parameters() override = default;
+
+  void registerParameters( MultiFluidBase * fluid ) override;
+  void postProcessInput( MultiFluidBase const * fluid ) override;
+
+  struct viewKeyStruct
+  {
+    static constexpr char const * constantPhaseViscosityString() { return "constantPhaseViscosity"; }
+  };
+
+  array1d< real64 > m_constantPhaseViscosity;
+};
+private:
+real64 m_constantPhaseViscosity{defaultViscosity};
 };
 
 template< integer USD1, integer USD2 >
@@ -88,7 +115,7 @@ void ConstantViscosityUpdate::compute( ComponentProperties::KernelWrapper const 
   GEOS_UNUSED_VAR( phaseComposition );
   GEOS_UNUSED_VAR( density, dDensity );
 
-  viscosity = 0.001;
+  viscosity = m_constantPhaseViscosity;
 
   LvArray::forValuesInSlice( dViscosity, setZero );
 }
