+++ src/coreComponents/constitutive/fluid/multifluid/compositional/models/CompositionalDensity.hpp
@@ -63,9 +63,12 @@ class CompositionalDensity : public FunctionBase
 {
 public:
   CompositionalDensity( string const & name,
-                        ComponentProperties const & componentProperties )
+                        ComponentProperties const & componentProperties,
+                     integer const phaseIndex,
+                             ModelParameters const & modelParameters )
     : FunctionBase( name, componentProperties )
   {
+    GEOS_UNUSED_VAR(phaseIndex, modelParameters);
     // Calculate the dimensional volume shift
     m_componentDimensionalVolumeShift.resize( componentProperties.getNumberOfComponents());
     EOS_TYPE::calculateDimensionalVolumeShift( componentProperties,

[…]
