+++ src/coreComponents/constitutive/fluid/multifluid/compositional/models/NegativeTwoPhaseFlashModel.cpp
@@ -38,10 +38,10 @@ template< typename EOS_TYPE_LIQUID, typename EOS_TYPE_VAPOUR >
 NegativeTwoPhaseFlashModel< EOS_TYPE_LIQUID, EOS_TYPE_VAPOUR >::
 NegativeTwoPhaseFlashModel( string const & name,
                             ComponentProperties const & componentProperties,
-                             ModelParameters const & modelParameters ):
+                            ModelParameters const & modelParameters ):
   FunctionBase( name, componentProperties )
 {
-    GEOS_UNUSED_VAR(modelParameters);
+  GEOS_UNUSED_VAR( modelParameters );
 }
 
 template< typename EOS_TYPE_LIQUID, typename EOS_TYPE_VAPOUR >
