### 2024-06-03 eb33b36
Fix LBC test
