### 2024-06-04 9e2fefd
Fix unused variable
