+++ src/coreComponents/schema/docs/CellElementRegion.rst
@@ -1,22 +1,22 @@
 
 
-======================== ================== ======== ==================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================== 
-Name                     Type               Default  Description                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
-======================== ================== ======== ==================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================== 
-cellBlockAttributeValues integer_array      {0}      | List of regionAttribute values for which we want to add the cells in this CellElementRegion.                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
-                                                     | I.e. { 1, 2 } selects the { 1_tetrahedra, 1_pyramid, 2_tetrahedra, 2_pyramid... } cellBlocks.                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
-                                                     | This setting cannot be used simultaneously with cellBlocks nor cellBlocksMatch.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
-cellBlocks               groupNameRef_array {}       | List of the desired cellBlocks names to contain in this CellElementRegion.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
-                                                     | The form of this attribute is of "regionAttribute_elementType", so "1_tetrahedra" select the cellBlock that contains the tetrahedric elements for which the regionAttribute is 1.                                                                                                                                                                                                                                                                                                                                                                                    
-                                                     | The element types are: tetrahedra, pyramids, wedges, hexahedra, pentagonalPrisms, hexagonalPrisms, heptagonalPrisms, octagonalPrisms, nonagonalPrisms, decagonalPrisms, hendecagonalPrisms, polyhedra.                                                                                                                                                                                                                                                                                                                                                               
-                                                     | This setting cannot be used simultaneously with cellBlockAttributeValues nor cellBlocksMatch.                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
-cellBlocksMatch          string_array       {}       | List of fnmatch pattern to match cellBlock names to add them in this CellElementRegion.                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
-                                                     | I.e. "{ * }" selects every elements, { [1-5]_* } selects the { 1_tetrahedra, 2_tetrahedra, ..., 5_tetrahedra, 1_pyramid... } cellBlocks.                                                                                                                                                                                                                                                                                                                                                                                                                             
-                                                     | This setting cannot be used simultaneously with cellBlocks nor cellBlockAttributeValues.                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
-coarseningRatio          real64             0        (no description available)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
-materialList             groupNameRef_array required List of materials present in this region                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
-meshBody                 groupNameRef                Mesh body that contains this region                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
-name                     groupName          required A name is required for any non-unique nodes                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
-======================== ================== ======== ==================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================== 
+======================== ================== ======== ====================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================== 
[…]
