### 2024-06-04 d598c0b
worked on clarifying the code as much as possible
