+++ src/coreComponents/mesh/CellElementRegion.cpp
@@ -15,6 +15,7 @@
 #include "CellElementRegion.hpp"
 #include "CellElementSubRegion.hpp"
 #include "mesh/generators/CellBlockABC.hpp"
+#include "mesh/CellElementRegionSelector.hpp"
 
 #include <fnmatch.h>
 
@@ -37,9 +38,9 @@ CellElementRegion::CellElementRegion( string const & name, Group * const parent 
   registerWrapper( viewKeyStruct::sourceCellBlockNamesString(), &m_cellBlockNames ).
     setRTTypeName( rtTypes::CustomTypes::groupNameRefArray ).
     setInputFlag( InputFlags::OPTIONAL ).
-    setDescription( GEOS_FMT( "List of the desired cellBlocks names to contain in this {}.\n"
+    setDescription( GEOS_FMT( "List of the desired cell-blocks names to contain in this {}.\n"
                               "The form of this attribute is of \"regionAttribute_elementType\", so \"1_tetrahedra\" select the "
-                              "cellBlock that contains the tetrahedric elements for which the regionAttribute is 1.\n"
+                              "cell-block that contains the tetrahedric elements for which the regionAttribute is 1.\n"
                               "The element types are: {}.\n"
                               "This setting cannot be used simultaneously with {} nor {}.",
                               catalogName(), stringutilities::join( elementNames, ", " ),
@@ -57,7 +58,7 @@ CellElementRegion::CellElementRegion( string const & name, Group * const parent 
 
   registerWrapper( viewKeyStruct::cellBlockMatchPatternsString(), &m_cellBlockMatchPatterns ).
     setInputFlag( InputFlags::OPTIONAL ).
-    setDescription( GEOS_FMT( "List of fnmatch pattern to match cellBlock names to add them in this {}.\n"
+    setDescription( GEOS_FMT( "List of fnmatch pattern to match cell-block names to add them in this {}.\n"
                               "I.e. \"{{ * }}\" selects every elements, {{ [1-5]_* }} selects the "
                               "{{ 1_tetrahedra, 2_tetrahedra, ..., 5_tetrahedra, 1_pyramid... }} cellBlocks.\n"
                               "This setting cannot be used simultaneously with {} nor {}.",
@@ -77,159 +78,29 @@ CellElementRegion::~CellElementRegion()
 void CellElementRegion::registerSubRegion( CellBlockABC const & cellBlock )
 {
   Group & elementSubRegions = this->getGroup( viewKeyStruct::elementSubRegions() );
-  // For now, subRegion name must be the same as the cellBlock (so we can match them and reference them in errors).
+  // For now, subRegion name must be the same as the cell-block (so we can match them and reference them in errors).
   CellElementSubRegion & subRegion =
     elementSubRegions.registerGroup< CellElementSubRegion >( cellBlock.getName() );
 
   subRegion.copyFromCellBlock( cellBlock );
 }
 
-string CellElementRegion::getCellBlockAttributeValue( string_view cellBlockName )
+
+void CellElementRegion::generateMesh( CellElementRegionSelector & cellBlockSelector )
 {
-  return string( stringutilities::removeStringAndFollowingContent( cellBlockName, "_" ) );
-}
-
-std::set< string > getAvailableAttributeValues( std::set< string > const & cellBlocksNames )
-{
-  std::set< string > attributeValues;
-  for( auto const & cellBlockName : cellBlocksNames )
-  {
-    attributeValues.insert( CellElementRegion::getCellBlockAttributeValue( cellBlockName ) );
-  }
-  return attributeValues;
-}
-
-std::set< string >
-CellElementRegion::computeSelectedCellBlocks( std::set< string > const & cellBlocksNames ) const
-{
-  std::set< string > selectedCellBlocks;
-
[…]
