### 2024-06-04 f83b297
Fix bug when we take minTime < 0 in wave solvers (#3158)
* Fix bug when we take minTime < 0

* Remove not necessary test when we compute the source

* Fix unused variable
