### 2024-06-04 #6829 @rrsettgast
Hello @kthyng,
Can you help me with some questions I have.
1. the error is because I attempted to specify a dual affiliation for some of the authors. For example:
```
  - name: Rasim Hasanzade
    affiliation: 3,4
```
When GitHub actions generates the pdf from the PR in my repository it works fine. Is there a way to do this in the official generation?

2. Some of the references do not have a DOI. Is this OK?
3. The word count is reported as `1489`. This includes every word in the markdown file, but the number of words in the text is closer to `1000`. Is this ok?
