### 2024-06-04 b051ed1
by default, SourceFluxStat targets al SourceFlux + adapted test
