+++ src/coreComponents/schema/docs/SourceFluxStatistics.rst
@@ -4,7 +4,7 @@
 Name           Type               Default  Description                                                                                                                                                                           
 ============== ================== ======== ===================================================================================================================================================================================== 
 flowSolverName groupNameRef       required Name of the flow solver                                                                                                                                                               
-fluxNames      groupNameRef_array required Name(s) array of the SourceFlux(s) for which we want the statistics. Use "*" to target all SourceFlux.                                                                                
+fluxNames      groupNameRef_array {*}      Name(s) array of the SourceFlux(s) for which we want the statistics. Use "*" to target all SourceFlux.                                                                                
 logLevel       integer            0        | Log level                                                                                                                                                                             
                                            | - Log Level 1 outputs the sum of all SourceFlux(s) produced rate & mass,                                                                                                              
                                            | - Log Level 2 details values for each SourceFlux,                                                                                                                                     
