+++ src/coreComponents/common/Units.cpp
@@ -37,6 +37,11 @@ TimeFormatInfo::TimeFormatInfo( double const totalSeconds, int const years, int 
 string TimeFormatInfo::toString() const
 {
   std::ostringstream oss;
+  constexpr double negativeTime = -1.0 + LvArray::NumericLimits< double >::epsilon;
+  if( m_totalSeconds < negativeTime )
+  {
+    oss << "-(";
+  }
   if( m_years != 0 )
   {
     oss << m_years << "y, " << m_days << "d, ";
@@ -45,8 +50,12 @@ string TimeFormatInfo::toString() const
   {
     oss << m_days << "d, ";
   }
-  oss << GEOS_FMT( "{:0>2}h{:0>2}m{:0>2}s ({} s)",
-                   m_hours, m_minutes, m_seconds, m_totalSeconds );
+  oss << GEOS_FMT( "{:0>2}h{:0>2}m{:0>2}s", m_hours, m_minutes, m_seconds );
+  if( m_totalSeconds < negativeTime )
+  {
+    oss << ")";
+  }
+  oss << GEOS_FMT( " ({} s)", m_totalSeconds );
   return oss.str();
 }
 
@@ -77,11 +86,16 @@ template TimeFormatInfo TimeFormatInfo::fromDuration< SystemClock::duration >( S
 
 TimeFormatInfo TimeFormatInfo::fromSeconds( double const seconds )
 {
-  int totalYears = int(   seconds / YearSeconds );
-  int daysOut = int(    ( seconds - totalYears * YearSeconds ) / DaySeconds );
-  int hoursOut = int(   ( seconds - totalYears * YearSeconds - daysOut * DaySeconds ) / HourSeconds );
-  int minutesOut = int( ( seconds - totalYears * YearSeconds - daysOut * DaySeconds - hoursOut * HourSeconds ) / MinuteSeconds );
-  int secondsOut = int(   seconds - totalYears * YearSeconds - daysOut * DaySeconds - hoursOut * HourSeconds - minutesOut * MinuteSeconds );
+  double remainingSeconds = seconds < 0.0 ? -seconds : seconds;
+  int const totalYears = int( remainingSeconds / YearSeconds );
+  remainingSeconds -= totalYears * YearSeconds;
+  int const daysOut = int( remainingSeconds / DaySeconds );
+  remainingSeconds -= daysOut * DaySeconds;
+  int const hoursOut = int( remainingSeconds / HourSeconds );
+  remainingSeconds -= hoursOut * HourSeconds;
+  int const minutesOut = int( remainingSeconds / MinuteSeconds );
+  remainingSeconds -= minutesOut * MinuteSeconds;
+  int const secondsOut = int( remainingSeconds );
 
   return TimeFormatInfo( seconds, totalYears, daysOut, hoursOut, minutesOut, secondsOut );
 }
