### 2024-06-04 ddebea4
Change component properties
