+++ src/coreComponents/constitutive/fluid/multifluid/compositional/CompositionalMultiphaseFluid.hpp
@@ -122,14 +122,8 @@ private:
   std::unique_ptr< PHASE2 > m_phase2;
   std::unique_ptr< PHASE3 > m_phase3;
 
+  // Standard EOS component input
   std::unique_ptr< compositional::ComponentProperties > m_componentProperties{};
-
-  // standard EOS component input
-  array1d< real64 > m_componentCriticalPressure;
-  array1d< real64 > m_componentCriticalTemperature;
-  array1d< real64 > m_componentAcentricFactor;
-  array1d< real64 > m_componentVolumeShift;
-  array2d< real64 > m_componentBinaryCoeff;
 
   // Extra parameters for specific to this model
   std::unique_ptr< compositional::ModelParameters > m_parameters{};
