+++ src/coreComponents/constitutive/fluid/multifluid/compositional/models/ConstantViscosity.hpp
@@ -85,7 +85,7 @@ public:
     ~Parameters() override = default;
 
     void registerParameters( MultiFluidBase * fluid ) override;
-    void postProcessInput( MultiFluidBase const * fluid ) override;
+    void postProcessInput( MultiFluidBase const * fluid, ComponentProperties const & componentProperties ) override;
 
     struct viewKeyStruct
     {

[…]
