### 2024-06-04 b2f3e04
Log refactor - Table logging features (#2984)
