### 2024-06-04 776713e
Log refactor - Table logging features (#2984)
