+++ src/coreComponents/physicsSolvers/fluidFlow/IsothermalCompositionalMultiphaseBaseKernels.hpp
@@ -1156,7 +1156,9 @@ public:
   /**
    * @brief Create a new kernel instance
    * @param[in] maxRelativePresChange the max allowed relative pressure change
+   * @param[in] maxAbsolutePresChange the max allowed absolute pressure change
    * @param[in] maxCompFracChange the max allowed comp fraction change
+   * @param[in] maxRelativeCompDensChange the max allowed comp density change
    * @param[in] rankOffset the rank offset
    * @param[in] numComp the number of components
    * @param[in] dofKey the dof key to get dof numbers
@@ -1170,6 +1172,7 @@ public:
   ScalingForSystemSolutionKernel( real64 const maxRelativePresChange,
                                   real64 const maxAbsolutePresChange,
                                   real64 const maxCompFracChange,
+                                  real64 const maxRelativeCompDensChange,
                                   globalIndex const rankOffset,
                                   integer const numComp,
                                   string const dofKey,
@@ -1190,7 +1193,8 @@ public:
             compDensScalingFactor ),
     m_maxRelativePresChange( maxRelativePresChange ),
     m_maxAbsolutePresChange( maxAbsolutePresChange ),
-    m_maxCompFracChange( maxCompFracChange )
+    m_maxCompFracChange( maxCompFracChange ),
+    m_maxRelativeCompDensChange( maxRelativeCompDensChange )
   {}
 
   /**
@@ -1379,14 +1383,13 @@ public:
         stack.localMaxDeltaCompDens = absCompDensChange;
       }
 
-      real64 const maxAbsCompDensChange = m_maxCompFracChange * prevTotalDens;
-
       // This actually checks the change in component fraction, using a lagged total density
       // Indeed we can rewrite the following check as:
       //    | prevCompDens / prevTotalDens - newCompDens / prevTotalDens | > maxCompFracChange
       // Note that the total density in the second term is lagged (i.e, we use prevTotalDens)
       // because I found it more robust than using directly newTotalDens (which can vary also
       // wildly when the compDens change is large)
+      real64 const maxAbsCompDensChange = m_maxCompFracChange * prevTotalDens;
       if( absCompDensChange > maxAbsCompDensChange && absCompDensChange > eps )
       {
         real64 const compScalingFactor = maxAbsCompDensChange / absCompDensChange;
@@ -1400,6 +1403,24 @@ public:
           stack.localMinCompDensScalingFactor = compScalingFactor;
         }
       }
+
+        // switch from relative to absolute when value is < 1.0
+        real64 const maxRelCompDensChange = m_maxRelativeCompDensChange * LvArray::math::max(m_compDens[ei][ic], 1.0);
+        if( absCompDensChange > maxRelCompDensChange && absCompDensChange > eps )
+        {
+          real64 const compScalingFactor = maxRelCompDensChange / absCompDensChange;
+          std::cout  << ei << " " << ic << " compScalingFactor = " << compScalingFactor << " maxRelCompDensChange = " << maxRelCompDensChange
+          << " absCompDensChange = " << absCompDensChange << " m_compDens[ei][ic] = " << m_compDens[ei][ic] << std::endl;
+          m_compDensScalingFactor[ei] = LvArray::math::min( m_compDensScalingFactor[ei], compScalingFactor );
+          if( stack.localMinVal > compScalingFactor )
+          {
+            stack.localMinVal = compScalingFactor;
+          }
+          if( stack.localMinCompDensScalingFactor > compScalingFactor )
+          {
+            stack.localMinCompDensScalingFactor = compScalingFactor;
+          }
+        }
     }
 
     // compute the scaling factor for other vars, such as temperature
@@ -1412,6 +1433,7 @@ protected:
[…]
