+++ src/coreComponents/physicsSolvers/fluidFlow/CompositionalMultiphaseBase.hpp
@@ -248,6 +248,7 @@ public:
     static constexpr char const * targetRelativePresChangeString() { return "targetRelativePressureChangeInTimeStep"; }
     static constexpr char const * targetRelativeTempChangeString() { return "targetRelativeTemperatureChangeInTimeStep"; }
     static constexpr char const * targetPhaseVolFracChangeString() { return "targetPhaseVolFractionChangeInTimeStep"; }
+    static constexpr char const * targetRelativeCompDensChangeString() { return "targetRelativeCompDensChangeInTimeStep"; }
     static constexpr char const * targetFlowCFLString() { return "targetFlowCFL"; }
 
 
@@ -256,6 +257,7 @@ public:
     static constexpr char const * maxCompFracChangeString() { return "maxCompFractionChange"; }
     static constexpr char const * maxRelativePresChangeString() { return "maxRelativePressureChange"; }
     static constexpr char const * maxRelativeTempChangeString() { return "maxRelativeTemperatureChange"; }
+    static constexpr char const * maxRelativeCompDensChangeString() { return "maxRelativeCompDensChange"; }
     static constexpr char const * allowLocalCompDensChoppingString() { return "allowLocalCompDensityChopping"; }
     static constexpr char const * useTotalMassEquationString() { return "useTotalMassEquation"; }
     static constexpr char const * useSimpleAccumulationString() { return "useSimpleAccumulation"; }
@@ -443,6 +445,9 @@ protected:
   /// maximum (relative) change in temperature in a Newton iteration
   real64 m_maxRelativeTempChange;
 
+  /// maximum (relative) change in component density in a Newton iteration
+  real64 m_maxRelativeCompDensChange;
+
   /// damping factor for solution change targets
   real64 m_solutionChangeScalingFactor;
 
@@ -454,6 +459,9 @@ protected:
 
   /// target (absolute) change in phase volume fraction in a time step
   real64 m_targetPhaseVolFracChange;
+
+  /// target (relative) change in component density in a time step
+  real64 m_targetRelativeCompDensChange;
 
   /// minimum value of the scaling factor obtained by enforcing maxCompFracChange
   real64 m_minScalingFactor;
