### 2024-06-04 fec8f63
Update scaling figure
