+++ src/coreComponents/events/EventManager.cpp
@@ -242,7 +242,7 @@ void EventManager::outputTime() const
   if( m_maxCycle < std::numeric_limits< integer >::max() )
   {
     GEOS_LOG_RANK_0( GEOS_FMT( "    - Cycle:      {} out of {} ({}% completed)",
-                                m_cycle, m_maxCycle, fmt::format( "{:.{}f}", 100.0 * m_cycle / m_maxCycle, 0 ) ) );
+                               m_cycle, m_maxCycle, fmt::format( "{:.{}f}", 100.0 * m_cycle / m_maxCycle, 0 ) ) );
   }
   else
   {
