### 2024-06-04 7652709
Grammar + increase font sizes in the new figure
