### 2024-06-04 c616848
Remove helper class
