+++ src/coreComponents/constitutive/unitTests/testLohrenzBrayClarkViscosity.cpp
@@ -67,8 +67,8 @@ public:
     : m_fluid( FluidData< NC >::createFluid() )
   {
     ComponentProperties const & componentProperties = this->m_fluid->getComponentProperties();
-    m_parameters = std::make_unique< LohrenzBrayClarkViscosity::Parameters >();
-    auto * parameters = dynamic_cast< LohrenzBrayClarkViscosity::Parameters * >(m_parameters.get());
+    m_parameters = LohrenzBrayClarkViscosity::createParameters( nullptr );
+    auto * parameters = const_cast< LohrenzBrayClarkViscosity::Parameters * >(m_parameters->getParameters< LohrenzBrayClarkViscosity::Parameters >());
     parameters->m_componentCriticalVolume.resize( NC );
     TestFluid< 9 >::populateArray( parameters->m_componentCriticalVolume, this->m_fluid->criticalVolume );
     m_density = std::make_unique< CompositionalDensity< CubicEOSPhaseModel< PengRobinsonEOS > > >( "PhaseDensity", componentProperties, 0, *m_parameters );

[…]
