### 2024-06-04 ce89903
update notes file
