### 2024-06-04 1a4a7bc
Activate fluid models (#3123)
Co-authored-by: Brian Han <han12@llnl.gov>
