+++ src/coreComponents/constitutive/fluid/multifluid/MultiFluidSelector.hpp
@@ -41,18 +41,19 @@ void constitutiveUpdatePassThru( MultiFluidBase const & fluid,
 {
   ConstitutivePassThruHandler< DeadOilFluid,
                                BlackOilFluid,
-                               CompositionalTwoPhasePengRobinsonConstantViscosity,
-                               CompositionalTwoPhaseSoaveRedlichKwongConstantViscosity,
-//                               CompositionalTwoPhasePengRobinsonLBCViscosity,
-//                               CompositionalTwoPhaseSoaveRedlichKwongLBCViscosity,
 #ifdef GEOSX_USE_PVTPackage
                                CompositionalMultiphaseFluidPVTPackage,
 #endif
                                CO2BrinePhillipsFluid,
                                CO2BrineEzrokhiFluid,
-                               CO2BrinePhillipsThermalFluid
-//                               ,CO2BrineEzrokhiThermalFluid   "Uncommenting this will lead to compiler segfault. Need to split compilation
-// units for all the options"
+                               CO2BrinePhillipsThermalFluid,
+#if !defined(GEOS_DEVICE_COMPILE)
+                               CO2BrineEzrokhiThermalFluid,
+                               CompositionalTwoPhasePengRobinsonLBCViscosity,
+                               CompositionalTwoPhaseSoaveRedlichKwongLBCViscosity,
+#endif
+                               CompositionalTwoPhasePengRobinsonConstantViscosity,
+                               CompositionalTwoPhaseSoaveRedlichKwongConstantViscosity
                                >::execute( fluid, std::forward< LAMBDA >( lambda ) );
 }
 
@@ -62,18 +63,19 @@ void constitutiveUpdatePassThru( MultiFluidBase & fluid,
 {
   ConstitutivePassThruHandler< DeadOilFluid,
                                BlackOilFluid,
-                               CompositionalTwoPhasePengRobinsonConstantViscosity,
-                               CompositionalTwoPhaseSoaveRedlichKwongConstantViscosity,
-//                               CompositionalTwoPhasePengRobinsonLBCViscosity,
-//                               CompositionalTwoPhaseSoaveRedlichKwongLBCViscosity,
 #ifdef GEOSX_USE_PVTPackage
                                CompositionalMultiphaseFluidPVTPackage,
 #endif
                                CO2BrinePhillipsFluid,
                                CO2BrineEzrokhiFluid,
-                               CO2BrinePhillipsThermalFluid
-                               //,CO2BrineEzrokhiThermalFluid   "Uncommenting this will lead to compiler segfault. Need to split compilation
-// units for all the options"
+                               CO2BrinePhillipsThermalFluid,
+#if !defined(GEOS_DEVICE_COMPILE)
+                               CO2BrineEzrokhiThermalFluid,
+                               CompositionalTwoPhasePengRobinsonLBCViscosity,
+                               CompositionalTwoPhaseSoaveRedlichKwongLBCViscosity,
+#endif
+                               CompositionalTwoPhasePengRobinsonConstantViscosity,
+                               CompositionalTwoPhaseSoaveRedlichKwongConstantViscosity
                                >::execute( fluid, std::forward< LAMBDA >( lambda ) );
 }
 
