### 2024-06-04 5b300f8
Activate fluid models (#3123)
Co-authored-by: Brian Han <han12@llnl.gov>
