+++ src/coreComponents/constitutive/fluid/multifluid/compositional/models/NegativeTwoPhaseFlashModel.hpp
@@ -74,9 +74,9 @@ public:
       phaseFraction.value[m_vapourIndex],
       phaseCompFraction.value[m_liquidIndex],
       phaseCompFraction.value[m_vapourIndex] );
-    GEOS_UNUSED_VAR( flashStatus );
-    //GEOS_ERROR_IF( !flashStatus,
-    //                 GEOS_FMT( "Negative two phase flash failed to converge {:.5e} {:.3f}", pressure, temperature ));
+    GEOS_ERROR_IF( !flashStatus,
+                   GEOS_FMT( "Negative two phase flash failed to converge at pressure {:.5e} and temperature {:.3f}",
+                             pressure, temperature ));
 
     // Calculate derivatives
     NegativeTwoPhaseFlash::computeDerivatives< EOS_TYPE_LIQUID, EOS_TYPE_VAPOUR >(
