+++ src/coreComponents/mesh/generators/include/FaceMgr.hpp
@@ -50,6 +50,10 @@ public:
    * In case the face only belongs to one single element, the second value of the table is -1.
    */
   [[nodiscard]] virtual ToCellRelation< array2d< localIndex > > getFaceToElements() const = 0;
+
+  [[nodiscard]] virtual array1d< integer > getGhostRank() const = 0;
+
+  [[nodiscard]] virtual array1d< globalIndex > getLocalToGlobal() const = 0;
 };
 
 }
