+++ src/coreComponents/physicsSolvers/fluidFlow/SinglePhaseHybridFVM.cpp
@@ -284,8 +284,8 @@ void SinglePhaseHybridFVM::assembleStabilizedFluxTerms( real64 const dt,
                                                         arrayView1d< real64 > const & localRhs )
 {
   // pressure stabilization not implemented
-  GEOS_UNUSED_VAR(dt, domain, dofManager, localMatrix, localRhs);
-  GEOS_ERROR("Stabilized flux not available for this flow solver");
+  GEOS_UNUSED_VAR( dt, domain, dofManager, localMatrix, localRhs );
+  GEOS_ERROR( "Stabilized flux not available for this flow solver" );
 }
 
 
