### 2024-06-04 0ed550f
Separating g2l and l2g.
