+++ src/coreComponents/mesh/generators/BuildPods.cpp
@@ -26,22 +26,21 @@
 namespace geos::ghosting
 {
 
-// TODO Do we want 2 separated `g2l` and `l2g` structs?
-struct GlobalNumberings
+struct GlobalToLocal
 {
-  std::map< NodeGlbIdx, NodeLocIdx > ng2l;
-  std::map< NodeLocIdx, NodeGlbIdx > nl2g;
-
-  std::map< EdgeGlbIdx, EdgeLocIdx > eg2l;
-  std::map< EdgeLocIdx, EdgeGlbIdx > el2g;  // TODO Do we want to make this a vector already? It's surely the most efficient way to build it.
-
-  std::map< FaceGlbIdx, FaceLocIdx > fg2l;
-  std::map< FaceLocIdx, FaceGlbIdx > fl2g;
-
-  std::map< CellGlbIdx, CellLocIdx > cg2l;
-  std::map< CellLocIdx, CellGlbIdx > cl2g;
+  std::map< NodeGlbIdx, NodeLocIdx > nodes;
+  std::map< EdgeGlbIdx, EdgeLocIdx > edges;
+  std::map< FaceGlbIdx, FaceLocIdx > faces;
+  std::map< CellGlbIdx, CellLocIdx > cells;
 };
 
+struct LocalToGlobal
+{
+  std::map< NodeLocIdx, NodeGlbIdx > nodes;
+  std::map< EdgeLocIdx, EdgeGlbIdx > edges;  // TODO Do we want to make this a vector already? It's surely the most efficient way to build it.
+  std::map< FaceLocIdx, FaceGlbIdx > faces;
+  std::map< CellLocIdx, CellGlbIdx > cells;
+};
 
 template< class GI, class LI >
 void buildL2GMappings( std::set< GI > const & gis,
@@ -62,9 +61,10 @@ void buildL2GMappings( std::set< GI > const & gis,
   // TODO add a check to see if we have the full range of local indices.
 }
 
-GlobalNumberings buildL2GMappings( MeshGraph const & graph )
+std::tuple< GlobalToLocal, LocalToGlobal > buildL2GMappings( MeshGraph const & graph )
 {
-  GlobalNumberings result;
+  GlobalToLocal g2l;
+  LocalToGlobal l2g;
 
   // Use `std::ranges::views::keys` when switching to C++20
   auto const keys = []( auto const & m )
@@ -72,12 +72,12 @@ GlobalNumberings buildL2GMappings( MeshGraph const & graph )
     return mapKeys< std::set >( m );
   };
 
-  buildL2GMappings( graph.n, result.ng2l, result.nl2g );
-  buildL2GMappings( keys( graph.e2n ), result.eg2l, result.el2g );
-  buildL2GMappings( keys( graph.f2e ), result.fg2l, result.fl2g );
-  buildL2GMappings( keys( graph.c2f ), result.cg2l, result.cl2g );
+  buildL2GMappings( graph.n, g2l.nodes, l2g.nodes );
+  buildL2GMappings( keys( graph.e2n ), g2l.edges, l2g.edges );
+  buildL2GMappings( keys( graph.f2e ), g2l.faces, l2g.faces );
+  buildL2GMappings( keys( graph.c2f ), g2l.cells, l2g.cells );
 
-  return result;
+  return { g2l, l2g };
 }
 
 template< class GI, class LI >
@@ -282,25 +282,23 @@ CellBlkImpl makeFlavorlessCellBlkImpl( std::size_t const & numCells,
   return CellBlkImpl( intConv< localIndex >( numCells ), ghostRank_, convertToA2d( c2n, 8 ), convertToA2d( c2e, 12 ), convertToA2d( c2f, 6 ), std::move( l2g ) );
 }
 
-DownwardMappings buildDownwardMappings( GlobalNumberings const & numberings,
[…]
