### 2024-06-04 93f5169
Split the construction of the upward matrix information from the ghosting algorithm.
