+++ src/coreComponents/mesh/generators/NewGhosting.cpp
@@ -457,37 +457,38 @@ std::array< std::size_t, N > decode( std::size_t const & basis,
 }
 
 
-std::tuple< MeshGraph, GhostRecv, GhostSend > assembleAdjacencyMatrix( MeshGraph const & owned,
-                                                                       MeshGraph const & present,
-                                                                       MaxGlbIdcs const & gis,
-                                                                       MpiRank curRank )
+struct Adjacency
 {
-  FindGeometricalType const convert( gis.nodes, gis.edges, gis.faces, gis.cells );
-  using Geom = FindGeometricalType::Geom;
-
-  std::size_t const n = convert.numEntries();  // Total number of entries in the graph.
-
-  std::size_t const numOwnedNodes = std::size( owned.n );
-  std::size_t const numOwnedEdges = std::size( owned.e2n );
-  std::size_t const numOwnedFaces = std::size( owned.f2e );
-  std::size_t const numOwnedCells = std::size( owned.c2f );
-  std::size_t const numOwned = numOwnedNodes + numOwnedEdges + numOwnedFaces + numOwnedCells;
-
-  std::size_t const numOtherNodes = std::size( present.n );
-  std::size_t const numOtherEdges = std::size( present.e2n );
-  std::size_t const numOtherFaces = std::size( present.f2e );
-  std::size_t const numOther = numOtherNodes + numOtherEdges + numOtherFaces;
-
-  std::vector< int > ownedGlbIdcs, numEntriesPerRow;  // TODO I couldn't use a vector of `std::size_t`
+  std::vector< int > ownedGlbIdcs;  // TODO Use some strongly typed ints.
+  std::vector< int > otherGlbIdcs;  // TODO Use some strongly typed ints.
+  std::vector< int > numEntriesPerRow;
   std::vector< std::vector< int > > indices;
   std::vector< std::vector< double > > values;
+};
+
+
+Adjacency buildAdjacency( MeshGraph const & owned,
+                          MeshGraph const & present,
+                          FindGeometricalType const & convert )
+{
+  Adjacency adjacency;
+
+  std::size_t const numOwned = std::size( owned.n ) + std::size( owned.e2n ) + std::size( owned.f2e ) + std::size( owned.c2f );
+  std::size_t const numOther = std::size( present.n ) + std::size( present.e2n ) + std::size( present.f2e );
+
+  // Aliases
+  std::vector< int > & ownedGlbIdcs = adjacency.ownedGlbIdcs;
+  std::vector< int > & otherGlbIdcs = adjacency.otherGlbIdcs;
+  std::vector< int > & numEntriesPerRow = adjacency.numEntriesPerRow;
+  std::vector< std::vector< int > > & indices = adjacency.indices;
+  std::vector< std::vector< double > > & values = adjacency.values;
+
   ownedGlbIdcs.reserve( numOwned );
+  otherGlbIdcs.reserve( numOther );
   numEntriesPerRow.reserve( numOwned );
   indices.reserve( numOwned );
-  indices.reserve( numOwned );
+  values.reserve( numOwned );
 
-  std::vector< int > otherGlbIdcs;  // TODO I couldn't use a vector of `std::size_t`
-  otherGlbIdcs.reserve( numOther );
   for( NodeGlbIdx const & ngi: present.n )
   {
     otherGlbIdcs.emplace_back( convert.fromNodeGlbIdx( ngi ) );
@@ -592,6 +593,30 @@ std::tuple< MeshGraph, GhostRecv, GhostSend > assembleAdjacencyMatrix( MeshGraph
   {
     GEOS_ASSERT_EQ( indices[i].size(), std::size_t( numEntriesPerRow[i] ) );
   }
+
[…]
