### 2024-06-05 b3b1b5f
Use one single pair of vectors when copying the rows.
