+++ src/coreComponents/mesh/generators/NewGhosting.cpp
@@ -724,14 +724,19 @@ std::tuple< MeshGraph, GhostRecv, GhostSend > performGhosting( MeshGraph const &
   }
   EpetraExt::RowMatrixToMatrixMarketFile( "/tmp/matrices/ghostInfo.mat", ghostingFootprint );
 
-  int extracted2 = 0;
-  std::vector< double > extractedValues2( commSize );  // TODO improve with resize...
-  std::vector< int > extractedIndices2( commSize );
+  int extracted = 0;
+  std::vector< double > extractedValues;
+  std::vector< int > extractedIndices;
 
   GhostSend send;
   for( int const & index: ownedGlbIdcs )
   {
-    ghostingFootprint.ExtractGlobalRowCopy( index, commSize, extracted2, extractedValues2.data(), extractedIndices2.data() );
+    int const length = ghostingFootprint.NumGlobalEntries( index );
+    extractedValues.resize( length );
+    extractedIndices.resize( length );
+    ghostingFootprint.ExtractGlobalRowCopy( index, length, extracted, extractedValues.data(), extractedIndices.data() );
+    GEOS_ASSERT_EQ( extracted, length );
+
     std::set< MpiRank > * sendingTo = nullptr;
     switch( convert.getGeometricalType( index ) )
     {
@@ -755,11 +760,15 @@ std::tuple< MeshGraph, GhostRecv, GhostSend > performGhosting( MeshGraph const &
         sendingTo = &send.cells[convert.toCellGlbIdx( index )];
         break;
       }
+      default:
+      {
+        GEOS_ERROR( "Internal error" );
+      }
     }
 
-    for( int ii = 0; ii < extracted2; ++ii )
+    for( int ii = 0; ii < extracted; ++ii )
     {
-      MpiRank const rank{ extractedIndices2[ii] };
+      MpiRank const rank{ extractedIndices[ii] };
       if( rank != curRank )
       {
         sendingTo->insert( rank );
@@ -767,12 +776,11 @@ std::tuple< MeshGraph, GhostRecv, GhostSend > performGhosting( MeshGraph const &
     }
   }
 
-  int extracted = 0;
-  int const length = ghostExchange.NumGlobalEntries( curRank.get() );
-  std::vector< double > extractedValues( length );
-  std::vector< int > extractedIndices( length );
-  ghostExchange.ExtractGlobalRowCopy( curRank.get(), length, extracted, extractedValues.data(), extractedIndices.data() );
-  GEOS_ASSERT_EQ( extracted, length );
+  int const numNeededIndices = ghostExchange.NumGlobalEntries( curRank.get() );
+  extractedValues.resize( numNeededIndices );
+  extractedIndices.resize( numNeededIndices );
+  ghostExchange.ExtractGlobalRowCopy( curRank.get(), numNeededIndices, extracted, extractedValues.data(), extractedIndices.data() );
+  GEOS_ASSERT_EQ( extracted, numNeededIndices );
 
   std::set< int > const allNeededIndices( std::cbegin( extractedIndices ), std::cend( extractedIndices ) );
   std::set< int > receivedIndices;  // The graph nodes that my neighbors will send me.
@@ -783,6 +791,7 @@ std::tuple< MeshGraph, GhostRecv, GhostSend > performGhosting( MeshGraph const &
   std::set_difference( std::cbegin( receivedIndices ), std::cend( receivedIndices ),
                        std::cbegin( otherGlbIdcs ), std::cend( otherGlbIdcs ),
                        std::back_inserter( notPresentIndices ) );
[…]
