### 2024-06-05 58116e7
discard for now dHalWeight_dVar
