+++ src/coreComponents/finiteVolume/CellElementStencilTPFA.hpp
@@ -325,25 +325,13 @@ CellElementStencilTPFAWrapper::averageWeights( const localIndex iconn,
   real64 const harmonicWeight = sum > 0 ? product / sum : 0.0;
   real64 const arithmeticWeight = sum / 2;
 
-
-  real64 dHarmonicWeight_dVar[2];
-  real64 dArithmeticWeight_dVar[2];
-
-  dHarmonicWeight_dVar[0] = sum > 0 ? (dHalfWeight_dVar[0]*sum*halfWeight[1] - dHalfWeight_dVar[0]*halfWeight[0]*halfWeight[1]) / ( sum*sum ) : 0.0;
-  dHarmonicWeight_dVar[1] = sum > 0 ? (dHalfWeight_dVar[1]*sum*halfWeight[0] - dHalfWeight_dVar[1]*halfWeight[1]*halfWeight[0]) / ( sum*sum ) : 0.0;
-
-  dArithmeticWeight_dVar[0] = dHalfWeight_dVar[0] / 2;
-  dArithmeticWeight_dVar[1] = dHalfWeight_dVar[1] / 2;
-
   real64 const meanPermCoeff = 1.0; //TODO make it a member if it is really necessary
 
   real64 const value = meanPermCoeff * harmonicWeight + (1 - meanPermCoeff) * arithmeticWeight;
   for( localIndex ke = 0; ke < 2; ++ke )
   {
     weight[0][ke] = m_transMultiplier[iconn] * value * (ke == 0 ? 1 : -1);
-
-    real64 const dValue_dVar = meanPermCoeff * dHarmonicWeight_dVar[ke] + (1 - meanPermCoeff) * dArithmeticWeight_dVar[ke];
-    dWeight_dVar[0][ke] = m_transMultiplier[iconn] * dValue_dVar;
+    dWeight_dVar[0][ke] = 0;
   }
 }
 
