### 2024-06-05 b1ec0d0
begin refacto in table function for more clarity
