+++ src/coreComponents/fileIO/Table/TableFormatter.hpp
@@ -139,12 +139,10 @@ private:
    * @param columns The vector containing all table columns
    * @param msgTableError A vector containg all error related to the table
    * @param sectionSeparatingLine An empty string for building the section separator
-   * @param sectionSeparatingLine An empty string for building the section separator
    */
   void outputLayout( std::ostringstream & tableOutput,
                      std::vector< TableLayout::Column > & columns,
                      std::vector< string > const & msgTableError,
-                     string & sectionSeparatingLine ) const;
                      string & sectionSeparatingLine ) const;
 
   /**
@@ -175,27 +173,20 @@ private:
   /**
    * @brief Compute the max table line length, taking into account the length of : title, error, columns header and content
    * Increase the size of the columns if necessary
-   * @brief Compute the max table line length, taking into account the length of : title, error, columns header and content
-   * Increase the size of the columns if necessary
    * @param columns Vector of column containing containing the largest string for each column
    * @param msgTableError Vector containing all error messages
    */
   void computeTableWidth( std::vector< TableLayout::Column > & columns,
                           std::vector< string > const & msgTableError ) const;
-  void computeTableWidth( std::vector< TableLayout::Column > & columns,
-                          std::vector< string > const & msgTableError ) const;
 
   /**
-   * @brief Build all separators needed from content length contained in the columns vector
    * @brief Build all separators needed from content length contained in the columns vector
    * @param columns Vector containing all table columns
    * @param topSeparator Top separator to be built
    * @param sectionSeparatingLine Line section separator to be built
-   * @param sectionSeparatingLine Line section separator to be built
    */
   void buildTableSeparators( std::vector< TableLayout::Column > const & columns,
                              string & topSeparator,
-                             string & sectionSeparatingLine ) const;
                              string & sectionSeparatingLine ) const;
 
   /**
@@ -214,14 +205,12 @@ private:
    * @brief Output a section by specifying it's type ( header or section )
    * @param columns Vector containing all table columns
    * @param sectionSeparatingLine Line separator between sections
-   * @param sectionSeparatingLine Line separator between sections
    * @param rows A section row
    * @param nbRows Indicates the number of lines in a section
    * @param section The section to be built
    * @note Add the ending line if there are one or more rows
    */
   void outputSectionRows( std::vector< TableLayout::Column > const & columns,
-                          string_view sectionSeparatingLine,
                           string_view sectionSeparatingLine,
                           std::ostringstream & rows,
                           integer const nbRows,
