+++ src/coreComponents/functions/TableFunction.cpp
@@ -20,9 +20,6 @@
 #include "codingUtilities/Parsing.hpp"
 #include "common/DataTypes.hpp"
 #include "fileIO/Outputs/OutputBase.hpp"
-#include "fileIO/Table/TableLayout.hpp"
-#include "fileIO/Table/TableData.hpp"
-#include "fileIO/Table/TableFormatter.hpp"
 
 
 #include <algorithm>
@@ -188,6 +185,56 @@ void TableFunction::checkCoord( real64 const coord, localIndex const dim ) const
                  SimulationError );
 }
 
+void TableFunction::printCSVHeader( std::ofstream & os, integer const numDimensions ) const
+{
+  for( integer d = 0; d < numDimensions; d++ )
+  {
+    os << units::getDescription( getDimUnit( d )) << ",";
+  }
+  os << units::getDescription( m_valueUnit ) << "\n";
+}
+
+void TableFunction::printCSVValues( std::ofstream & os, integer const numDimensions ) const
+{
+  // prepare dividers
+  std::vector< integer > div( numDimensions );
+  div[0] = 1;
+  for( integer d = 1; d < numDimensions; d++ )
+  {
+    div[d] = div[d-1] * m_coordinates[d-1].size();
+  }
+  // loop through all the values
+  for( integer v = 0; v < m_values.size(); v++ )
+  {
+    // find coords indices
+    std::vector< integer > idx( numDimensions );
+    integer r = v;
+    for( integer d = numDimensions-1; d >= 0; d-- )
+    {
+      idx[d] = r / div[d];
+      r = r % div[d];
+    }
+    // finally print out in right order
+
+    for( integer d = 0; d < numDimensions; d++ )
+    {
+      arraySlice1d< real64 const > const coords = m_coordinates[d];
+      os << coords[idx[d]] << ",";
+    }
+    os << m_values[v] << "\n";
+  }
+}
+
+void TableFunction::convertTable2D( TableData2D::Conversion1D & tableConverted ) const
+{
+  TableData2D tableData2D;
+  tableData2D.collect2DData( m_coordinates[0], m_coordinates[1], m_values );
+
+  tableConverted = tableData2D.convert2DData( m_valueUnit,
+                                              units::getDescription( getDimUnit( 0 ) ),
+                                              units::getDescription( getDimUnit( 1 ) ));
+}
+
 void TableFunction::printInCSV( string const & filename ) const
 {
   std::ofstream os( joinPath( OutputBase::getOutputDirectory(), filename + ".csv" ) );
@@ -198,70 +245,19 @@ void TableFunction::printInCSV( string const & filename ) const
 
   if( numDimensions != 2 )
   {
-    // print header
-
-    for( integer d = 0; d < numDimensions; d++ )
-    {
-      os << units::getDescription( getDimUnit( d )) << ",";
-    }
-    os << units::getDescription( m_valueUnit ) << "\n";
-
-    // print values
-
-    // prepare dividers
-    std::vector< integer > div( numDimensions );
-    div[0] = 1;
-    for( integer d = 1; d < numDimensions; d++ )
-    {
-      div[d] = div[d-1] * m_coordinates[d-1].size();
-    }
-    // loop through all the values
-    for( integer v = 0; v < m_values.size(); v++ )
-    {
-      // find coords indices
-      std::vector< integer > idx( numDimensions );
-      integer r = v;
-      for( integer d = numDimensions-1; d >= 0; d-- )
-      {
-        idx[d] = r / div[d];
-        r = r % div[d];
-      }
-      // finally print out in right order
-
-      for( integer d = 0; d < numDimensions; d++ )
-      {
-        arraySlice1d< real64 const > const coords = m_coordinates[d];
-        os << coords[idx[d]] << ",";
-      }
-      os << m_values[v] << "\n";
-    }
[…]

[…]
