+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3050-5325-9f50d94
+  baseline: integratedTests/baseline_integratedTests-pr3113-5435-b794227
 
 allow_fail:
   all: ''
