### 2024-06-05 32ae6b9
1) add Aww numerical derivative unit tests, 2) thermal index fix
