+++ src/coreComponents/unitTests/wellsTests/testThermalReservoirCompositionalMultiphaseMSWells.cpp
@@ -1,0 +1,797 @@
+/*
+ * ------------------------------------------------------------------------------------------------------------
+ * SPDX-License-Identifier: LGPL-2.1-only
+ *
+ * Copyright (c) 2018-2020 Lawrence Livermore National Security LLC
+ * Copyright (c) 2018-2020 The Board of Trustees of the Leland Stanford Junior University
+ * Copyright (c) 2018-2020 TotalEnergies
+ * Copyright (c) 2019-     GEOSX Contributors
+ * All rights reserved
+ *
+ * See top level LICENSE, COPYRIGHT, CONTRIBUTORS, NOTICE, and ACKNOWLEDGEMENTS files for details.
+ * ------------------------------------------------------------------------------------------------------------
+ */
+
+#include "unitTests/fluidFlowTests/testCompFlowUtils.hpp"
+
+#include "common/DataTypes.hpp"
+#include "mainInterface/initialization.hpp"
+#include "constitutive/fluid/multifluid/MultiFluidBase.hpp"
+#include "mainInterface/ProblemManager.hpp"
+#include "mesh/DomainPartition.hpp"
+#include "mainInterface/GeosxState.hpp"
+#include "mesh/WellElementSubRegion.hpp"
+#include "physicsSolvers/PhysicsSolverManager.hpp"
+#include "physicsSolvers/multiphysics/CompositionalMultiphaseReservoirAndWells.hpp"
+#include "physicsSolvers/fluidFlow/CompositionalMultiphaseFVM.hpp"
+#include "physicsSolvers/fluidFlow/wells/CompositionalMultiphaseWell.hpp"
+#include "physicsSolvers/fluidFlow/wells/CompositionalMultiphaseWellKernels.hpp"
+#include "physicsSolvers/fluidFlow/wells/CompositionalMultiphaseWellFields.hpp"
+#include "physicsSolvers/fluidFlow/wells/WellSolverBaseFields.hpp"
+
+using namespace geos;
+using namespace geos::dataRepository;
+using namespace geos::constitutive;
+using namespace geos::testing;
+
+CommandLineOptions g_commandLineOptions;
+
+char const * xmlInput =
+  R"xml(
+
+<?xml version="1.0" ?>
+<Problem>
+  <Solvers>
+    <CompositionalMultiphaseReservoir
+      name="reservoirSystem"
+      flowSolverName="compflow"
+      wellSolverName="compositionalMultiphaseWell"
+      logLevel="4"
+      initialDt="1e4"
+      targetRegions="{ region, injwell }">
+      <NonlinearSolverParameters
+         logLevel="0"
+        newtonTol="1.0e-5"
+        lineSearchAction="None"
+        newtonMaxIter="40"/>
+      <LinearSolverParameters
+        logLevel="4"
+        directParallel="0"/>
+    </CompositionalMultiphaseReservoir>
+
+    <CompositionalMultiphaseFVM
+      name="compflow"
+      logLevel="4"
+      discretization="fluidTPFA"
+      temperature="368.15"
+      useMass="1"
+      isThermal="1"
+      initialDt="1e4"
+      targetRelativePressureChangeInTimeStep="1"
+      targetRelativeTemperatureChangeInTimeStep="1"
+      targetPhaseVolFractionChangeInTimeStep="1"      
+      maxCompFractionChange="0.5"
+      targetRegions="{ region }">
+    </CompositionalMultiphaseFVM>
+
+    <CompositionalMultiphaseWell
+      name="compositionalMultiphaseWell"
+      targetRegions="{ injwell }"
+      isThermal="1"
+      logLevel="1"
+      initialDt="1e4"
+      useMass="1">
+      <WellControls
+        name="WC_CO2_INJ"
[…]
