### 2024-06-05 1e11446
getParameters -> get
