### 2024-06-05 6407ca1
Fix updateConfiguration (#3113)
Co-authored-by: Matteo Cusini <matteo.cusini@gmail.com>
Co-authored-by: Jian HUANG <jian.huang@totalenergies.com>
