### 2024-06-05 bd245ed
Update .integrated_tests.yaml
