+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3113-5435-b794227
+  baseline: integratedTests/baseline_integratedTests-pr3120-5441-d1e1daf
 
 allow_fail:
   all: ''
