+++ src/coreComponents/physicsSolvers/NonlinearSolverParameters.cpp
@@ -181,7 +181,7 @@ void NonlinearSolverParameters::postProcessInput()
   GEOS_ERROR_IF_LE_MSG( m_lineSearchResidualFactor, 0.0,
                         getWrapperDataContext( viewKeysStruct::lineSearchResidualFactorString() ) << ": should be positive" );
 
-  //if( getLogLevel() > 0 )
+  if( getLogLevel() > 0 )
   {
     print();
   }
