+++ src/coreComponents/unitTests/wellsTests/testIsothermalReservoirCompositionalMultiphaseMSWells.cpp
@@ -531,7 +531,7 @@ void testNumericalJacobian( CompositionalMultiphaseReservoirAndWells< Compositio
   residual.zero();
   jacobian.zero();
   assembleFunction( jacobian.toViewConstSizes(), residual.toView() );
-  printCompareLocalMatrices( jacobian.toViewConst(), jacobianFD.toViewConst());
+  //printCompareLocalMatrices( jacobian.toViewConst(), jacobianFD.toViewConst());
   compareLocalMatrices( jacobian.toViewConst(), jacobianFD.toViewConst(), relTol );
 }
 
