### 2024-06-05 d10b58c
Add missing tests into ATS and adjust output naming (#3120)
