### 2024-06-05 be43b71
Add missing tests into ATS and adjust output naming (#3120)
