### 2024-06-05 accbc8c
Update .integrated_tests.yaml
