+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3120-5441-d1e1daf
+  baseline: integratedTests/baseline_integratedTests-pr3075-5458-8cde9da
 
 allow_fail:
   all: ''
