+++ src/coreComponents/mesh/generators/NewGhosting.cpp
@@ -138,7 +138,22 @@ void to_json( json & j,
   j = json{ { "c2f", v.f2e },
             { "f2e", v.f2e },
             { "e2n", v.e2n },
-            { "n", v.n } };
+            { "n2pos", v.n2pos } };
+}
+
+
+std::map< NodeGlbIdx, vtkIdType > buildNgiToVtk( vtkSmartPointer< vtkDataSet > mesh )
+{
+  std::map< NodeGlbIdx, vtkIdType > res;
+
+  vtkIdTypeArray const * globalPtIds = vtkIdTypeArray::FastDownCast( mesh->GetPointData()->GetGlobalIds() );
+
+  for( vtkIdType i = 0; i < mesh->GetNumberOfPoints(); ++i )
+  {
+    res[NodeGlbIdx{ globalPtIds->GetValue( i ) }] = i;
+  }
+
+  return res;
 }
 
 /**
@@ -162,10 +177,15 @@ std::tuple< MeshGraph, MeshGraph > buildMeshGraph( vtkSmartPointer< vtkDataSet >
     return curRank == *std::min_element( std::cbegin( ranks ), std::cend( ranks ) );
   };
 
+  std::map< NodeGlbIdx, vtkIdType > const n2vtk = buildNgiToVtk( mesh );
   for( auto const & [ranks, nodes]: buckets.nodes )
   {
-    std::set< NodeGlbIdx > & n = isCurrentRankOwning( ranks ) ? owned.n : present.n;
-    n.insert( std::cbegin( nodes ), std::cend( nodes ) );
+    std::map< NodeGlbIdx, std::array< double, 3 > > & n2pos = isCurrentRankOwning( ranks ) ? owned.n2pos : present.n2pos;
+    for( NodeGlbIdx const & ngi: nodes )
+    {
+      double const * pos = mesh->GetPoint( n2vtk.at( ngi ) );
+      n2pos[ngi] = { pos[0], pos[1], pos[2] };
+    }
   }
 
   for( auto const & [ranks, edges]: buckets.edges )
@@ -309,6 +329,7 @@ Epetra_CrsMatrix multiply( int commSize,
 
   // Downward (c -> f -> e -> n)
   auto tDownward = makeTranspose( upward );  // TODO check the algorithm to understand what's more relevant.
+  // TODO why do we have to perform the transposition ourselves instead of using the flag from `EpetraExt::MatrixMatrix::Multiply`.
 
   Epetra_CrsMatrix result_d0_0( Epetra_DataAccess::Copy, ownedMap, commSize, false );
   EpetraExt::MatrixMatrix::Multiply( *tDownward, false, result_u0_2, false, result_d0_0, false );
@@ -456,9 +477,9 @@ std::array< std::size_t, N > decode( std::size_t const & basis,
   return result;
 }
 
-
 struct Adjacency
 {
+  std::vector< int > ownedNodesIdcs;  // TODO Use some strongly typed ints.
   std::vector< int > ownedGlbIdcs;  // TODO Use some strongly typed ints.
   std::vector< int > otherGlbIdcs;  // TODO Use some strongly typed ints.
   std::vector< int > numEntriesPerRow;
@@ -466,30 +487,31 @@ struct Adjacency
   std::vector< std::vector< double > > values;
 };
 
-
 Adjacency buildAdjacency( MeshGraph const & owned,
                           MeshGraph const & present,
                           FindGeometricalType const & convert )
 {
   Adjacency adjacency;
 
-  std::size_t const numOwned = std::size( owned.n ) + std::size( owned.e2n ) + std::size( owned.f2e ) + std::size( owned.c2f );
[…]
