### 2024-06-06 a799687
Give the node positions to the NodeMgr.
