+++ src/coreComponents/unitTests/wellsTests/testThermalReservoirCompositionalMultiphaseSSWells.cpp
@@ -340,150 +340,151 @@ void testNumericalJacobian( CompositionalMultiphaseReservoirAndWells< Compositio
   ////////////////////////////////////////////////
   // Step 1) Compute the terms in J_RR and J_WR //
   ////////////////////////////////////////////////
-  if (1)
+  if( 1 )
     domain.forMeshBodies( [&] ( MeshBody & meshBody )
     {
       meshBody.forMeshLevels( [&] ( MeshLevel & mesh )
       {
-      std::cout << mesh.getName() << " " ;
+        std::cout << mesh.getName() << " ";
         ElementRegionManager & elemManager = mesh.getElemManager();
-      std::cout << elemManager.getName() << std::endl;
+        std::cout << elemManager.getName() << std::endl;
         for( localIndex er = 0; er < elemManager.numRegions(); ++er )
         {
           ElementRegionBase & elemRegion = elemManager.getRegion( er );
           elemRegion.forElementSubRegionsIndex< CellElementSubRegion >( [&]( localIndex const, CellElementSubRegion & subRegion )
           {
-          std::cout << " " << subRegion.getName() << " " << subRegion.size() << std::endl;
+            std::cout << " " << subRegion.getName() << " " << subRegion.size() << std::endl;
             // get the degrees of freedom and ghosting information
             arrayView1d< globalIndex const > const & dofNumber =
               subRegion.getReference< array1d< globalIndex > >( resDofKey );
 
             // get the primary variables on the reservoir elements
             arrayView1d< real64 > const & pres =
-            subRegion.getField< fields::flow::pressure >();
+              subRegion.getField< fields::flow::pressure >();
             pres.move( hostMemorySpace, false );
 
             arrayView2d< real64, compflow::USD_COMP > const & compDens =
-            subRegion.getField< fields::flow::globalCompDensity >();
+              subRegion.getField< fields::flow::globalCompDensity >();
             compDens.move( hostMemorySpace, false );
 
-          arrayView1d< real64 > const & temp =
-            subRegion.getField< fields::flow::temperature >();
-          temp.move( hostMemorySpace, false );
+            arrayView1d< real64 > const & temp =
+              subRegion.getField< fields::flow::temperature >();
+            temp.move( hostMemorySpace, false );
 
             // a) compute all the derivatives wrt to the pressure in RESERVOIR elem ei
             for( localIndex ei = 0; ei < subRegion.size(); ++ei )
             {
-            if ( ei !=0 )
-              break;
-              real64 totalDensity = 0.0;
-              for( localIndex ic = 0; ic < NC; ++ic )
+              if( ei ==0 )
               {
-                totalDensity += compDens[ei][ic];
-              }
-
-              {
-                solver.resetStateToBeginningOfStep( domain );
-
-                // here is the perturbation in the pressure of the element
-                real64 const dP = perturbParameter * (pres[ei] + perturbParameter);
-                pres.move( hostMemorySpace, true );
-                pres[ei] += dP;
-
-                // after perturbing, update the pressure-dependent quantities in the reservoir
-                flowSolver.forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&] ( string const &,
-                                                                                         MeshLevel & mesh2,
-                                                                                         arrayView1d< string const > const & regionNames2 )
+                real64 totalDensity = 0.0;
+                for( localIndex ic = 0; ic < NC; ++ic )
                 {
-                  mesh2.getElemManager().forElementSubRegions( regionNames2,
-                                                               [&]( localIndex const,
-                                                                    ElementSubRegionBase & subRegion2 )
+                  totalDensity += compDens[ei][ic];
+                }
+
+                {
[…]
