### 2024-06-06 d811b45
Update text and figure with GPU-aware results
