### 2024-06-06 ba28c76
Fix LaTeX syntax
