### 2024-06-06 aafe671
Fix LaTeX syntax
