### 2024-06-06 3ccdb41
Start bringing the new mappings to the core of geos.
