### 2024-06-06 #6829 @kthyng
@rrsettgast We encourage authors to keep closer to the 1000 word limit and move content to docs, etc, where possible, but it is a soft limit. Don't worry about it too much.
