### 2024-06-06 #6829 @kthyng
- [ ] Clear research application/in scope
