### 2024-06-06 #6829 @kthyng
> Some of the references do not have a DOI. Is this OK?

Yes, if a reference has no possible DOI, that is fine.
