### 2024-06-06 #6829 @kthyng
@rrsettgast
* I see in your paper you differentiate from another GEOS. You might want to differentiate with the major python package https://libgeos.org/ too! (it's a pretty major library). Though, it's not necessary if you have a specific reason for listing that other package.
* Your paper needs spaces before some references, for example on lines 57 and 62, please check for others.
* I've run out of time at the moment but it would help if you can point me to the way users can verify that the code is working properly when they're using it (tests or example/demo code).
