### 2024-06-07 #6829 @rrsettgast
Hello @kthyng. I have put some responses inline:

> * I see in your paper you differentiate from another GEOS. You might want to differentiate with the major python package https://libgeos.org/ too! (it's a pretty major library). Though, it's not necessary if you have a specific reason for listing that other package.

I did this because I am the author of the previous GEOS, and that GEOS had a similar application space. I had to rewrite from scratch when we targeted exascale hardware, and had called the new code GEOSX. Due to some complaints we received that the pronunciation sounded like "GEO-SEX", we decided to take the GEOS name from the abandoned earlier code.

> * Your paper needs spaces before some references, for example on lines 57 and 62, please check for others.

got them. I think we are good on this now.

> * I've run out of time at the moment but it would help if you can point me to the way users can verify that the code is working properly when they're using it (tests or example/demo code).

There is some information on testing contained in our [quick start guide](https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/QuickStart.html#testing). There is a [tutorials section](https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/tutorials/Index.html). And finally there is a section on [integrated tests](https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/tutorials/Index.html). I didn't really refer to testing in the document as I thought their existence would be implied. Should I add some text in the main document to highlight it?
