+++ src/docs/JOSS/paper.bib
@@ -9,6 +9,15 @@
   doi = {10.1002/nag.2557}
 }
 
+@Manual{libgeos,
+  title = {{GEOS} computational geometry library},
+  author = {{GEOS contributors}},
+  organization = {Open Source Geospatial Foundation},
+  year = {2021},
+  url = {https://libgeos.org/},
+  doi = {10.5281/zenodo.11396894}
+}
+
 @InProceedings{Beckingsale:2019,
   author={Beckingsale, David A. and Burmark, Jason and Hornung, Rich and Jones, Holger and Killian, William and Kunen, Adam J. and Pearce, Olga and Robinson, Peter and Ryujin, Brian S. and Scogland, Thomas R. W.},
   booktitle={2019 IEEE/ACM International Workshop on Performance, Portability and Productivity in HPC (P3HPC)}, 
