+++ src/coreComponents/functions/TableFunction.cpp
@@ -20,8 +20,6 @@
 #include "codingUtilities/Parsing.hpp"
 #include "common/DataTypes.hpp"
 #include "fileIO/Outputs/OutputBase.hpp"
-
-
 #include <algorithm>
 
 namespace geos
@@ -185,133 +183,6 @@ void TableFunction::checkCoord( real64 const coord, localIndex const dim ) const
                  SimulationError );
 }
 
-void TableFunction::printCSVHeader( std::ofstream & logStream, integer const numDimensions ) const
-{
-  for( integer d = 0; d < numDimensions; d++ )
-  {
-    logStream << units::getDescription( getDimUnit( d )) << ",";
-  }
-  logStream << units::getDescription( m_valueUnit ) << "\n";
-}
-
-void TableFunction::printCSVValues( std::ofstream & logStream, integer const numDimensions ) const
-{
-  // prepare dividers
-  std::vector< integer > div( numDimensions );
-  div[0] = 1;
-  for( integer d = 1; d < numDimensions; d++ )
-  {
-    div[d] = div[d-1] * m_coordinates[d-1].size();
-  }
-  // loop through all the values
-  for( integer v = 0; v < m_values.size(); v++ )
-  {
-    // find coords indices
-    std::vector< integer > idx( numDimensions );
-    integer r = v;
-    for( integer d = numDimensions-1; d >= 0; d-- )
-    {
-      idx[d] = r / div[d];
-      r = r % div[d];
-    }
-    // finally print out in right order
-    for( integer d = 0; d < numDimensions; d++ )
-    {
-      arraySlice1d< real64 const > const coords = m_coordinates[d];
-      logStream << coords[idx[d]] << ",";
-    }
-    logStream << m_values[v] << "\n";
-  }
-}
-
-void TableFunction::convertTable2D( TableData2D::Conversion1D & tableConverted ) const
-{
-  TableData2D tableData2D;
-  tableData2D.collect2DData( m_coordinates[0], m_coordinates[1], m_values );
-  tableConverted = tableData2D.convert2DData( m_valueUnit,
-                                              units::getDescription( getDimUnit( 0 ) ),
-                                              units::getDescription( getDimUnit( 1 ) ));
-}
-
-void TableFunction::printInCSV( string const & filename ) const
-{
-  std::ofstream logStream( joinPath( OutputBase::getOutputDirectory(), filename + ".csv" ) );
-  GEOS_LOG_RANK_0( GEOS_FMT( "CSV Generated to inputFiles/compositionalMultiphaseWell/{}/{}.csv \n",
-                             OutputBase::getOutputDirectory(),
-                             filename ));
-  integer const numDimensions = LvArray::integerConversion< integer >( m_coordinates.size() );
-
-  if( numDimensions != 2 )
-  {
-    printCSVHeader( logStream, numDimensions );
-    printCSVValues( logStream, numDimensions );
-  }
-  else // numDimensions == 2
-  {
-    //1.
-    TableData2D::Conversion1D tableConverted;
-    convertTable2D( tableConverted );
-    //2.
-    TableLayout tableLayout( tableConverted.headerNames );
-    //3.
-    TableCSVFormatter csvFormat( tableLayout );
-    logStream << csvFormat.headerToString() << csvFormat.dataToString( tableConverted.tableData );
-  }
-  logStream.close();
-}
-
-void TableFunction::printInLog( string const & title ) const
-{
-  integer const numDimensions = LvArray::integerConversion< integer >( m_coordinates.size() );
-
-  GEOS_LOG_RANK_0( GEOS_FMT( "Values in the table are represented by : {}", units::getDescription( m_valueUnit )));
-
[…]
