+++ src/coreComponents/fileIO/Table/TableFormatter.hpp
@@ -21,6 +21,10 @@
 
 #include "TableData.hpp"
 #include "TableLayout.hpp"
+#include "common/Units.hpp"
+#include <numeric>
+#include "codingUtilities/StringUtilities.hpp"
+#include "fileIO/Outputs/OutputBase.hpp"
 
 namespace geos
 {
@@ -35,6 +39,8 @@ protected:
 
   /// Layout for a table
   TableLayout m_tableLayout;
+
+  TableFormatter() = default;
 
   /**
    * @brief Construct a new Table Formatter from a tableLayout
@@ -54,6 +60,11 @@ protected:
 class TableCSVFormatter : public TableFormatter
 {
 public:
+
+  /**
+   * @brief Construct a new Table Formatter
+   */
+  TableCSVFormatter(): TableFormatter( TableLayout()) {}
 
   /**
    * @brief Construct a new Table Formatter from a tableLayout
@@ -79,27 +90,43 @@ public:
   string dataToString( TableData const & tableData ) const;
 
   /**
-   * @brief Convert the TableData to a table string.
-   * @param tableData The TableData to convert.
-   * @return The table string representation of the TableData.
+   * @brief Convert a data source to a CSV string.
+   * @tparam DATASOURCE The soruce to convert
+   * @return The CSV string representation of a data source.
    */
-  string toString( TableData const & tableData ) const;
+  template< typename DATASOURCE >
+  string toString( DATASOURCE const & tableData ) const;
 
 };
+
+/**
+ * @brief Convert the TableData to a table string.
+ * @param tableData The TableData to convert.
+ * @return The CSV string representation of the TableData.
+ */
+template<>
+string TableCSVFormatter::toString< TableData >( TableData const & tableData ) const;
+
 
 /**
  * @brief class for log formatting
  */
 class TableTextFormatter : public TableFormatter
 {
-
 public:
+
+
+  /**
+   * @brief Construct a new TableFormatter
+   */
+  TableTextFormatter(): TableFormatter( TableLayout()) {}
 
   /**
    * @brief Construct a new TableFormatter from a tableLayout
    * @param tableLayout Contain all column names and optionnaly the table title
    */
   TableTextFormatter( TableLayout const & tableLayout );
+
 
   /**
    * @brief Destroy the Table Text Formatter object
@@ -112,11 +139,12 @@ public:
   string layoutToString() const;
 
   /**
-   * @brief Convert the TableData to a table string.
-   * @param tableData The TableData to convert.
+   * @brief Convert a data source to a table string.
+   * @param tableData The data source to convert.
    * @return The table string representation of the TableData.
    */
-  string toString( TableData const & tableData ) const;
+  template< typename DATASOURCE >
+  string toString( DATASOURCE const & tableData ) const;
 
 private:
 
@@ -125,7 +153,7 @@ private:
   ///  for the extremity of a row
   static constexpr char m_horizontalLine = '-';
 
-  /**F
+  /**
    * @brief Fill the vector (m_column) in tableData with values from rows stored in tableData.
    * @param columns Vector of columns to be filled.
    * @param tableData Vector containing all rows filled with values
@@ -216,6 +244,14 @@ private:
                           integer const nbRows,
                           TableLayout::Section const section ) const;
 };
+
+/**
+ * @brief Convert a TableData to a table string.
+ * @param tableData The TableData to convert.
+ * @return The table string representation of the TableData.
+ */
+template<>
+string TableTextFormatter::toString< TableData >( TableData const & tableData ) const;
 }
 
 #endif
