### 2024-06-07 63ad939
Revert "Oleg s non intersecting mog"
This reverts commit 4fe3e71204c73a21ae3c158a2a64905b473f00c5.
