+++ src/coreComponents/functions/TableFunction.hpp
@@ -23,8 +23,6 @@
 
 #include "codingUtilities/EnumStrings.hpp"
 #include "LvArray/src/tensorOps.hpp"
-#include "fileIO/Table/TableData.hpp"
-#include "fileIO/Table/TableLayout.hpp"
 #include "fileIO/Table/TableFormatter.hpp"
 #include "common/Units.hpp"
 
