+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/functions/CO2Solubility.hpp
@@ -24,7 +24,7 @@
 #include "constitutive/fluid/multifluid/CO2Brine/functions/PVTFunctionHelpers.hpp"
 #include "constitutive/fluid/multifluid/Layouts.hpp"
 #include "constitutive/fluid/multifluid/MultiFluidUtils.hpp"
-#include "fileIO/Table/TableFormatter.hpp"
+#include "fileIO/Outputs/OutputBase.hpp"
 #include "functions/TableFunction.hpp"
 
 namespace geos
