### 2024-06-07 cbd1d9e
trying to improve daptive linear solver tolerance selection
