+++ src/coreComponents/physicsSolvers/SolverBase.cpp
@@ -672,22 +672,24 @@ bool SolverBase::lineSearchWithParabolicInterpolation( real64 const & time_n,
  */
 real64 SolverBase::eisenstatWalker( real64 const newNewtonNorm,
                                     real64 const oldNewtonNorm,
-                                    real64 const weakestTol )
+                                    LinearSolverParameters::Krylov const & krylovParams,
+                                    integer const logLevel )
 {
-  real64 const strongestTol = 1e-8;
-  real64 const exponent = 2.0;
-  real64 const gamma = 0.9;
-
-  real64 normRatio = newNewtonNorm / oldNewtonNorm;
-  if( normRatio > 1 )
-    normRatio = 1;
-
-  real64 newKrylovTol = gamma*std::pow( normRatio, exponent );
-  real64 altKrylovTol = gamma*std::pow( oldNewtonNorm, exponent );
+  real64 normRatio = std::min( newNewtonNorm / oldNewtonNorm, 1.0 );
+  real64 newKrylovTol = krylovParams.adaptiveGamma * std::pow( normRatio, krylovParams.adaptiveExponent );
+  // the following is a safeguard to avoid too sharp tolerance reduction
+  // the bound is the quadratic reduction wrt previous value
+  real64 altKrylovTol = std::pow( krylovParams.relTolerance, 2.0 );
 
   real64 krylovTol = std::max( newKrylovTol, altKrylovTol );
-  krylovTol = std::min( krylovTol, weakestTol );
-  krylovTol = std::max( krylovTol, strongestTol );
+  krylovTol = std::min( krylovTol, krylovParams.weakestTol );
+  krylovTol = std::max( krylovTol, krylovParams.strongestTol );
+
+  if( logLevel > 0 )
+  {
+    GEOS_LOG_RANK_0( GEOS_FMT( "        Adaptive linear tolerance = {:4.2e} (norm ratio = {:4.2e}, old tolerance = {:4.2e}, new tolerance = {:4.2e}, safeguard = {:4.2e})",
+                               krylovTol, normRatio, krylovParams.relTolerance, newKrylovTol, altKrylovTol ) );
+  }
 
   return krylovTol;
 }
@@ -944,15 +946,15 @@ bool SolverBase::solveNonlinearSystem( real64 const & time_n,
       }
     }
 
-    // if using adaptive Krylov tolerance scheme, update tolerance.
-    LinearSolverParameters::Krylov & krylovParams = m_linearSolverParameters.get().krylov;
-    if( krylovParams.useAdaptiveTol )
-    {
-      krylovParams.relTolerance = eisenstatWalker( residualNorm, lastResidual, krylovParams.weakestTol );
-    }
-
     {
       Timer timer( m_timers["linear solver total"] );
+
+      // if using adaptive Krylov tolerance scheme, update tolerance.
+      LinearSolverParameters::Krylov & krylovParams = m_linearSolverParameters.get().krylov;
+      if( krylovParams.useAdaptiveTol )
+      {
+        krylovParams.relTolerance = newtonIter > 0 ? eisenstatWalker( residualNorm, lastResidual, krylovParams, m_linearSolverParameters.getLogLevel() ) : krylovParams.weakestTol;
+      }
 
       // TODO: Trilinos currently requires this, re-evaluate after moving to Tpetra-based solvers
       if( m_precond )
