+++ src/coreComponents/physicsSolvers/LinearSolverParameters.hpp
@@ -100,6 +100,12 @@ public:
     static constexpr char const * krylovAdaptiveTolString() { return "krylovAdaptiveTol"; }
     /// Krylov weakest tolerance key
     static constexpr char const * krylovWeakTolString() { return "krylovWeakestTol"; }
+    /// Krylov strongest tolerance key
+    static constexpr char const * krylovStrongTolString() { return "krylovStrongestTol"; }
+    /// Adaptive gamma parameter key
+    static constexpr char const * adaptiveGammaString() { return "adaptiveGamma"; }
+    /// Adaptive exponent parameter key
+    static constexpr char const * adaptiveExponentString() { return "adaptiveExponent"; }
 
     /// AMG number of sweeps key
     static constexpr char const * amgNumSweepsString() { return "amgNumSweeps"; }
