+++ src/coreComponents/schema/docs/SolidInternalEnergy.rst
@@ -3,7 +3,7 @@
 ==================================== ========= ======== ================================================================================= 
 Name                                 Type      Default  Description                                                                       
 ==================================== ========= ======== ================================================================================= 
-dVolumetricHeatCapacity_dTemperature real64    required Derivative of the solid volumetric heat capacity w.r.t. temperature [J/(m^3.K^2)] 
+dVolumetricHeatCapacity_dTemperature real64    0        Derivative of the solid volumetric heat capacity w.r.t. temperature [J/(m^3.K^2)] 
 name                                 groupName required A name is required for any non-unique nodes                                       
 referenceInternalEnergy              real64    required Internal energy at the reference temperature [J/kg]                               
 referenceTemperature                 real64    required Reference temperature [K]                                                         
