### 2024-06-10 ae2f847
Add dane and ruby host configs
