### 2024-06-10 f8f800f
Added methods to choose which is the desired time format
