### 2024-06-10 5aebfad
edited timestep format to match user feedback & devs need
