### 2024-06-10 86ed597
lets try all combinations of file input path
