### 2024-06-10 6215d75
error if ones tries to use 3d plot.
