### 2024-06-11 df986fc
add linear solver params table output
