+++ src/coreComponents/physicsSolvers/LinearSolverParameters.cpp
@@ -17,6 +17,9 @@
  */
 
 #include "LinearSolverParameters.hpp"
+#include "fileIO/Table/TableLayout.hpp"
+#include "fileIO/Table/TableData.hpp"
+#include "fileIO/Table/TableFormatter.hpp"
 
 namespace geos
 {
@@ -257,6 +260,79 @@ void LinearSolverParametersInput::postProcessInput()
                         ": Invalid value." );
 
   // TODO input validation for other AMG parameters ?
+
+  if( getLogLevel() > 0 )
+    print();
+}
+
+void LinearSolverParametersInput::print()
+{
+  TableData tableData;
+  tableData.addRow( "Log level", getLogLevel());
+  tableData.addRow( "Linear solver type", EnumStrings< LinearSolverParameters::SolverType >::toString( m_parameters.solverType ) );
+  tableData.addRow( "Preconditioner type", EnumStrings< LinearSolverParameters::PreconditionerType >::toString( m_parameters.preconditionerType ) );
+  tableData.addRow( "Stop if error", m_parameters.stopIfError );
+  if( m_parameters.solverType == LinearSolverParameters::SolverType::direct )
+  {
+    tableData.addRow( "Check residual", m_parameters.direct.checkResidual );
+    tableData.addRow( "Scale rows and columns", m_parameters.direct.equilibrate );
+    tableData.addRow( "Columns permutation", EnumStrings< LinearSolverParameters::Direct::ColPerm >::toString( m_parameters.direct.colPerm ) );
+    tableData.addRow( "Rows permutation", EnumStrings< LinearSolverParameters::Direct::RowPerm >::toString( m_parameters.direct.rowPerm ) );
+    tableData.addRow( "Replace tiny pivots", m_parameters.direct.replaceTinyPivot );
+    tableData.addRow( "Perform iterative refinement", m_parameters.direct.iterativeRefine );
+    tableData.addRow( "Use parallel solver", m_parameters.direct.parallel );
+  }
+  else
+  {
+    tableData.addRow( "Maximum iterations", m_parameters.krylov.maxIterations );
+    if( m_parameters.solverType == LinearSolverParameters::SolverType::gmres ||
+        m_parameters.solverType == LinearSolverParameters::SolverType::fgmres )
+    {
+      tableData.addRow( "Maximum iterations before restart", m_parameters.krylov.maxRestart );
+    }
+    tableData.addRow( "Use adaptive tolerance", m_parameters.krylov.useAdaptiveTol );
+    if( m_parameters.krylov.useAdaptiveTol )
+    {
+      tableData.addRow( "Weakest-allowed tolerance", m_parameters.krylov.weakestTol );
+    }
+    else
+    {
+      tableData.addRow( "Relative convergence tolerance", m_parameters.krylov.relTolerance );
+    }
+  }
+  if( m_parameters.preconditionerType == LinearSolverParameters::PreconditionerType::amg )
+  {
+    tableData.addRow( "AMG", "" );
+    tableData.addRow( "  Smoother sweeps", m_parameters.amg.numSweeps );
+    tableData.addRow( "  Smoother type", EnumStrings< LinearSolverParameters::AMG::SmootherType >::toString( m_parameters.amg.smootherType ) );
+    tableData.addRow( "  Relaxation factor for the smoother", m_parameters.amg.relaxWeight );
+    tableData.addRow( "  Coarsest level solver/smoother type", EnumStrings< LinearSolverParameters::AMG::CoarseType >::toString( m_parameters.amg.coarseType ) );
+    tableData.addRow( "  Coarsening algorithm", EnumStrings< LinearSolverParameters::AMG::CoarseningType >::toString( m_parameters.amg.coarseningType ) );
[…]
