### 2024-06-11 7051142
Compilation failure on pangea-3 if ENABLE_CUDA is OFF (#3129)
