### 2024-06-11 8210a1a
Revert .integrated_tests.yaml to investigate why some (unrelated) integratedTests are failing
