+++ src/coreComponents/physicsSolvers/wavePropagation/shared/WaveSolverBase.hpp
@@ -121,6 +121,10 @@ public:
     static constexpr char const * receiverElemString() { return "receiverElem"; }
     static constexpr char const * receiverRegionString() { return "receiverRegion"; }
     static constexpr char const * freeSurfaceString() { return "FreeSurface"; }
+
+    static constexpr char const * attenuationTypeString() { return "attenuationType"; }
+    static constexpr char const * slsReferenceAngularFrequenciesString() { return "slsReferenceAngularFrequencies"; }
+    static constexpr char const * slsAnelasticityCoefficientsString() { return "slsAnelasticityCoefficients"; }
   };
 
   /**
@@ -270,7 +274,7 @@ protected:
   /// Amount of seismoTrace that will be recorded for each receiver
   localIndex m_nsamplesSeismoTrace;
 
-  /// Flag to indicate which DAS type  will be modeled
+  /// Flag to indicate which DAS type will be modeled
   WaveSolverUtils::DASType m_useDAS;
 
   /// Number of points used for strain integration for dipole DAS
@@ -287,6 +291,15 @@ protected:
 
   /// Z component of the linear DAS direction vector
   array1d< real32 > m_linearDASVectorZ;
+
+  /// Flag to indicate which attenuation type will be modeled
+  WaveSolverUtils::AttenuationType m_attenuationType;
+
+  /// Vector containing the reference frequencies for the standard-linear-solid (SLS) anelasticity model.
+  array1d< real32 > m_slsReferenceAngularFrequencies;
+
+  /// Vector containing the anelasticity coefficients for the standard-linear-solid (SLS) anelasticity model.
+  array1d< real32 > m_slsAnelasticityCoefficients;
 
   /// Indicate if we want to compute forward ou backward
   localIndex m_forward;
