+++ src/coreComponents/physicsSolvers/wavePropagation/sem/elastic/secondOrderEqn/isotropic/ElasticWaveEquationSEM.cpp
@@ -156,6 +156,20 @@ void ElasticWaveEquationSEM::registerDataOnMesh( Group & meshBodies )
                                elasticfields::StiffnessVectorz,
                                elasticfields::ElasticFreeSurfaceNodeIndicator >( getName() );
 
+    if( m_attenuationType == WaveSolverUtils::AttenuationType::sls )
+    {
+      integer l = m_slsReferenceAngularFrequencies.size( 0 );
+      nodeManager.registerField< elasticfields::DivPsix,
+                                 elasticfields::DivPsiy,
+                                 elasticfields::DivPsiz,
+                                 elasticfields::StiffnessVectorAx,
+                                 elasticfields::StiffnessVectorAy,
+                                 elasticfields::StiffnessVectorAz >( getName() );
+      nodeManager.getField< elasticfields::DivPsix >().resizeDimension< 1 >( l );
+      nodeManager.getField< elasticfields::DivPsiy >().resizeDimension< 1 >( l );
+      nodeManager.getField< elasticfields::DivPsiz >().resizeDimension< 1 >( l );
+    }
+
     FaceManager & faceManager = mesh.getFaceManager();
     faceManager.registerField< elasticfields::ElasticFreeSurfaceFaceIndicator >( getName() );
 
@@ -166,6 +180,11 @@ void ElasticWaveEquationSEM::registerDataOnMesh( Group & meshBodies )
       subRegion.registerField< elasticfields::ElasticVelocityVp >( getName() );
       subRegion.registerField< elasticfields::ElasticVelocityVs >( getName() );
       subRegion.registerField< elasticfields::ElasticDensity >( getName() );
+      if( m_attenuationType == WaveSolverUtils::AttenuationType::sls )
+      {
+        subRegion.registerField< elasticfields::ElasticQualityFactorP >( getName() );
+        subRegion.registerField< elasticfields::ElasticQualityFactorS >( getName() );
+      }
 
       if( m_useVTI )
       {
@@ -173,7 +192,6 @@ void ElasticWaveEquationSEM::registerDataOnMesh( Group & meshBodies )
         subRegion.registerField< elasticvtifields::Epsilon >( getName() );
         subRegion.registerField< elasticvtifields::Delta >( getName() );
       }
-
     } );
 
   } );
@@ -415,12 +433,53 @@ void ElasticWaveEquationSEM::initializePostInitialConditionsPreSubGroups()
                                                                              dampingz );
       } );
     } );
+
+    // check anelasticity coefficient and/or compute it if needed
+    if( m_attenuationType == WaveSolverUtils::AttenuationType::sls )
+    {
+      real32 minQVal = computeGlobalMinQFactor();
+      if( m_slsAnelasticityCoefficients.size( 0 ) == 1 && m_slsAnelasticityCoefficients[ 0 ] < 0 )
+      {
+        m_slsAnelasticityCoefficients[ 0 ] = 2.0 * minQVal / ( minQVal - 1.0 );
+      }
+      // test if anelasticity is too high and artifacts could appear
+      real32 ySum = 0.0;
+      for( integer l = 0; l < m_slsAnelasticityCoefficients.size( 0 ); l++ )
+      {
+        ySum += m_slsAnelasticityCoefficients[ l ];
+      }
+      GEOS_WARNING_IF( ySum > minQVal, "The anelasticity parameters are too high for the given quality factor. This could lead to solution artifacts such as zero-velocity waves." );
+    }
+
   } );
 
[…]
