+++ src/coreComponents/unitTests/wavePropagationTests/testWavePropagationAttenuation.cpp
@@ -1,0 +1,240 @@
+/*
+ * ------------------------------------------------------------------------------------------------------------
+ * SPDX-License-Identifier: LGPL-2.1-only
+ *
+ * Copyright (c) 2018-2020 Lawrence Livermore National Security LLC
+ * Copyright (c) 2018-2020 The Board of Trustees of the Leland Stanford Junior University
+ * Copyright (c) 2018-2020 Total, S.A
+ * Copyright (c) 2020-     GEOSX Contributors
+ * All right reserved
+ *
+ * See top level LICENSE, COPYRIGHT, CONTRIBUTORS, NOTICE, and ACKNOWLEDGEMENTS files for details.
+ * ------------------------------------------------------------------------------------------------------------
+ */
+
+// using some utility classes from the following unit test
+#include "unitTests/fluidFlowTests/testCompFlowUtils.hpp"
+
+#include "common/DataTypes.hpp"
+#include "mainInterface/initialization.hpp"
+#include "mainInterface/ProblemManager.hpp"
+#include "mesh/DomainPartition.hpp"
+#include "mainInterface/GeosxState.hpp"
+#include "physicsSolvers/PhysicsSolverManager.hpp"
+#include "physicsSolvers/wavePropagation/shared/WaveSolverBase.hpp"
+#include "physicsSolvers/wavePropagation/sem/elastic/secondOrderEqn/isotropic/ElasticWaveEquationSEM.hpp"
+
+#include <gtest/gtest.h>
+
+using namespace geos;
+using namespace geos::dataRepository;
+using namespace geos::testing;
+
+CommandLineOptions g_commandLineOptions;
+
+// This unit test checks the interpolation done to extract seismic traces from a wavefield.
+// It computes a seismogram at a receiver co-located with the source and compares it to the surrounding receivers.
+char const * xmlInput =
+  R"xml(
+  <Problem>
+    <Solvers>
+      <ElasticSEM
+        name="elasticSolver"
+        cflFactor="0.25"
+        discretization="FE1"
+        targetRegions="{ Region }"
+        sourceCoordinates="{ { 50, 50, 50 } }"
+        timeSourceFrequency="2"
+        useDAS="strainIntegration"
+        attenuationType="sls"
+        slsReferenceAngularFrequencies="{ 69.6283, 592.177 }" 
+        slsAnelasticityCoefficients="{ 1.63675, 1.75153 }" 
+        linearDASSamples="5"
+        linearDASGeometry="{ { 0, 0, 10 }, { 0, 0, 10 }, { 0, 0, 10 }, { 0, 0, 10 },
+                                { 0, 0, 10 }, { 0, 0, 10 }, { 0, 0, 10 }, { 0, 0, 10 },
+                                { 0, 0, 10 } }"
+        receiverCoordinates="{ { 0.1, 0.1, 0.1 }, { 0.1, 0.1, 99.9 }, { 0.1, 99.9, 0.1 }, { 0.1, 99.9, 99.9 },
+                                { 99.9, 0.1, 0.1 }, { 99.9, 0.1, 99.9 }, { 99.9, 99.9, 0.1 }, { 99.9, 99.9, 99.9 },
+                                { 50, 50, 50 } }"
+        outputSeismoTrace="0"
+        dtSeismoTrace="0.1"/>
+    </Solvers>
+    <Mesh>
+      <InternalMesh
+        name="mesh"
+        elementTypes="{ C3D8 }"
+        xCoords="{ 0, 100 }"
+        yCoords="{ 0, 100 }"
+        zCoords="{ 0, 100 }"
+        nx="{ 1 }"
+        ny="{ 1 }"
+        nz="{ 1 }"
[…]
