### 2024-06-11 010fc34
Viscoelastic wave propagator (#3080)
* first version of viscoelastic propagator. Anelasticity coefficient are not automatically calculated.

* removed unused variables

* added default computation of anelasticity parameter and check on Q vs. y values

* added unit test and integrated test. Also added the forgotten DAS integrated test to the ats config file

* corrected bug in input string

* bugfix

* uncrustify + unused var

* split unit test for code coverage

* getting aroung compilation error in debug mode

* typo

* change after merge

* leftover modifications from merge

* switched VTI a
[…]
