+++ src/coreComponents/physicsSolvers/LinearSolverParameters.cpp
@@ -274,8 +274,8 @@ void LinearSolverParametersInput::print()
   {
     tableData.addRow( "Check residual", m_parameters.direct.checkResidual );
     tableData.addRow( "Scale rows and columns", m_parameters.direct.equilibrate );
-    tableData.addRow( "Columns permutation", EnumStrings< LinearSolverParameters::Direct::ColPerm >::toString( m_parameters.direct.colPerm ) );
-    tableData.addRow( "Rows permutation", EnumStrings< LinearSolverParameters::Direct::RowPerm >::toString( m_parameters.direct.rowPerm ) );
+    tableData.addRow( "Columns permutation", m_parameters.direct.colPerm );
+    tableData.addRow( "Rows permutation", m_parameters.direct.rowPerm );
     tableData.addRow( "Replace tiny pivots", m_parameters.direct.replaceTinyPivot );
     tableData.addRow( "Perform iterative refinement", m_parameters.direct.iterativeRefine );
     tableData.addRow( "Use parallel solver", m_parameters.direct.parallel );
@@ -302,19 +302,19 @@ void LinearSolverParametersInput::print()
   {
     tableData.addRow( "AMG", "" );
     tableData.addRow( "  Smoother sweeps", m_parameters.amg.numSweeps );
-    tableData.addRow( "  Smoother type", EnumStrings< LinearSolverParameters::AMG::SmootherType >::toString( m_parameters.amg.smootherType ) );
+    tableData.addRow( "  Smoother type", m_parameters.amg.smootherType );
     tableData.addRow( "  Relaxation factor for the smoother", m_parameters.amg.relaxWeight );
-    tableData.addRow( "  Coarsest level solver/smoother type", EnumStrings< LinearSolverParameters::AMG::CoarseType >::toString( m_parameters.amg.coarseType ) );
-    tableData.addRow( "  Coarsening algorithm", EnumStrings< LinearSolverParameters::AMG::CoarseningType >::toString( m_parameters.amg.coarseningType ) );
-    tableData.addRow( "  Interpolation algorithm", EnumStrings< LinearSolverParameters::AMG::InterpType >::toString( m_parameters.amg.interpolationType ) );
+    tableData.addRow( "  Coarsest level solver/smoother type", m_parameters.amg.coarseType );
+    tableData.addRow( "  Coarsening algorithm", m_parameters.amg.coarseningType );
+    tableData.addRow( "  Interpolation algorithm", m_parameters.amg.interpolationType );
     tableData.addRow( "  Interpolation maximum number of nonzeros per row", m_parameters.amg.interpolationMaxNonZeros );
     tableData.addRow( "  Number of functions", m_parameters.amg.numFunctions );
     tableData.addRow( "  Number of paths for aggressive coarsening", m_parameters.amg.aggressiveNumPaths );
     tableData.addRow( "  Number of levels for aggressive coarsening", m_parameters.amg.aggressiveNumLevels );
-    tableData.addRow( "  Aggressive interpolation algorithm", EnumStrings< LinearSolverParameters::AMG::AggInterpType >::toString( m_parameters.amg.aggressiveInterpType ) );
+    tableData.addRow( "  Aggressive interpolation algorithm", m_parameters.amg.aggressiveInterpType );
     tableData.addRow( "  Strength-of-connection threshold", m_parameters.amg.threshold );
     tableData.addRow( "  Apply separate component filter for multi-variable problems", m_parameters.amg.separateComponents );
-    tableData.addRow( "  Near null space approximation", EnumStrings< LinearSolverParameters::AMG::NullSpaceType >::toString( m_parameters.amg.nullSpaceType ) );
[…]
