+++ src/coreComponents/physicsSolvers/NonlinearSolverParameters.cpp
@@ -14,6 +14,7 @@
 
 #include "NonlinearSolverParameters.hpp"
 #include "common/Logger.hpp"
+#include "fileIO/Table/TableFormatter.hpp"
 
 namespace geos
 {
@@ -187,45 +188,50 @@ void NonlinearSolverParameters::postProcessInput()
                         getWrapperDataContext( viewKeysStruct::lineSearchResidualFactorString() ) << ": should be positive" );
 
   if( getLogLevel() > 0 )
-  {
     print();
-  }
 }
 
 void NonlinearSolverParameters::print() const
 {
-  GEOS_LOG_RANK_0( GEOS_FMT( "{}: nonlinear solver parameters", getParent().getName() ) );
-  GEOS_LOG_RANK_0( GEOS_FMT( "  Log level = {}", getLogLevel() ) );
-  GEOS_LOG_RANK_0( GEOS_FMT( "  Line search action = {}", EnumStrings< LineSearchAction >::toString( m_lineSearchAction ) ) );
+  TableData tableData;
+  tableData.addRow( "Log level", getLogLevel());
+  tableData.addRow( "Line search", "" );
+  tableData.addRow( "  Action", m_lineSearchAction );
   if( m_lineSearchAction != LineSearchAction::None )
   {
-    GEOS_LOG_RANK_0( GEOS_FMT( "  Line search interpolation type = {}", EnumStrings< LineSearchInterpolationType >::toString( m_lineSearchInterpType ) ) );
-    GEOS_LOG_RANK_0( GEOS_FMT( "  Line search maximum number of cuts = {}", m_lineSearchMaxCuts ) );
-    GEOS_LOG_RANK_0( GEOS_FMT( "  Line search cut factor = {}", m_lineSearchCutFactor ) );
-    GEOS_LOG_RANK_0( GEOS_FMT( "  Line search starting iteration = {}", m_lineSearchStartingIteration ) );
-    GEOS_LOG_RANK_0( GEOS_FMT( "  Line search residual increase factor = {}", m_lineSearchResidualFactor ) );
+    tableData.addRow( "  Interpolation type", m_lineSearchInterpType );
+    tableData.addRow( "  Maximum number of cuts", m_lineSearchMaxCuts );
+    tableData.addRow( "  Cut factor", m_lineSearchCutFactor );
+    tableData.addRow( "  Starting iteration", m_lineSearchStartingIteration );
+    tableData.addRow( "  Residual increase factor", m_lineSearchResidualFactor );
   }
-  GEOS_LOG_RANK_0( GEOS_FMT( "  Norm type (flow solver) = {}", EnumStrings< solverBaseKernels::NormType >::toString( m_normType ) ) );
-  GEOS_LOG_RANK_0( GEOS_FMT( "  Minimum residual normalizer = {}", m_minNormalizer ) );
-  GEOS_LOG_RANK_0( GEOS_FMT( "  Convergence tolerance = {}", m_newtonTol ) );
-  GEOS_LOG_RANK_0( GEOS_FMT( "  Maximum iterations = {}", m_maxIterNewton ) );
-  GEOS_LOG_RANK_0( GEOS_FMT( "  Minimum iterations = {}", m_minIterNewton ) );
-  GEOS_LOG_RANK_0( GEOS_FMT( "  Maximum allowed residual norm = {}", m_maxAllowedResidualNorm ) );
-  GEOS_LOG_RANK_0( GEOS_FMT( "  Allow non-converged = {}", m_allowNonConverged ) );
-  GEOS_LOG_RANK_0( GEOS_FMT( "  Time step decrease iterations limit = {}", m_timeStepDecreaseIterLimit ) );
[…]
