+++ src/coreComponents/constitutive/fluid/multifluid/MultiFluidSelector.hpp
@@ -47,6 +47,8 @@ void constitutiveUpdatePassThru( MultiFluidBase const & fluid,
                                CO2BrinePhillipsFluid,
                                CO2BrineEzrokhiFluid,
                                CO2BrinePhillipsThermalFluid,
+// Including these in a CUDA build will lead to compiler segfault.
+// Need to split compilation units for all the options
 #if !defined(GEOS_DEVICE_COMPILE)
                                CO2BrineEzrokhiThermalFluid,
                                CompositionalTwoPhasePengRobinsonLBCViscosity,
@@ -69,6 +71,8 @@ void constitutiveUpdatePassThru( MultiFluidBase & fluid,
                                CO2BrinePhillipsFluid,
                                CO2BrineEzrokhiFluid,
                                CO2BrinePhillipsThermalFluid,
+// Including these in a CUDA build will lead to compiler segfault.
+// Need to split compilation units for all the options"
 #if !defined(GEOS_DEVICE_COMPILE)
                                CO2BrineEzrokhiThermalFluid,
                                CompositionalTwoPhasePengRobinsonLBCViscosity,
