+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3080-5511-8210a1a
+  baseline: integratedTests/baseline_integratedTests-pr3140-5514-fe26f80
 
 allow_fail:
   all: ''
