+++ src/coreComponents/unitTests/constitutiveTests/MultiFluidTest.hpp
@@ -1,0 +1,529 @@
+/*
+ * ------------------------------------------------------------------------------------------------------------
+ * SPDX-License-Identifier: LGPL-2.1-only
+ *
+ * Copyright (c) 2018-2020 Lawrence Livermore National Security LLC
+ * Copyright (c) 2018-2020 The Board of Trustees of the Leland Stanford Junior University
+ * Copyright (c) 2018-2020 TotalEnergies
+ * Copyright (c) 2019-     GEOSX Contributors
+ * All rights reserved
+ *
+ * See top level LICENSE, COPYRIGHT, CONTRIBUTORS, NOTICE, and ACKNOWLEDGEMENTS files for details.
+ * ------------------------------------------------------------------------------------------------------------
+ */
+
+/**
+ * @file MultiFluidTest.hpp
+ */
+
+#ifndef GEOS_UNITTESTS_CONSTITUTIVETESTS_MULTIFLUIDTEST_HPP_
+#define GEOS_UNITTESTS_CONSTITUTIVETESTS_MULTIFLUIDTEST_HPP_
+
+#include "constitutiveTestHelpers.hpp"
+#include "constitutive/fluid/multifluid/MultiFluidBase.hpp"
+#include "constitutive/fluid/multifluid/MultiFluidFields.hpp"
+#include "constitutive/fluid/multifluid/MultiFluidSelector.hpp"
+
+using namespace geos::dataRepository;
+using namespace geos::constitutive;
+using namespace geos::constitutive::multifluid;
+
+namespace geos
+{
+namespace testing
+{
+
+template< typename FLUID_TYPE >
+struct UsePVTPackage
+{
+  static constexpr bool value = false;
+};
+
+template< integer N >
+using Feed = std::array< real64, N >;
+
+template< integer N, typename ARRAY1, typename ARRAY2 >
+void copy( ARRAY1 const & inputArray, ARRAY2 & outputArray )
+{
+  for( integer i = 0; i < N; ++i )
+  {
+    outputArray[i] = inputArray[i];
+  }
+}
+
+template< integer N, typename T >
+void fill( array1d< T > & outputArray, std::array< T const, N > const inputArray )
+{
+  for( integer i = 0; i < N; ++i )
+  {
+    outputArray.emplace_back( inputArray[i] );
+  }
+}
+
+template< integer N, integer U, typename T >
+void fill( arraySlice1d< T, U > const & outputArray, std::array< T const, N > const inputArray )
+{
+  for( integer i = 0; i < N; ++i )
+  {
+    outputArray[i] = inputArray[i];
+  }
+}
+
+template< integer NUM_PHASE, integer NUM_COMP >
+struct MultiFluidTestData
+{
+  MultiFluidTestData( real64 const pressure_,
+                      real64 const temperature_,
+                      Feed< NUM_COMP > && composition_ );
+
+  MultiFluidTestData( real64 const pressure_,
+                      real64 const temperature_,
+                      arraySlice1d< real64 const > const & composition_ );
+
+  MultiFluidTestData( real64 const pressure_,
+                      real64 const temperature_,
+                      Feed< NUM_COMP > && composition_,
+                      real64 const totalDensity_,
+                      Feed< NUM_PHASE > && phaseFraction_,
+                      Feed< NUM_PHASE > && phaseDensity_,
+                      Feed< NUM_PHASE > && phaseMassDensity_,
+                      Feed< NUM_PHASE > && phaseViscosity_,
+                      Feed< NUM_PHASE > && phaseEnthalpy_,
[…]
