+++ BASELINE_NOTES.md
@@ -6,6 +6,10 @@ This file is designed to track changes to the integrated test baselines.
 Any developer who updates the baseline ID in the .integrated_tests.yaml file is expected to create an entry in this file with the pull request number, date, and their justification for rebaselining.
 These notes should be in reverse-chronological order, and use the following time format: (YYYY-MM-DD).
 
+PR #3140 (2024-06-11)
+======================
+Fixed derivative in EzrokhiBrineDensity
+
 PR #3080 (2024-06-07)
 =====================
 Rebaseline after adding viscoelastic wave propagator.
