### 2024-06-11 bba7659
Added variables for future features, refactored nodes to nodeManager in solver, added option to disable explicit normals and positions on particle damage
