+++ src/coreComponents/physicsSolvers/solidMechanics/SolidMechanicsMPM.cpp
@@ -68,30 +68,6 @@ bool compareFloat( real64 a, real64 b, real64 epsilon){
   return std::fabs( a - b ) < epsilon;
 }
 
-// template< typename T >
-// void generateCombinationsUtil( std::vector<std::vector<T>>& result, 
-//                                std::vector<T>& current,
-//                                std::vector<std::vector<T>>& arrays, int depth) {
-//     if (depth == arrays.size()) {
-//         result.push_back(current);
-//         return;
-//     }
-
-//     for (int i = 0; i < arrays[depth].size(); ++i) {
-//         current.push_back(arrays[depth][i]);
-//         generateCombinationsUtil(result, current, arrays, depth + 1);
-//         current.pop_back();
-//     }
-// }
-
-// template< typename T >
-// vector<vector<T>> generateCombinations(std::vector<std::vector<T>>& arrays) {
-//     std::vector<std::vector<T>> result;
-//     std::vector<T> current;
-//     generateCombinationsUtil(result, current, arrays, 0);
-//     return result;
-// }
-
 // Flattened combinations function to avoid performance hit from recursive calls
 GEOS_HOST_DEVICE
 array2d< int > generateCombinations(array1d< array1d< int > > sets)
@@ -201,6 +177,7 @@ SolidMechanicsMPM::SolidMechanicsMPM( const string & name,
   m_cpdiDomainScaling( 0 ),
   m_subdivideParticles( 0 ),
   m_disableSurfaceNormalsAndPositionsOnCPDIScaling( 0 ),
+  m_disableSurfaceNormalsAndPositionsOnDamage( 0 ),
   m_smallMass( DBL_MAX ),
   m_numContactGroups( 0 ),
   m_numContactFlags(),
@@ -692,6 +669,12 @@ SolidMechanicsMPM::SolidMechanicsMPM( const string & name,
     setRestartFlags( RestartFlags::NO_WRITE ).
     setDescription( "Option for disabling explicit surface normals and positions when a particle has severely deformed" );
 
+  registerWrapper( "disableSurfaceNormalsAndPositionsOnDamage", &m_disableSurfaceNormalsAndPositionsOnDamage ).
+    setInputFlag( InputFlags::OPTIONAL ).
+    setDefaultValue( m_disableSurfaceNormalsAndPositionsOnDamage ).
+    setRestartFlags( RestartFlags::NO_WRITE ).
+    setDescription( "Option for disabling explicit surface normals and positions when a particle has been severely damaged" );
+
   registerWrapper( "generalizedVortexMMS", &m_generalizedVortexMMS ).
     setInputFlag( InputFlags::OPTIONAL ).
     setDefaultValue( m_generalizedVortexMMS ).
@@ -909,6 +892,12 @@ SolidMechanicsMPM::SolidMechanicsMPM( const string & name,
     setRestartFlags( RestartFlags::WRITE_AND_READ ).
     setDescription( "number of elements along partition directions" );
   
+  registerWrapper( "implicitContinuumFluidPressure", &m_implicitContinuumFluidPressure).
+    setInputFlag( InputFlags::OPTIONAL ).
+    setApplyDefaultValue( 0.0 ).
+    setRestartFlags( RestartFlags::NO_WRITE ).
+    setDescription( "Pressure of implicit continuum fluid" );
+
   m_mpmEventManager = &registerGroup< MPMEventManager >( groupKeys.mpmEventManager );
 }
 
@@ -1280,182 +1269,192 @@ void SolidMechanicsMPM::registerDataOnMesh( Group & meshBodies )
     }
     else // Background grid field registration
     {
-      NodeManager & nodes = meshLevel.getNodeManager();
[…]
