+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3140-5514-fe26f80
+  baseline: integratedTests/baseline_integratedTests-pr3138-5523-ffe12f3
 
 allow_fail:
   all: ''
