+++ src/coreComponents/finiteElement/elementFormulations/H1_Hexahedron_Lagrange1_GaussLegendre2.hpp
@@ -76,7 +76,7 @@ public:
   /// The number of nodes/support points per element.
   constexpr static localIndex numNodes = LagrangeBasis1::TensorProduct3D::numSupportPoints;
 
-  /// The number of nodes/support points per element.
+  /// The number of faces/support points for bubble functions per element.
   constexpr static localIndex numFaces = LagrangeBasis1::TensorProduct3D::numSupportFaces;
 
   /// The maximum number of support points per element.
@@ -214,7 +214,7 @@ public:
   }
 
   /**
-   * @brief Calculate shape functions values for each support face at a
+   * @brief Calculate face bubble functions values for each face at a
    *   given point in the parent space.
    * @param pointCoord coordinates of the given point.
    * @param N An array to pass back the shape function values for each support
@@ -229,7 +229,7 @@ public:
   }
 
   /**
-   * @brief Calculate shape functions values for each support face at a
+   * @brief Calculate face bubble functions values for each face at a
    *   quadrature point.
    * @param q Index of the quadrature point.
    * @param N An array to pass back the shape function values for each support
@@ -281,7 +281,7 @@ public:
                            real64 ( &gradN )[numNodes][3] );
 
   /**
-   * @brief Calculate the shape bubble function derivatives wrt the physical
+   * @brief Calculate the bubble function derivatives wrt the physical
    *   coordinates.
    * @param q Index of the quadrature point.
    * @param X Array containing the coordinates of the support points.
