+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3138-5523-ffe12f3
+  baseline: integratedTests/baseline_integratedTests-pr3138-5526-f053f0a
 
 allow_fail:
   all: ''
