### 2024-06-11 ed3cbce
Properly sync nonlinear solver params for coupled solver (#3138)
Co-authored-by: Dickson Kachuma <81433670+dkachuma@users.noreply.github.com>
Co-authored-by: Ryan Aronson <raronson0031@gmail.com>
