+++ BASELINE_NOTES.md
@@ -6,13 +6,20 @@ This file is designed to track changes to the integrated test baselines.
 Any developer who updates the baseline ID in the .integrated_tests.yaml file is expected to create an entry in this file with the pull request number, date, and their justification for rebaselining.
 These notes should be in reverse-chronological order, and use the following time format: (YYYY-MM-DD).
 
+PR #3138 (2024-06-11)
+======================
+Properly sync nonlinear solver params for coupled solver. Baseline update mostly due to number of iterations change in baseline files.
+
+
 PR #3140 (2024-06-11)
 ======================
 Fixed derivative in EzrokhiBrineDensity
 
+
 PR #3080 (2024-06-07)
 =====================
 Rebaseline after adding viscoelastic wave propagator.
+
 
 PR #3075 (2024-06-05)
 =====================
