### 2024-06-11 c1d12a9
Properly sync nonlinear solver params for coupled solver (#3138)
Co-authored-by: Dickson Kachuma <81433670+dkachuma@users.noreply.github.com>
Co-authored-by: Ryan Aronson <raronson0031@gmail.com>
