+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3140-5514-fe26f80
+  baseline: integratedTests/baseline_integratedTests-pr3138-5526-f053f0a
 
 allow_fail:
   all: ''
