+++ src/coreComponents/physicsSolvers/multiphysics/SinglePhasePoromechanicsConformingFractures.cpp
@@ -142,8 +142,6 @@ void SinglePhasePoromechanicsConformingFractures::assembleSystem( real64 const t
 
   solidMechanicsSolver()->synchronizeFractureState( domain );
 
-  synchronizeNonLinearParameters();
-
   assembleElementBasedContributions( time_n,
                                      dt,
                                      domain,
