+++ src/coreComponents/physicsSolvers/wavePropagation/shared/PrecomputeSourcesAndReceiversKernel.hpp
@@ -33,13 +33,13 @@ struct PreComputeSourcesAndReceivers
    * @tparam FE_TYPE finite element type
    * @param[in] size the number of cells in the subRegion
    * @param[in] numFacesPerElem number of faces per element
+   * @param[in] facesToNodes face to node map
    * @param[in] nodeCoords coordinates of the nodes
+   * @param[in] nodeLocalToGlobal local to global index map for nodes
    * @param[in] elemGhostRank rank of the ghost element
    * @param[in] elemsToNodes map from element to nodes
    * @param[in] elemsToFaces map from element to faces
    * @param[in] elemCenter coordinates of the element centers
-   * @param[in] faceNormal normal of each faces
-   * @param[in] faceCenter coordinates of the center of a face
    * @param[in] sourceCoordinates coordinates of the source terms
    * @param[out] sourceIsAccessible flag indicating whether the source is accessible or not
    * @param[out] sourceNodeIds indices of the nodes of the element where the source is located
@@ -58,13 +58,13 @@ struct PreComputeSourcesAndReceivers
   static void
   Compute1DSourceAndReceiverConstants( localIndex const size,
                                        localIndex const numFacesPerElem,
+                                          ArrayOfArraysView< localIndex const > const facesToNodes,
                                        arrayView2d< WaveSolverBase::wsCoordType const, nodes::REFERENCE_POSITION_USD > const nodeCoords,
+                                              arrayView1d< globalIndex const > const nodeLocalToGlobal,
                                        arrayView1d< integer const > const elemGhostRank,
                                        arrayView2d< localIndex const, cells::NODE_MAP_USD > const & elemsToNodes,
                                        arrayView2d< localIndex const > const elemsToFaces,
                                        arrayView2d< real64 const > const & elemCenter,
-                                       arrayView2d< real64 const > const faceNormal,
-                                       arrayView2d< real64 const > const faceCenter,
                                        arrayView2d< real64 const > const sourceCoordinates,
                                        arrayView1d< localIndex > const sourceIsAccessible,
                                        arrayView2d< localIndex > const sourceNodeIds,
@@ -97,14 +97,13 @@ struct PreComputeSourcesAndReceivers
           real64 const coords[3] = { sourceCoordinates[isrc][0],
                                      sourceCoordinates[isrc][1],
                                      sourceCoordinates[isrc][2] };
-
-          bool const sourceFound =
-            WaveSolverUtils::locateSourceElement( numFacesPerElem,
-                                                  center,
-                                                  faceNormal,
-                                                  faceCenter,
-                                                  elemsToFaces[k],
-                                                  coords );
+          bool const sourceFound = 
+              computationalGeometry::isPointInsideConvexPolyhedronRobust( nodeCoords,
+                                                                          elemsToFaces[ k ],
+                                                                          facesToNodes,
+                                                                          nodeLocalToGlobal,
+                                                                  center,  
+                                                                          coords ); 
           if( sourceFound )
           {
             real64 coordsOnRefElem[3]{};
@@ -146,13 +145,13 @@ struct PreComputeSourcesAndReceivers
                                      receiverCoordinates[ircv][2] };
 
           real64 coordsOnRefElem[3]{};
-          bool const receiverFound =
-            WaveSolverUtils::locateSourceElement( numFacesPerElem,
-                                                  center,
-                                                  faceNormal,
-                                                  faceCenter,
-                                                  elemsToFaces[k],
-                                                  coords );
+          bool const receiverFound = 
+              computationalGeometry::isPointInsideConvexPolyhedronRobust( nodeCoords,
+                                                                    elemsToFaces[ k ],
+                                                                    facesToNodes,
+                                                                    nodeLocalToGlobal,
+                                                                  center,  
[…]
