+++ src/coreComponents/physicsSolvers/fluidFlow/CompositionalMultiphaseStatistics.hpp
@@ -78,47 +78,166 @@ private:
     /// String for the region statistics
     constexpr static char const * regionStatisticsString() { return "regionStatistics"; }
     /// String for the relperm threshold
-    constexpr static char const * relpermThresholdString() { return "relpermThreshold"; }
+    constexpr static char const * relpermThresholdString() { return "relpermThreshold";}
   };
 
-  struct RegionStatistics
+  class RegionStatistics : public dataRepository::Group 
   {
+    public:
+      RegionStatistics( string const & name,
+                        Group * const parent )
+      : Group( name, parent )
+    {
+      registerWrapper( viewKeyStruct::averagePressureString(), &m_averagePressure ).
+        setApplyDefaultValue( 0 ).
+        //setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+        setDescription( "average region pressure" );
+
+      registerWrapper( viewKeyStruct::minPressureString(), &m_minPressure ).
+        setApplyDefaultValue( 0 ).
+        //setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+        setDescription( "minimum region pressure" );
+
+      registerWrapper( viewKeyStruct::maxPressureString(), &m_maxPressure ).
+        setApplyDefaultValue( 0 ).
+        //setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+        setDescription( "maximum region pressure" );
+
+
+      registerWrapper( viewKeyStruct::minDeltaPressureString(), &m_minDeltaPressure ).
+        setApplyDefaultValue( 0 ).
+        //setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+        setDescription( "minimum region delta pressure" );
+
+      registerWrapper( viewKeyStruct::maxDeltaPressureString(), &m_maxDeltaPressure ).
+        setApplyDefaultValue( 0 ).
+        //setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+        setDescription( "maximum region delta pressure" );
+
+
+      registerWrapper( viewKeyStruct::averageTemperatureString(), &m_averageTemperature ).
+        setApplyDefaultValue( 0 ).
+        //setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+        setDescription( "average region temperature" );
+
+      registerWrapper( viewKeyStruct::minTemperatureString(), &m_minTemperature ).
+        setApplyDefaultValue( 0 ).
+        //setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+        setDescription( "minimum region temperature" );
+
+      registerWrapper( viewKeyStruct::maxTemperatureString(), &m_maxTemperature ).
+        setApplyDefaultValue( 0 ).
+        //setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+        setDescription( "maximum region temperature" );
+
+
+      registerWrapper( viewKeyStruct::totalPoreVolumeString(), &m_totalPoreVolume ).
+        setApplyDefaultValue( 0 ).
+        //setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+        setDescription( "total region pore volume" );
+
+      registerWrapper( viewKeyStruct::totalUncompactedPoreVolumeString(), &m_totalUncompactedPoreVolume ).
+        setApplyDefaultValue( 0 ).
+        //setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+        setDescription( "total region uncompacted pore volume" );
+
+
+      registerWrapper( viewKeyStruct::phasePoreVolumeString(), &m_phasePoreVolume ).
+        setApplyDefaultValue( 0 ).
+        //setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+        setDescription( "Phase region phase pore volume" );
+
+
[…]
