+++ src/coreComponents/constitutive/unitTests/testLohrenzBrayClarkViscosity.cpp
@@ -68,7 +68,7 @@ public:
   {
     ComponentProperties const & componentProperties = this->m_fluid->getComponentProperties();
     m_parameters = LohrenzBrayClarkViscosity::createParameters( nullptr );
-    auto * parameters = const_cast< LohrenzBrayClarkViscosity::Parameters * >(m_parameters->getParameters< LohrenzBrayClarkViscosity::Parameters >());
+    auto * parameters = const_cast< LohrenzBrayClarkViscosity::Parameters * >(m_parameters->get< LohrenzBrayClarkViscosity::Parameters >());
     parameters->m_componentCriticalVolume.resize( NC );
     TestFluid< 9 >::populateArray( parameters->m_componentCriticalVolume, this->m_fluid->criticalVolume );
     m_density = std::make_unique< CompositionalDensity< CubicEOSPhaseModel< PengRobinsonEOS > > >( "PhaseDensity", componentProperties, 0, *m_parameters );
