+++ src/coreComponents/constitutive/unitTests/testCompositionalDensity.cpp
@@ -53,7 +53,7 @@ struct FluidData< 9 >
   }
 };
 
-template< int NC, typename EOS_TYPE >
+template< int NC, EquationOfStateType EOS_TYPE >
 class CompositionalDensityTestFixture :  public ::testing::TestWithParam< DensityData< NC > >
 {
   static constexpr real64 relTol = 1.0e-5;
@@ -66,8 +66,11 @@ public:
     : m_fluid( FluidData< NC >::createFluid() )
   {
     ComponentProperties const & componentProperties = this->m_fluid->getComponentProperties();
-    m_parameters = CompositionalDensity< EOS_TYPE >::createParameters( nullptr );
-    m_density = std::make_unique< CompositionalDensity< EOS_TYPE > >( "PhaseDensity", componentProperties, 0, *m_parameters );
+    m_parameters = CompositionalDensity::createParameters( nullptr );
+    auto * equationOfState = const_cast< EquationOfState * >(m_parameters->get< EquationOfState >());
+    equationOfState->m_equationsOfStateNames.resize( 1 );
+    equationOfState->m_equationsOfStateNames[0] = EnumStrings< EquationOfStateType >::toString( EOS_TYPE );
+    m_density = std::make_unique< CompositionalDensity >( "PhaseDensity", componentProperties, 0, *m_parameters );
   }
 
   ~CompositionalDensityTestFixture() = default;
@@ -198,27 +201,28 @@ public:
   }
 
 protected:
-  std::unique_ptr< CompositionalDensity< EOS_TYPE > > m_density{};
+  std::unique_ptr< CompositionalDensity > m_density{};
   std::unique_ptr< TestFluid< NC > > m_fluid{};
   std::unique_ptr< ModelParameters > m_parameters{};
 };
 
-using CompositionalDensity9CompPR = CompositionalDensityTestFixture< 9, CubicEOSPhaseModel< PengRobinsonEOS > >;
-using CompositionalDensity9CompSRK = CompositionalDensityTestFixture< 9, CubicEOSPhaseModel< SoaveRedlichKwongEOS > >;
+using PengRobinson = CompositionalDensityTestFixture< 9, EquationOfStateType::PengRobinson >;
+using SoaveRedlichKwong = CompositionalDensityTestFixture< 9, EquationOfStateType::SoaveRedlichKwong >;
 
-TEST_P( CompositionalDensity9CompPR, testDensityDerivatives )
+TEST_P( PengRobinson, testDensityDerivatives )
 {
   testDensityDerivatives( GetParam() );
 }
 
-TEST_P( CompositionalDensity9CompSRK, testDensityDerivatives )
+TEST_P( SoaveRedlichKwong, testDensityDerivatives )
 {
   testDensityDerivatives( GetParam() );
 }
+
+/* UNCRUSTIFY-OFF */
 
 INSTANTIATE_TEST_SUITE_P(
-  CompositionalDensityTest,
-  CompositionalDensity9CompPR,
+  CompositionalDensityTest, PengRobinson,
   ::testing::ValuesIn( {
       DensityData< 9 >{1.839590e+06, 2.971500e+02, {0.009000, 0.003000, 0.534700, 0.114600, 0.087900, 0.045600, 0.020900, 0.015100, 0.169200}, 8.355571e+03, 4.559906e+02 },
[…]
