### 2024-06-12 2670fd5
Zeroed reference normals and positions so deactivated explicit normals and positions would be visible in plotting
