+++ src/coreComponents/physicsSolvers/fluidFlow/IsothermalCompositionalMultiphaseBaseKernels.hpp
@@ -1409,8 +1409,6 @@ public:
       if( absCompDensChange > maxRelCompDensChange && absCompDensChange > eps )
       {
         real64 const compScalingFactor = maxRelCompDensChange / absCompDensChange;
-        std::cout << ei << " " << ic << " compScalingFactor = " << compScalingFactor << " maxRelCompDensChange = " << maxRelCompDensChange
-                  << " absCompDensChange = " << absCompDensChange << " m_compDens[ei][ic] = " << m_compDens[ei][ic] << std::endl;
         m_compDensScalingFactor[ei] = LvArray::math::min( m_compDensScalingFactor[ei], compScalingFactor );
         if( stack.localMinVal > compScalingFactor )
         {
