+++ src/docs/JOSS/paper.md
@@ -102,6 +102,8 @@ authors:
     affiliation: 4
   - name: Lionel Untereiner
     affiliation: 8
+  - name: Arturo Vargas
+    affiliation: 1
   - name: Sohail Waziri
     affiliation: "3,4"
   - name: Xianhuan Wen
