### 2024-06-12 e86a05f
lagrange contact solver needs to save traction fields for restart. bug resolved
