+++ src/coreComponents/physicsSolvers/multiphysics/SinglePhaseReservoirAndWells.hpp
@@ -83,38 +83,6 @@ public:
                                       CRSMatrixView< real64, globalIndex const > const & localMatrix,
                                       arrayView1d< real64 > const & localRhs ) override;
 
-  void
-  assembleFluxTerms( real64 const dt,
-                     DomainPartition const & domain,
-                     DofManager const & dofManager,
-                     CRSMatrixView< real64, globalIndex const > const & localMatrix,
-                     arrayView1d< real64 > const & localRhs ) const
-  { flowSolver()->assembleFluxTerms( dt, domain, dofManager, localMatrix, localRhs );  }
-
-  void
-  assembleStabilizedFluxTerms( real64 const dt,
-                               DomainPartition const & domain,
-                               DofManager const & dofManager,
-                               CRSMatrixView< real64, globalIndex const > const & localMatrix,
-                               arrayView1d< real64 > const & localRhs ) const
-  { flowSolver()->assembleStabilizedFluxTerms( dt, domain, dofManager, localMatrix, localRhs );  }
-
-  void setKeepFlowVariablesConstantDuringInitStep( bool const keepFlowVariablesConstantDuringInitStep )
-  { flowSolver()->setKeepFlowVariablesConstantDuringInitStep( keepFlowVariablesConstantDuringInitStep ); }
-
-  void updateFluidState( ElementSubRegionBase & subRegion ) const
-  { flowSolver()->updateFluidState( subRegion ); }
-  void updatePorosityAndPermeability( CellElementSubRegion & subRegion ) const
-  { flowSolver()->updatePorosityAndPermeability( subRegion ); }
-  void updateSolidInternalEnergyModel( ObjectManagerBase & dataGroup ) const
-  { flowSolver()->updateSolidInternalEnergyModel( dataGroup ); }
-
-  integer & isThermal() { return flowSolver()->isThermal(); }
-
-  void enableFixedStressPoromechanicsUpdate() { flowSolver()->enableFixedStressPoromechanicsUpdate(); }
-
-  void enableJumpStabilization() {flowSolver()->enableJumpStabilization();}
-
   virtual void saveSequentialIterationState( DomainPartition & domain ) override { flowSolver()->saveSequentialIterationState( domain ); }
 
 protected:
