### 2024-06-12 4239500
remove RESQML support from GEOS (#2969)
