### 2024-06-12 3028fcb
saveSequentialIterationState
