+++ src/coreComponents/physicsSolvers/multiphysics/CoupledReservoirAndWellsBase.hpp
@@ -255,6 +255,8 @@ public:
   void setKeepFlowVariablesConstantDuringInitStep( bool const keepFlowVariablesConstantDuringInitStep )
   { reservoirSolver()->setKeepFlowVariablesConstantDuringInitStep( keepFlowVariablesConstantDuringInitStep ); }
 
+  virtual void saveSequentialIterationState( DomainPartition & domain ) override { reservoirSolver()->saveSequentialIterationState( domain ); }
+
 protected:
 
   /**
