### 2024-06-13 14e96e0
Refactor the StencilOutput so the PackCollection can read it. Point is to allow to do an HDF output, and remove the CSV one.
