+++ src/coreComponents/physicsSolvers/fluidFlow/StencilDataCollection.hpp
@@ -1,0 +1,201 @@
+/*
+ * ------------------------------------------------------------------------------------------------------------
+ * SPDX-License-Identifier: LGPL-2.1-only
+ *
+ * Copyright (c) 2018-2020 Lawrence Livermore National Security LLC
+ * Copyright (c) 2018-2020 The Board of Trustees of the Leland Stanford Junior University
+ * Copyright (c) 2018-2020 TotalEnergies
+ * Copyright (c) 2019-     GEOSX Contributors
+ * All rights reserved
+ *
+ * See top level LICENSE, COPYRIGHT, CONTRIBUTORS, NOTICE, and ACKNOWLEDGEMENTS files for details.
+ * ------------------------------------------------------------------------------------------------------------
+ */
+
+/**
+ * @file StencilDataCollection.hpp
+ */
+
+#ifndef SRC_CORECOMPONENTS_PHYSICSSOLVERS_FLUIDFLOW_CELLTOCELLDATACOLLECTOR_HPP_
+#define SRC_CORECOMPONENTS_PHYSICSSOLVERS_FLUIDFLOW_CELLTOCELLDATACOLLECTOR_HPP_
+
+#include "events/tasks/TaskBase.hpp"
+#include "physicsSolvers/fluidFlow/FlowSolverBase.hpp"
+
+namespace geos
+{
+
+/**
+ * @class StencilDataCollection
+ * Task class allowing the ouput of communicating elements data.
+ * Only current usage is to output the cell-cell transmissibility (= "CellToCellDataCollection"), but we could add more
+ * implementations to output the transmissibility of boundary or surface connections (or everything that is managed
+ * by the stencils).
+ */
+class StencilDataCollection : public TaskBase
+{
+public:
+
+  /**
+   * @brief Constructor for the statistics class
+   * @param[in] name the name of the task coming from the xml
+   * @param[in] parent the parent group of the task
+   */
+  StencilDataCollection( const string & name,
+                         dataRepository::Group * const parent );
+
+  /// Accessor for the catalog name
+  static string catalogName() { return "CellToCellDataCollection"; }
+
+  /**
+   * @defgroup Tasks Interface Functions
+   *
+   * This function implements the interface defined by the abstract TaskBase class
+   */
+  /**@{*/
+
+  virtual bool execute( real64 const time_n,
+                        real64 const dt,
+                        integer const cycleNumber,
+                        integer const eventCounter,
+                        real64 const eventProgress,
+                        DomainPartition & domain ) override;
+
+  /**@}*/
+
+private:
+
+  using Base = TaskBase;
+  using LocalToGlobalMap = ElementRegionManager::ElementViewAccessor< arrayView1d< globalIndex const > >;
+
+  class Kernel;
+
+  /**
+   * @brief Element-element connection data extracted from the kernel.
+   */
+  struct KernelConnectionData
+  {
+    real64 m_transmissibility[2];
+    localIndex m_regionId[2];
+    localIndex m_subRegionId[2];
+    localIndex m_elementId[2];
+  };
+
+  /**
+   * @brief Element-element connection data to be output by the class.
+   */
+  struct ConnectionData
+  {
+    real64 m_transmissibility[2];
+    globalIndex m_globalId[2];
+
+
+    /**
+     * @return Construct a ConnectionData from the given kernel data.
+     * @param kernelData the kernel extracted data to be converted.
+     * @param localToGlobalMap the local to global id map, useful to convert the kernel extracted ids of the elements.
+     */
+    static ConnectionData fromKernel( KernelConnectionData const & kernelData,
+                                      LocalToGlobalMap const & localToGlobalMap );
+
+    /**
+     * @brief minus operator to be able to sort instances before outputing them.
+     * The sorting is done by the first global id, then by the second global id.
[…]
