+++ src/coreComponents/schema/docs/Tasks.rst
@@ -3,7 +3,7 @@
 =========================================================== ==== ======= ====================================================================== 
 Name                                                        Type Default Description                                                            
 =========================================================== ==== ======= ====================================================================== 
-CellToCellOutput                                            node         :ref:`XML_CellToCellOutput`                                            
+CellToCellDataCollection                                    node         :ref:`XML_CellToCellDataCollection`                                    
 CompositionalMultiphaseReservoirPoromechanicsInitialization node         :ref:`XML_CompositionalMultiphaseReservoirPoromechanicsInitialization` 
 CompositionalMultiphaseStatistics                           node         :ref:`XML_CompositionalMultiphaseStatistics`                           
 MultiphasePoromechanicsInitialization                       node         :ref:`XML_MultiphasePoromechanicsInitialization`                       
