+++ src/coreComponents/schema/docs/CellToCellOutput.rst
@@ -1,12 +1,0 @@
-
-
-============== ============ ======== ========================================================================================================== 
-Name           Type         Default  Description                                                                                                
-============== ============ ======== ========================================================================================================== 
-flowSolverName groupNameRef required Name of the flow solver                                                                                    
-logLevel       integer      0        When higher than 1: Display output events details.                                                         
-name           groupName    required A name is required for any non-unique nodes                                                                
-writeCSV       integer      0        Enable the CSV output for each targeted solver discretisation scheme under the folder defined by writeCSV. 
-============== ============ ======== ========================================================================================================== 
-
-

[…]
