+++ src/coreComponents/schema/docs/CompositionalMultiphaseHybridFVM.rst
@@ -18,6 +18,7 @@ maxSequentialCompDensChange               real64             1        Maximum (a
 maxSequentialPressureChange               real64             100000   Maximum (absolute) pressure change in a sequential iteration, used for outer loop convergence check                                                                                                                                                                                                                      
 maxSequentialTemperatureChange            real64             0.1      Maximum (absolute) temperature change in a sequential iteration, used for outer loop convergence check                                                                                                                                                                                                                   
 minCompDens                               real64             1e-10    Minimum allowed global component density                                                                                                                                                                                                                                                                                 
+minScalingFactor                          real64             0.01     Minimum value for solution scaling factor                                                                                                                                                                                                                                                                                
 name                                      groupName          required A name is required for any non-unique nodes                                                                                                                                                                                                                                                                              
 solutionChangeScalingFactor               real64             0.5      Damping factor for solution change targets                                                                                                                                                                                                                                                                               
 targetFlowCFL                             real64             -1       Target CFL condition `CFL condition <http://en.wikipedia.org/wiki/Courant-Friedrichs-Lewy_condition>`_when computing the next timestep.                                                                                                                                                                                  
