+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3138-5526-f053f0a
+  baseline: integratedTests/baseline_integratedTests-pr3152-5558-5c6e97f
 
 allow_fail:
   all: ''
