+++ src/coreComponents/physicsSolvers/LinearSolverParameters.cpp
@@ -327,7 +327,7 @@ void LinearSolverParametersInput::print()
   }
   TableLayout const tableLayout = TableLayout( {
       TableLayout::ColumnParam{"Parameter", TableLayout::Alignment::left},
-      TableLayout::ColumnParam{"Value", TableLayout::Alignment::center},
+      TableLayout::ColumnParam{"Value", TableLayout::Alignment::left},
     }, GEOS_FMT( "{}: linear solver", getParent().getName() ) );
   TableTextFormatter const tableFormatter( tableLayout );
   GEOS_LOG_RANK_0( tableFormatter.toString( tableData ));
