+++ src/coreComponents/physicsSolvers/fluidFlow/wells/SinglePhaseWell.hpp
@@ -274,8 +274,8 @@ private:
    * @param subRegion the well subRegion
    */
   virtual void validateWellConstraints( real64 const & time_n,
-                                real64 const & dt,
-                                WellElementSubRegion const & subRegion ) override;
+                                        real64 const & dt,
+                                        WellElementSubRegion const & subRegion ) override;
 
 };
 
