### 2024-06-13 91acdaf
Implement linear solve
