### 2024-06-13 b2d55d8
run integrated tests on dedicated runner to avoid overloading of streak2 (#3171)
