+++ src/coreComponents/mesh/NodeManager.cpp
@@ -193,6 +193,36 @@ void NodeManager::setGeometricalRelations( CellBlockManagerABC const & cellBlock
   elemRegionManager.forElementRegionsComplete< SurfaceElementRegion >( connectNodesTo2dElements );
 }
 
+void NodeManager::setGeometricalRelations( generators::MeshMappings const & meshMappings,
+                                           ElementRegionManager const & elemRegionManager )
+{
+  GEOS_MARK_FUNCTION;
+  auto const & nodeMgr = meshMappings.getNodeMgr();
+
+  resize( nodeMgr.numNodes() );
+
+  m_referencePosition = nodeMgr.getNodePositions();
+
+  // TODO I add the copy of the g2l and l2g mappings here. This is new
+  m_localToGlobalMap = nodeMgr.getLocalToGlobal();
+  m_globalToLocalMap = nodeMgr.getGlobalToLocal();
+
+  // TODO not for there, but it's convenient
+  copyExchangeInfo( meshMappings.getNeighbors(), nodeMgr.getGhostRank(), nodeMgr.getSend(), nodeMgr.getRecv() );
+
+  m_toEdgesRelation.base().assimilate< parallelHostPolicy >( nodeMgr.getNodeToEdges(),
+                                                             LvArray::sortedArrayManipulation::UNSORTED_NO_DUPLICATES );
+  m_toFacesRelation.base().assimilate< parallelHostPolicy >( nodeMgr.getNodeToFaces(),
+                                                             LvArray::sortedArrayManipulation::UNSORTED_NO_DUPLICATES );
+
+  ToCellRelation< ArrayOfArrays< localIndex > > const toCellBlock = nodeMgr.getNodeToElements();
+  array2d< localIndex > const blockToSubRegion = elemRegionManager.getCellBlockToSubRegionMap( meshMappings.getCellMgr().getCellBlks() ); // TODO This already exists in FaceManager
+  meshMapUtilities::transformCellBlockToRegionMap< parallelHostPolicy >( blockToSubRegion.toViewConst(),
+                                                                         toCellBlock,
+                                                                         m_toElements );
+  // TODO add the fracture component
+}
+
 void NodeManager::setupRelatedObjectsInRelations( EdgeManager const & edgeManager,
                                                   FaceManager const & faceManager,
                                                   ElementRegionManager const & elemRegionManager )
