### 2024-06-14 8705452
remove elemRegionManager from setGeometricalRelations and use the cell block to sub region mapping directly.
