+++ src/coreComponents/mesh/NodeManager.hpp
@@ -169,10 +169,10 @@ public:
   /**
    * @brief Initialise the current @c NodeManager with the information from @c meshMappings.
    * @param[in] meshMappings Provides all the geometrical mappings.
-   * @param[in] elemRegionManager Provides the mapping from the incoming geometrical cell blocks to their sub-region twin.
+   * @param[in] cb2sr The mapping from the incoming geometrical cell blocks to their sub-region twin.
    */
   void setGeometricalRelations( generators::MeshMappings const & meshMappings,
-                                ElementRegionManager const & elemRegionManager );
+                                arrayView2d< localIndex const > const & cb2sr );
 
   /**
    * @brief Link the current manager to other managers.
