### 2024-06-14 b72f7b2
Update BASELINE_NOTES.md
