### 2024-06-14 c3c2904
Preparatory work for fractures + wells (#3021)
Co-authored-by: Jian HUANG <jian.huang@totalenergies.com>
