+++ src/coreComponents/physicsSolvers/multiphysics/MultiphasePoromechanics.cpp
@@ -28,6 +28,8 @@
 #include "physicsSolvers/solidMechanics/SolidMechanicsFields.hpp"
 #include "physicsSolvers/solidMechanics/SolidMechanicsLagrangianFEM.hpp"
 #include "physicsSolvers/solidMechanics/kernels/ImplicitSmallStrainQuasiStatic.hpp"
+//#include "physicsSolvers/contact/SolidMechanicsLagrangeContact.hpp"
+//#include "physicsSolvers/contact/SolidMechanicsEmbeddedFractures.hpp"
 
 namespace geos
 {
@@ -37,9 +39,9 @@ using namespace constitutive;
 using namespace fields;
 using namespace stabilization;
 
-template< typename FLOW_SOLVER >
-MultiphasePoromechanics< FLOW_SOLVER >::MultiphasePoromechanics( const string & name,
-                                                                 Group * const parent )
+template< typename FLOW_SOLVER, typename MECHANICS_SOLVER >
+MultiphasePoromechanics< FLOW_SOLVER, MECHANICS_SOLVER >::MultiphasePoromechanics( const string & name,
+                                                                                   Group * const parent )
   : Base( name, parent )
 {
   this->registerWrapper( viewKeyStruct::stabilizationTypeString(), &m_stabilizationType ).
@@ -66,8 +68,8 @@ MultiphasePoromechanics< FLOW_SOLVER >::MultiphasePoromechanics( const string & 
   linearSolverParameters.dofsPerNode = 3;
 }
 
-template< typename FLOW_SOLVER >
-void MultiphasePoromechanics< FLOW_SOLVER >::postProcessInput()
+template< typename FLOW_SOLVER, typename MECHANICS_SOLVER >
+void MultiphasePoromechanics< FLOW_SOLVER, MECHANICS_SOLVER >::postProcessInput()
 {
   Base::postProcessInput();
 
@@ -79,8 +81,8 @@ void MultiphasePoromechanics< FLOW_SOLVER >::postProcessInput()
                            ));
 }
 
-template< typename FLOW_SOLVER >
-void MultiphasePoromechanics< FLOW_SOLVER >::registerDataOnMesh( Group & meshBodies )
+template< typename FLOW_SOLVER, typename MECHANICS_SOLVER >
+void MultiphasePoromechanics< FLOW_SOLVER, MECHANICS_SOLVER >::registerDataOnMesh( Group & meshBodies )
 {
   Base::registerDataOnMesh( meshBodies );
 
@@ -110,22 +112,22 @@ void MultiphasePoromechanics< FLOW_SOLVER >::registerDataOnMesh( Group & meshBod
   }
 }
 
-template< typename FLOW_SOLVER >
-void MultiphasePoromechanics< FLOW_SOLVER >::setupCoupling( DomainPartition const & GEOS_UNUSED_PARAM( domain ),
-                                                            DofManager & dofManager ) const
+template< typename FLOW_SOLVER, typename MECHANICS_SOLVER >
+void MultiphasePoromechanics< FLOW_SOLVER, MECHANICS_SOLVER >::setupCoupling( DomainPartition const & GEOS_UNUSED_PARAM( domain ),
+                                                                              DofManager & dofManager ) const
 {
   dofManager.addCoupling( solidMechanics::totalDisplacement::key(),
                           CompositionalMultiphaseBase::viewKeyStruct::elemDofFieldString(),
                           DofManager::Connector::Elem );
 }
 
-template< typename FLOW_SOLVER >
-void MultiphasePoromechanics< FLOW_SOLVER >::assembleSystem( real64 const time,
-                                                             real64 const dt,
-                                                             DomainPartition & domain,
-                                                             DofManager const & dofManager,
-                                                             CRSMatrixView< real64, globalIndex const > const & localMatrix,
[…]
