+++ src/coreComponents/physicsSolvers/multiphysics/CompositionalMultiphaseReservoirAndWells.hpp
@@ -27,7 +27,7 @@ namespace geos
 {
 
 /// @tparam RESERVOIR_SOLVER compositional flow or compositional poromechanics solver
-template< typename RESERVOIR_SOLVER >
+template< typename RESERVOIR_SOLVER = CompositionalMultiphaseBase >
 class CompositionalMultiphaseReservoirAndWells : public CoupledReservoirAndWellsBase< RESERVOIR_SOLVER,
                                                                                       CompositionalMultiphaseWell >
 {
