### 2024-06-14 e592755
Preparatory work for fractures + wells (#3021)
Co-authored-by: Jian HUANG <jian.huang@totalenergies.com>
