+++ src/coreComponents/physicsSolvers/multiphysics/SinglePhaseReservoirAndWells.hpp
@@ -27,7 +27,7 @@ namespace geos
 {
 
 /// @tparam RESERVOIR_SOLVER single-phase flow or single-phase poromechanics solver
-template< typename RESERVOIR_SOLVER >
+template< typename RESERVOIR_SOLVER = SinglePhaseBase >
 class SinglePhaseReservoirAndWells : public CoupledReservoirAndWellsBase< RESERVOIR_SOLVER,
                                                                           SinglePhaseWell >
 {
@@ -90,6 +90,22 @@ public:
                      CRSMatrixView< real64, globalIndex const > const & localMatrix,
                      arrayView1d< real64 > const & localRhs ) const
   { flowSolver()->assembleFluxTerms( dt, domain, dofManager, localMatrix, localRhs );  }
+  void
+  assembleHydrofracFluxTerms( real64 const time_n,
+                              real64 const dt,
+                              DomainPartition const & domain,
+                              DofManager const & dofManager,
+                              CRSMatrixView< real64, globalIndex const > const & localMatrix,
+                              arrayView1d< real64 > const & localRhs,
+                              CRSMatrixView< real64, localIndex const > const & dR_dAper )
+  { flowSolver()->assembleHydrofracFluxTerms( time_n, dt, domain, dofManager, localMatrix, localRhs, dR_dAper ); }
+
+  template< typename SUBREGION_TYPE >
+  void accumulationAssemblyLaunch( DofManager const & dofManager,
+                                   SUBREGION_TYPE const & subRegion,
+                                   CRSMatrixView< real64, globalIndex const > const & localMatrix,
+                                   arrayView1d< real64 > const & localRhs )
+  { flowSolver()->accumulationAssemblyLaunch( dofManager, subRegion, localMatrix, localRhs ); }
 
   void setKeepFlowVariablesConstantDuringInitStep( bool const keepFlowVariablesConstantDuringInitStep )
   { flowSolver()->setKeepFlowVariablesConstantDuringInitStep( keepFlowVariablesConstantDuringInitStep ); }
@@ -106,6 +122,11 @@ public:
   void enableFixedStressPoromechanicsUpdate() { flowSolver()->enableFixedStressPoromechanicsUpdate(); }
 
   virtual void saveSequentialIterationState( DomainPartition & domain ) override { flowSolver()->saveSequentialIterationState( domain ); }
+
+  void prepareStencilWeights( DomainPartition & domain ) const
+  { flowSolver()->prepareStencilWeights( domain ); }
+  void updateStencilWeights( DomainPartition & domain ) const
+  { flowSolver()->updateStencilWeights( domain ); }
 
 protected:
 
