+++ src/coreComponents/unitTests/wellsTests/testReservoirSinglePhaseMSWells.cpp
@@ -136,7 +136,7 @@ char const * PostXmlInput =
   )xml";
 
 template< typename LAMBDA >
-void testNumericalJacobian( SinglePhaseReservoirAndWells< SinglePhaseBase > & solver,
+void testNumericalJacobian( SinglePhaseReservoirAndWells<> & solver,
                             DomainPartition & domain,
                             real64 const perturbParameter,
                             real64 const relTol,
@@ -345,7 +345,7 @@ protected:
 
   void SetUp() override
   {
-    solver = &state.getProblemManager().getPhysicsSolverManager().getGroup< SinglePhaseReservoirAndWells< SinglePhaseBase > >( "reservoirSystem" );
+    solver = &state.getProblemManager().getPhysicsSolverManager().getGroup< SinglePhaseReservoirAndWells<> >( "reservoirSystem" );
 
     DomainPartition & domain = state.getProblemManager().getDomainPartition();
 
@@ -423,7 +423,7 @@ protected:
   static real64 constexpr EPS = std::numeric_limits< real64 >::epsilon();
 
   GeosxState state;
-  SinglePhaseReservoirAndWells< SinglePhaseBase > * solver;
+  SinglePhaseReservoirAndWells<> * solver;
 };
 
 real64 constexpr SinglePhaseReservoirSolverTest::TIME;
