+++ BASELINE_NOTES.md
@@ -6,6 +6,11 @@ This file is designed to track changes to the integrated test baselines.
 Any developer who updates the baseline ID in the .integrated_tests.yaml file is expected to create an entry in this file with the pull request number, date, and their justification for rebaselining.
 These notes should be in reverse-chronological order, and use the following time format: (YYYY-MM-DD).
 
+PR #3021 (2024-06-13)
+======================
+Preparatory work for fractures + wells. New test case added.
+
+
 PR #3152 (2024-06-13)
 ======================
 Some random things. Baseline update because of the new parameter (minScalingFactor).
@@ -39,7 +44,6 @@ Add missing compositionalMultiphaseFlow tests into ATS and adjust output naming.
 PR #3113 (2024-06-05)
 =====================
 Add general version updateConfiguration. Rebaseline of some edfm cases is needed.
-
 
 PR #3050 (2024-05-20)
 =====================
