+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3152-5558-5c6e97f
+  baseline: integratedTests/baseline_integratedTests-pr3021-5589-3c278fb
 
 allow_fail:
   all: ''
