+++ src/coreComponents/common/DataTypes.cpp
@@ -185,8 +185,8 @@ rtTypes::RegexMapType rtTypes::createBasicTypesRegexMap()
 
   string_view const groupNameDesc = "Input value must be a string that cannot be empty and contains only upper/lower letters, digits, and the characters  . - _";
   string_view const groupNameRegex = "[a-zA-Z0-9.\\-_]+";
-  string_view const groupNameRefDesc = "Input value must be a string that can contain only upper/lower letters, digits, and the characters  . - _ /";
-  string_view const groupNameRefRegex = "[a-zA-Z0-9.\\-_/]*";
+  string_view const groupNameRefDesc = "Input value must be a string that can contain only upper/lower letters, digits, and the characters  . - _ / *";
+  string_view const groupNameRefRegex = "[a-zA-Z0-9.\\-_/*]*";
 
 
   // Build master list of regexes
