+++ src/coreComponents/unitTests/fluidFlowTests/CMakeLists.txt
@@ -1,10 +1,9 @@
 # Specify list of tests
-
-set(gtest_geosx_tests
-        testSinglePhaseBaseKernels.cpp
-        testThermalCompMultiphaseFlow.cpp
-        testThermalSinglePhaseFlow.cpp
-        )
+set( gtest_geosx_tests
+     testSinglePhaseBaseKernels.cpp
+     testThermalCompMultiphaseFlow.cpp
+     testThermalSinglePhaseFlow.cpp
+     testFlowStatistics.cpp )
 
 set( dependencyList ${parallelDeps} gtest )
 
@@ -14,6 +13,7 @@ else()
   list( APPEND dependencyList ${geosx_core_libs} )
 endif()
 
+set( dependencyList testingUtilities )
 
 if( ENABLE_PVTPackage )
     list( APPEND gtest_geosx_tests

[…]
