+++ src/coreComponents/schema/docs/CompositionalMultiphaseFVM.rst
@@ -36,7 +36,7 @@ targetRelativePressureChangeInTimeStep    real64                                
 targetRelativeTemperatureChangeInTimeStep real64                                      0.2      Target (relative) change in temperature in a time step (expected value between 0 and 1)                                                                                                                                                                                                                                  
 temperature                               real64                                      required Temperature                                                                                                                                                                                                                                                                                                              
 useDBC                                    integer                                     0        Enable Dissipation-based continuation flux                                                                                                                                                                                                                                                                               
-useMass                                   integer                                     0        Use mass formulation instead of molar                                                                                                                                                                                                                                                                                    
+useMass                                   integer                                     0        Use mass formulation instead of molar. Warning : Affects SourceFlux rates units.                                                                                                                                                                                                                                         
 useSimpleAccumulation                     integer                                     1        Flag indicating whether simple accumulation form is used                                                                                                                                                                                                                                                                 
 useTotalMassEquation                      integer                                     1        Flag indicating whether total mass equation is used                                                                                                                                                                                                                                                                      
 LinearSolverParameters                    node                                        unique   :ref:`XML_LinearSolverParameters`                                                                                                                                                                                                                                                                                        
