+++ src/coreComponents/schema/docs/SinglePhaseFVM.rst
@@ -1,22 +1,26 @@
 
 
-============================== ================== ======== ======================================================================================================================================================================================================================================================================================================================== 
-Name                           Type               Default  Description                                                                                                                                                                                                                                                                                                              
-============================== ================== ======== ======================================================================================================================================================================================================================================================================================================================== 
-allowNegativePressure          integer            1        Flag indicating if negative pressure is allowed                                                                                                                                                                                                                                                                          
-cflFactor                      real64             0.5      Factor to apply to the `CFL condition <http://en.wikipedia.org/wiki/Courant-Friedrichs-Lewy_condition>`_ when calculating the maximum allowable time step. Values should be in the interval (0,1]                                                                                                                        
-discretization                 groupNameRef       required Name of discretization object (defined in the :ref:`NumericalMethodsManager`) to use for this solver. For instance, if this is a Finite Element Solver, the name of a :ref:`FiniteElement` should be specified. If this is a Finite Volume Method, the name of a :ref:`FiniteVolume` discretization should be specified. 
-initialDt                      real64             1e+99    Initial time-step value required by the solver to the event manager.                                                                                                                                                                                                                                                     
-isThermal                      integer            0        Flag indicating whether the problem is thermal or not.                                                                                                                                                                                                                                                                   
-logLevel                       integer            0        Log level                                                                                                                                                                                                                                                                                                                
-maxAbsolutePressureChange      real64             -1       Maximum (absolute) pressure change in a Newton iteration                                                                                                                                                                                                                                                                 
-maxSequentialPressureChange    real64             100000   Maximum (absolute) pressure change in a sequential iteration, used for outer loop convergence check                                                                                                                                                                                                                      
-maxSequentialTemperatureChange real64             0.1      Maximum (absolute) temperature change in a sequential iteration, used for outer loop convergence check                                                                                                                                                                                                                   
-name                           groupName          required A name is required for any non-unique nodes                                                                                                                                                                                                                                                                              
-targetRegions                  groupNameRef_array required Allowable regions that the solver may be applied to. Note that this does not indicate that the solver will be applied to these regions, only that allocation will occur such that the solver may be applied to these regions. The decision about what regions this solver will beapplied to rests in the EventManager.   
-temperature                    real64             0        Temperature                                                                                                                                                                                                                                                                                                              
-LinearSolverParameters         node               unique   :ref:`XML_LinearSolverParameters`                                                                                                                                                                                                                                                                                        
-NonlinearSolverParameters      node               unique   :ref:`XML_NonlinearSolverParameters`                                                                                                                                                                                                                                                                                     
-============================== ================== ======== ======================================================================================================================================================================================================================================================================================================================== 
+============================== ================== ======== ======================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================== 
+Name                           Type               Default  Description                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
+============================== ================== ======== ======================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================== 
+allowNegativePressure          integer            1        Flag indicating if negative pressure is allowed                                                                                                                                                                                                                                                                                                                                                                                                                                                          
+cflFactor                      real64             0.5      Factor to apply to the `CFL condition <http://en.wikipedia.org/wiki/Courant-Friedrichs-Lewy_condition>`_ when calculating the maximum allowable time step. Values should be in the interval (0,1]                                                                                                                                                                                                                                                                                                        
[…]
