+++ src/coreComponents/constitutive/fluid/multifluid/reactive/ReactiveBrineFluid.cpp
@@ -204,12 +204,14 @@ void ReactiveBrineFluid< PHASE > ::createPVTModels( bool isClone )
                  GEOS_FMT( "{}: PVT model {} not found in input files", getFullName(), PHASE::Enthalpy::catalogName() ),
                  InputError );
 
-  bool const writeCSV = !isClone && m_writeCSV;
-  bool const writeInLog = !isClone && (getLogLevel() > 0 && logger::internal::rank==0);
+  PVTFunctionBase::PVTOutputOptions pvtOpts = {
+    !isClone && m_writeCSV,// writeCSV
+    !isClone && (getLogLevel() >= 0 && logger::internal::rank==0), // writeInLog
+  };
 
   // then, we are ready to instantiate the phase models
   m_phase = std::make_unique< PHASE >( getName() + "_phaseModel1", phase1InputParams, m_componentNames, m_componentMolarWeight,
-                                       writeCSV, writeInLog );
+                                       pvtOpts );
 }
 
 template< typename PHASE >
