+++ src/coreComponents/dataRepository/ConduitRestart.hpp
@@ -55,7 +55,6 @@ struct conduitTypeInfo
 {};
 
 // Native integer types
-CONDUIT_TYPE_INFO( bool, CONDUIT_NATIVE_UNSIGNED_CHAR );
 CONDUIT_TYPE_INFO( char, CONDUIT_NATIVE_CHAR );
 CONDUIT_TYPE_INFO( signed char, CONDUIT_NATIVE_SIGNED_CHAR );
 CONDUIT_TYPE_INFO( unsigned char, CONDUIT_NATIVE_UNSIGNED_CHAR );
