+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/functions/PhillipsBrineViscosity.hpp
@@ -83,8 +83,7 @@ public:
                           string_array const & inputPara,
                           string_array const & componentNames,
                           array1d< real64 > const & componentMolarWeight,
-                          bool const printInCsv,
-                          bool const printInLog );
+                          PVTOutputOptions pvtOpts );
 
   virtual ~PhillipsBrineViscosity() override = default;
 

[…]
