### 2024-06-14 7f1b8ab
pvtOpt struct renaming + fct move to FlashModel
