+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/functions/PhillipsBrineDensity.hpp
@@ -85,7 +85,7 @@ public:
                         string_array const & inputParams,
                         string_array const & componentNames,
                         array1d< real64 > const & componentMolarWeight,
-                        PVTOutputOptions pvtOpts );
+                        TableOutputOptions pvtOpts );
 
   static string catalogName() { return "PhillipsBrineDensity"; }
 
