+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/functions/FenghourCO2Viscosity.hpp
@@ -75,7 +75,7 @@ public:
                         string_array const & inputParams,
                         string_array const & componentNames,
                         array1d< real64 > const & componentMolarWeight,
-                        PVTOutputOptions pvtOpts );
+                        TableOutputOptions pvtOpts );
 
   virtual ~FenghourCO2Viscosity() override = default;
 
