### 2024-06-14 7c70795
fix linear interpolation of velocities
