+++ src/coreComponents/fileIO/Outputs/VTKOutput.cpp
@@ -122,8 +122,8 @@ void VTKOutput::postProcessInput()
                         catalogName(), getDataContext(),
                         std::to_string( m_fieldNames.size() ), fieldNamesString, m_plotLevel ) );
 
-  GEOS_ERROR_IF( m_writeFaceElementsAs3D ), GEOS_FMT( "{} `{}`: 3D vtk plot of faceElements is not yet supported.",
-                                                      catalogName(), getDataContext() );
+  GEOS_ERROR_IF( m_writeFaceElementsAs3D, GEOS_FMT( "{} `{}`: 3D vtk plot of faceElements is not yet supported.",
+                                                    catalogName(), getDataContext() );
 }
 
 
