+++ src/coreComponents/constitutive/unitTests/testKValueInitialization.cpp
@@ -83,7 +83,6 @@ private:
       discarded,
       criticalPressure,
       criticalTemperature,
-      discarded,
       acentricFactor,
       discarded,
       discarded2d );
