+++ src/coreComponents/constitutive/fluid/multifluid/compositional/models/ConstantViscosity.cpp
@@ -17,6 +17,7 @@
  */
 
 #include "ConstantViscosity.hpp"
+#include "constitutive/fluid/multifluid/MultiFluidBase.hpp"
 
 namespace geos
 {
@@ -28,14 +29,65 @@ namespace compositional
 {
 
 ConstantViscosity::ConstantViscosity( string const & name,
-                                      ComponentProperties const & componentProperties ):
+                                      ComponentProperties const & componentProperties,
+                                      integer const phaseIndex,
+                                      ModelParameters const & modelParameters ):
   FunctionBase( name, componentProperties )
+{
+  Parameters const * parameters = modelParameters.get< Parameters >();
+  m_constantPhaseViscosity = parameters->m_constantPhaseViscosity[phaseIndex];
+}
+
+ConstantViscosityUpdate::ConstantViscosityUpdate( real64 const constantPhaseViscosity ):
+  m_constantPhaseViscosity( constantPhaseViscosity )
 {}
 
 ConstantViscosity::KernelWrapper
 ConstantViscosity::createKernelWrapper() const
 {
-  return KernelWrapper( );
+  return KernelWrapper( m_constantPhaseViscosity );
+}
+
+std::unique_ptr< ModelParameters >
+ConstantViscosity::createParameters( std::unique_ptr< ModelParameters > parameters )
+{
+  if( parameters && parameters->get< Parameters >() != nullptr )
+  {
+    return parameters;
+  }
+  return std::make_unique< Parameters >( std::move( parameters ) );
+}
+
+ConstantViscosity::Parameters::Parameters( std::unique_ptr< ModelParameters > parameters ):
+  ModelParameters( std::move( parameters ) )
+{}
+
+void ConstantViscosity::Parameters::registerParametersImpl( MultiFluidBase * fluid )
+{
+  fluid->registerWrapper( viewKeyStruct::constantPhaseViscosityString(), &m_constantPhaseViscosity ).
+    setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+    setDescription( "Constant phase viscosity" );
+}
+
+void ConstantViscosity::Parameters::postProcessInputImpl( MultiFluidBase const * fluid,
+                                                          ComponentProperties const & componentProperties )
+{
+  GEOS_UNUSED_VAR( componentProperties );
+
+  integer const numPhase = fluid->numFluidPhases();
+
+  if( m_constantPhaseViscosity.empty() )
+  {
+    m_constantPhaseViscosity.resize( numPhase );
+    for( integer ip = 0; ip < numPhase; ++ip )
+    {
+      m_constantPhaseViscosity[ip] = ConstantViscosity::defaultViscosity;
+    }
+  }
+  GEOS_THROW_IF_NE_MSG( m_constantPhaseViscosity.size(), numPhase,
+                        GEOS_FMT( "{}: invalid number of values in attribute '{}'", fluid->getFullName(),
+                                  viewKeyStruct::constantPhaseViscosityString() ),
+                        InputError );
 }
 
 } // end namespace compositional

[…]
