+++ src/coreComponents/fileIO/Outputs/VTKOutput.cpp
@@ -123,7 +123,7 @@ void VTKOutput::postProcessInput()
                         std::to_string( m_fieldNames.size() ), fieldNamesString, m_plotLevel ) );
 
   GEOS_ERROR_IF( m_writeFaceElementsAs3D, GEOS_FMT( "{} `{}`: 3D vtk plot of faceElements is not yet supported.",
-                                                    catalogName(), getDataContext() );
+                                                    catalogName(), getDataContext() ) );
 }
 
 
