+++ src/coreComponents/unitTests/wellsTests/testThermalReservoirCompositionalMultiphaseSSWells.cpp
@@ -202,8 +202,8 @@ char const * xmlInput =
       phaseNames="{ gas, water }"
       componentNames="{ co2, water }"
       componentMolarWeight="{ 44e-3, 18e-3 }"
-      phasePVTParaFiles="{ /Users/byer3/GEOS-DEV/mass_inj_const/src/coreComponents/unitTests/wellsTests/pvtgas.txt, /Users/byer3/GEOS-DEV/mass_inj_const/src/coreComponents/unitTests/wellsTests/pvtliquid.txt }"
-      flashModelParaFile="/Users/byer3/GEOS-DEV/mass_inj_const/src/coreComponents/unitTests/wellsTests/co2flash.txt"/>
+      phasePVTParaFiles="{ pvtgas.txt, pvtliquid.txt }"
+      flashModelParaFile="co2flash.txt"/>
 
     <BrooksCoreyRelativePermeability
       name="relperm"
@@ -345,15 +345,12 @@ void testNumericalJacobian( CompositionalMultiphaseReservoirAndWells< Compositio
     {
       meshBody.forMeshLevels( [&] ( MeshLevel & mesh )
       {
-        std::cout << mesh.getName() << " ";
         ElementRegionManager & elemManager = mesh.getElemManager();
-        std::cout << elemManager.getName() << std::endl;
         for( localIndex er = 0; er < elemManager.numRegions(); ++er )
         {
           ElementRegionBase & elemRegion = elemManager.getRegion( er );
           elemRegion.forElementSubRegionsIndex< CellElementSubRegion >( [&]( localIndex const, CellElementSubRegion & subRegion )
           {
-            std::cout << " " << subRegion.getName() << " " << subRegion.size() << std::endl;
             // get the degrees of freedom and ghosting information
             arrayView1d< globalIndex const > const & dofNumber =
               subRegion.getReference< array1d< globalIndex > >( resDofKey );
