+++ src/coreComponents/physicsSolvers/fluidFlow/wells/WellSolverBase.hpp
@@ -180,12 +180,12 @@ public:
    * @param matrix the system matrix
    * @param rhs the system right-hand side vector
    */
-  virtual void assembleSystem( real64 const time ,
-                               real64 const dt ,
-                               DomainPartition &  domain ,
-                               DofManager const &  dofManager ,
-                               CRSMatrixView< real64, globalIndex const > const &  localMatrix ,
-                               arrayView1d< real64 > const & localRhs ) override; 
+  virtual void assembleSystem( real64 const time,
+                               real64 const dt,
+                               DomainPartition & domain,
+                               DofManager const & dofManager,
+                               CRSMatrixView< real64, globalIndex const > const & localMatrix,
+                               arrayView1d< real64 > const & localRhs ) override;
 
   /**
    * @brief assembles the flux terms for all connections between well elements
