+++ src/coreComponents/unitTests/wellsTests/testThermalReservoirCompositionalMultiphaseMSWells.cpp
@@ -340,312 +340,312 @@ void testNumericalJacobian( CompositionalMultiphaseReservoirAndWells< Compositio
   ////////////////////////////////////////////////
   // Step 1) Compute the terms in J_RR and J_WR //
   ////////////////////////////////////////////////
-  if (1)
-  domain.forMeshBodies( [&] ( MeshBody & meshBody )
-  {
-    meshBody.forMeshLevels( [&] ( MeshLevel & mesh )
+  if( 1 )
+    domain.forMeshBodies( [&] ( MeshBody & meshBody )
     {
-      ElementRegionManager & elemManager = mesh.getElemManager();
-      for( localIndex er = 0; er < elemManager.numRegions(); ++er )
+      meshBody.forMeshLevels( [&] ( MeshLevel & mesh )
       {
-        ElementRegionBase & elemRegion = elemManager.getRegion( er );
-        elemRegion.forElementSubRegionsIndex< CellElementSubRegion >( [&]( localIndex const, CellElementSubRegion & subRegion )
+        ElementRegionManager & elemManager = mesh.getElemManager();
+        for( localIndex er = 0; er < elemManager.numRegions(); ++er )
         {
-          // get the degrees of freedom and ghosting information
-          arrayView1d< globalIndex const > const & dofNumber =
-            subRegion.getReference< array1d< globalIndex > >( resDofKey );
-
-          // get the primary variables on the reservoir elements
-          arrayView1d< real64 > const & pres =
-            subRegion.getField< fields::flow::pressure >();
-          pres.move( hostMemorySpace, false );
-
-          arrayView2d< real64, compflow::USD_COMP > const & compDens =
-            subRegion.getField< fields::flow::globalCompDensity >();
-          compDens.move( hostMemorySpace, false );
-
-          arrayView1d< real64 > const & temp =
-            subRegion.getField< fields::flow::temperature >();
-          temp.move( hostMemorySpace, false );
-
-          // a) compute all the derivatives wrt to the pressure in RESERVOIR elem ei
-          for( localIndex ei = 0; ei < subRegion.size(); ++ei )
+          ElementRegionBase & elemRegion = elemManager.getRegion( er );
+          elemRegion.forElementSubRegionsIndex< CellElementSubRegion >( [&]( localIndex const, CellElementSubRegion & subRegion )
           {
-            if ( ei !=0 )
-              break;
-            real64 totalDensity = 0.0;
-            for( localIndex ic = 0; ic < NC; ++ic )
+            // get the degrees of freedom and ghosting information
+            arrayView1d< globalIndex const > const & dofNumber =
+              subRegion.getReference< array1d< globalIndex > >( resDofKey );
+
+            // get the primary variables on the reservoir elements
+            arrayView1d< real64 > const & pres =
+              subRegion.getField< fields::flow::pressure >();
+            pres.move( hostMemorySpace, false );
+
+            arrayView2d< real64, compflow::USD_COMP > const & compDens =
+              subRegion.getField< fields::flow::globalCompDensity >();
+            compDens.move( hostMemorySpace, false );
+
+            arrayView1d< real64 > const & temp =
+              subRegion.getField< fields::flow::temperature >();
+            temp.move( hostMemorySpace, false );
+
+            // a) compute all the derivatives wrt to the pressure in RESERVOIR elem ei
+            for( localIndex ei = 0; ei < subRegion.size(); ++ei )
             {
-              totalDensity += compDens[ei][ic];
-            }
-
-            {
-              solver.resetStateToBeginningOfStep( domain );
-
-              // here is the perturbation in the pressure of the element
-              real64 const dP = perturbParameter * (pres[ei] + perturbParameter);
-              pres.move( hostMemorySpace, true );
-              pres[ei] += dP;
-
[…]
