+++ src/coreComponents/physicsSolvers/fluidFlow/wells/CompositionalMultiphaseWell.cpp
@@ -1220,11 +1220,11 @@ void CompositionalMultiphaseWell::assembleAccumulationTerms( real64 const & time
 
 
 real64
-CompositionalMultiphaseWell::calculateResidualNorm1( real64 const & time_n,
-                                                     real64 const & dt,
-                                                     DomainPartition const & domain,
-                                                     DofManager const & dofManager,
-                                                     arrayView1d< real64 const > const & localRhs )
+CompositionalMultiphaseWell::calculateResidualNorm( real64 const & time_n,
+                                                    real64 const & dt,
+                                                    DomainPartition const & domain,
+                                                    DofManager const & dofManager,
+                                                    arrayView1d< real64 const > const & localRhs )
 {
   GEOS_MARK_FUNCTION;
 
@@ -1346,79 +1346,7 @@ CompositionalMultiphaseWell::calculateResidualNorm1( real64 const & time_n,
   }
   return resNorm;
 }
-real64
-CompositionalMultiphaseWell::calculateResidualNorm( real64 const & time_n,
-                                                    real64 const & dt,
-                                                    DomainPartition const & domain,
-                                                    DofManager const & dofManager,
-                                                    arrayView1d< real64 const > const & localRhs )
-{
 
-  return calculateResidualNorm1( time_n, dt, domain, dofManager, localRhs );
-
-  GEOS_MARK_FUNCTION;
-
-  real64 localResidualNorm = 0.0;
-
-  globalIndex const rankOffset = dofManager.rankOffset();
-  string const wellDofKey = dofManager.getKey( wellElementDofName() );
-
-  forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&] ( string const &,
-                                                                MeshLevel const & mesh,
-                                                                arrayView1d< string const > const & regionNames )
-  {
-
-
-    ElementRegionManager const & elemManager = mesh.getElemManager();
-
-    elemManager.forElementSubRegions< WellElementSubRegion >( regionNames,
-                                                              [&]( localIndex const,
-                                                                   WellElementSubRegion const & subRegion )
-    {
-      real64 subRegionResidualNorm[1]{};
-
-      string const & fluidName = subRegion.getReference< string >( viewKeyStruct::fluidNamesString() );
-      MultiFluidBase const & fluid = subRegion.getConstitutiveModel< MultiFluidBase >( fluidName );
-
-      WellControls const & wellControls = getWellControls( subRegion );
-
-      // step 1: compute the norm in the subRegion
-
-      ResidualNormKernelFactory::
-        createAndLaunch< parallelDevicePolicy<> >( m_numComponents,
-                                                   numDofPerWellElement(),
-                                                   m_targetPhaseIndex,
-                                                   rankOffset,
-                                                   wellDofKey,
-                                                   localRhs,
-                                                   subRegion,
-                                                   fluid,
-                                                   wellControls,
-                                                   time_n + dt,
-                                                   dt,
-                                                   m_nonlinearSolverParameters.m_minNormalizer,
-                                                   subRegionResidualNorm );
-
-      // step 2: reduction across meshBodies/regions/subRegions
-
-      if( subRegionResidualNorm[0] > localResidualNorm )
-      {
-        localResidualNorm = subRegionResidualNorm[0];
-      }
-
-    } );
-  } );
-
-  // step 3: second reduction across MPI ranks
-
[…]

[…]
