+++ src/coreComponents/physicsSolvers/fluidFlow/wells/WellElementKernelUtilities.hpp
@@ -1,283 +1,0 @@
-/*
- * ------------------------------------------------------------------------------------------------------------
- * SPDX-License-Identifier: LGPL-2.1-only
- *
- * Copyright (c) 2018-2020 Lawrence Livermore National Security LLC
- * Copyright (c) 2018-2020 The Board of Trustees of the Leland Stanford Junior University
- * Copyright (c) 2018-2020 TotalEnergies
- * Copyright (c) 2019-     GEOSX Contributors
- * All rights reserved
- *
- * See top level LICENSE, COPYRIGHT, CONTRIBUTORS, NOTICE, and ACKNOWLEDGEMENTS files for details.
- * ------------------------------------------------------------------------------------------------------------
- */
-
-/**
- * @file WellElementKernellUtilities.hpp
- */
-
-#ifndef GEOS_PHYSICSSOLVERS_FLUIDFLOW_WELLS_WELLELEMENTKERNELS_HPP
-#define GEOS_PHYSICSSOLVERS_FLUIDFLOW_WELLS_WELLELEMENTKERNELS_HPP
-
-#include "common/DataLayouts.hpp"
-#include "common/DataTypes.hpp"
-#include "constitutive/fluid/multifluid/Layouts.hpp"
-#include "constitutive/capillaryPressure/layouts.hpp"
-
-
-
-namespace geos
-{
-
-namespace wellElementKernelUtilities
-{
-
-// TODO make input parameter
-static constexpr real64 epsC1PPU = 5000;
-
-
-using Deriv = constitutive::multifluid::DerivativeOffset;
-
-struct PotGrad
-{
-  template< integer NC, integer ISTHERMAL >
-  GEOS_HOST_DEVICE
-  static void
-  compute ( real64 const & gravCoef,
-            real64 const & gravCoefNext,
-            real64 const & pres,
-            real64 const & presNext,
-            real64 const & totalMassDens,
-            real64 const & totalMassDensNext,
-            arraySlice1d< real64 const, compflow::USD_FLUID_DC - 1 > const & dTotalMassDens,
-            arraySlice1d< real64 const, compflow::USD_FLUID_DC - 1 > const & dTotalMassDensNext,
-            arraySlice1d< real64 const, compflow::USD_FLUID_DC - 1 > const & dTotalMassDens_dCompDens,
-            arraySlice1d< real64 const, compflow::USD_FLUID_DC - 1 > const & dTotalMassDens_dCompDensNext,
-            real64 & potDiff,
-            real64 ( & dPotDiff)[2][NC+1+IS_THERMAL] )
-
-  {
-    // local working variables and arrays
-    real64 pres[2]{};
-    real64 dPres_dP[2]{};
-    real64 dPres_dC[2][NC]{};
-    real64 dFlux_dP[2]{};
-    real64 dFlux_dC[2][NC]{};
-    real64 dMult_dP[2]{};
-    real64 dMult_dC[2][NC]{};
-    real64 dPotDiff_dP[2]{};
-    real64 dPotDiff_dC[2][NC]{};
-
-    real64 multiplier[2]{};
-    multiplier[ElemTag::CURRENT] = 1.0;
-    multiplier[ElemTag::NEXT]    = -1.0;
-
-    pres[ElemTag::CURRENT] = pres;
-    pres[ElemTag::NEXT]    = presNext;
-
-
-    // local working variables and arrays
-    real64 dAvgMassDens_dCompCurrent[NC]{};
-    real64 dAvgMassDens_dCompNext[NC]{};
-
-    // compute the average density at the interface between well elements
-    real64 const avgMassDens = 0.5 * ( totalMassDensNext + totalMassDens );
-
[…]
