+++ src/coreComponents/physicsSolvers/fluidFlow/wells/ThermalCompositionalMultiphaseWellKernels.hpp
@@ -998,8 +998,8 @@ public:
         }
         stack.localEnergyFlux[TAG::NEXT   ]   =  m_dt * eflux * currentConnRate;
         stack.localEnergyFlux[TAG::CURRENT  ] = -m_dt * eflux * currentConnRate;
-        stack.localEnergyFluxJacobian_dQ [TAG::NEXT   ][0] =  m_dt * eflux;
-        stack.localEnergyFluxJacobian_dQ [TAG::CURRENT][0] =  -m_dt * eflux;
+        stack.localEnergyFluxJacobian_dQ [TAG::NEXT   ][0] =  m_dt * eflux_dq;
+        stack.localEnergyFluxJacobian_dQ [TAG::CURRENT][0] =  -m_dt * eflux_dq;
         for( integer dof=0; dof < CP_Deriv::nDer; dof++ )
         {
           stack.localEnergyFluxJacobian[TAG::NEXT      ][dof]  *=  m_dt*currentConnRate;
