+++ BASELINE_NOTES.md
@@ -6,10 +6,13 @@ This file is designed to track changes to the integrated test baselines.
 Any developer who updates the baseline ID in the .integrated_tests.yaml file is expected to create an entry in this file with the pull request number, date, and their justification for rebaselining.
 These notes should be in reverse-chronological order, and use the following time format: (YYYY-MM-DD).
 
+PR #3133 (2024-06-14)
+======================
+Fix node ordering for faceElements.
+
 PR #3021 (2024-06-13)
 ======================
 Preparatory work for fractures + wells. New test case added.
-
 
 PR #3152 (2024-06-13)
 ======================
