+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3021-5589-3c278fb
+  baseline: integratedTests/baseline_integratedTests-pr3133-5610-b42c02a
 
 allow_fail:
   all: ''
