### 2024-06-14 fbe4bb3
Extract the ghostRank from the send/recv instead of duplicating it.
