+++ src/coreComponents/mesh/generators/include/GhostExchange.hpp
@@ -24,12 +24,6 @@ class GhostExchange
 {
 public:
   /**
-   * @brief Get the ghost rank mapping. Index is an edge index local to the MPI rank.
-   * @return A array matching the size of the contained geometrical quantity (nodes, edges...).
-   */
-  [[nodiscard]] virtual array1d< integer > getGhostRank() const = 0;
-
-  /**
    * @brief Get local to global map for the contained geometrical quantity (nodes, edges...).
    * @return The mapping relationship as an array (local indexing is contiguous).
    */
