+++ src/coreComponents/mesh/ObjectManagerBase.hpp
@@ -927,12 +927,10 @@ protected:
   /**
    * @brief Copy the ghosting exchange information for the entities (e.g. nodes, edges...) held by the current manager.
    * @param[in] neighbors The ranks of the neighbors (does not include the current MPI rank)
-   * @param[in] ghostRank The ghost ranks of the entities own by the current manager.
    * @param[in] send The entities to send to which neighbors.
    * @param[in] recv The entities that will be received from which neighbor.
    */
   void copyExchangeInfo( std::set< integer > const & neighbors,
-                         array1d< integer > const & ghostRank,
                          std::map< integer, array1d< localIndex > > const & send,
                          std::map< integer, array1d< localIndex > > const & recv );
   /// Group that holds object sets.
