+++ src/coreComponents/mesh/generators/Pods.hpp
@@ -30,7 +30,6 @@ namespace geos
 
 struct GhostMapping
 {
-  array1d< integer > m_ghostRank;
   array1d< globalIndex > m_l2g;
   unordered_map< globalIndex, localIndex > m_g2l;
   std::map< integer, array1d< localIndex > > m_send;
@@ -44,7 +43,6 @@ public:
 
   NodeMgrImpl( localIndex numNodes,
                array2d< real64, nodes::REFERENCE_POSITION_PERM > && positions,
-               array1d< integer > && ghostRank,
                ArrayOfArrays< localIndex > const & n2e,
                ArrayOfArrays< localIndex > const & n2f,
                ArrayOfArrays< localIndex > const & n2c,
@@ -81,11 +79,6 @@ public:
   }
 
   // Diamond
-  [[nodiscard]] array1d< integer > getGhostRank() const override
-  {
-    return m_ghost.m_ghostRank;
-  }
-
   [[nodiscard]] array1d< globalIndex > getLocalToGlobal() const override
   {
     return m_ghost.m_l2g;
@@ -121,7 +114,6 @@ public:
   EdgeMgrImpl() = default;
 
   EdgeMgrImpl( std::size_t numEdges,
-               array1d< integer > && ghostRank,
                array2d< localIndex > && e2n,
                ArrayOfArrays< localIndex > && e2f,
                unordered_map< globalIndex, localIndex > && g2l,
@@ -145,11 +137,6 @@ public:
   }
 
   // Diamond
-  [[nodiscard]] array1d< integer > getGhostRank() const override
-  {
-    return m_ghost.m_ghostRank;
-  }
-
   [[nodiscard]] array1d< globalIndex > getLocalToGlobal() const override
   {
     return m_ghost.m_l2g;
@@ -183,7 +170,6 @@ public:
   FaceMgrImpl() = default;
 
   FaceMgrImpl( std::size_t numFaces,
-               array1d< integer > && ghostRank,
                ArrayOfArrays< localIndex > && f2n,
                ArrayOfArrays< localIndex > && f2e,
                array2d< localIndex > && f2c,
@@ -210,11 +196,6 @@ public:
   [[nodiscard]] ToCellRelation< array2d< localIndex > > getFaceToElements() const override;
 
   // Diamond
-  [[nodiscard]] array1d< integer > getGhostRank() const override
-  {
-    return m_ghost.m_ghostRank;
-  }
-
   [[nodiscard]] array1d< globalIndex > getLocalToGlobal() const override
   {
     return m_ghost.m_l2g;
@@ -248,7 +229,6 @@ public:
   CellBlkImpl() = default;
 
   CellBlkImpl( localIndex numCells,
-               array1d< integer > && ghostRank,
                array2d< localIndex, cells::NODE_MAP_PERMUTATION > const & c2n,
                array2d< localIndex > const & c2e,
                array2d< localIndex > const & c2f,
@@ -286,11 +266,6 @@ public:
   }
 
   // Diamond
-  [[nodiscard]] array1d< integer > getGhostRank() const override
-  {
-    return m_ghost.m_ghostRank;
-  }
-
   [[nodiscard]] array1d< globalIndex > getLocalToGlobal() const override
   {
     return m_ghost.m_l2g;
