### 2024-06-14 2a97f46
Compute the neighbors from the send/recv information.
