+++ src/coreComponents/mesh/EdgeManager.cpp
@@ -105,8 +105,7 @@ void EdgeManager::setGeometricalRelations( CellBlockManagerABC const & cellBlock
                                                              LvArray::sortedArrayManipulation::UNSORTED_NO_DUPLICATES );
 }
 
-void EdgeManager::setGeometricalRelations( std::set< integer > const & neighbors,
-                                           generators::EdgeMgr const & edgeMgr )
+void EdgeManager::setGeometricalRelations( generators::EdgeMgr const & edgeMgr )
 {
   GEOS_MARK_FUNCTION;
   resize( edgeMgr.numEdges() );
@@ -119,7 +118,7 @@ void EdgeManager::setGeometricalRelations( std::set< integer > const & neighbors
   m_globalToLocalMap = edgeMgr.getGlobalToLocal();
 
   // TODO not for there, but it's convenient
-  copyExchangeInfo( neighbors, edgeMgr.getSend(), edgeMgr.getRecv() );
+  copyExchangeInfo( edgeMgr.getSend(), edgeMgr.getRecv() );
 }
 
 void EdgeManager::setupRelatedObjectsInRelations( NodeManager const & nodeManager,
