+++ src/coreComponents/mesh/NodeManager.cpp
@@ -193,12 +193,10 @@ void NodeManager::setGeometricalRelations( CellBlockManagerABC const & cellBlock
   elemRegionManager.forElementRegionsComplete< SurfaceElementRegion >( connectNodesTo2dElements );
 }
 
-void NodeManager::setGeometricalRelations( generators::MeshMappings const & meshMappings,
+void NodeManager::setGeometricalRelations( generators::NodeMgr const & nodeMgr,
                                            arrayView2d< localIndex const > const & cb2sr )
 {
   GEOS_MARK_FUNCTION;
-  auto const & nodeMgr = meshMappings.getNodeMgr();
-
   resize( nodeMgr.numNodes() );
 
   m_referencePosition = nodeMgr.getNodePositions();
@@ -208,7 +206,7 @@ void NodeManager::setGeometricalRelations( generators::MeshMappings const & mesh
   m_globalToLocalMap = nodeMgr.getGlobalToLocal();
 
   // TODO not for there, but it's convenient
-  copyExchangeInfo( meshMappings.getNeighbors(), nodeMgr.getSend(), nodeMgr.getRecv() );
+  copyExchangeInfo( nodeMgr.getSend(), nodeMgr.getRecv() );
 
   m_toEdgesRelation.base().assimilate< parallelHostPolicy >( nodeMgr.getNodeToEdges(),
                                                              LvArray::sortedArrayManipulation::UNSORTED_NO_DUPLICATES );

[…]
