### 2024-06-15 87df323
Do not provide the l2g map anymore and let the client compute it from g2l.
