+++ src/coreComponents/mesh/generators/BuildPods.cpp
@@ -34,50 +34,25 @@ struct GlobalToLocal
   std::map< CellGlbIdx, CellLocIdx > cells;
 };
 
-struct LocalToGlobal
+/**
+ * @brief Computes a global to local map from the global indices provided in @c gis.
+ * @tparam GI The global index type.
+ * @param gis The container of global indices.
+ * @return The map instance.
+ */
+template< class GI >
+std::map< GI, toLocIdx_t< GI > > buildGlobalToLocalMap( std::set< GI > const & gis )
 {
-  std::map< NodeLocIdx, NodeGlbIdx > nodes;
-  std::map< EdgeLocIdx, EdgeGlbIdx > edges;  // TODO Do we want to make this a vector already? It's surely the most efficient way to build it.
-  std::map< FaceLocIdx, FaceGlbIdx > faces;
-  std::map< CellLocIdx, CellGlbIdx > cells;
-};
-
-template< class GI, class LI >
-void buildL2GMappings( std::set< GI > const & gis,
-                       std::map< GI, LI > & g2l,
-                       std::map< LI, GI > & l2g ) // TODO we can make it a vector -> simple push_backs will be enough to build it
-{
-  g2l.clear();
-  l2g.clear();
+  using LI = toLocIdx_t< GI >;
+  std::map< GI, LI > g2l;
 
   LI li{ 0 };
   for( GI const & gi: gis )
   {
-    g2l.insert( { gi, li } );
-    l2g.insert( { li, gi } );
-    ++li;
+    g2l.insert( { gi, li++ } );
   }
 
-  // TODO add a check to see if we have the full range of local indices.
-}
-
-std::tuple< GlobalToLocal, LocalToGlobal > buildL2GMappings( MeshGraph const & graph )
-{
-  GlobalToLocal g2l;
-  LocalToGlobal l2g;
-
-  // Use `std::ranges::views::keys` when switching to C++20
-  auto const keys = []( auto const & m )
-  {
-    return mapKeys< std::set >( m );
-  };
-
-  buildL2GMappings( keys( graph.n2pos ), g2l.nodes, l2g.nodes );
-  buildL2GMappings( keys( graph.e2n ), g2l.edges, l2g.edges );
-  buildL2GMappings( keys( graph.f2e ), g2l.faces, l2g.faces );
-  buildL2GMappings( keys( graph.c2f ), g2l.cells, l2g.cells );
-
-  return { g2l, l2g };
+  return g2l;
 }
 
 MeshGraph mergeMeshGraph( MeshGraph const & owned,
@@ -104,7 +79,6 @@ struct DownwardMappings
   std::map< CellLocIdx, std::vector< FaceLocIdx > > c2f;
   std::map< CellLocIdx, std::vector< EdgeLocIdx > > c2e;
 };
-
 
 struct UpwardMappings
 {
@@ -141,7 +115,7 @@ ArrayOfArrays< localIndex > convertToAoA( std::map< T, std::vector< U > > const 
   return t2u_;
 }
 
-template< class T, class U, class P=camp::make_idx_seq_t< 2 > >
+template< class T, class U, class P=typename array2d< localIndex >::Permutation >
 array2d< localIndex, P > convertToA2d( std::map< T, std::vector< U > > const & t2u,
                                        int dimU,
                                        bool check = true )
@@ -164,21 +138,16 @@ array2d< localIndex, P > convertToA2d( std::map< T, std::vector< U > > const & t
   return t2u_;
 }
 
-template< class GI, class LI >
-std::tuple< array1d< globalIndex >, unordered_map< globalIndex, localIndex > >
-convertGlbLoc( std::map< GI, LI > const & g2l )
[…]
