+++ src/coreComponents/mesh/generators/Pods.hpp
@@ -30,7 +30,6 @@ namespace geos
 
 struct GhostMapping
 {
-  array1d< globalIndex > m_l2g;
   unordered_map< globalIndex, localIndex > m_g2l;
   std::map< integer, array1d< localIndex > > m_send;
   std::map< integer, array1d< localIndex > > m_recv;
@@ -46,7 +45,6 @@ public:
                ArrayOfArrays< localIndex > const & n2e,
                ArrayOfArrays< localIndex > const & n2f,
                ArrayOfArrays< localIndex > const & n2c,
-               array1d< globalIndex > && l2g,
                unordered_map< globalIndex, localIndex > && g2l,
                std::map< integer, array1d< localIndex > > && send,
                std::map< integer, array1d< localIndex > > && recv );
@@ -79,11 +77,6 @@ public:
   }
 
   // Diamond
-  [[nodiscard]] array1d< globalIndex > getLocalToGlobal() const override
-  {
-    return m_ghost.m_l2g;
-  }
-
   [[nodiscard]] unordered_map< globalIndex, localIndex > getGlobalToLocal() const override
   {
     return m_ghost.m_g2l;
@@ -117,7 +110,6 @@ public:
                array2d< localIndex > && e2n,
                ArrayOfArrays< localIndex > && e2f,
                unordered_map< globalIndex, localIndex > && g2l,
-               array1d< globalIndex > && l2g,
                std::map< integer, array1d< localIndex > > && send,
                std::map< integer, array1d< localIndex > > && recv );
 
@@ -137,11 +129,6 @@ public:
   }
 
   // Diamond
-  [[nodiscard]] array1d< globalIndex > getLocalToGlobal() const override
-  {
-    return m_ghost.m_l2g;
-  }
-
   [[nodiscard]] unordered_map< globalIndex, localIndex > getGlobalToLocal() const override
   {
     return m_ghost.m_g2l;
@@ -174,7 +161,6 @@ public:
                ArrayOfArrays< localIndex > && f2e,
                array2d< localIndex > && f2c,
                unordered_map< globalIndex, localIndex > && g2l,
-               array1d< globalIndex > && l2g,
                std::map< integer, array1d< localIndex > > && send,
                std::map< integer, array1d< localIndex > > && recv );
 
@@ -196,11 +182,6 @@ public:
   [[nodiscard]] ToCellRelation< array2d< localIndex > > getFaceToElements() const override;
 
   // Diamond
-  [[nodiscard]] array1d< globalIndex > getLocalToGlobal() const override
-  {
-    return m_ghost.m_l2g;
-  }
-
   [[nodiscard]] unordered_map< globalIndex, localIndex > getGlobalToLocal() const override
   {
     return m_ghost.m_g2l;
@@ -232,7 +213,6 @@ public:
                array2d< localIndex, cells::NODE_MAP_PERMUTATION > const & c2n,
                array2d< localIndex > const & c2e,
                array2d< localIndex > const & c2f,
-               array1d< globalIndex > && l2g,
                unordered_map< globalIndex, localIndex > && m_g2l,
                std::map< integer, array1d< localIndex > > && send,
                std::map< integer, array1d< localIndex > > && recv );
@@ -266,11 +246,6 @@ public:
   }
 
   // Diamond
-  [[nodiscard]] array1d< globalIndex > getLocalToGlobal() const override
-  {
-    return m_ghost.m_l2g;
-  }
-
   [[nodiscard]] unordered_map< globalIndex, localIndex > getGlobalToLocal() const override
   {
     return m_ghost.m_g2l;
