+++ src/coreComponents/physicsSolvers/multiphysics/PoromechanicsSolver.hpp
@@ -20,6 +20,7 @@
 #ifndef GEOS_PHYSICSSOLVERS_MULTIPHYSICS_POROMECHANICSSOLVER_HPP_
 #define GEOS_PHYSICSSOLVERS_MULTIPHYSICS_POROMECHANICSSOLVER_HPP_
 
+#include "physicsSolvers/fluidFlow/FlowSolverBaseFields.hpp"
 #include "physicsSolvers/multiphysics/CoupledSolver.hpp"
 #include "physicsSolvers/multiphysics/PoromechanicsFields.hpp"
 #include "physicsSolvers/solidMechanics/SolidMechanicsLagrangianFEM.hpp"
@@ -31,6 +32,26 @@
 
 namespace geos
 {
+
+namespace stabilization
+{
+enum class StabilizationType : integer
+{
+  None,
+  Global,
+  Local,
+};
+
+ENUM_STRINGS( StabilizationType,
+              "None",
+              "Global",
+              "Local" );
+}
+
+using namespace stabilization;
+using namespace fields;
+using namespace constitutive;
+using namespace dataRepository;
 
 template< typename FLOW_SOLVER, typename MECHANICS_SOLVER = SolidMechanicsLagrangianFEM >
 class PoromechanicsSolver : public CoupledSolver< FLOW_SOLVER, MECHANICS_SOLVER >
@@ -72,6 +93,23 @@ public:
       setApplyDefaultValue( false ).
       setInputFlag( dataRepository::InputFlags::FALSE ).
       setDescription( "Flag to indicate that the solver is going to perform stress initialization" );
+
+    this->registerWrapper( viewKeyStruct::stabilizationTypeString(), &m_stabilizationType ).
+      setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+      setDescription( "StabilizationType. Options are:\n" +
+                      toString( StabilizationType::None ) + "- Add no stabilization to mass equation \n" +
+                      toString( StabilizationType::Global ) + "- Add jump stabilization to all faces \n" +
+                      toString( StabilizationType::Local ) + "- Add jump stabilization on interior of macro elements" );
+
+    this->registerWrapper( viewKeyStruct::stabilizationRegionNamesString(), &m_stabilizationRegionNames ).
+      setRTTypeName( rtTypes::CustomTypes::groupNameRefArray ).
+      setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+      setDescription( "Regions where stabilization is applied." );
+
+    this->registerWrapper( viewKeyStruct::stabilizationMultiplierString(), &m_stabilizationMultiplier ).
+      setApplyDefaultValue( 1.0 ).
+      setInputFlag( dataRepository::InputFlags::OPTIONAL ).
+      setDescription( "Constant multiplier of stabilization strength" );
   }
 
   virtual void initializePostInitialConditionsPreSubGroups() override
@@ -87,6 +125,11 @@ public:
   virtual void initializePreSubGroups() override
   {
     Base::initializePreSubGroups();
+
+    GEOS_THROW_IF( m_stabilizationType == StabilizationType::Local,
+                   this->getWrapperDataContext( viewKeyStruct::stabilizationTypeString() ) <<
+                   ": Local stabilization has been temporarily disabled",
+                   InputError );
 
     DomainPartition & domain = this->template getGroupByPath< DomainPartition >( "/Problem/domain" );
 
@@ -135,6 +178,11 @@ public:
       flowSolver()->enableFixedStressPoromechanicsUpdate();
     }
 
+    if( m_stabilizationType == StabilizationType::Global || m_stabilizationType == StabilizationType::Local )
+    {
[…]
