+++ src/coreComponents/physicsSolvers/multiphysics/SinglePhasePoromechanics.cpp
@@ -145,11 +145,22 @@ void SinglePhasePoromechanics< FLOW_SOLVER, MECHANICS_SOLVER >::assembleSystem( 
                              localRhs );
 
   // Step 3: compute the fluxes (face-based contributions)
-  this->flowSolver()->assembleFluxTerms( dt,
-                                         domain,
-                                         dofManager,
-                                         localMatrix,
-                                         localRhs );
+  if( m_stabilizationType == StabilizationType::Global || m_stabilizationType == StabilizationType::Local )
+  {
+    this->flowSolver()->assembleStabilizedFluxTerms( dt,
+                                                     domain,
+                                                     dofManager,
+                                                     localMatrix,
+                                                     localRhs );
+  }
+  else
+  {
+    this->flowSolver()->assembleFluxTerms( dt,
+                                           domain,
+                                           dofManager,
+                                           localMatrix,
+                                           localRhs );
+  }
 
 }
 
