+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/CO2BrineFluid.hpp
@@ -169,6 +169,10 @@ protected:
 
 private:
 
+  /**
+   * @brief Create a PVT Model and output them
+   * @param isClone Check if we have a clone of this constitutive model. If so don't regenarate.output
+   */
   void createPVTModels( bool isClone );
 
   /// Names of the files defining the viscosity and density models
