+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/functions/PhillipsBrineViscosity.cpp
@@ -44,7 +44,7 @@ PhillipsBrineViscosity::PhillipsBrineViscosity( string const & name,
   m_waterViscosityTable = PureWaterProperties::makeSaturationViscosityTable( m_functionName, FunctionManager::getInstance() );
   makeCoefficients( inputPara );
 
-  handleTableOutputOptions( m_waterViscosityTable, pvtOpts );
+  outputPVTTableData( m_waterViscosityTable, pvtOpts );
 }
 
 void PhillipsBrineViscosity::makeCoefficients( string_array const & inputPara )
