+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/functions/EzrokhiBrineDensity.cpp
@@ -52,8 +52,8 @@ EzrokhiBrineDensity::EzrokhiBrineDensity( string const & name,
   m_waterSatDensityTable = PureWaterProperties::makeSaturationDensityTable( m_functionName, FunctionManager::getInstance() );
   m_waterSatPressureTable = PureWaterProperties::makeSaturationPressureTable( m_functionName, FunctionManager::getInstance() );
 
-  handleTableOutputOptions( m_waterSatPressureTable, pvtOpts );
-  handleTableOutputOptions( m_waterSatDensityTable, pvtOpts );
+  outputPVTTableData( m_waterSatPressureTable, pvtOpts );
+  outputPVTTableData( m_waterSatDensityTable, pvtOpts );
 }
 
 void EzrokhiBrineDensity::makeCoefficients( string_array const & inputPara )
