### 2024-06-17 a6e7fba
Add some unit tests
