+++ src/coreComponents/denseLinearAlgebra/unitTests/testSolveLinearSystem.cpp
@@ -1,0 +1,499 @@
+/*
+ * ------------------------------------------------------------------------------------------------------------
+ * SPDX-License-Identifier: LGPL-2.1-only
+ *
+ * Copyright (c) 2018-2020 Lawrence Livermore National Security LLC
+ * Copyright (c) 2018-2020 The Board of Trustees of the Leland Stanford Junior University
+ * Copyright (c) 2018-2020 TotalEnergies
+ * Copyright (c) 2019-     GEOSX Contributors
+ * All rights reserved
+ *
+ * See top level LICENSE, COPYRIGHT, CONTRIBUTORS, NOTICE, and ACKNOWLEDGEMENTS files for details.
+ * ------------------------------------------------------------------------------------------------------------
+ */
+
+/**
+ * @file testSolveLinearSystem.cpp
+ */
+
+// Source includes
+#include "denseLinearAlgebra/interfaces/blaslapack/BlasLapackLA.hpp"
+
+#include "gtest/gtest.h"
+
+using namespace geos;
+
+using INDEX_TYPE = std::ptrdiff_t;
+
+constexpr INDEX_TYPE MAX_SIZE = 20;
+constexpr real64 machinePrecision = 1.0e2 * LvArray::NumericLimits< real64 >::epsilon;
+
+// Test matrices
+enum TestMatrixType
+{
+  LAPLACE,
+  SAMPLING,
+  REDHEFFER,
+  GRCAR
+};
+
+void random_permutation( arraySlice1d< INDEX_TYPE > const & pivot )
+{
+  INDEX_TYPE const N = LvArray::integerConversion< INDEX_TYPE >( pivot.size() );
+  GEOS_ASSERT( 2 < N );
+
+  for( INDEX_TYPE j = 0; j < N; j++ )
+  {
+    pivot[j] = j;
+  }
+
+  StackArray< real64, 1, MAX_SIZE > ordering( N );
+  BlasLapackLA::vectorRand( ordering.toSlice() );
+
+  for( INDEX_TYPE j = 0; j < N; j++ )
+  {
+    INDEX_TYPE const i = static_cast< INDEX_TYPE >(ordering[j]*N);
+    INDEX_TYPE const pi = pivot[i];
+    pivot[i] = pivot[j];
+    pivot[j] = pi;
+  }
+}
+
+template< TestMatrixType MATRIX_TYPE >
+struct TestMatrix {};
+
+template<>
+struct TestMatrix< TestMatrixType::LAPLACE >
+{
+  template< int USD >
+  static void create( arraySlice2d< real64, USD > const & A )
+  {
+    INDEX_TYPE const N = LvArray::integerConversion< INDEX_TYPE >( A.size( 0 ) );
+    GEOS_ASSERT( 2 < N );
+    LvArray::forValuesInSlice( A, []( real64 & a ){ a = 0.0; } );
+    for( INDEX_TYPE i = 0; i < N-1; i++ )
+    {
+      A( i+1, i ) = -1.0;
+      A( i+1, i+1 ) = 2.0;
+      A( i, i+1 ) = -1.0;
+    }
+    A( 0, 0 ) = 2.0;
+  }
+};
+
+template<>
+struct TestMatrix< TestMatrixType::SAMPLING >
+{
+  template< int USD >
+  static void create( arraySlice2d< real64, USD > const & A )
+  {
+    INDEX_TYPE const N = LvArray::integerConversion< INDEX_TYPE >( A.size( 0 ) );
+    GEOS_ASSERT( 2 < N );
+    StackArray< INDEX_TYPE, 1, MAX_SIZE > pivot( N );
+    random_permutation( pivot.toSlice() );
+
+    real64 minajj = LvArray::NumericLimits< real64 >::max;
[…]
