### 2024-06-17 5599af2
Added 2d averaging kernel
