### 2024-06-17 31819fc
Implement Temperature-dependent Solid Volumetric Heat Capacity (#3088)
Co-authored-by: Jian HUANG <jian.huang@totalenergies.com>
