+++ src/coreComponents/unitTests/fluidFlowTests/testThermalSinglePhaseFlow.cpp
@@ -94,7 +94,7 @@ char const * xmlInput =
                         referencePressure="0.0"
                         compressibility="1.0e-9" />
       <SolidInternalEnergy name="rockInternalEnergy"
-                           volumetricHeatCapacity="1.95e6"
+                           referenceVolumetricHeatCapacity="1.95e6"
                            referenceTemperature="368.15"
                            referenceInternalEnergy="0" />
       <ConstantPermeability name="rockPerm"
