+++ host-configs/LLNL/llnl-cpu-base.cmake
@@ -1,6 +1,6 @@
 ###############################################################################
 #
-# Base configuration for LC Ruby builds
+# Base configuration for LC cpu builds
 # Calling configuration file must define the following CMAKE variables:
 #
 # MPI_HOME
@@ -34,14 +34,6 @@ set(ENABLE_PYGEOSX ON CACHE BOOL "")
 set(Python3_ROOT_DIR /usr/apps/python-3.11.5 CACHE PATH "")
 set(Python3_EXECUTABLE ${Python3_ROOT_DIR}/bin/python3 CACHE PATH "")
 
-# YAPF python formatting
-#set(YAPF_EXECUTABLE /usr/gapps/GEOSX/thirdPartyLibs/python/quartz-gcc-python/python/bin/yapf CACHE PATH "" FORCE)
-
-# Sphinx
-#set(SPHINX_EXECUTABLE /usr/gapps/GEOSX/thirdPartyLibs/python/quartz-gcc-python/python/bin/sphinx-build CACHE PATH "" FORCE)
-
-set(ENABLE_FESAPI OFF CACHE BOOL "" FORCE)
-
 # caliper
 set(ENABLE_CALIPER ON CACHE BOOL "" FORCE)
 set(ENABLE_CALIPER_HYPRE ON CACHE BOOL "" FORCE)
@@ -56,6 +48,7 @@ set(MKL_LIBRARIES ${MKL_ROOT}/lib/intel64/libmkl_intel_lp64.so
                   CACHE STRING "")
 
 # ATS
-set(ATS_ARGUMENTS "--machine slurm56"  CACHE STRING "")
-
+set(USER $ENV{USER} CACHE STRING "")
+set(ATS_WORKING_DIR "/p/lustre2/${USER}/integratedTestsGEOS/${CONFIG_NAME}"  CACHE PATH "")
+set(ATS_BASELINE_DIR "/p/lustre2/${USER}/integratedTestsGEOS/baselines"  CACHE PATH "")
 include(${CMAKE_CURRENT_LIST_DIR}/../tpls.cmake)
