### 2024-06-18 7326339
prototype kernel formulation for linear elasticity
