+++ src/coreComponents/physicsSolvers/solidMechanics/kernels/StrainHelper.hpp
@@ -13,130 +13,29 @@
  */
 
 /**
- * @file AverageOverQuadraturePointsKernel.hpp
+ * @file StrainHelper.hpp
  */
 
-#ifndef GEOS_MESH_UTILITIES_AVERAGEOVERQUADRATUREPOINTSKERNEL_HPP_
-#define GEOS_MESH_UTILITIES_AVERAGEOVERQUADRATUREPOINTSKERNEL_HPP_
+#ifndef GEOS_PHYSICSSOLVERS_SOLIDMECHANICS_KERNELS_STRAINHELPER_HPP_
+#define GEOS_PHYSICSSOLVERS_SOLIDMECHANICS_KERNELS_STRAINHELPER_HPP_
 
 #include "common/DataTypes.hpp"
 #include "common/GEOS_RAJA_Interface.hpp"
 #include "finiteElement/FiniteElementDispatch.hpp"
 #include "mesh/CellElementSubRegion.hpp"
+#include "mesh/utilities/AverageOverQuadraturePointsKernel.hpp"
+#include "physicsSolvers/solidMechanics/SolidMechanicsFields.hpp"
 
 namespace geos
 {
-
 /**
- * @class AverageOverQuadraturePointsBase
+ * @class AverageStrainOverQuadraturePoints
  * @tparam SUBREGION_TYPE the subRegion type
  * @tparam FE_TYPE the finite element type
  */
 template< typename SUBREGION_TYPE,
           typename FE_TYPE >
-class AverageOverQuadraturePointsBase
-{
-public:
-
-  /**
-   * @brief Constructor for the class
-   * @param nodeManager the node manager
-   * @param edgeManager the edge manager
-   * @param faceManager the face manager
-   * @param elementSubRegion the element subRegion
-   * @param finiteElementSpace the finite element space
-   */
-  AverageOverQuadraturePointsBase( NodeManager & nodeManager,
-                                   EdgeManager const & edgeManager,
-                                   FaceManager const & faceManager,
-                                   SUBREGION_TYPE const & elementSubRegion,
-                                   FE_TYPE const & finiteElementSpace ):
-    m_finiteElementSpace( finiteElementSpace ),
-    m_elemsToNodes( elementSubRegion.nodeList().toViewConst() ),
-    m_X( nodeManager.referencePosition() ),
-    m_elementVolume( elementSubRegion.getElementVolume() )
-  {
-    finiteElement::FiniteElementBase::
-      initialize< FE_TYPE >( nodeManager,
-                             edgeManager,
-                             faceManager,
-                             elementSubRegion,
-                             m_meshData );
-  }
-
-  //*****************************************************************************
-  /**
-   * @copydoc finiteElement::KernelBase::StackVariables
-   */
-  struct StackVariables
-  {
-public:
-
-    /**
-     * Default constructor
-     */
-    GEOS_HOST_DEVICE
-    StackVariables():
-      xLocal()
-    {}
-
-    /// C-array stack storage for element local the nodal positions.
-    real64 xLocal[ FE_TYPE::maxSupportPoints ][ 3 ];
-
-    /// Stack variables needed for the underlying FEM type
-    typename FE_TYPE::StackVariables feStack;
-  };
-  //***************************************************************************
-
-  /**
-   * @brief Performs the setup phase for the kernel.
-   * @param k The element index.
-   * @param stack The StackVariable object that hold the stack variables.
-   */
-  GEOS_HOST_DEVICE
-  void setup( localIndex const k,
-              StackVariables & stack ) const
-  {
-    m_finiteElementSpace.template setup< FE_TYPE >( k, m_meshData, stack.feStack );
-
-    for( localIndex a = 0; a < FE_TYPE::maxSupportPoints; ++a )
[…]
