+++ src/coreComponents/physicsSolvers/solidMechanics/kernels/ImplicitSmallStrainQuasiStatic.hpp
@@ -236,7 +236,7 @@ protected:
   arrayView2d< real64 const, nodes::INCR_DISPLACEMENT_USD > const m_uhat;
 
   /// The rank-global average strain increment array
-  arrayView2d< real64, cells::STRAIN_USD > & m_incStrain;
+ // arrayView2d< real64, cells::STRAIN_USD > & m_incStrain;
   /// The rank-global element volume
   arrayView1d< real64 const > const m_elementVolume;
 
