+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/functions/EzrokhiBrineDensity.hpp
@@ -108,7 +108,7 @@ public:
                        string_array const & inputPara,
                        string_array const & componentNames,
                        array1d< real64 > const & componentMolarWeight,
-                       TableOutputOptions pvtOpts );
+                       TableOutputOptions pvtOutputOpts );
 
   virtual ~EzrokhiBrineDensity() override = default;
 
