### 2024-06-18 062620a
correct abs not true in derivatives
