### 2024-06-18 f508369
move mpi com outside of subregion loop
