+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3088-5630-5ecff59
+  baseline: integratedTests/baseline_integratedTests-pr3165-5641-568d6e7
 
 allow_fail:
   all: ''
