### 2024-06-18 a1b70ac
build fix for ubunnto etal t2
