### 2024-06-18 72f0230
Update BASELINE_NOTES.md
