+++ src/coreComponents/denseLinearAlgebra/unitTests/testSolveLinearSystem.cpp
@@ -23,9 +23,7 @@
 
 using namespace geos;
 
-using INDEX_TYPE = std::ptrdiff_t;
-
-constexpr INDEX_TYPE MAX_SIZE = 20;
+constexpr int MAX_SIZE = 20;
 constexpr real64 machinePrecision = 1.0e2 * LvArray::NumericLimits< real64 >::epsilon;
 
 // Test matrices
@@ -37,28 +35,6 @@ enum TestMatrixType
   GRCAR
 };
 
-void random_permutation( arraySlice1d< INDEX_TYPE > const & pivot )
-{
-  INDEX_TYPE const N = LvArray::integerConversion< INDEX_TYPE >( pivot.size() );
-  GEOS_ASSERT( 2 < N );
-
-  for( INDEX_TYPE j = 0; j < N; j++ )
-  {
-    pivot[j] = j;
-  }
-
-  StackArray< real64, 1, MAX_SIZE > ordering( N );
-  BlasLapackLA::vectorRand( ordering.toSlice() );
-
-  for( INDEX_TYPE j = 0; j < N; j++ )
-  {
-    INDEX_TYPE const i = static_cast< INDEX_TYPE >(ordering[j]*N);
-    INDEX_TYPE const pi = pivot[i];
-    pivot[i] = pivot[j];
-    pivot[j] = pi;
-  }
-}
-
 template< TestMatrixType MATRIX_TYPE >
 struct TestMatrix {};
 
@@ -68,10 +44,10 @@ struct TestMatrix< TestMatrixType::LAPLACE >
   template< int USD >
   static void create( arraySlice2d< real64, USD > const & A )
   {
-    INDEX_TYPE const N = LvArray::integerConversion< INDEX_TYPE >( A.size( 0 ) );
-    GEOS_ASSERT( 2 < N );
+    int const N = LvArray::integerConversion< int >( A.size( 0 ) );
+    GEOS_ASSERT( 2 <= N );
     LvArray::forValuesInSlice( A, []( real64 & a ){ a = 0.0; } );
-    for( INDEX_TYPE i = 0; i < N-1; i++ )
+    for( int i = 0; i < N-1; i++ )
     {
       A( i+1, i ) = -1.0;
       A( i+1, i+1 ) = 2.0;
@@ -87,31 +63,27 @@ struct TestMatrix< TestMatrixType::SAMPLING >
   template< int USD >
   static void create( arraySlice2d< real64, USD > const & A )
   {
-    INDEX_TYPE const N = LvArray::integerConversion< INDEX_TYPE >( A.size( 0 ) );
-    GEOS_ASSERT( 2 < N );
-    StackArray< INDEX_TYPE, 1, MAX_SIZE > pivot( N );
-    random_permutation( pivot.toSlice() );
+    int const N = LvArray::integerConversion< int >( A.size( 0 ) );
+    GEOS_ASSERT( 2 <= N );
 
     real64 minajj = LvArray::NumericLimits< real64 >::max;
-    for( INDEX_TYPE j = 1; j <= N; j++ )
+    for( int j = 1; j <= N; j++ )
     {
       real64 ajj = 0.0;
-      INDEX_TYPE const pj = pivot[j-1];
-      for( INDEX_TYPE i = 1; i <= N; i++ )
+      for( int i = 1; i <= N; i++ )
       {
         if( i != j )
         {
-          A( i-1, pj ) = 0.0; //static_cast< real64 >(i)/static_cast< real64 >(i-j);
-          ajj += A( i-1, pj );
+          A( i-1, j-1 ) = static_cast< real64 >(i)/static_cast< real64 >(i-j);
+          ajj += A( i-1, j-1 );
         }
       }
[…]
