+++ src/coreComponents/denseLinearAlgebra/interfaces/blaslapack/BlasLapackLA.cpp
@@ -379,20 +379,8 @@ void matrixLeastSquaresSolutionSolve( arraySlice2d< real64, USD > const & A,
   GEOS_ERROR_IF( INFO != 0, "The algorithm computing matrix linear system failed to converge." );
 }
 
-template< int USD, bool INPLACE >
-struct MatrixInPlace
-{
-  using type = arraySlice2d< real64, USD >;
-};
-
-template< int USD >
-struct MatrixInPlace< USD, false >
-{
-  using type = arraySlice2d< real64 const, USD >;
-};
-
-template< bool INPLACE, int USD >
-void solveLinearSystem( typename MatrixInPlace< USD, INPLACE >::type const & A,
+template< typename T, int USD >
+void solveLinearSystem( arraySlice2d< T, USD > const & A,
                         arraySlice2d< real64 const, USD > const & B,
                         arraySlice2d< real64, USD > const & X )
 {
@@ -419,7 +407,7 @@ void solveLinearSystem( typename MatrixInPlace< USD, INPLACE >::type const & A,
 
   real64 * matrixData = nullptr;
   array2d< real64 > LU;   // Space for LU-factors
-  if constexpr ( INPLACE )
+  if constexpr ( !std::is_const< T >::value )
   {
     matrixData = A.dataIfContiguous();
   }
@@ -436,23 +424,13 @@ void solveLinearSystem( typename MatrixInPlace< USD, INPLACE >::type const & A,
 
   array1d< int > IPIV( N );
   int INFO;
-  char transA[3] = {'N','T','T'};
-  char matA[2] = {'L','U'};
-  char const TRANS = (USD == MatrixLayout::ROW_MAJOR) ? transA[0] : 'N';
+  char const TRANS = (USD == MatrixLayout::ROW_MAJOR) ? 'T' : 'N';
 
   GEOS_dgetrf( &N, &N, matrixData, &N, IPIV.data(), &INFO );
 
   GEOS_ASSERT_MSG( INFO == 0, "LAPACK dgetrf error code: " << INFO );
 
-  std::cout << "IPIV:";
-  for( integer i = 0; i < N; ++i )
-  {
-    std::cout << " " << IPIV[i];
-  }
-  std::cout << "\n";
-
-  // If INPLACE is false then it means that on entry B == X
-  if constexpr ( !INPLACE )
+  if constexpr ( std::is_const< T >::value )
   {
     int const INCX = 1;
     int const INCY = 1;
@@ -460,61 +438,69 @@ void solveLinearSystem( typename MatrixInPlace< USD, INPLACE >::type const & A,
     GEOS_dcopy( &K, B.dataIfContiguous(), &INCX, X.dataIfContiguous(), &INCY );
   }
 
-  auto printMatrix = []( auto const & AA )
+  // For row-major form, we need to reorder into col-major form
+  // This might require an extra allocation
+  real64 * solutionData = X.dataIfContiguous();
+  array2d< real64, MatrixLayout::COL_MAJOR_PERM > X0;
+  if constexpr ( USD == MatrixLayout::ROW_MAJOR )
   {
-    int const MA = LvArray::integerConversion< int >( AA.size( 0 ) );
-    int const NA = LvArray::integerConversion< int >( AA.size( 1 ) );
-    std::cout << std::fixed << std::setprecision( 4 );
-    for( int i = 0; i < MA; i++ )
+    if( 1 < M && M == N )
     {
-      for( int j = 0; j < NA; j++ )
+      // Square case: swap in place
+      for( int i = 0; i < N; i++ )
       {
-        std::cout << " " << std::setw( 10 ) << AA( i, j );
-      }
-      std::cout << "\n";
-    }
[…]
