### 2024-06-18 0b47c5a
Undo addition of BLAS function
