### 2024-06-18 1a4273e
arraySlice1d -> Vec
