### 2024-06-18 0ebf04f
Remove Redheffer matrix. It's singular!
