+++ src/coreComponents/mesh/particleGenerators/ParticleMeshGenerator.cpp
@@ -17,8 +17,8 @@
  */
 
 #include "ParticleMeshGenerator.hpp"
-#include "ParticleBlockManager.hpp"
 
+#include "ParticleBlockManager.hpp"
 #include "mesh/mpiCommunications/SpatialPartition.hpp"
 
 #include "common/DataTypes.hpp"
@@ -48,7 +48,12 @@ ParticleMeshGenerator::ParticleMeshGenerator( string const & name, Group * const
     setSizedFromParent( 0 ).
     setDescription( "Names of each particle block" );
 
-  registerWrapper( viewKeyStruct::particleTypesString(), &m_particleType ).
+  registerWrapper( viewKeyStruct::particleMaterialNamesString(), &m_materialNames ).
+    setInputFlag( InputFlags::REQUIRED ).
+    setSizedFromParent( 0 ).
+    setDescription( "Names of each particle material, index corresponds to material ID in particle file");
+
+  registerWrapper( viewKeyStruct::particleTypesString(), &m_particleTypes ).
     setInputFlag( InputFlags::REQUIRED ).
     setSizedFromParent( 0 ).
     setDescription( "Particle types of each particle block" );
@@ -70,11 +75,8 @@ void ParticleMeshGenerator::fillParticleBlockManager( ParticleBlockManager & par
   for( auto & particleBlockName : m_blockNames )
   {
     ParticleBlock & particleBlock = particleBlockManager.registerParticleBlock( particleBlockName );
-    particleBlock.setParticleType( EnumStrings< ParticleType >::fromString( m_particleType[aa++] ) );
+    particleBlock.setParticleType( EnumStrings< ParticleType >::fromString( m_particleTypes[aa++] ) );
   }
-
-  //GEOS_LOG_RANK_0( "MPM particle file path: " << m_particleFilePath );
-  //GEOS_LOG_RANK_0( "MPM header file path: " << m_headerFilePath );
 
   int numMaterials, numParticleTypes;
   map< std::string, std::vector< std::vector< double > > > particleData;
@@ -91,13 +93,17 @@ void ParticleMeshGenerator::fillParticleBlockManager( ParticleBlockManager & par
   std::string line; // initialize line variable
 
   // Read in number of materials and particle types
+  // TODO CC remove the header file, all of this should be read from the particle mesh generator xml block
   std::getline( headerFile, line ); // get a line
   std::istringstream iss1( line ); // turn the line into a stream
   iss1 >> numMaterials >> numParticleTypes;
   particleTypes.resize( numParticleTypes );
 
-  //GEOS_LOG_RANK_0( "Number of particle materials: " << numMaterials );
-  //GEOS_LOG_RANK_0( "Number of particle types: " << numParticleTypes );
+  // numParticleTypes = m_particleTypes.size();
+  // numMaterials = m_materialNames.size(); // As read directly from XML block
+
+  GEOS_LOG_RANK_0( "Number of particle materials: " << numMaterials );
+  GEOS_LOG_RANK_0( "Number of particle types: " << numParticleTypes );
 
   // Read in the material key
   for( int i=0; i<numMaterials; i++ )
@@ -108,7 +114,7 @@ void ParticleMeshGenerator::fillParticleBlockManager( ParticleBlockManager & par
     int value; // Material ID
     iss2 >> key >> value;
     materialMap[key] = value;
-    //GEOS_LOG_RANK_0( "Material name/ID: " + key + "/" << value );
+    // GEOS_LOG_RANK_0( "Material name/ID: " + key + "/" << value );
   }
 
   // Read in the particle type key
@@ -123,38 +129,105 @@ void ParticleMeshGenerator::fillParticleBlockManager( ParticleBlockManager & par
[…]
