### 2024-06-18 e4edc21
move input file data to char array, write out array to file, delete file
