+++ src/coreComponents/unitTests/wellsTests/testThermalReservoirCompositionalMultiphaseSSWells.cpp
@@ -36,6 +36,28 @@ using namespace geos::testing;
 
 CommandLineOptions g_commandLineOptions;
 
+void writeTableToFile( string const & filename, char const * str )
+{
+  std::ofstream os( filename );
+  ASSERT_TRUE( os.is_open() );
+  os << str;
+  os.close();
+}
+
+void removeFile( string const & filename )
+{
+  int const ret = std::remove( filename.c_str() );
+  ASSERT_TRUE( ret == 0 );
+}
+char const * co2flash = "FlashModel CO2Solubility  1e6 7.5e7 5e5 299.15 369.15 10 0";
+char const * pvtLiquid = "DensityFun PhillipsBrineDensity 1e6 7.5e7 5e5 299.15 369.15 10 0\n"
+                         "ViscosityFun PhillipsBrineViscosity 0\n"
+                         "EnthalpyFun BrineEnthalpy 1e6 7.5e7 5e5 299.15 369.15 10 0\n";
+
+char const * pvtGas = "DensityFun SpanWagnerCO2Density 1e6 7.5e7 5e5 299.15 369.15 10\n"
+                         "ViscosityFun FenghourCO2Viscosity 1e6 7.5e7 5e5 299.15 369.15 10\n"
+                          "EnthalpyFun CO2Enthalpy 1e6 7.5e7 5e5 299.15 369.15 10\n";
+ 
 char const * xmlInput =
   R"xml(
 
@@ -774,9 +796,16 @@ TEST_F( CompositionalMultiphaseReservoirSolverTest, jacobianNumericalCheck_Press
 #endif
 int main( int argc, char * * argv )
 {
+  writeTableToFile( "co2flash.txt", co2flash );
+  writeTableToFile( "pvtliquid.txt", pvtLiquid );
+  writeTableToFile( "pvtgas.txt", pvtGas );
   ::testing::InitGoogleTest( &argc, argv );
   g_commandLineOptions = *geos::basicSetup( argc, argv );
   int const result = RUN_ALL_TESTS();
   geos::basicCleanup();
+  removeFile( "co2flash.txt" );
+  removeFile("pvtliquid.txt");
+  removeFile("pvtgas.txt");
+ 
   return result;
 }
