### 2024-06-18 021597b
Fixed periodic boundaries
