### 2024-06-18 f7fad8d
Fixed periodic boundaries
