+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3165-5641-568d6e7
+  baseline: integratedTests/baseline_integratedTests-pr3160-5651-0cad939
 
 allow_fail:
   all: ''
