+++ HOW_TO_CONTRIBUTE.md
@@ -1,0 +1,39 @@
+How to contribute to the GEOS project
+-------------------------------------
+
+Architecture committee
+- Prior to opening a PR, a developer should present the idea to the Committee
+    - Go/No go decisions in reasonable time (1 week)
+- Any development which significantly modifies the architecture of GEOS must be validated by the committee before merging
+    - with review
+        - Logs
+        - Ghosting
+        - add/remove TPL dependency
+        - High-order methods / shiva
+    - without review
+        - Bugfixes
+        - documentation
+        - numerical methods tuning
+- bi-weekly + on-demand (it’s the developer’s duty to ask for a slot on the #dev-infrastructure slack channel)
+- who:
+    - 1 person of each entity + every person which skillset matches (+/- product owner of components).
+    - Everybody’s welcome to attend.
+
+Every day management.
+- Pull request management
+    - Reviewer selection confirmation for all the developments that went through the Architecture committee.
+        - Make sure the owner/s of the impacted components of GEOS are included
+        - Ensure a mix of junior and senior developers is involved in each PR
+    - Track progress
+- Issue (ticket) management
+    - Triaging by the committee
+        - Meeting frequency (weekly to take care of the current backlog, later possibly less frequent)
+    - Track progress
+- Assignments
+    - Each entity can assign tasks to their people.
+    - If resource needs to hop from one entity to another, collegial discussion will happen during this meeting
+- Every week.
+
+Tools
+- Everything currently happens on GitHub.
+- New tools will be evaluated depending on the needs.
