+++ HOW_TO_CONTRIBUTE.md
@@ -1,39 +1,112 @@
-How to contribute to the GEOS project
--------------------------------------
+How is the GEOS project organized
+=================================
 
-Architecture committee
-- Prior to opening a PR, a developer should present the idea to the Committee
-    - Go/No go decisions in reasonable time (1 week)
-- Any development which significantly modifies the architecture of GEOS must be validated by the committee before merging
-    - with review
-        - Logs
-        - Ghosting
-        - add/remove TPL dependency
-        - High-order methods / shiva
-    - without review
-        - Bugfixes
-        - documentation
-        - numerical methods tuning
-- bi-weekly + on-demand (it’s the developer’s duty to ask for a slot on the #dev-infrastructure slack channel)
-- who:
-    - 1 person of each entity + every person which skillset matches (+/- product owner of components).
-    - Everybody’s welcome to attend.
+The **Architecture Committee** oversees significant changes to the GEOS architecture, ensuring compliance with organizational standards and security measures. Developers must present their ideas to the committee for initial approval before opening a Pull Request (PR), with the committee aiming to provide decisions within a week. The committee meets bi-weekly and on-demand, including representatives from each entity and relevant stakeholders.
 
-Every day management.
-- Pull request management
-    - Reviewer selection confirmation for all the developments that went through the Architecture committee.
-        - Make sure the owner/s of the impacted components of GEOS are included
-        - Ensure a mix of junior and senior developers is involved in each PR
-    - Track progress
-- Issue (ticket) management
-    - Triaging by the committee
-        - Meeting frequency (weekly to take care of the current backlog, later possibly less frequent)
-    - Track progress
-- Assignments
-    - Each entity can assign tasks to their people.
-    - If resource needs to hop from one entity to another, collegial discussion will happen during this meeting
-- Every week.
+The **Operations Committee** handles the day-to-day management of the project, including managing pull requests (PRs), resolving tickets/issues, and assigning tasks. This committee ensures smooth development workflows and aligns daily operations with project goals.
 
-Tools
+Tools:
 - Everything currently happens on GitHub.
 - New tools will be evaluated depending on the needs.
+
+
+Architecture Committee
+----------------------
+
+## 1. General information
+
+- **Committee Overview:**
+  - The architecture committee is responsible for overseeing significant changes to the system architecture.
+  - Prior to opening a Pull Request (PR), developers must present their idea to the committee for initial feedback and approval.
+
+- **Meeting Schedule:**
+  - The committee meets bi-weekly and on-demand.
+  - It is the developer's responsibility to request a slot for their presentation via the ``#dev-infrastructure`` Slack channel.
+    If you do not have access to our Slack workspace, you can contact get in contact be mail (if you know a member of the organization) or simply by opening an issue on GitHub.
+
+- **Committee Members:**
+  - The committee consists of one representative from each entity involved in the project.
+  - Additional members with relevant skill sets and the product owner of affected components are included.
+  - Meetings are open for all interested parties to attend.
+
+- **Changes Requiring Committee Review:**
+  - Any development significantly modifying the architecture.
+  - Changes involving logs, ghosting, adding/removing TPL dependencies, or high-order methods (e.g., Shiva).
+  - Significant modifications to the architecture must be validated by the committee before merging.
+
+- **Changes Not Requiring Committee Review:**
+  - Bug fixes, documentation updates, and numerical methods tuning.
+
+## 2. Presenting your idea to the committee
+- **Description of Change:**
+  - Clearly describe the proposed change, including the motivation and problem it solves.
[…]
