+++ scripts/repositoryStats_condenseDev.py
@@ -1,0 +1,186 @@
+import sys
+
+#
+# git log --since="2010-01-01" --until="2026-11-31" --all --pretty="%an" | sort | uniq | tr -d " ' " | xargs python3 repositoryStats_condenseDev.py
+#
+
+# Define the username mapping dictionary
+username_mapping = {
+"AleksNovikov":"aleksnovikov",
+"AlexandreLapene":"alexandrelapene",
+"AndreCosta":"andrecosta",
+"AndreMacieiraBragaCosta":"andrecosta",
+"AndreaBorio":"andreaborio",
+"AndreaFranceschini":"andreafranceschini",
+"AntoineMazuyer":"antoinemazuyer",
+"AntoineMazuyer":"antoinemazuyer",
+"ArnaudDUDES":"arnauddudes",
+"ArturoVargas":"arturovargas",
+"Aurelien":"aurelien",
+"BESSETJulien":"julienrenethomasbesset",
+"Ben":"bencorbett",
+"BenCorbett":"bencorbett",
+"BenjaminCorbett":"bencorbett",
+"BenjaminCurticeCorbett":"bencorbett",
+"BenjaminT.Liu":"benjamintliu",
+"BertrandThierry":"bertrandthierry",
+"BrianHan":"brianhan",
+"BrianManhHienHan":"brianhan",
+"CameronCrook":"cameroncrook",
+"CameronMikelCrook":"cameroncrook",
+"Castelletto":"castelletto",
+"Castelletto1":"castelletto",
+"ChaoyiWang":"chaoyiwang",
+"ChrisWhite":"chriswhite",
+"ChristopherSherman":"christophersherman",
+"DanielOsei-Kuffuor":"danieloseikuffuor",
+"DicksonKachuma":"dicksonkachuma",
+"FanFei":"frankfeifan",
+"Fei":"frankfeifan",
+"frankfeifan":"frankfeifan",
+"FrancoisHamon":"francoishamon",
+"GuotongRen":"guotongren",
+"Guotong-Ren":"guotongren",
+"Gaetan":"gaetanfuss",
+"GaetanFuss":"gaetanfuss",
+"HerveGross":"hervegross",
+"HeweiTang":"heweitang",
+"HuiWu":"huiwu",
+"IbrahimIsraa":"ibrahimisraa",
+"IsaacJu":"isaacju",
+"IsaacJu-debug":"isaacju",
+"JacquesFranc":"jacquesfranc",
+"JaisreeIyer":"jaisreeiyer",
+"JamesCorbett":"jamescorbett",
+"JeannePellerin":"jeannepellerin",
+"JianHUANG":"jianhuang",
+"JianHuang":"jianhuang",
+"JieMeng":"jiaminjiang",
+"JixiangHuang":"jixianghuang",
+"JoshuaWhite":"joshuawhite",
+"Julien":"julienrenethomasbesset",
+"JulienReneThomasBesset":"julienrenethomasbesset",
+"KevinJ.Dugan":"kevinjdugan",
+"LionelUntereiner":"lioneluntereiner",
+"LucasVALARCHER":"lucasvalarcher",
+"MadonnaYoder":"madonnayoder",
+"MamadouN'diaye":"mamadoundiaye",
+"MamadouNdiaye":"mamadoundiaye",
+"MarkKhait":"markkhait",
+"Matteo":"matteocusini",
+"MatteoCusini":"matteocusini",
+"Matthias":"matthiascremon",
+"MatthiasCremon":"matthiascremon",
+"MelReyCG":"melreycg",
+"MichaelHomel":"michaelhomel",
+"MiladBader":"miladbader",
+"MohammadKarimi-Fard":"mohammadkarimifard",
+"NicolaCastelletto":"castelletto",
+"Ouassim":"ouassim",
[…]
