+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/functions/BrineEnthalpy.hpp
@@ -91,7 +91,7 @@ public:
                  string_array const & inputParams,
                  string_array const & componentNames,
                  array1d< real64 > const & componentMolarWeight,
-                 TableOutputOptions pvtOutputOpts );
+                 TableOutputOptions const pvtOutputOpts );
 
   static string catalogName() { return "BrineEnthalpy"; }
 
