+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/functions/WaterDensity.hpp
@@ -75,7 +75,7 @@ public:
                 string_array const & inputParams,
                 string_array const & componentNames,
                 array1d< real64 > const & componentMolarWeight,
-                TableOutputOptions pvtOutputOpts );
+                TableOutputOptions const pvtOutputOpts );
 
   static string catalogName() { return "WaterDensity"; }
   virtual string getCatalogName() const final { return catalogName(); }
