+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/functions/CO2Solubility.hpp
@@ -116,7 +116,7 @@ public:
                  string_array const & phaseNames,
                  string_array const & componentNames,
                  array1d< real64 > const & componentMolarWeight,
-                 PVTProps::FlashModelBase::TableOutputOptions pvtOutputOpts );
+                 PVTProps::FlashModelBase::TableOutputOptions const pvtOutputOpts );
 
   static string catalogName() { return "CO2Solubility"; }
 
