+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/functions/PhillipsBrineDensity.hpp
@@ -85,7 +85,7 @@ public:
                         string_array const & inputParams,
                         string_array const & componentNames,
                         array1d< real64 > const & componentMolarWeight,
-                        TableOutputOptions pvtOutputOpts );
+                        TableOutputOptions const pvtOutputOpts );
 
   static string catalogName() { return "PhillipsBrineDensity"; }
 
