+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/PhaseModel.hpp
@@ -61,7 +61,7 @@ struct PhaseModel
               array1d< array1d< string > > const & inputParams,
               string_array const & componentNames,
               array1d< real64 > const & componentMolarWeight,
-              PVTProps::PVTFunctionBase::TableOutputOptions const  pvtOutputOpts )
+              PVTProps::PVTFunctionBase::TableOutputOptions const pvtOutputOpts )
     : density( phaseModelName + "_" + Density::catalogName(),
                inputParams[InputParamOrder::DENSITY],
                componentNames,
