+++ src/coreComponents/schema/docs/LinearSolverParameters.rst
@@ -1,44 +1,86 @@
 
 
-============================= ============================================== ============= ======================================================================================================================================================================================================================================================================================================================= 
-Name                          Type                                           Default       Description                                                                                                                                                                                                                                                                                                             
-============================= ============================================== ============= ======================================================================================================================================================================================================================================================================================================================= 
-amgAggressiveCoarseningLevels integer                                        0             AMG number of levels for aggressive coarsening                                                                                                                                                                                                                                                                          
-amgAggressiveCoarseningPaths  integer                                        1             AMG number of paths for aggressive coarsening                                                                                                                                                                                                                                                                           
-amgAggressiveInterpType       geos_LinearSolverParameters_AMG_AggInterpType  multipass     AMG aggressive interpolation algorithm. Available options are: ``default\|extendedIStage2\|standardStage2\|extendedStage2\|multipass\|modifiedExtended\|modifiedExtendedI\|modifiedExtendedE\|modifiedMultipass``                                                                                                       
-amgCoarseSolver               geos_LinearSolverParameters_AMG_CoarseType     direct        AMG coarsest level solver/smoother type. Available options are: ``default\|jacobi\|l1jacobi\|fgs\|sgs\|l1sgs\|chebyshev\|direct\|bgs``                                                                                                                                                                                  
-amgCoarseningType             geos_LinearSolverParameters_AMG_CoarseningType HMIS          AMG coarsening algorithm. Available options are: ``default\|CLJP\|RugeStueben\|Falgout\|PMIS\|HMIS``                                                                                                                                                                                                                    
-amgInterpolationMaxNonZeros   integer                                        4             AMG interpolation maximum number of nonzeros per row                                                                                                                                                                                                                                                                    
-amgInterpolationType          geos_LinearSolverParameters_AMG_InterpType     extendedI     AMG interpolation algorithm. Available options are: ``default\|modifiedClassical\|direct\|multipass\|extendedI\|standard\|extended\|directBAMG\|modifiedExtended\|modifiedExtendedI\|modifiedExtendedE``                                                                                                                
-amgNullSpaceType              geos_LinearSolverParameters_AMG_NullSpaceType  constantModes AMG near null space approximation. Available options are:``constantModes\|rigidBodyModes``                                                                                                                                                                                                                              
-amgNumFunctions               integer                                        1             AMG number of functions                                                                                                                                                                                                                                                                                                 
-amgNumSweeps                  integer                                        1             AMG smoother sweeps                                                                                                                                                                                                                                                                                                     
-amgRelaxWeight                real64                                         1             AMG relaxation factor for the smoother                                                                                                                                                                                                                                                                                  
-amgSeparateComponents         integer                                        0             AMG apply separate component filter for multi-variable problems                                                                                                                                                                                                                                                         
-amgSmootherType               geos_LinearSolverParameters_AMG_SmootherType   l1sgs         AMG smoother type. Available options are: ``default\|jacobi\|l1jacobi\|fgs\|bgs\|sgs\|l1sgs\|chebyshev\|ilu0\|ilut\|ic0\|ict``                                                                                                                                                                                          
-amgThreshold                  real64                                         0             AMG strength-of-connection threshold                                                                                                                                                                                                                                                                                    
-directCheckResidual           integer                                        0             Whether to check the linear system solution residual                                                                                                                                                                                                                                                                    
-directColPerm                 geos_LinearSolverParameters_Direct_ColPerm     metis         How to permute the columns. Available options are: ``none\|MMD_AtplusA\|MMD_AtA\|colAMD\|metis\|parmetis``                                                                                                                                                                                                              
-directEquil                   integer                                        1             Whether to scale the rows and columns of the matrix                                                                                                                                                                                                                                                                     
[…]
