+++ src/coreComponents/schema/docs/SurfaceGenerator.rst
@@ -1,37 +1,46 @@
 
 
-========================= ================== ======== ============================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================ 
-Name                      Type               Default  Description                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
-========================= ================== ======== ============================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================================ 
-cflFactor                 real64             0.5      Factor to apply to the `CFL condition <http://en.wikipedia.org/wiki/Courant-Friedrichs-Lewy_condition>`_ when calculating the maximum allowable time step. Values should be in the interval (0,1]                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
-fractureRegion            groupNameRef       Fracture (no description available)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
-initialDt                 real64             1e+99    Initial time-step value required by the solver to the event manager.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
-isPoroelastic             integer            0        Flag that defines whether the material is poroelastic or not.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
-logLevel                  integer            1        | Sets the level of information to write in the standard output (the console typically).                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
-                                                      | A level of 0 outputs minimal information, higher levels require more.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
-                                                      | 1 :                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
-                                                      | - Display information about line search                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
-                                                      | - Display global solution scaling factor                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
-                                                      | 1 and step remaining :                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
-                                                      | - Display informations about timestep                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
-                                                      | only with loglevel = 2 :                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
-                                                      | - Display informations about linear algebra objects (matrices and vectors)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
-                                                      | 1 :                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
-                                                      | - Display informations about splitting node/edge/faces                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
-                                                      |                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
-                                                      | 2 :                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
-                                                      | - Display informations about all elements attached to the nodeID                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
-                                                      | - Display informations about all faces on their element                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
-                                                      |                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
[…]
