+++ src/coreComponents/schema/docs/SinglePhasePoromechanicsInitialization.rst
@@ -1,13 +1,15 @@
 
 
-=========================== ============ ======== ============================================================================================================================================================= 
-Name                        Type         Default  Description                                                                                                                                                   
-=========================== ============ ======== ============================================================================================================================================================= 
-logLevel                    integer      1        | Sets the level of information to write in the standard output (the console typically).                                                                        
-                                                  | A level of 0 outputs minimal information, higher levels require more.                                                                                         
-name                        groupName    required A name is required for any non-unique nodes                                                                                                                   
-performStressInitialization integer      required Flag to indicate that the solver is going to perform stress initialization                                                                                    
-poromechanicsSolverName     groupNameRef required Name of the poromechanics solver                                                                                                                              
-=========================== ============ ======== ============================================================================================================================================================= 
+============================ ============ ======== ======================================================================================================================================================================================================================================= 
+Name                         Type         Default  Description                                                                                                                                                                                                                             
+============================ ============ ======== ======================================================================================================================================================================================================================================= 
+logLevel                     integer      0        | Sets the level of information to write in the standard output (the console typically).                                                                                                                                                  
+                                                   | A level of 0 outputs minimal information, higher levels require more.                                                                                                                                                                   
+                                                   | logLevel >= 1                                                                                                                                                                                                                           
+                                                   |  - Information about physics solver stress initialization                                                                                                                                                                               
+name                         groupName    required A name is required for any non-unique nodes                                                                                                                                                                                             
+poromechanicsSolverName      groupNameRef required Name of the poromechanics solver                                                                                                                                                                                                        
+solidMechanicsStatisticsName groupNameRef          Name of the solid mechanics statistics                                                                                                                                                                                                  
+============================ ============ ======== ======================================================================================================================================================================================================================================= 
 
 

[…]
