+++ src/coreComponents/physicsSolvers/fluidFlow/wells/CompositionalMultiphaseWellKernels.cpp
@@ -473,118 +473,6 @@ INST_PressureRelationKernel( 4, 1 );
 INST_PressureRelationKernel( 5, 0 );
 INST_PressureRelationKernel( 5, 1 );
 
-
-/******************************** VolumeBalanceKernel ********************************/
-
-template< integer NC >
-GEOS_HOST_DEVICE
-void
-VolumeBalanceKernel::
-  compute( integer const numPhases,
-           real64 const & volume,
-           arraySlice1d< real64 const, compflow::USD_PHASE - 1 > const & phaseVolFrac,
-           arraySlice2d< real64 const, compflow::USD_PHASE_DC - 1 > const & dPhaseVolFrac,
-           real64 & localVolBalance,
-           real64 ( & localVolBalanceJacobian )[NC+1] )
-{
-  using Deriv = multifluid::DerivativeOffset;
-
-  localVolBalance = 1.0;
-  for( integer ic = 0; ic < NC+1; ++ic )
-  {
-    localVolBalanceJacobian[ic] = 0.0;
-  }
-
-  // sum contributions to component accumulation from each phase
-  for( integer ip = 0; ip < numPhases; ++ip )
-  {
-    localVolBalance -= phaseVolFrac[ip];
-    localVolBalanceJacobian[0] -= dPhaseVolFrac[ip][Deriv::dP];
-
-    for( integer jc = 0; jc < NC; ++jc )
-    {
-      localVolBalanceJacobian[jc + 1] -= dPhaseVolFrac[ip][Deriv::dC+jc];
-    }
-  }
-
-  // scale saturation-based volume balance by pore volume (for better scaling w.r.t. other equations)
-  for( integer idof = 0; idof < NC+1; ++idof )
-  {
-    localVolBalanceJacobian[idof] *= volume;
-  }
-  localVolBalance *= volume;
-}
-
-template< integer NC >
-void
-VolumeBalanceKernel::
-  launch( localIndex const size,
-          integer const numPhases,
-          globalIndex const rankOffset,
-          arrayView1d< globalIndex const > const & wellElemDofNumber,
-          arrayView1d< integer const > const & wellElemGhostRank,
-          arrayView2d< real64 const, compflow::USD_PHASE > const & wellElemPhaseVolFrac,
-          arrayView3d< real64 const, compflow::USD_PHASE_DC > const & dWellElemPhaseVolFrac,
-          arrayView1d< real64 const > const & wellElemVolume,
-          CRSMatrixView< real64, globalIndex const > const & localMatrix,
-          arrayView1d< real64 > const & localRhs )
-{
-  forAll< parallelDevicePolicy<> >( size, [=] GEOS_HOST_DEVICE ( localIndex const iwelem )
-  {
-
-    if( wellElemGhostRank[iwelem] >= 0 )
-    {
-      return;
-    }
-
-    real64 localVolBalance = 1.0;
-    real64 localVolBalanceJacobian[NC+1]{};
-
-    compute< NC >( numPhases,
-                   wellElemVolume[iwelem],
-                   wellElemPhaseVolFrac[iwelem],
-                   dWellElemPhaseVolFrac[iwelem],
-                   localVolBalance,
-                   localVolBalanceJacobian );
-
-    // get equation/dof indices
-    localIndex const localVolBalanceEqnIndex = wellElemDofNumber[iwelem] - rankOffset + ROFFSET::MASSBAL + NC;
-    globalIndex localVolBalanceDOF[NC+1]{};
-    for( integer jdof = 0; jdof < NC+1; ++jdof )
-    {
-      localVolBalanceDOF[jdof] = wellElemDofNumber[iwelem] + COFFSET::DPRES + jdof;
-    }
-
[…]
