### 2024-06-19 be211f8
Augmented Lagrangian (stick mode) - InterfaceKernelBase - Bubble Functions (#3130)
