### 2024-06-19 e275829
Augmented Lagrangian (stick mode) - InterfaceKernelBase - Bubble Functions (#3130)
