### 2024-06-19 949a758
Augmented Lagrangian (stick mode) - InterfaceKernelBase - Bubble Functions (#3130)
