### 2024-06-19 cf8d23f
Augmented Lagrangian (stick mode) - InterfaceKernelBase - Bubble Functions (#3130)
