+++ src/coreComponents/physicsSolvers/contact/SolidMechanicsALMKernelsBase.hpp
@@ -1,0 +1,273 @@
+/*
+ * ------------------------------------------------------------------------------------------------------------
+ * SPDX-License-Identifier: LGPL-2.1-only
+ *
+ * Copyright (c) 2018-2020 Lawrence Livermore National Security LLC
+ * Copyright (c) 2018-2020 The Board of Trustees of the Leland Stanford Junior University
+ * Copyright (c) 2018-2020 Total, S.A
+ * Copyright (c) 2019-     GEOSX Contributors
+ * All rights reserved
+ *
+ * See top level LICENSE, COPYRIGHT, CONTRIBUTORS, NOTICE, and ACKNOWLEDGEMENTS files for details.
+ * ------------------------------------------------------------------------------------------------------------
+ */
+
+/**
+ * @file SolidMechanicsALMKernelsBase.hpp
+ */
+
+#ifndef GEOS_PHYSICSSOLVERS_CONTACT_SOLIDMECHANICSALMKERNELSBASE_HPP_
+#define GEOS_PHYSICSSOLVERS_CONTACT_SOLIDMECHANICSALMKERNELSBASE_HPP_
+
+#include "finiteElement/kernelInterface/InterfaceKernelBase.hpp"
+#include "SolidMechanicsALMKernelsHelper.hpp"
+
+namespace geos
+{
+
+namespace solidMechanicsALMKernels
+{
+
+/**
+ * @brief Implements kernels for ALM.
+ * @copydoc geos::finiteElement::InterfaceKernelBase
+ *
+ */
+template< typename CONSTITUTIVE_TYPE,
+          typename FE_TYPE >
+class ALMKernelsBase :
+  public finiteElement::InterfaceKernelBase< CONSTITUTIVE_TYPE,
+                                             FE_TYPE,
+                                             3, 3 >
+{
+public:
+  /// Alias for the base class;
+  using Base = finiteElement::InterfaceKernelBase< CONSTITUTIVE_TYPE,
+                                                   FE_TYPE,
+                                                   3, 3 >;
+
+  /// Number of nodes per element...which is equal to the
+  /// numTestSupportPointPerElem and numTrialSupportPointPerElem by definition.
+  static constexpr int numNodesPerElem = Base::maxNumTestSupportPointsPerElem;
+
+  /// Compile time value for the number of quadrature points per element.
+  static constexpr int numQuadraturePointsPerElem = FE_TYPE::numQuadraturePoints;
+
+  using Base::m_dofNumber;
+  using Base::m_dofRankOffset;
+  using Base::m_finiteElementSpace;
+  using Base::m_matrix;
+  using Base::m_rhs;
+
+  /**
+   * @brief Constructor
+   * @copydoc geos::finiteElement::InterfaceKernelBase::InterfaceKernelBase
+   */
+  ALMKernelsBase( NodeManager const & nodeManager,
+                  EdgeManager const & edgeManager,
+                  FaceManager const & faceManager,
+                  localIndex const targetRegionIndex,
+                  FaceElementSubRegion & elementSubRegion,
+                  FE_TYPE const & finiteElementSpace,
+                  CONSTITUTIVE_TYPE & inputConstitutiveType,
+                  arrayView1d< globalIndex const > const uDofNumber,
+                  arrayView1d< globalIndex const > const bDofNumber,
+                  globalIndex const rankOffset,
+                  CRSMatrixView< real64, globalIndex const > const inputMatrix,
+                  arrayView1d< real64 > const inputRhs,
+                  real64 const inputDt,
+                  arrayView1d< localIndex const > const & faceElementList ):
+    Base( nodeManager,
+          edgeManager,
+          faceManager,
+          targetRegionIndex,
+          elementSubRegion,
[…]
