+++ src/coreComponents/finiteElement/elementFormulations/LagrangeBasis1.hpp
@@ -107,6 +107,18 @@ public:
     return 0.5 + 0.5 * xi;
   }
 
+  /**
+   * @brief The value of the bubble basis function.
+   * @param xi The coordinate at which to evaluate the basis.
+   * @return The value of the basis.
+   */
+  GEOS_HOST_DEVICE
+  GEOS_FORCE_INLINE
+  constexpr static real64 valueBubble( const real64 xi )
+  {
+    return 1.0 - pow( xi, 2 );
+  }
+
 
   /**
    * @brief The gradient of the basis function for a support point evaluated at
@@ -151,6 +163,20 @@ public:
   {
     GEOS_UNUSED_VAR( xi );
     return 0.5;
+  }
+
+  /**
+   * @brief The gradient of the bubble basis function for support point 1 evaluated at
+   *   a point along the axes.
+   * @param xi The coordinate at which to evaluate the gradient.
+   * @return The gradient of basis function
+   */
+  GEOS_HOST_DEVICE
+  GEOS_FORCE_INLINE
+  constexpr static real64 gradientBubble( const real64 xi )
+  {
+    GEOS_UNUSED_VAR( xi );
+    return -0.5*xi;
   }
 
   /**
@@ -247,6 +273,22 @@ public:
     }
 
     /**
+     * @brief The value of the bubble basis function evaluated at a
+     *   point along the axes.
+     *
+     * @param coords The coordinates (in the parent frame) at which to evaluate the basis
+     * @param N Array to hold the value of the basis functions.
+     */
+    GEOS_HOST_DEVICE
+    GEOS_FORCE_INLINE
+    static void valueBubble( real64 const (&coords)[2],
+                             real64 (& N)[1] )
+    {
+      N[0] = LagrangeBasis1::valueBubble( coords[0] ) *
+             LagrangeBasis1::valueBubble( coords[1] );
+    }
+
+    /**
      * @brief The parent coordinates for a support point in the xi0 direction.
      * @param linearIndex The linear index of the support point
      * @return
@@ -297,6 +339,9 @@ public:
   {
     /// The number of support points in the basis.
     constexpr static localIndex numSupportPoints = 8;
+
+    /// The number of support faces in the basis.
+    constexpr static localIndex numSupportFaces = 6;
 
     /**
      * @brief Calculates the linear index for support/quadrature points from ijk
@@ -359,6 +404,80 @@ public:
                           LagrangeBasis1::value( c, coords[2] );
           }
         }
+      }
+    }
+
+    /**
+     * @brief The value of the bubble basis function for a support face evaluated at a
+     *   point along the axes.
+     *
+     * @param coords The coordinates (in the parent frame) at which to evaluate the basis
+     * @param N Array to hold the value of the basis functions at each support face.
+     */
+    GEOS_HOST_DEVICE
+    GEOS_FORCE_INLINE
+    static void valueFaceBubble( real64 const (&coords)[3],
+                                 real64 (& N)[numSupportFaces] )
+    {
+      for( int a=0; a<2; ++a )
+      {
+        N[ a*5 ]   = LagrangeBasis1::valueBubble( coords[0] ) *
+                     LagrangeBasis1::valueBubble( coords[1] ) *
+                     LagrangeBasis1::value( a, coords[2] );
+
+        N[ a*3+1 ] = LagrangeBasis1::valueBubble( coords[0] ) *
+                     LagrangeBasis1::value( a, coords[1] ) *
+                     LagrangeBasis1::valueBubble( coords[2] );
+
+        N[ a+2 ]   = LagrangeBasis1::value( a, coords[0] ) *
+                     LagrangeBasis1::valueBubble( coords[1] ) *
[…]
