### 2024-06-19 b448aed
Minor typo fixes and now check if surface normals are zeroed instead of particle flags during mapping of surface mass
