### 2024-06-19 3bf12d2
Fix tutorial example for thermal debonding wellbore problem (#3170)
