+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3130-5671-b19ba86
+  baseline: integratedTests/baseline_integratedTests-pr3170-5687-d6e16c7
 
 allow_fail:
   all: ''
