### 2024-06-19 44b4666
Fix tutorial example for thermal debonding wellbore problem (#3170)
