### 2024-06-19 25e5f8a
Removed particle header input files and add support for reading particle type from particle file
