### 2024-06-19 0f84a74
Removed particle header input files and add support for reading particle type from particle file
