+++ src/coreComponents/mesh/particleGenerators/ParticleMeshGenerator.cpp
@@ -38,20 +38,15 @@ ParticleMeshGenerator::ParticleMeshGenerator( string const & name, Group * const
     setRestartFlags( RestartFlags::NO_WRITE ).
     setDescription( "path to the particle file" );
 
-  registerWrapper( viewKeyStruct::headerFilePathString(), &m_headerFilePath ).
-    setInputFlag( InputFlags::REQUIRED ).
-    setRestartFlags( RestartFlags::NO_WRITE ).
-    setDescription( "path to the header file" );
+  // registerWrapper( viewKeyStruct::headerFilePathString(), &m_headerFilePath ).
+  //   setInputFlag( InputFlags::REQUIRED ).
+  //   setRestartFlags( RestartFlags::NO_WRITE ).
+  //   setDescription( "path to the header file" );
 
   registerWrapper( viewKeyStruct::particleBlockNamesString(), &m_blockNames ).
     setInputFlag( InputFlags::REQUIRED ).
     setSizedFromParent( 0 ).
     setDescription( "Names of each particle block" );
-
-  registerWrapper( viewKeyStruct::particleMaterialNamesString(), &m_materialNames ).
-    setInputFlag( InputFlags::REQUIRED ).
-    setSizedFromParent( 0 ).
-    setDescription( "Names of each particle material, index corresponds to material ID in particle file");
 
   registerWrapper( viewKeyStruct::particleTypesString(), &m_particleTypes ).
     setInputFlag( InputFlags::REQUIRED ).
@@ -65,69 +60,74 @@ Group * ParticleMeshGenerator::createChild( string const & GEOS_UNUSED_PARAM( ch
   return nullptr;
 }
 
+void ParticleMeshGenerator::postProcessInput()
+{
+  ParticleMeshGeneratorBase::postProcessInput();
+
+  GEOS_ERROR_IF(m_blockNames.size() == 0, "No particle blocks were specified! Must specify at least one particle block.");
+  GEOS_ERROR_IF(m_blockNames.size() != m_particleTypes.size(), "The particle block and type lists must have the same size.");
+
+  for( int i = 0; i < m_particleTypes.size(); i++)
+  {
+    for( int j = 0; j < EnumSize<ParticleType>; j++ )
+    {
+      if( m_particleTypes[i] == EnumStrings< ParticleType >::toString( static_cast< ParticleType >( j ) ) )
+      {
+        break;
+      }
+
+      if( j == EnumSize<ParticleType>-1)
+      {
+        GEOS_ERROR( "No particle type of " << m_particleTypes[i] << " in particleTypes input. Available options are SinglePoint (0), SinglePointBSpline (1), CPDI (2), CPTI (3), CPDI2 (4)."); // TODO Make available options dynamic
+      }
+    }
+  }
+}
 
 void ParticleMeshGenerator::fillParticleBlockManager( ParticleBlockManager & particleBlockManager, ParticleManager & particleManager, SpatialPartition const & partition )
 {
   GEOS_MARK_FUNCTION;
 
-  // This should probably handled elsewhere:
-  int aa = 0;
-  for( auto & particleBlockName : m_blockNames )
+  int numBlocks = m_blockNames.size();
+
+  // This should probably be handled elsewhere:
+  std::vector<int> blockMaterialIndices( numBlocks );
+  for(int b = 0; b < numBlocks; b++ )
   {
-    ParticleBlock & particleBlock = particleBlockManager.registerParticleBlock( particleBlockName );
-    particleBlock.setParticleType( EnumStrings< ParticleType >::fromString( m_particleTypes[aa++] ) );
+    ParticleBlock & particleBlock = particleBlockManager.registerParticleBlock( m_blockNames[b] );
[…]
