### 2024-06-20 b47ffe8
fix and updates rst
