+++ src/coreComponents/physicsSolvers/multiphysics/SinglePhasePoromechanicsConformingFractures.cpp
@@ -42,10 +42,9 @@ SinglePhasePoromechanicsConformingFractures< FLOW_SOLVER >::SinglePhasePoromecha
                                                                                                          Group * const parent )
   : Base( name, parent )
 {
-  appendLogLevelDescription( "logLevel >= 3", "The summary of declared fields and coupling" );
+  this->appendLogLevelDescription( "logLevel >= 3", "The summary of declared fields and coupling" );
   
   LinearSolverParameters & params = this->m_linearSolverParameters.get();
-  LinearSolverParameters & params = m_linearSolverParameters.get();
   params.mgr.strategy = LinearSolverParameters::MGR::StrategyType::singlePhasePoromechanicsConformingFractures;
   params.mgr.separateComponents = false;
   params.mgr.displacementFieldName = solidMechanics::totalDisplacement::key();
