### 2024-06-20 736b706
loglevel viewkeystruct adjustments
