+++ src/coreComponents/dataRepository/Group.hpp
@@ -1414,15 +1414,6 @@ public:
    */
   void setInputFlags( InputFlags flags ) { m_input_flags = flags; }
 
-  /**
-   * @struct viewKeyStruct holds char strings and viewKeys for fast lookup
-   */
-  struct viewKeyStruct
-  {
-    /// String for the logLevel wrapper
-    constexpr static char const * logLevelString() { return "logLevel"; }
-  };
-
   ///@}
 
   /**
@@ -1541,9 +1532,16 @@ protected:
   virtual void postRestartInitialization()
   {}
 
-
-
   ///@}
+
+  /**
+   * @brief viewKeyStruct holds char strings and viewKeys for fast lookup
+   */
+  struct viewKeyStruct
+  {
+    /// @return String for the logLevel wrapper
+    constexpr static char const * logLevelString() { return "logLevel"; }
+  };
 
 private:
   /**
