### 2024-06-20 a9c2fab
Solved a bug where datasets were duplicated in the hdf5
