+++ src/coreComponents/physicsSolvers/SolverBase.hpp
@@ -52,7 +52,7 @@ public:
   /**
    * @return Get the final class Catalog name
    */
-  virtual string getCatalogName() const = 0;
+  virtual string const getCatalogName() const = 0;
 
 
   virtual void registerDataOnMesh( Group & MeshBodies ) override;
