+++ src/coreComponents/constitutive/solid/ElasticOrthotropic.hpp
@@ -357,7 +357,7 @@ public:
    */
   static string catalogName() { return m_catalogNameString; }
 
-  virtual string getCatalogName() const override { return catalogName(); }
+  virtual string const getCatalogName() const override { return catalogName(); }
   ///@}
 
   /**
@@ -729,7 +729,7 @@ public:
   }
 
 protected:
-  virtual void postProcessInput() override;
+  virtual void postInputInitialization() override;
 
   /// The default value of the Young's modulus E1 for any new
   /// allocations.
