+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/functions/CO2Enthalpy.hpp
@@ -83,7 +83,7 @@ public:
 
   static string catalogName() { return "CO2Enthalpy"; }
 
-  virtual string getCatalogName() const final { return catalogName(); }
+  virtual string const getCatalogName() const final { return catalogName(); }
 
   /**
    * @copydoc PVTFunctionBase::checkTablesParameters( real64 pressure, real64 temperature )
