+++ src/coreComponents/physicsSolvers/solidMechanics/kernels/ImplicitSmallStrainQuasiStatic_impl.hpp
@@ -36,7 +36,7 @@ ImplicitSmallStrainQuasiStatic( NodeManager const & nodeManager,
                                 EdgeManager const & edgeManager,
                                 FaceManager const & faceManager,
                                 localIndex const targetRegionIndex,
-                                SUBREGION_TYPE & elementSubRegion,
+                                SUBREGION_TYPE const & elementSubRegion,
                                 FE_TYPE const & finiteElementSpace,
                                 CONSTITUTIVE_TYPE & inputConstitutiveType,
                                 arrayView1d< globalIndex const > const inputDofNumber,
@@ -60,8 +60,6 @@ ImplicitSmallStrainQuasiStatic( NodeManager const & nodeManager,
   m_X( nodeManager.referencePosition()),
   m_disp( nodeManager.getField< fields::solidMechanics::totalDisplacement >() ),
   m_uhat( nodeManager.getField< fields::solidMechanics::incrementalDisplacement >() ),
-  //m_incStrain( elementSubRegion.template getField< fields::solidMechanics::incrementalStrain >() ),
-  m_elementVolume( elementSubRegion.getElementVolume() ),
   m_gravityVector{ inputGravityVector[0], inputGravityVector[1], inputGravityVector[2] },
   m_density( inputConstitutiveType.getDensity() )
 {}
@@ -98,7 +96,6 @@ setup( localIndex const k,
       stack.localColDofIndex[a*3+i] = m_dofNumber[localNodeIndex]+i;
     }
   }
-
   // Add stabilization to block diagonal parts of the local jacobian
   // (this is a no-operation with FEM classes)
   real64 const stabilizationScaling = computeStabilizationScaling( k );
@@ -128,15 +125,6 @@ void ImplicitSmallStrainQuasiStatic< SUBREGION_TYPE, CONSTITUTIVE_TYPE, FE_TYPE 
   typename CONSTITUTIVE_TYPE::KernelWrapper::DiscretizationOps stiffness;
 
   FE_TYPE::symmetricGradient( dNdX, stack.uhat_local, strainInc );
-
-  //for( int is = 0; is < 6; ++is )
-  //{
-  //  if( q == 0 )
-  //  {
-  //    m_incStrain[k][is] = 0.0;
-  //  }
-  //  m_incStrain[k][is] += strainInc[is]*detJxW/m_elementVolume[k];
- // }
 
   m_constitutiveUpdate.smallStrainUpdate( k, q, m_dt, strainInc, stress, stiffness );
 
