+++ src/coreComponents/physicsSolvers/multiphysics/poromechanicsKernels/ThermalMultiphasePoromechanics.hpp
@@ -88,8 +88,6 @@ public:
   using Base::m_dFluidPhaseCompFrac;
   using Base::m_dGlobalCompFraction_dGlobalCompDensity;
   using Base::m_dt;
-  using Base::m_incStrain;
-  using Base::m_elementVolume;
 
   /**
    * @brief Constructor
@@ -100,7 +98,7 @@ public:
                                   EdgeManager const & edgeManager,
                                   FaceManager const & faceManager,
                                   localIndex const targetRegionIndex,
-                                  SUBREGION_TYPE & elementSubRegion,
+                                  SUBREGION_TYPE const & elementSubRegion,
                                   FE_TYPE const & finiteElementSpace,
                                   CONSTITUTIVE_TYPE & inputConstitutiveType,
                                   arrayView1d< globalIndex const > const inputDispDofNumber,
