+++ src/coreComponents/physicsSolvers/solidMechanics/SolidMechanicsStateReset.cpp
@@ -96,7 +96,6 @@ bool SolidMechanicsStateReset::execute( real64 const time_n,
       nodeManager.getField< solidMechanics::incrementalDisplacement >().zero();
     }
 
-
     // Option 2: enable / disable inelastic behavior
     ElementRegionManager & elementRegionManager = mesh.getElemManager();
     elementRegionManager.forElementSubRegions< CellElementSubRegion >( regionNames,
