### 2024-06-20 4181f6c
try to fix cuda compile error
