### 2024-06-20 04319db
try to fix cuda compile error
