### 2024-06-20 f10de97
try to fix cuda compile error
