+++ host-configs/LLNL/old/quartz-gcc-12.cmake
@@ -1,4 +1,4 @@
-include(${CMAKE_CURRENT_LIST_DIR}/../../src/coreComponents/LvArray/host-configs/LLNL/quartz-gcc-12.cmake)
+include(${CMAKE_CURRENT_LIST_DIR}../../../src/coreComponents/LvArray/host-configs/LLNL/old/quartz-gcc-12.cmake)
 
 # C++
 # The "-march=native -mtune=native" which LvArray adds breaks the PVT package.
