### 2024-06-21 b3e928f
Update BASELINE_NOTES.md
