+++ BASELINE_NOTES.md
@@ -6,6 +6,10 @@ This file is designed to track changes to the integrated test baselines.
 Any developer who updates the baseline ID in the .integrated_tests.yaml file is expected to create an entry in this file with the pull request number, date, and their justification for rebaselining.
 These notes should be in reverse-chronological order, and use the following time format: (YYYY-MM-DD).
 
+PR #3142 (2024-06-20)
+======================
+Adding output of total strain. Rebaseline because of new inclusion of strain in output.
+
 PR #3170 (2024-06-19)
 ======================
 Fix tutorial example for thermal debonding wellbore problem. Test case modified.
