### 2024-06-21 d6b6e0b
Adding field to output total strain (#3142)
