### 2024-06-21 30428a6
Adding field to output total strain (#3142)
