+++ .integrated_tests.yaml
@@ -1,6 +1,6 @@
 baselines:
   bucket: geosx
-  baseline: integratedTests/baseline_integratedTests-pr3170-5687-d6e16c7
+  baseline: integratedTests/baseline_integratedTests-pr3142-5705-ebdd3cb
 
 allow_fail:
   all: ''
