### 2024-06-21 14b9646
Adding field to output total strain (#3142)
