### 2024-06-21 adbd3c2
decouple debug matrix output from logLevel
