+++ src/coreComponents/physicsSolvers/SolverBase.hpp
@@ -584,7 +584,7 @@ public:
     static constexpr char const * discretizationString() { return "discretization"; }
     static constexpr char const * targetRegionsString() { return "targetRegions"; }
     static constexpr char const * meshTargetsString() { return "meshTargets"; }
-
+    static constexpr char const * writeMatrixString() { return "writeMatrix"; }
   };
 
   struct groupKeyStruct
@@ -803,6 +803,9 @@ protected:
 
   /// Custom preconditioner for the "native" iterative solver
   std::unique_ptr< PreconditionerBase< LAInterface > > m_precond;
+
+  /// flag for debug output of matrix, rhs, and solution
+  integer m_writeMatrix;
 
   /// Linear solver parameters
   LinearSolverParametersInput m_linearSolverParameters;
