+++ src/conf.py
@@ -18,11 +18,11 @@ import shutil
 
 # Add python modules to be documented
 python_root = './coreComponents/python/modules'
-python_modules = ('geosx_mesh_tools_package',
-                  'geosx_xml_tools_package',
-                  'hdf5_wrapper_package',
-                  'pygeosx_tools_package',
-                  'timehistory_package')
+python_modules = ('geos-mesh-tools',
+                  'geos-xml-tools',
+                  'hdf5-wrapper',
+                  'pygeos-tools',
+                  'geos-timehistory')
 for m in python_modules:
     sys.path.insert(0, os.path.abspath(os.path.join(python_root, m)))
 
