### 2024-06-21 371eb04
mpi parallel fix for closed wells
