### 2024-06-22 4b10843
Making it work with tpetra and integer matrices.
