+++ src/coreComponents/mesh/generators/NewGhosting.cpp
@@ -20,19 +20,16 @@
 
 #include "Pods.hpp"
 
+#include "Indices.hpp"
+
 #include "common/MpiWrapper.hpp"
 #include "common/DataTypes.hpp"
 
-#include <EpetraExt_MatrixMatrix.h>
-#include <EpetraExt_RowMatrixOut.h>
-#include <EpetraExt_VectorOut.h>
-#include <Epetra_CrsMatrix.h>
-#include <Epetra_Map.h>
-#include <Epetra_MpiComm.h>
-#include <Epetra_RowMatrixTransposer.h>
-#include <Epetra_Vector.h>
-
-#include "Indices.hpp"
+#include <Tpetra_Map.hpp>
+#include <Tpetra_CrsMatrix.hpp>
+#include <Teuchos_Comm.hpp>
+#include <TpetraExt_MatrixMatrix.hpp>
+#include <MatrixMarket_Tpetra.hpp>
 
 #include <vtkCellData.h>
 #include <vtkPointData.h>
@@ -272,16 +269,29 @@ std::tuple< MeshGraph, MeshGraph > buildMeshGraph( vtkSmartPointer< vtkDataSet >
   return { std::move( owned ), std::move( present ) };
 }
 
-std::unique_ptr< Epetra_CrsMatrix > makeTranspose( Epetra_CrsMatrix & input,
-                                                   bool makeDataContiguous = true )
-{
-  Epetra_RowMatrixTransposer transposer( &input );  // This process does not modify the original `ghosted` matrix.
-  Epetra_CrsMatrix * tr = nullptr;  // The transposer returns a pointer we must handle ourselves.
-  transposer.CreateTranspose( makeDataContiguous, tr );
+//using TriLocIdx = fluent::NamedType< localIndex, struct TagTriLocIdx, fluent::Arithmetic, fluent::ImplicitlyConvertibleTo >;
+//using TriGlbIdx = fluent::NamedType< globalIndex, struct TagTriGlbIdx, fluent::Arithmetic, fluent::ImplicitlyConvertibleTo >;
+//template<>
+//struct ::std::is_signed< TriLocIdx > : ::std::is_signed< typename TriLocIdx::UnderlyingType >
+//{
+//};
 
-  std::unique_ptr< Epetra_CrsMatrix > ptr;
-  ptr.reset( tr );
-  return ptr;
+using TriLocIdx = localIndex;
+using TriGlbIdx = globalIndex;
+
+using TriMap = Tpetra::Map< TriLocIdx, TriGlbIdx >;
+// We use matrices of integers for some of our usages
+// But it seems that Tpetra wants *signed* integers (at least the way we install it).
+// Consider using *unsigned* integers if you manage to make it work/link.
+using TriScalarInt = std::int32_t;
+using TriCrsMatrix = Tpetra::CrsMatrix< TriScalarInt , TriLocIdx, TriGlbIdx >;
+using TriDblCrsMatrix = Tpetra::CrsMatrix< double, TriLocIdx, TriGlbIdx >;
+using TriComm = Teuchos::Comm< int >;
+
+template< typename T, typename... ARGS >
+Teuchos::RCP< T > make_rcp( ARGS && ... args )
+{
+  return Teuchos::rcp( new T( std::forward< ARGS >( args )... ) );
 }
 
 void to_json( json & j,
@@ -303,48 +313,48 @@ void to_json( json & j,
 }
 
 
-Epetra_CrsMatrix multiply( int commSize,
-                           Epetra_CrsMatrix const & indicator,
-                           Epetra_CrsMatrix & upward )
+TriCrsMatrix multiply( int commSize,
[…]
