### 2024-06-22 77eef65
extractedValues is now a collection of integers, no more float
