+++ src/coreComponents/physicsSolvers/wavePropagation/shared/WaveSolverBase.hpp
@@ -119,11 +119,10 @@ public:
     static constexpr char const * parametersPMLString() { return "parametersPML"; }
 
     static constexpr char const * useTaperString() {return "useTaper";}
-    static constexpr char const * xMinTaperString() {return "xMinTaper" ;}
-    static constexpr char const * xMaxTaperString() {return "xMaxTaper" ;}
-    static constexpr char const * thicknessMinXYZTaperString() {return "thicknessMinXYZTaper" ;}
-    static constexpr char const * thicknessMaxXYZTaperString() {return "thicknessMaxXYZTaper" ;}
-    static constexpr char const * taperConstantString() {return "taperConstant" ;}
+    static constexpr char const * xMinTaperString() {return "xMinTaper";}
+    static constexpr char const * xMaxTaperString() {return "xMaxTaper";}
+    static constexpr char const * thicknessMinXYZTaperString() {return "thicknessMinXYZTaper";}
+    static constexpr char const * thicknessMaxXYZTaperString() {return "thicknessMaxXYZTaper";}
 
     static constexpr char const * receiverElemString() { return "receiverElem"; }
     static constexpr char const * receiverRegionString() { return "receiverRegion"; }
@@ -365,18 +364,15 @@ protected:
   /// A set of target nodes IDs that will be handled by the current solver
   SortedArray< localIndex > m_solverTargetNodesSet;
 
-   /// Mininum (x,y,z) coordinates of inner PML boundaries
-    R1Tensor32 m_xMinTaper;
+  /// Mininum (x,y,z) coordinates of inner PML boundaries
+  R1Tensor32 m_xMinTaper;
 
-    /// Maximum (x,y,z) coordinates of inner PML boundaries
-    R1Tensor32 m_xMaxTaper;
+  /// Maximum (x,y,z) coordinates of inner PML boundaries
+  R1Tensor32 m_xMaxTaper;
 
-    /// Desired constant for the taper, used to compute the damping profile
-    real32 m_taperConstant;
-
-    /// Thickness of the PML region, used to compute the damping profile
-    R1Tensor32 m_thicknessMinXYZTaper;
-    R1Tensor32 m_thicknessMaxXYZTaper;
+  /// Thickness of the PML region, used to compute the damping profile
+  R1Tensor32 m_thicknessMinXYZTaper;
+  R1Tensor32 m_thicknessMaxXYZTaper;
 
   struct parametersPML
   {
