+++ src/coreComponents/physicsSolvers/wavePropagation/shared/TaperKernel.hpp
@@ -24,187 +24,89 @@
 namespace geos
 {
 
-  using wsCoordType = WaveSolverUtils::wsCoordType;
+using wsCoordType = WaveSolverUtils::wsCoordType;
 
-  struct TaperKernel
+struct TaperKernel
+{
+
+  /**
+   * @brief Launches the computation of field gradients and divergence for PML region
+   * @tparam EXEC_POLICY the execution policy
+   * @tparam ATOMIC_POLICY the atomic policy
+   * @param[in] targetSet list of cells in the target set
+   * @param[in] nodeCoords coordinates of the nodes
+   * @param[in] elemToNodes constant array view of map from element to nodes
+   * @param[in] velocity cell-wise velocity
+   * @param[in] p_n pressure field at time n
+   * @param[in] v_n PML auxiliary field at time n
+   * @param[in] u_n PML auxiliary field at time n
+   * @param[in] xMin coordinate limits of the inner PML boundaries, left-front-top
+   * @param[in] xMax coordinate limits of the inner PML boundaries, right-back-bottom
+   * @param[in] xMin PML thickness, left-front-top
+   * @param[in] xMax PML thickness, right-back-bottom
+   * @param[in] cMin PML wave speed, left-front-top
+   * @param[in] cMax PML wave speed, right-back-bottom
+   * @param[in] r desired reflectivity of the PML
+   * @param[out] grad_n array holding the gradients at time n
+   * @param[out] divV_n array holding the divergence at time n
+   */
+  template< typename EXEC_POLICY >
+  static void
+  computeTaperCoeff( localIndex const size,
+                     arrayView2d< wsCoordType const, nodes::REFERENCE_POSITION_USD > const nodeCoords,
+                     R1Tensor32 const xMin,
+                     R1Tensor32 const xMax,
+                     R1Tensor32 const dMin,
+                     R1Tensor32 const dMax,
+                     arrayView1d< real32 > const taperCoeff )
   {
-
-    /**
-     * @brief Launches the computation of field gradients and divergence for PML region
-     * @tparam EXEC_POLICY the execution policy
-     * @tparam ATOMIC_POLICY the atomic policy
-     * @param[in] targetSet list of cells in the target set
-     * @param[in] nodeCoords coordinates of the nodes
-     * @param[in] elemToNodes constant array view of map from element to nodes
-     * @param[in] velocity cell-wise velocity
-     * @param[in] p_n pressure field at time n
-     * @param[in] v_n PML auxiliary field at time n
-     * @param[in] u_n PML auxiliary field at time n
-     * @param[in] xMin coordinate limits of the inner PML boundaries, left-front-top
-     * @param[in] xMax coordinate limits of the inner PML boundaries, right-back-bottom
-     * @param[in] xMin PML thickness, left-front-top
-     * @param[in] xMax PML thickness, right-back-bottom
-     * @param[in] cMin PML wave speed, left-front-top
-     * @param[in] cMax PML wave speed, right-back-bottom
-     * @param[in] r desired reflectivity of the PML
-     * @param[out] grad_n array holding the gradients at time n
-     * @param[out] divV_n array holding the divergence at time n
-     */
-    template< typename EXEC_POLICY>
-    static void
-    computeTaperCoeff( localIndex const size,
[…]
