### 2024-06-24 dec32be
Change taper coefficient
