+++ src/coreComponents/common/GeosxConfig.hpp.in
@@ -65,6 +65,9 @@
 /// Enables use of Hypre library (CMake option ENABLE_HYPRE)
 #cmakedefine GEOSX_USE_HYPRE
 
+/// Enables use of GRPC library (CMake option ENABLE_GRPC)
+#cmakedefine GEOSX_USE_GRPC
+
 #if defined( GEOSX_USE_HYPRE )
   /// Parsed hypre version information
   #define HYPRE_VERSION_MAJOR @HYPRE_VERSION_MAJOR@
