+++ src/coreComponents/schema/docs/CompositionalMultiphaseWell.rst
@@ -1,23 +1,25 @@
 
 
-============================= ================== ======== ====================================================================================================================================================================================================================================================================================================================== 
-Name                          Type               Default  Description                                                                                                                                                                                                                                                                                                            
-============================= ================== ======== ====================================================================================================================================================================================================================================================================================================================== 
-allowLocalCompDensityChopping integer            1        Flag indicating whether local (cell-wise) chopping of negative compositions is allowed                                                                                                                                                                                                                                 
-cflFactor                     real64             0.5      Factor to apply to the `CFL condition <http://en.wikipedia.org/wiki/Courant-Friedrichs-Lewy_condition>`_ when calculating the maximum allowable time step. Values should be in the interval (0,1]                                                                                                                      
-initialDt                     real64             1e+99    Initial time-step value required by the solver to the event manager.                                                                                                                                                                                                                                                   
-logLevel                      integer            0        | Sets the level of information to write in the standard output (the console typically).                                                                                                                                                                                                                                 
-                                                          | A level of 0 outputs minimal information, higher levels require more.                                                                                                                                                                                                                                                  
-maxAbsolutePressureChange     real64             -1       Maximum (absolute) pressure change in a Newton iteration                                                                                                                                                                                                                                                               
-maxCompFractionChange         real64             1        Maximum (absolute) change in a component fraction between two Newton iterations                                                                                                                                                                                                                                        
-maxRelativePressureChange     real64             1        Maximum (relative) change in pressure between two Newton iterations (recommended with rate control)                                                                                                                                                                                                                    
-name                          groupName          required A name is required for any non-unique nodes                                                                                                                                                                                                                                                                            
-targetRegions                 groupNameRef_array required Allowable regions that the solver may be applied to. Note that this does not indicate that the solver will be applied to these regions, only that allocation will occur such that the solver may be applied to these regions. The decision about what regions this solver will beapplied to rests in the EventManager. 
-useMass                       integer            0        Use total mass equation                                                                                                                                                                                                                                                                                                
-writeCSV                      integer            0        Write rates into a CSV file                                                                                                                                                                                                                                                                                            
-LinearSolverParameters        node               unique   :ref:`XML_LinearSolverParameters`                                                                                                                                                                                                                                                                                      
-NonlinearSolverParameters     node               unique   :ref:`XML_NonlinearSolverParameters`                                                                                                                                                                                                                                                                                   
-WellControls                  node                        :ref:`XML_WellControls`                                                                                                                                                                                                                                                                                                
-============================= ================== ======== ====================================================================================================================================================================================================================================================================================================================== 
+============================= ================== ============ ====================================================================================================================================================================================================================================================================================================================== 
+Name                          Type               Default      Description                                                                                                                                                                                                                                                                                                            
+============================= ================== ============ ====================================================================================================================================================================================================================================================================================================================== 
+allowLocalCompDensityChopping integer            1            Flag indicating whether local (cell-wise) chopping of negative compositions is allowed                                                                                                                                                                                                                                 
+cflFactor                     real64             0.5          Factor to apply to the `CFL condition <http://en.wikipedia.org/wiki/Courant-Friedrichs-Lewy_condition>`_ when calculating the maximum allowable time step. Values should be in the interval (0,1]                                                                                                                      
+initialDt                     real64             1e+99        Initial time-step value required by the solver to the event manager.                                                                                                                                                                                                                                                   
+logLevel                      integer            0            | Sets the level of information to write in the standard output (the console typically).                                                                                                                                                                                                                                 
+                                                              | A level of 0 outputs minimal information, higher levels require more.                                                                                                                                                                                                                                                  
[…]
