+++ src/coreComponents/schema/docs/SolidMechanicsAugmentedLagrangianContact.rst
@@ -1,0 +1,29 @@
+
+
+========================= ====================================================== =============== ======================================================================================================================================================================================================================================================================================================================== 
+Name                      Type                                                   Default         Description                                                                                                                                                                                                                                                                                                              
+========================= ====================================================== =============== ======================================================================================================================================================================================================================================================================================================================== 
+cflFactor                 real64                                                 0.5             Factor to apply to the `CFL condition <http://en.wikipedia.org/wiki/Courant-Friedrichs-Lewy_condition>`_ when calculating the maximum allowable time step. Values should be in the interval (0,1]                                                                                                                        
+discretization            groupNameRef                                           required        Name of discretization object (defined in the :ref:`NumericalMethodsManager`) to use for this solver. For instance, if this is a Finite Element Solver, the name of a :ref:`FiniteElement` should be specified. If this is a Finite Volume Method, the name of a :ref:`FiniteVolume` discretization should be specified. 
+initialDt                 real64                                                 1e+99           Initial time-step value required by the solver to the event manager.                                                                                                                                                                                                                                                     
+logLevel                  integer                                                0               | Sets the level of information to write in the standard output (the console typically).                                                                                                                                                                                                                                   
+                                                                                                 | A level of 0 outputs minimal information, higher levels require more.                                                                                                                                                                                                                                                    
+massDamping               real64                                                 0               Value of mass based damping coefficient.                                                                                                                                                                                                                                                                                 
+maxNumResolves            integer                                                10              Value to indicate how many resolves may be executed after some other event is executed. For example, if a SurfaceGenerator is specified, it will be executed after the mechanics solve. However if a new surface is generated, then the mechanics solve must be executed again due to the change in topology.            
+name                      groupName                                              required        A name is required for any non-unique nodes                                                                                                                                                                                                                                                                              
+newmarkBeta               real64                                                 0.25            Value of :math:`\beta` in the Newmark Method for Implicit Dynamic time integration option. This should be pow(newmarkGamma+0.5,2.0)/4.0 unless you know what you are doing.                                                                                                                                              
+newmarkGamma              real64                                                 0.5             Value of :math:`\gamma` in the Newmark Method for Implicit Dynamic time integration option                                                                                                                                                                                                                               
+stiffnessDamping          real64                                                 0               Value of stiffness based damping coefficient.                                                                                                                                                                                                                                                                            
+strainTheory              integer                                                0               | Indicates whether or not to use `Infinitesimal Strain Theory <https://en.wikipedia.org/wiki/Infinitesimal_strain_theory>`_, or `Finite Strain Theory <https://en.wikipedia.org/wiki/Finite_strain_theory>`_. Valid Inputs are:                                                                                           
+                                                                                                 |  0 - Infinitesimal Strain                                                                                                                                                                                                                                                                                                
+                                                                                                 |  1 - Finite Strain                                                                                                                                                                                                                                                                                                       
+targetRegions             groupNameRef_array                                     required        Allowable regions that the solver may be applied to. Note that this does not indicate that the solver will be applied to these regions, only that allocation will occur such that the solver may be applied to these regions. The decision about what regions this solver will beapplied to rests in the EventManager.   
+timeIntegrationOption     geos_SolidMechanicsLagrangianFEM_TimeIntegrationOption ExplicitDynamic | Time integration method. Options are:                                                                                                                                                                                                                                                                                    
+                                                                                                 | * QuasiStatic                                                                                                                                                                                                                                                                                                            
+                                                                                                 | * ImplicitDynamic                                                                                                                                                                                                                                                                                                        
+                                                                                                 | * ExplicitDynamic                                                                                                                                                                                                                                                                                                        
[…]

[…]
