### 2024-06-24 0b53b88
added comments for global numbering (Ryan's version)
