+++ src/coreComponents/mesh/generators/NewGlobalNumbering.cpp
@@ -31,583 +31,632 @@
 namespace geos::ghosting
 {
 
-struct Exchange
-{
-  std::set< NodeGlbIdx > nodes;
-  std::set< Edge > edges;
-  std::set< Face > faces;
-};
-
-void to_json( json & j,
-              const Exchange & v )
-{
-  j = json{ { "nodes", v.nodes },
-            { "edges", v.edges },
-            { "faces", v.faces } };
-}
-
-void from_json( const json & j,
-                Exchange & v )
-{
-  v.nodes = j.at( "nodes" ).get< std::set< NodeGlbIdx > >();
-  v.edges = j.at( "edges" ).get< std::set< Edge > >();
-  v.faces = j.at( "faces" ).get< std::set< Face > >();
-}
-
-array1d< std::uint8_t > convertExchange( Exchange const & exchange )
-{
-  std::vector< std::uint8_t > const tmp = json::to_cbor( json( exchange ) );
-  array1d< std::uint8_t > result;
-  result.reserve( tmp.size() );
-  for( std::uint8_t const & t: tmp )
+  struct Exchange
   {
-    result.emplace_back( t );
-  }
-  return result;
-}
+    std::set<NodeGlbIdx> nodes;
+    std::set<Edge> edges;
+    std::set<Face> faces;
+  };
 
-Exchange convertExchange( array1d< std::uint8_t > const & input )
-{
-  std::vector< std::uint8_t > const tmp( std::cbegin( input ), std::cend( input ) );
-  return json::from_cbor( tmp, false ).template get< Exchange >();
-}
-
-/**
- * @brief Extract the cells at the boundary of the mesh.
- * @param mesh The vtk mesh.
- * @return The vtk cell ids.
- */
-std::set< vtkIdType > extractBoundaryCells( vtkSmartPointer< vtkDataSet > mesh )
-{
-  // TODO Better handle the boundary information, forgetting about the 3d cells and simply handling the outside shell.
-  auto f = vtkDataSetSurfaceFilter::New();
-  f->PassThroughCellIdsOn();
-  f->PassThroughPointIdsOff();
-  f->FastModeOff();
-
-  string const originalCellsKey = "ORIGINAL_CELLS";
-  f->SetOriginalCellIdsName( originalCellsKey.c_str() );
-  auto boundaryMesh = vtkPolyData::New();
-  f->UnstructuredGridExecute( mesh, boundaryMesh );
-  vtkIdTypeArray const * originalCells = vtkIdTypeArray::FastDownCast( boundaryMesh->GetCellData()->GetArray( originalCellsKey.c_str() ) );
-
-  std::set< vtkIdType > boundaryCellIdxs;
-  for( auto i = 0; i < originalCells->GetNumberOfTuples(); ++i )
+  void to_json(json &j,
+               const Exchange &v)
   {
-    boundaryCellIdxs.insert( originalCells->GetValue( i ) );
+    j = json{{"nodes", v.nodes},
+             {"edges", v.edges},
+             {"faces", v.faces}};
   }
 
-  return boundaryCellIdxs;
-}
-
-Face reorderFaceNodes( std::vector< NodeGlbIdx > const & nodes, bool & isFlipped, std::uint8_t & start )  // TODO unit test
-{
-  std::size_t const n = nodes.size();
-
-  std::vector< NodeGlbIdx > f( nodes );
-
-  auto const cit = std::min_element( std::cbegin( nodes ), std::cend( nodes ) );
-  auto const minIdx = std::distance( std::cbegin( nodes ), cit );
-
-  std::size_t const prevIdx = minIdx == 0 ? n - 1 : minIdx - 1;
[…]
