### 2024-06-24 550f744
Added individual flags for applying transverse boundary velocities to each domain face
