+++ src/coreComponents/constitutive/docs/CompositionalMultiphaseFluid.rst
@@ -8,11 +8,116 @@ Overview
 =========================
 
 This model represents a full composition description of a multiphase multicomponent fluid.
-Phase behavior is modeled by an Equation of State (EOS) and partitioning of components into
+Phase behavior is modeled by an equation of state (EOS) and partitioning of components into
 phases is computed based on instantaneous chemical equilibrium via a two- or three-phase flash.
 Each component (species) is characterized by molar weight and critical properties that
 serve as input parameters for the EOS.
 See `Petrowiki`_ for more information.
+
+In this model the fluid is described by :math:`N_c` components with :math:`z_c` being the total
+mole fration of component :math:`c`. The fluid can partition into a liquid phase, denoted :math:`\ell`
+and a vapor phase denoted by :math:`v`. Therefore, by taking into account the molar phase component
+fractions, (which is the fraction of the molar mass of phase :math:`p` represented by component 
+:math:`c`), the following partition matrix establishes the component distribution within the two
+phases:
+
+.. math::
+    \begin{bmatrix}
+    x_{1} & x_{2} & x_{3} & \cdots & x_{N_c} \\
+    y_{1} & y_{2} & y_{3} & \cdots & y_{N_c} \\
+    \end{bmatrix}
+
+where :math:`x_c` is the mole fraction of component :math:`c` in the liquid phase and :math:`y_c`
+is the mole fraction of component :math:`c` in the vapor phase.
+
+The fluid properties are updated through the following steps:
+
+1) The phase fractions (:math:`\nu_p`) and phase component fractions (:math:`x_c` and :math:`y_c`) are
+computed as a function of pressure (:math:`p`), temperature (:math:`T`) and total component fractions
+(:math:`z_c`).
+
+2) The phase densities (:math:`\rho_p`) and phase viscosities (:math:`\mu_p`) are computed as a function
+of pressure, temperature and the updated phase component fractions.
+
+After calculating the phase fractions, phase component fractions, phase densities, phase viscosities,
+and their derivatives with respect to pressure, temperature, and component fractions, the
+:ref:`CompositionalMultiphaseFlow` then moves on to assembling the accumulation and flux terms.
+
+Step 1: Computation of the phase fractions and phase component fractions (flash)
+================================================================================
+Stability test
+-------------------------------
+The first step is to determine if provided mixture with total molar fractions :math:`z_c` is stable
+as a single phase at the current pressure :math:`p` and temperature :math:`T`. However, this can only
+be confirmed with through stability testing.
+
+The stability of a mixture is traditionally assessed using the Tangent Plane Distance (TPD) criterion
+developed by Michelsen (1982a). This criterion states that a phase with composition :math:`z` is stable
+at a specified pressure :math:`p` and temperature :math:`T` if and only if 
+
+.. math::
+  g(y) = \sum_{i=1}^{N_c}y_i\left(\ln y_i + \ln \phi_i(y) - \ln z_i - \ln \phi_i(z) \right) \geq 0
+
+for any permissible trial composition :math:`y`, where :math:`\phi_i` denotes the fugacity
+coefficient of component :math:`i`. 
+
+To determine stability of the mixture this testing in initiated from a basic starting point, based on
+Wilson K-values, to get both a lighter and a heavier trial mixture. The two trial mixtures are
+calculated as :math:`y_i = z_i/K_i` and :math:`y_i = z_iK_i` where :math:`K_i` are defined by
+
+.. math::
[…]
