### 2024-06-24 8fb37d7
comments for constructing owned and present mesh data (Ryan's version)
