+++ src/coreComponents/mesh/generators/NewGhosting.cpp
@@ -84,6 +84,9 @@ MaxGlbIdcs gatherOffset( vtkSmartPointer< vtkDataSet > mesh,
                          EdgeGlbIdx const & maxEdgeId,
                          FaceGlbIdx const & maxFaceId )
 {
+  // MaxGlbIdcs is just a struct of the 4 index types (ints under the hood)
+  // edge and face max (global) index values are passed in, 
+  // we use a little lambda below to get the node and cell max indices
   MaxGlbIdcs offsets{ NodeGlbIdx{ 0 }, maxEdgeId, maxFaceId, CellGlbIdx{ 0 } };
 
   auto const extract = []( vtkDataArray * globalIds ) -> vtkIdType
@@ -96,6 +99,9 @@ MaxGlbIdcs gatherOffset( vtkSmartPointer< vtkDataSet > mesh,
   offsets.nodes = NodeGlbIdx{ extract( mesh->GetPointData()->GetGlobalIds() ) };
   offsets.cells = CellGlbIdx{ extract( mesh->GetCellData()->GetGlobalIds() ) };
 
+  // I would have to look in detail about what these MPI lines do,
+  // But im fairly sure they are just handling type stuff so that you can do communications
+  // with the MaxGlbIdcs type
   // Otherwise, use `MPI_Type_create_struct`.
   static_assert( std::is_same_v< NodeGlbIdx::UnderlyingType, EdgeGlbIdx::UnderlyingType > );
   static_assert( std::is_same_v< NodeGlbIdx::UnderlyingType, FaceGlbIdx::UnderlyingType > );
@@ -110,6 +116,8 @@ MaxGlbIdcs gatherOffset( vtkSmartPointer< vtkDataSet > mesh,
   MPI_Op op;
   MPI_Op_create( g, true, &op );
 
+  // Do an allReduce so that every rank knows the max index for (nodes, edges, faces, cells)
+  // g is the custom reduction function which takes max for each (nodes, edges, faces, cells)
   MaxGlbIdcs result( offsets );
 
   MPI_Allreduce( &offsets, &result, 1, t, op, MPI_COMM_WORLD );
@@ -170,13 +178,25 @@ std::tuple< MeshGraph, MeshGraph > buildMeshGraph( vtkSmartPointer< vtkDataSet >
                                                    BucketOffsets const & offsets,
                                                    MpiRank curRank )
 {
+  // MeshGraph (defined in BuildPods.hpp) is just a struct of maps which map: 
+  // cell global index to vector of entries of type (face global index, extra stuff defining orientation of face) for each of the adjacent faces
+  // face global index to vector of entries of type (edge global index plus orientation) for each of the adjacent edges
+  // edge global index to tuple of node global indices
+  // node global index to 3d array of position
+  // We distinguish between the owned data and the present data
+  // present data is stuff that is owned by another rank, but is already present on the current rank because of the boundary exchange done previously.
+  // Note that present just includes the 2D boundary between ranks, we still need the cell info, the other faces, edges, nodes which make up that cell
+  // (and further if we wanted deeper ghosting)
   MeshGraph owned, present;
 
+  // Any (node, edge, face, cell) is owned by the lowest rank on which it appears
+  // so we define a lambda to compute if the current rank owns some piece of data
   auto const isCurrentRankOwning = [&curRank]( std::set< MpiRank > const & ranks ) -> bool
   {
     return curRank == *std::min_element( std::cbegin( ranks ), std::cend( ranks ) );
   };
 
+  // Get the nodal positions for nodes on current rank mesh and store in appropriate MeshGraph (owned/present)
   std::map< NodeGlbIdx, vtkIdType > const n2vtk = buildNgiToVtk( mesh );
[…]
