### 2024-06-24 #6829 @kthyng
@rrsettgast In my track (Ecology and Earth Sciences) I don't think any more recent than yours have moved off, though I'm not sure — it depends on relevant editor availability. However, there are several that have been waiting quite a bit longer than yours too. We're all doing our best based on the time we can dedicate to JOSS.
