+++ src/coreComponents/mesh/generators/VTKEmbeddedSurfaceBlockUtilities.cpp
@@ -14,32 +14,120 @@
 
 
 #include "VTKEmbeddedSurfaceBlockUtilities.hpp"
+#include <vtkUnstructuredGrid.h>
+#include <vtkDataArray.h>
+#include <vtkCellData.h>
 #include "CellBlockManager.hpp"
 
 namespace geos::vtk
 {
 
-void importFractureNetwork( string const & embeddedSurfaceBlockName,
+void importEmbeddedFractureNetwork( string const & embeddedSurfaceBlockName,
                             vtkSmartPointer< vtkDataSet > embeddedSurfaceMesh,
                             vtkSmartPointer< vtkDataSet > mesh,
                             CellBlockManager & cellBlockManager ){
     
-// prepare the data
-vtkIdType const numEdfmSurf = embeddedSurfaceMesh->GetNumberOfCells();
+  // Ouassim change
+  vtkIdType numEdfmFracs = embeddedSurfaceMesh->GetNumberOfCells();
+  //1.Get EDFM vertices
+  vtkUnstructuredGrid * grid = vtkUnstructuredGrid::SafeDownCast(embeddedSurfaceMesh);
+  vtkPoints* const nodes = grid->GetPoints();
+  vtkIdType numNodes = nodes->GetNumberOfPoints();
 
+  ArrayOfArrays< localIndex> elem2dToNodes(  numEdfmFracs );
+  for (vtkIdType i = 0; i < numEdfmFracs; ++i ){
+    auto  val = grid->GetCell(i)->GetPointIds();
+    elem2dToNodes.emplaceBack(i, val->GetId(0));
+    elem2dToNodes.emplaceBack(i, val->GetId(1));
+    elem2dToNodes.emplaceBack(i, val->GetId(2));
+    elem2dToNodes.emplaceBack(i, val->GetId(3));
+  }
+  
+  
+  ArrayOfArrays< real64 > embeddedSurfaceElemNodes( LvArray::integerConversion< localIndex >( numNodes ) );
+  for (vtkIdType i = 0; i < numNodes; ++i ){
+    double point[3];
+    nodes->GetPoint(i, point);
+    embeddedSurfaceElemNodes.emplaceBack(i, point[0]);
+    embeddedSurfaceElemNodes.emplaceBack(i, point[1]);
+    embeddedSurfaceElemNodes.emplaceBack(i, point[2]);
+  }
+  
+  //2. Get EDFM normal vectors
+  ArrayOfArrays< real64 > embeddedSurfaceNormalVectors(  numEdfmFracs );
+  const char * const normal_str = "normal_vectors";
+  vtkDataArray  * normals =grid->GetCellData()->GetArray(normal_str);
+  for (vtkIdType i = 0; i < numEdfmFracs; ++i ){
+    auto val = normals->GetTuple3(i);
+    embeddedSurfaceNormalVectors.emplaceBack(i, val[0]);
+    embeddedSurfaceNormalVectors.emplaceBack(i, val[1]);
+    embeddedSurfaceNormalVectors.emplaceBack(i, val[2]);
+  }
+  
+  //3. Get EDFM  tangential length vectors (horizontal)
+  ArrayOfArrays< real64 > embeddedSurfaceTangentialLengthVectors( numEdfmFracs );
+  const char * const tl_str = "tangential_length_vectors";
+  vtkDataArray  * tangential_len =grid->GetCellData()->GetArray(tl_str);
+  for (vtkIdType i = 0; i < numEdfmFracs; ++i ){
+    auto val = tangential_len->GetTuple3(i);
+    embeddedSurfaceTangentialLengthVectors.emplaceBack(i, val[0]);
[…]
