+++ src/coreComponents/physicsSolvers/solidMechanics/SolidMechanicsMPM.cpp
@@ -980,7 +980,7 @@ void SolidMechanicsMPM::postProcessInput()
   if( m_enablePrescribedBoundaryTransverseVelocities.size() == 0 )
   {
     m_enablePrescribedBoundaryTransverseVelocities.resize( 6 );
-    LvArray::tensorOps::fill< 6 >( m_boundaryConditionTypes, 0 );
+    LvArray::tensorOps::fill< 6 >( m_enablePrescribedBoundaryTransverseVelocities, 0 );
   }
   else
   {
@@ -7434,10 +7434,18 @@ void SolidMechanicsMPM::computeCohesiveTraction( int g,
                               normalStress,
                               shearStress );
       break;
+    case CohesiveLawOption::Polymer:
+      polymerCohesiveLaw( totalNormalDisplacement,
+                          totalTangentialDisplacement,
+                          normalStress,
+                          shearStress );
+      break;
     default:
       GEOS_ERROR( "No cohesive law of the kind specified!" );
       break;
   }
+
+  // GEOS_LOG_RANK("dn: " << totalNormalDisplacement << ", " << "dt: " << totalTangentialDisplacement << ", " << "sigma_n: " << normalStress << ", " << "sigma_t: " << shearStress );
 
   // Mass-weighted average of the projected area
   // Do we want to add choice of weighting as we do in contact calculations for normals?
@@ -7539,6 +7547,55 @@ void SolidMechanicsMPM::needlemanXuCohesiveLaw( real64 normalDisplacement,
   shearStress = -( shearWorkOfSeparation / tangentialCharacteristicDisplacement ) * ( 2 * normalCharacteristicDisplacement / tangentialCharacteristicDisplacement ) * normalizedTangentialDisplacement * 
                         ( q + ( r - q ) / ( r - 1.0 ) * normalizedNormalDisplacement ) *  
                         exp( -normalizedNormalDisplacement ) * exp( -pow( normalizedTangentialDisplacement, 2) );
+}
+
+void SolidMechanicsMPM::polymerCohesiveLaw( real64 normalDisplacement,
+                                            real64 tangentialDisplacement,
+                                            real64 & normalStress,
+                                            real64 & shearStress )
+{
+  real64 t = m_characteristicNormalDisplacement;
+  real64 tSqr = t * t;
+
+  real64 normalDisplacementSqr = normalDisplacement * normalDisplacement;
+  real64 shearDisplacementSqr = tangentialDisplacement * tangentialDisplacement;
+
+  // These should eventually be user inputs
+  real64 k = 0.8; // Polymer bulk modulus
+  real64 g = 0.2; // Polymer shear modulus
+  real64 gSqr = g * g;
+  real64 yieldStrength0 = 0.003;
+  real64 r0 = 0.0225;
+  real64 r1 = 0.0108;
+  real64 r2 = 0.177;
+  real64 Gr = 0.00174;
+  real64 lambdaMax = 2.14; // Polymer max stretch
+
+  real64 lambda = sqrt(pow(normalDisplacement + t,2) + shearDisplacementSqr)/t; // Stretch
+
+  real64 sigma_H = Gr*pow(lambda * lambda - 1/lambda,2);
+
+  real64 tau = sqrt(gSqr*(4*t*normalDisplacement*(normalDisplacementSqr + shearDisplacementSqr)+pow(normalDisplacementSqr+shearDisplacementSqr,2)+tSqr * (4*normalDisplacementSqr+3*shearDisplacementSqr))/(pow(t*(t+normalDisplacement),2)));
+  
[…]
