### 2024-06-24 ecabd59
Fixed bug with initialization of transverse BC velocity flags
