### 2024-06-24 785957c
add code coverage...reduce number of cores from 16 to 8 for streak2 builds
