### 2024-06-24 a05c002
add new cuda12 job.
