+++ host-configs/LLNL/old/quartz-clang-14.cmake
@@ -1,4 +1,4 @@
-include(${CMAKE_CURRENT_LIST_DIR}../../../src/coreComponents/LvArray/host-configs/LLNL/quartz-clang-14.cmake)
+include(${CMAKE_CURRENT_LIST_DIR}../../../src/coreComponents/LvArray/host-configs/LLNL/old/quartz-clang-14.cmake)
 
 # Fortran
 set(CMAKE_Fortran_COMPILER /usr/tce/packages/gcc/gcc-12.1.1-magic/bin/gfortran CACHE PATH "")
