+++ .github/workflows/ci_tests.yml
@@ -236,6 +236,7 @@ jobs:
       ENABLE_HYPRE: ON
       ENABLE_TRILINOS: OFF
       GCP_BUCKET: geosx/ubuntu22.04-gcc11
+      RUNS_ON: streak2
       NPROC: 8
       DOCKER_RUN_ARGS: "--cpus=8 --memory=128g -v /etc/pki/ca-trust/source/anchors/:/etc/pki/ca-trust/source/anchors/llnl:ro"
       DOCKER_CERTS_DIR: "/etc/pki/ca-trust/source/anchors"
