+++ .github/workflows/ci_tests.yml
@@ -238,9 +238,9 @@ jobs:
       GCP_BUCKET: geosx/ubuntu22.04-gcc11
       RUNS_ON: streak2
       NPROC: 8
-      DOCKER_RUN_ARGS: "--cpus=8 --memory=128g -v /etc/pki/ca-trust/source/anchors/:/etc/pki/ca-trust/source/anchors/llnl:ro"
-      DOCKER_CERTS_DIR: "/etc/pki/ca-trust/source/anchors"
-      DOCKER_CERTS_UPDATE_COMMAND: "update-ca-trust"
+      DOCKER_RUN_ARGS: "--cpus=8 --memory=128g -v /etc/pki/ca-trust/source/anchors/:/usr/local/share/ca-certificates/llnl:ro"
+      DOCKER_CERTS_DIR: "/usr/local/share/ca-certificates"
+      DOCKER_CERTS_UPDATE_COMMAND: "update-ca-certificates"
 
   # If the 'ci: run CUDA builds' PR label is found, the cuda jobs run immediately along side linux jobs.
   # Note: CUDA jobs should only be run if PR is ready to merge.
