### 2024-06-24 70634d7
Updated boundary condition logic and fixed contact boundary condition treatment
