+++ src/coreComponents/physicsSolvers/solidMechanics/SolidMechanicsMPM.cpp
@@ -3383,152 +3383,161 @@ void SolidMechanicsMPM::applyEssentialBCs( const real64 dt,
   real64 localFaceReactions[6] = {0.0};
   for( int face = 0; face < 6; face++ )
   {
-    if( m_boundaryConditionTypes[face] == 1 )
+    // TODO Eventually perform cast to BC enum type!
+    switch(m_boundaryConditionTypes[face])
     {
-      // CC: TODO add gridDVelocity update to theses for XPIC
-      singleFaceVectorFieldSymmetryBC( face, gridVelocity, gridDVelocity, gridPosition, nodeSets );
-
-      //Dumby paramter, we do not need the change in grid acceleration
-      array3d< real64 > gridDAcceleration( gridAcceleration.size(0), gridAcceleration.size(1), gridAcceleration.size(2)); //CC: TODO Probably want to avoid allocating a lot of memory just for the dummy variable
-      singleFaceVectorFieldSymmetryBC( face, gridAcceleration, gridDAcceleration, gridPosition, nodeSets );
-    }
-    else if( ( m_boundaryConditionTypes[face] == 2 || m_boundaryConditionTypes[face] == 3 )  && ( m_prescribedBoundaryFTable == 1 || m_stressControl[0] == 1 || m_stressControl[1] == 1 || m_stressControl[2] == 1) ) // Double check stress control (think we only need to check if stress control for the direction of faces are on not all them)
-    {
-      for( int fieldIndex = 0; fieldIndex < m_numVelocityFields; fieldIndex++ )
-      {
-        // Face-associated quantities
-        int dir0 = face / 2;           // 0, 0, 1, 1, 2, 2 (x-, x+, y-, y+, z-, z+)
-        int dir1 = (dir0 + 1) % 3;     // 1, 1, 2, 2, 0, 0
-        int dir2 = (dir0 + 2) % 3;     // 2, 2, 0, 0, 1, 1
-        int positiveNormal = face % 2; // even => (-) => 0, odd => (+) => 1
-
-        // Enforce BCs on boundary nodes using F-table
-        SortedArrayView< localIndex const > const boundaryNodes = m_boundaryNodes[face].toView();
-        int const numBoundaryNodes = boundaryNodes.size();
-        forAll< serialPolicy >( numBoundaryNodes, [&, gridPosition, gridVelocity, gridDVelocity, gridMass] GEOS_HOST ( localIndex const gg ) // Probably
-                                                                                                                                             // not a big
-                                                                                                                                             // enough loop
-                                                                                                                                             // to warrant
-                                                                                                                                             // parallelization
+      case 0: // Outflow
+        break; // Do nothing
+      case 1: // Symmetry 
         {
-          int const g = boundaryNodes[gg];
+          // CC: TODO add gridDVelocity update to theses for XPIC
+          singleFaceVectorFieldSymmetryBC( face, gridVelocity, gridDVelocity, gridPosition, nodeSets );
 
-          // If boundary condition type is 3 (e.g. contact, default is sticky), performs a check if normal grid component of grid velocity is moving into plane, if so flips component and adds domain velocity
-          real64 prescribedVelocity = gridVelocity[g][fieldIndex][dir0];
-          if( m_boundaryConditionTypes[face] == 3 )
+          //Dumby paramter, we do not need the change in grid acceleration
[…]
