### 2024-06-25 3fabd34
Remove template parameter
