+++ src/coreComponents/constitutive/unitTests/CMakeLists.txt
@@ -13,6 +13,7 @@ set( gtest_geosx_tests
      testNegativeTwoPhaseFlash9Comp.cpp
      testParticleFluidEnums.cpp
      testPropertyConversions.cpp
+     testStabilityTest2Comp.cpp
      testStabilityTest9Comp.cpp
      testRachfordRice.cpp )
 
