### 2024-06-25 c091ef4
debug code coverage on streak2
