+++ scripts/ci_build_and_test_in_container.sh
@@ -18,6 +18,7 @@ free -g
 # The or_die function run the passed command line and
 # exits the program in case of non zero error code
 function or_die () {
+    echo $@
     "$@"
     local status=$?
 
@@ -263,7 +264,7 @@ if [[ ! -z "${SCCACHE_CREDS}" ]]; then
 fi
 
 if [[ "${CODE_COVERAGE}" = true ]]; then
-  or_die ninja coreComponents_coverage
+  or_die ninja -j1 coreComponents_coverage
   cp -r ${GEOSX_BUILD_DIR}/coreComponents_coverage.info.cleaned ${GEOS_SRC_DIR}/geos_coverage.info.cleaned
 fi
 
