### 2024-06-25 c97a966
Update some typing issues
