+++ scripts/ci_build_and_test_in_container.sh
@@ -18,7 +18,7 @@ free -g
 # The or_die function run the passed command line and
 # exits the program in case of non zero error code
 function or_die () {
-    echo "$@"
+    echo "Executing: $@"
     "$@"
     local status=$?
 
