### 2024-06-25 ba15483
debug code coverage on streak2
