### 2024-06-25 8c4b3c4
debug code coverage on streak2
