+++ scripts/ci_build_and_test_in_container.sh
@@ -18,8 +18,9 @@ free -g
 # The or_die function run the passed command line and
 # exits the program in case of non zero error code
 function or_die () {
-    echo "Executing: $@"
-    eval "$@"
+    local cmd=("$@")
+    echo "Executing: $cmd"
+    eval "$cmd"
     local status=$?
 
     if [[ $status != 0 ]] ; then
