### 2024-06-25 e47f30c
debug code coverage on streak2
