+++ src/coreComponents/constitutive/solid/PorousSolid.hpp
@@ -24,6 +24,9 @@
 #include "constitutive/solid/porosity/BiotPorosity.hpp"
 #include "constitutive/solid/SolidBase.hpp"
 #include "constitutive/permeability/ConstantPermeability.hpp"
+#include "dataRepository/ObjectCatalog.hpp"
+#include "common/DataTypes.hpp"
+#include "dataRepository/Group.hpp"
 
 namespace geos
 {
@@ -48,8 +51,12 @@ public:
    */
   PorousSolidUpdates( SOLID_TYPE const & solidModel,
                       BiotPorosity const & porosityModel,
-                      ConstantPermeability const & permModel ):
-    CoupledSolidUpdates< SOLID_TYPE, BiotPorosity, ConstantPermeability >( solidModel, porosityModel, permModel )
+                      ConstantPermeability const & permModel,
+                      real64 const dThermalExpansionCoefficient_dTemperature,
+                      real64 const referenceTemperature ):
+    CoupledSolidUpdates< SOLID_TYPE, BiotPorosity, ConstantPermeability >( solidModel, porosityModel, permModel ),
+    m_dThermalExpansionCoefficient_dTemperature( dThermalExpansionCoefficient_dTemperature ),
+    m_referenceTemperature( referenceTemperature )
   {}
 
   GEOS_HOST_DEVICE
@@ -161,7 +168,11 @@ public:
     real64 const bulkModulus = m_solidUpdate.getBulkModulus( k );
     real64 const meanEffectiveStressIncrement = bulkModulus * ( strainIncrement[0] + strainIncrement[1] + strainIncrement[2] );
     real64 const biotCoefficient = m_porosityUpdate.getBiotCoefficient( k );
-    real64 const thermalExpansionCoefficient = m_solidUpdate.getThermalExpansionCoefficient( k );
+    //real64 const thermalExpansionCoefficient = m_solidUpdate.getThermalExpansionCoefficient( k );
+
+    real64 const referenceThermalExpansionCoefficient = m_solidUpdate.getThermalExpansionCoefficient( k );
+	real64 const thermalExpansionCoefficient  = referenceThermalExpansionCoefficient + m_dThermalExpansionCoefficient_dTemperature * (temperature - m_referenceTemperature);
+
     real64 const meanTotalStressIncrement = meanEffectiveStressIncrement - biotCoefficient * ( pressure - pressure_n )
                                             - 3 * thermalExpansionCoefficient * bulkModulus * ( temperature - temperature_n );
     m_porosityUpdate.updateMeanTotalStressIncrement( k, q, meanTotalStressIncrement );
@@ -289,13 +300,22 @@ private:
 
     // Add the contributions of pressure and temperature to the total stress
     real64 const biotCoefficient = m_porosityUpdate.getBiotCoefficient( k );
-    real64 const thermalExpansionCoefficient = m_solidUpdate.getThermalExpansionCoefficient( k );
+    real64 const referenceThermalExpansionCoefficient = m_solidUpdate.getThermalExpansionCoefficient( k );
+
+	real64 const thermalExpansionCoefficient  = referenceThermalExpansionCoefficient + m_dThermalExpansionCoefficient_dTemperature * (temperature - m_referenceTemperature);
+
+
     real64 const bulkModulus = m_solidUpdate.getBulkModulus( k );
     real64 const thermalExpansionCoefficientTimesBulkModulus = thermalExpansionCoefficient * bulkModulus;
 
[…]
