### 2024-06-25 42a677d
Replace paid runners with streak2 (#3184)
Replace paid runners with streak2 for all builds except code coverage.
