### 2024-06-25 f3033c2
Fix unit tests on GPU CI platform. (#2985)
* re-enable testLifoStorage
* adding fix for unit test testLifoStorage

Co-authored-by: Jie Meng <jie.meng@total.com>
Co-authored-by: Stefano Frambati <stefano.frambati@totalenergies.com>
Co-authored-by: j0405284 <stefano.frambati@total.com>
