+++ .github/workflows/ci_tests.yml
@@ -41,7 +41,7 @@ jobs:
         
     # The TPL tag is contained in the codespaces configuration to avoid duplications.
     - name: Checkout .devcontainer/devcontainer.json
-      uses: actions/checkout@v4.1.4
+      uses: actions/checkout@v4.1.7
       with:
         sparse-checkout: |
           .devcontainer/devcontainer.json
@@ -77,7 +77,7 @@ jobs:
     # The integrated test submodule repository contains large data (using git lfs).
     # To save time (and money) we do not let Github Actions automatically clone all our (lfs) subrepositories and do it by hand.
     - name: Checkout Repository
-      uses: actions/checkout@v4.1.4
+      uses: actions/checkout@v4.1.7
       with:
         # Let script update submodules; Github Actions submodule history causes error
         submodules: false
@@ -212,7 +212,7 @@ jobs:
     runs-on: ubuntu-22.04
     steps:
     - name: Checkout Repository
-      uses: actions/checkout@v4.1.1
+      uses: actions/checkout@v4.1.7
       with:
         submodules: false
         lfs: false
