### 2024-06-25 8266f4b
define dTEC_dT at SolidBase level
