### 2024-06-25 901b1c0
add dTEC_dT getter and update all the solid constitutive classes
