+++ src/coreComponents/constitutive/solid/DelftEgg.hpp
@@ -66,10 +66,19 @@ public:
                    arrayView1d< real64 const > const & bulkModulus,
                    arrayView1d< real64 const > const & shearModulus,
                    arrayView1d< real64 const > const & thermalExpansionCoefficient,
+                   real64 const & dThermalExpansionCoefficient_dTemperature,
+                   real64 const & referenceTemperature,
                    arrayView3d< real64, solid::STRESS_USD > const & newStress,
                    arrayView3d< real64, solid::STRESS_USD > const & oldStress,
                    const bool & disableInelasticity ):
-    ElasticIsotropicUpdates( bulkModulus, shearModulus, thermalExpansionCoefficient, newStress, oldStress, disableInelasticity ),
+    ElasticIsotropicUpdates( bulkModulus, 
+shearModulus, 
+thermalExpansionCoefficient, 
+dThermalExpansionCoefficient_dTemperature,
+                      referenceTemperature,
+newStress, 
+oldStress, 
+disableInelasticity ),
     m_recompressionIndex( recompressionIndex ),
     m_virginCompressionIndex( virginCompressionIndex ),
     m_cslSlope( cslSlope ),
@@ -551,6 +560,8 @@ public:
                             m_bulkModulus,
                             m_shearModulus,
                             m_thermalExpansionCoefficient,
+m_dThermalExpansionCoefficient_dTemperature,
+                                      m_referenceTemperature,
                             m_newStress,
                             m_oldStress,
                             m_disableInelasticity );
@@ -576,6 +587,8 @@ public:
                           m_bulkModulus,
                           m_shearModulus,
                           m_thermalExpansionCoefficient,
+ m_dThermalExpansionCoefficient_dTemperature,
+                                      m_referenceTemperature,
                           m_newStress,
                           m_oldStress,
                           m_disableInelasticity );
