+++ src/coreComponents/constitutive/solid/ElasticIsotropicPressureDependent.hpp
@@ -57,10 +57,17 @@ public:
                                             arrayView1d< real64 const > const & recompressionIndex,
                                             arrayView1d< real64 const > const & shearModulus,
                                             arrayView1d< real64 const > const & thermalExpansionCoefficient,
+real64 const & dThermalExpansionCoefficient_dTemperature,
+                           real64 const & referenceTemperature,
                                             arrayView3d< real64, solid::STRESS_USD > const & newStress,
                                             arrayView3d< real64, solid::STRESS_USD > const & oldStress,
                                             bool const & disableInelasticity ):
-    SolidBaseUpdates( newStress, oldStress, thermalExpansionCoefficient, disableInelasticity ),
+    SolidBaseUpdates( newStress, 
+oldStress, 
+thermalExpansionCoefficient, 
+dThermalExpansionCoefficient_dTemperature,
+                      referenceTemperature,
+disableInelasticity ),
     m_refPressure( refPressure ),
     m_refStrainVol( refStrainVol ),
     m_recompressionIndex( recompressionIndex ),
@@ -522,6 +529,8 @@ public:
                                                        m_recompressionIndex,
                                                        m_shearModulus,
                                                        m_thermalExpansionCoefficient,
+m_dThermalExpansionCoefficient_dTemperature,
+                          m_referenceTemperature,
                                                        m_newStress,
                                                        m_oldStress,
                                                        m_disableInelasticity );
@@ -533,6 +542,8 @@ public:
                                                        m_recompressionIndex,
                                                        m_shearModulus,
                                                        m_thermalExpansionCoefficient,
+m_dThermalExpansionCoefficient_dTemperature,
+                          m_referenceTemperature,
                                                        arrayView3d< real64, solid::STRESS_USD >(),
                                                        arrayView3d< real64, solid::STRESS_USD >(),
                                                        m_disableInelasticity );
@@ -556,6 +567,8 @@ public:
                           m_recompressionIndex,
                           m_shearModulus,
                           m_thermalExpansionCoefficient,
+m_dThermalExpansionCoefficient_dTemperature,
+                          m_referenceTemperature,
                           m_newStress,
                           m_oldStress,
                           m_disableInelasticity );
