+++ src/coreComponents/constitutive/solid/CeramicDamage.hpp
@@ -79,10 +79,19 @@ public:
                         arrayView1d< real64 const > const & bulkModulus,
                         arrayView1d< real64 const > const & shearModulus,
                         arrayView1d< real64 const > const & thermalExpansionCoefficient,
+real64 const & dThermalExpansionCoefficient_dTemperature,
+                           real64 const & referenceTemperature,
                         arrayView3d< real64, solid::STRESS_USD > const & newStress,
                         arrayView3d< real64, solid::STRESS_USD > const & oldStress,
                         bool const & disableInelasticity ):
-    ElasticIsotropicUpdates( bulkModulus, shearModulus, thermalExpansionCoefficient, newStress, oldStress, disableInelasticity ),
+    ElasticIsotropicUpdates( bulkModulus, 
+shearModulus, 
+thermalExpansionCoefficient, 
+dThermalExpansionCoefficient_dTemperature,
+                      referenceTemperature,
+newStress, 
+oldStress, 
+disableInelasticity ),
     m_damage( damage ),
     m_jacobian( jacobian ),
     m_lengthScale( lengthScale ),
@@ -492,6 +501,8 @@ public:
                                  m_bulkModulus,
                                  m_shearModulus,
                                  m_thermalExpansionCoefficient,
+m_dThermalExpansionCoefficient_dTemperature,
+                          m_referenceTemperature,
                                  m_newStress,
                                  m_oldStress,
                                  m_disableInelasticity );
@@ -518,6 +529,8 @@ public:
                           m_bulkModulus,
                           m_shearModulus,
                           m_thermalExpansionCoefficient,
+m_dThermalExpansionCoefficient_dTemperature,
+                          m_referenceTemperature,
                           m_newStress,
                           m_oldStress,
                           m_disableInelasticity );

[…]
