### 2024-06-25 29f63b1
use the dTEC_dT getter to get param from XML
