+++ src/coreComponents/constitutive/solid/ElasticTransverseIsotropic.hpp
@@ -59,17 +59,17 @@ public:
                                      arrayView1d< real64 const > const & c44,
                                      arrayView1d< real64 const > const & c66,
                                      arrayView1d< real64 const > const & thermalExpansionCoefficient,
-real64 const & dThermalExpansionCoefficient_dTemperature,
-                           real64 const & referenceTemperature,
+                                     real64 const & dThermalExpansionCoefficient_dTemperature,
+                                     real64 const & referenceTemperature,
                                      arrayView3d< real64, solid::STRESS_USD > const & newStress,
                                      arrayView3d< real64, solid::STRESS_USD > const & oldStress,
                                      bool const & disableInelasticity ):
-    SolidBaseUpdates( newStress, 
-oldStress, 
-thermalExpansionCoefficient, 
-dThermalExpansionCoefficient_dTemperature,
+    SolidBaseUpdates( newStress,
+                      oldStress,
+                      thermalExpansionCoefficient,
+                      dThermalExpansionCoefficient_dTemperature,
                       referenceTemperature,
-disableInelasticity ),
+                      disableInelasticity ),
     m_c11( c11 ),
     m_c13( c13 ),
     m_c33( c33 ),
@@ -552,8 +552,8 @@ public:
                                               m_c44,
                                               m_c66,
                                               m_thermalExpansionCoefficient,
-m_dThermalExpansionCoefficient_dTemperature,
-                          m_referenceTemperature,
+                                              m_dThermalExpansionCoefficient_dTemperature,
+                                              m_referenceTemperature,
                                               m_newStress,
                                               m_oldStress,
                                               m_disableInelasticity );
@@ -577,7 +577,7 @@ m_dThermalExpansionCoefficient_dTemperature,
                           m_c44,
                           m_c66,
                           m_thermalExpansionCoefficient,
-m_dThermalExpansionCoefficient_dTemperature,
+                          m_dThermalExpansionCoefficient_dTemperature,
                           m_referenceTemperature,
                           m_newStress,
                           m_oldStress,

[…]
