+++ src/coreComponents/physicsSolvers/wavePropagation/AcousticWaveEquationSEM.cpp
@@ -27,10 +27,6 @@
 #include "mesh/ElementType.hpp"
 #include "mesh/mpiCommunications/CommunicationTools.hpp"
 #include "WaveSolverUtils.hpp"
-#include "AcousticTimeSchemeSEMKernel.hpp"
-#include "AcousticMatricesSEMKernel.hpp"
-#include "AcousticPMLSEMKernel.hpp"
-#include "PrecomputeSourcesAndReceiversKernel.hpp"
 
 namespace geos
 {
@@ -179,9 +175,9 @@ void AcousticWaveEquationSEM::precomputeSourceAndReceiverTerm( MeshLevel & mesh,
 
       {
         GEOS_MARK_SCOPE( acousticWaveEquationSEMKernels::PrecomputeSourceAndReceiverKernel );
-        PreComputeSourcesAndReceivers::
-          Compute1DSourceAndReceiverConstants
-        < EXEC_POLICY, FE_TYPE >
+        acousticWaveEquationSEMKernels::
+          PrecomputeSourceAndReceiverKernel::
+          launch< EXEC_POLICY, FE_TYPE >
           ( elementSubRegion.size(),
           numFacesPerElem,
           nodeCoords32,
@@ -298,51 +294,24 @@ void AcousticWaveEquationSEM::initializePostInitialConditionsPreSubGroups()
       {
         using FE_TYPE = TYPEOFREF( finiteElement );
 
-        // acousticWaveEquationSEMKernels::MassMatrixKernel< FE_TYPE > kernelM( finiteElement );
-        // kernelM.template launch< EXEC_POLICY, ATOMIC_POLICY >( elementSubRegion.size(),
-        //                                                        nodeCoords,
-        //                                                        elemsToNodes,
-        //                                                        velocity,
-        //                                                        density,
-        //                                                        mass );
+        acousticWaveEquationSEMKernels::MassMatrixKernel< FE_TYPE > kernelM( finiteElement );
+        kernelM.template launch< EXEC_POLICY, ATOMIC_POLICY >( elementSubRegion.size(),
+                                                               nodeCoords,
+                                                               elemsToNodes,
+                                                               velocity,
+                                                               density,
+                                                               mass );
 
-        AcousticMatricesSEM::MassMatrix< FE_TYPE > kernelM( finiteElement );
-        kernelM.template computeMassMatrix< EXEC_POLICY, ATOMIC_POLICY >( elementSubRegion.size(),
-                                                                          nodeCoords,
-                                                                          elemsToNodes,
-                                                                          velocity,
-                                                                          density,
-                                                                          mass );
-
-        AcousticMatricesSEM::DampingMatrix< FE_TYPE > kernelD( finiteElement );
-        kernelD.template computeDampingMatrix< EXEC_POLICY, ATOMIC_POLICY >( elementSubRegion.size(),
-                                                                             nodeCoords,
-                                                                             elemsToFaces,
-                                                                             facesToNodes,
-                                                                             facesDomainBoundaryIndicator,
-                                                                             freeSurfaceFaceIndicator,
-                                                                             velocity,
-                                                                             density,
-                                                                             damping );
-
-        //AcousticMatricesSEM::computeDampingMatrix<FE_TYPE,EXEC_POLICY,ATOMIC_POLICY>(finiteElement,
-        //                                                                             elementSubRegion.size(),
-        //
-        //
-        //
-        //
-        //
-        //
[…]
