+++ src/coreComponents/physicsSolvers/wavePropagation/ElasticWaveEquationSEM.cpp
@@ -26,9 +26,6 @@
 #include "mesh/ElementType.hpp"
 #include "mesh/mpiCommunications/CommunicationTools.hpp"
 #include "WaveSolverUtils.hpp"
-#include "ElasticTimeSchemeSEMKernel.hpp"
-#include "ElasticMatricesSEMKernel.hpp"
-#include "PrecomputeSourcesAndReceiversKernel.hpp"
 
 namespace geos
 {
@@ -279,9 +276,9 @@ void ElasticWaveEquationSEM::precomputeSourceAndReceiverTerm( MeshLevel & mesh, 
       using FE_TYPE = TYPEOFREF( finiteElement );
 
       localIndex const numFacesPerElem = elementSubRegion.numFacesPerElement();
-      PreComputeSourcesAndReceivers::
-        Compute3DSourceAndReceiverConstantsWithDAS
-      < EXEC_POLICY, FE_TYPE >
+      elasticWaveEquationSEMKernels::
+        PrecomputeSourceAndReceiverKernel::
+        launch< EXEC_POLICY, FE_TYPE >
         ( elementSubRegion.size(),
         numFacesPerElem,
         X,
@@ -403,27 +400,27 @@ void ElasticWaveEquationSEM::initializePostInitialConditionsPreSubGroups()
       {
         using FE_TYPE = TYPEOFREF( finiteElement );
 
-        ElasticMatricesSEM::MassMatrix< FE_TYPE > kernelM( finiteElement );
-        kernelM.template computeMassMatrix< EXEC_POLICY, ATOMIC_POLICY >( elementSubRegion.size(),
-                                                                          nodeCoords,
-                                                                          elemsToNodes,
-                                                                          density,
-                                                                          mass );
+        elasticWaveEquationSEMKernels::MassMatrixKernel< FE_TYPE > kernelM( finiteElement );
+        kernelM.template launch< EXEC_POLICY, ATOMIC_POLICY >( elementSubRegion.size(),
+                                                               nodeCoords,
+                                                               elemsToNodes,
+                                                               density,
+                                                               mass );
 
-        ElasticMatricesSEM::DampingMatrix< FE_TYPE > kernelD( finiteElement );
-        kernelD.template computeDampingMatrix< EXEC_POLICY, ATOMIC_POLICY >( elementSubRegion.size(),
-                                                                             nodeCoords,
-                                                                             elemsToFaces,
-                                                                             facesToNodes,
-                                                                             facesDomainBoundaryIndicator,
-                                                                             freeSurfaceFaceIndicator,
-                                                                             faceNormal,
-                                                                             density,
-                                                                             velocityVp,
-                                                                             velocityVs,
-                                                                             dampingx,
-                                                                             dampingy,
-                                                                             dampingz );
+        elasticWaveEquationSEMKernels::DampingMatrixKernel< FE_TYPE > kernelD( finiteElement );
+        kernelD.template launch< EXEC_POLICY, ATOMIC_POLICY >( elementSubRegion.size(),
+                                                               nodeCoords,
+                                                               elemsToFaces,
+                                                               facesToNodes,
[…]
