+++ src/coreComponents/physicsSolvers/wavePropagation/ElasticFirstOrderWaveEquationSEMKernel.hpp
@@ -30,6 +30,300 @@ namespace geos
 namespace elasticFirstOrderWaveEquationSEMKernels
 {
 
+struct PrecomputeSourceAndReceiverKernel
+{
+
+  /**
+   * @brief Launches the precomputation of the source and receiver terms
+   * @tparam EXEC_POLICY execution policy
+   * @tparam FE_TYPE finite element type
+   * @param[in] size the number of cells in the subRegion
+   * @param[in] numFacesPerElem number of face on an element
+   * @param[in] nodeCoords coordinates of the nodes
+   * @param[in] elemGhostRank array containing the ghost rank
+   * @param[in] elemsToNodes map from element to nodes
+   * @param[in] elemsToFaces map from element to faces
+   * @param[in] elemCenter coordinates of the element centers
+   * @param[in] faceNormal array containing the normal of all faces
+   * @param[in] faceCenter array containing the center of all faces
+   * @param[in] sourceCoordinates coordinates of the source terms
+   * @param[in] receiverCoordinates coordinates of the receiver terms
+   * @param[in] dt time-step
+   * @param[in] timeSourceFrequency Peak frequency of the source
+   * @param[in] timeSourceDelay the time delay of the source
+   * @param[in] rickerOrder Order of the Ricker wavelet
+   * @param[out] sourceNodeIds indices of the nodes of the element where the source is located
+   * @param[out] sourceConstants constant part of the source terms
+   * @param[out] receiverIsLocal flag indicating whether the receiver is local or not
+   * @param[out] receiverNodeIds indices of the nodes of the element where the receiver is located
+   * @param[out] sourceValue array containing the value of the time dependent source (Ricker for e.g)
+   */
+  template< typename EXEC_POLICY, typename FE_TYPE >
+  static void
+  launch( localIndex const size,
+          localIndex const regionIndex,
+          localIndex const numFacesPerElem,
+          arrayView2d< WaveSolverBase::wsCoordType const, nodes::REFERENCE_POSITION_USD > const nodeCoords,
+          arrayView1d< integer const > const elemGhostRank,
+          arrayView2d< localIndex const, cells::NODE_MAP_USD > const & elemsToNodes,
+          arrayView2d< localIndex const > const elemsToFaces,
+          arrayView2d< real64 const > const & elemCenter,
+          arrayView2d< real64 const > const faceNormal,
+          arrayView2d< real64 const > const faceCenter,
+          arrayView2d< real64 const > const sourceCoordinates,
+          arrayView1d< localIndex > const sourceIsAccessible,
+          arrayView1d< localIndex > const sourceElem,
+          arrayView2d< localIndex > const sourceNodeIds,
+          arrayView2d< real64 > const sourceConstants,
+          arrayView1d< localIndex > const sourceRegion,
+          arrayView2d< real64 const > const receiverCoordinates,
+          arrayView1d< localIndex > const receiverIsLocal,
+          arrayView1d< localIndex > const rcvElem,
+          arrayView2d< localIndex > const receiverNodeIds,
+          arrayView2d< real64 > const receiverConstants,
+          arrayView1d< localIndex > const receiverRegion,
+          arrayView2d< real32 > const sourceValue,
+          real64 const dt,
+          real32 const timeSourceFrequency,
+          real32 const timeSourceDelay,
+          localIndex const rickerOrder )
+  {
+    constexpr localIndex numNodesPerElem = FE_TYPE::numNodes;
+
+    forAll< EXEC_POLICY >( size, [=] GEOS_HOST_DEVICE ( localIndex const k )
+    {
+      real64 const center[3] = { elemCenter[k][0],
[…]

[…]
