+++ src/coreComponents/physicsSolvers/wavePropagation/AcousticROMFrechet3.cpp
@@ -1,1865 +1,0 @@
-
-/*
- * ------------------------------------------------------------------------------------------------------------
- * SPDX-License-Identifier: LGPL-2.1-only
- *
- * Copyright (c) 2018-2020 Lawrence Livermore National Security LLC
- * Copyright (c) 2018-2020 The Board of Trustees of the Leland Stanford Junior University
- * Copyright (c) 2018-2020 TotalEnergies
- * Copyright (c) 2019-     GEOSX Contributors
- * All rights reserved
- *
- * See top level LICENSE, COPYRIGHT, CONTRIBUTORS, NOTICE, and ACKNOWLEDGEMENTS files for details.
- * ------------------------------------------------------------------------------------------------------------
- */
-
-
-/**
- * @file AcousticROMFrechet.cpp
- */
-
-#include "AcousticROMFrechet.hpp"
-#include "AcousticROMFrechetKernel.hpp"
-
-#include "finiteElement/FiniteElementDiscretization.hpp"
-#include "fieldSpecification/FieldSpecificationManager.hpp"
-#include "fieldSpecification/PerfectlyMatchedLayer.hpp"
-#include "mainInterface/ProblemManager.hpp"
-#include "mesh/ElementType.hpp"
-#include "mesh/mpiCommunications/CommunicationTools.hpp"
-#include "WaveSolverUtils.hpp"
-
-namespace geos
-{
-
-using namespace dataRepository;
-
-AcousticROMFrechet::AcousticROMFrechet( const std::string & name,
-                                                  Group * const parent ):
-  WaveSolverBase( name,
-                  parent )
-{
-
-  registerWrapper( viewKeyStruct::pressureNp1AtReceiversString(), &m_pressureNp1AtReceivers ).
-    setInputFlag( InputFlags::FALSE ).
-    setSizedFromParent( 0 ).
-    setDescription( "Pressure value at each receiver for each timestep" );
-
-  registerWrapper( viewKeyStruct::orderFrechetString(), &m_orderFrechet ).
-    setInputFlag( InputFlags::OPTIONAL ).
-    setApplyDefaultValue( -1 ).
-    setDescription( "Frechet derivative order computation" );
-
-  registerWrapper( viewKeyStruct::epsilonGSString(), &m_epsilonGS ).
-    setInputFlag( InputFlags::OPTIONAL ).
-    setApplyDefaultValue( 1e-1 ).
-    setDescription( "Precision threshold in gramSchmidtStiffness process" );
-
-  registerWrapper( viewKeyStruct::count_qString(), &m_count_q ).
-    setInputFlag( InputFlags::FALSE ).
-    setSizedFromParent( 0 ).
-    setDescription( "Number of basis function for each order" );
-
-  registerWrapper( viewKeyStruct::totcount_qString(), &m_totcount_q ).
-    setInputFlag( InputFlags::FALSE ).
-    setApplyDefaultValue( 0 ).
-    setDescription( "Total number of basis function" );
-
-  registerWrapper( viewKeyStruct::selectionOrderString(), &m_selectionOrder ).
-    setInputFlag( InputFlags::FALSE ).
-    setSizedFromParent( 0 ).
-    setDescription( "Order of selection of each basis functions" );
-  
-  registerWrapper( viewKeyStruct::cycleOrderString(), &m_cycleOrder ).
-    setInputFlag( InputFlags::FALSE ).
-    setSizedFromParent( 0 ).
-    setDescription( "Cycle at which the basis functions is selected" );
-}
-
-AcousticROMFrechet::~AcousticROMFrechet()
-{
-  // TODO Auto-generated destructor stub
-}
-
[…]
