### 2024-06-25 6b2d15f
add Python script for validation
