### 2024-06-25 5a9968f
add nvidia-smi call.
