+++ .github/workflows/ci_tests.yml
@@ -296,7 +296,7 @@ jobs:
             HOST_ARCH: ppc64le
             RUNS_ON: streak2
             NPROC: 32
-            DOCKER_RUN_ARGS: "--cpus=32 --memory=256g --runtime=nvidia -v /etc/pki/ca-trust/source/anchors/:/etc/pki/ca-trust/source/anchors/llnl:ro"
+            DOCKER_RUN_ARGS: "--cpus=32 --memory=256g --runtime=nvidia --gpus all -v /etc/pki/ca-trust/source/anchors/:/etc/pki/ca-trust/source/anchors/llnl:ro"
             DOCKER_CERTS_DIR: "/etc/pki/ca-trust/source/anchors"
             DOCKER_CERTS_UPDATE_COMMAND: "update-ca-trust"
 
