### 2024-06-25 6f5054c
Add helper functions
