### 2024-06-25 dcbc947
Do not call host only function
