### 2024-06-25 c5ebbb1
Decouple debug matrix output from logLevel (#3181)
