+++ BASELINE_NOTES.md
@@ -6,9 +6,15 @@ This file is designed to track changes to the integrated test baselines.
 Any developer who updates the baseline ID in the .integrated_tests.yaml file is expected to create an entry in this file with the pull request number, date, and their justification for rebaselining.
 These notes should be in reverse-chronological order, and use the following time format: (YYYY-MM-DD).
 
+PR #3181 (2024-06-25)
+======================
+Decouple debug matrix output from logLevel. Rebaseline because of the new input flag.
+
+ 
 PR #3142 (2024-06-20)
 ======================
 Adding output of total strain. Rebaseline because of new inclusion of strain in output.
+
 
 PR #3170 (2024-06-19)
 ======================
