### 2024-06-25 02e899e
comments for construction of alocal djacency matrix data by each rank (Ryan's version)
