+++ src/coreComponents/mesh/generators/NewGhosting.cpp
@@ -398,6 +398,9 @@ TriCrsMatrix multiply( int commSize,
   return result_d0_2;
 }
 
+// Helper class which converts between matrix indices and (node, edge, face, cell) indices. 
+// The matrix rows/cols represent all the nodes, then all the edges, then all the faces, then all the cells
+// so conversion just means adding/subtracting the proper offsets
 class FindGeometricalType
 {
 public:
@@ -491,6 +494,14 @@ private:
   TriGlbIdx const m_numEntries;
 };
 
+// Helper function to encode geometrical information into ints which can be entered into adjacency matrix
+// For example, a face is connected to multiple edges, so in the face row, there will be a nonzero entry in the col for each edge
+// But if we just put a 1 in each we lose the information like what order the edges were in and what order the nodes were in
+// So instead we put in a value such that we can reconstruct this info if needed later
+// For encoding the edges within a face, N = 2, basis = NumEdges, array = {start node (0,1), index of edge within face},
+// so that result = start_node*numEdges+index_in_face
+// For encoding faces within a cell, N = 3, basis = numFaces,  array =  { isFlipped, start_index, index of face in cell },
+// so that result = (isFlipped*numFaces + start_index)*numFaces + index_in_cell
 template< std::uint8_t N >
 std::size_t encode( std::size_t const & basis,
                     std::array< std::size_t, N > const & array )
@@ -526,6 +537,7 @@ std::array< std::size_t, N > decode( std::size_t const & basis,
   return result;
 }
 
+// Structure which describes the adjacency matrix for the mesh graph with the data on current rank
 struct Adjacency
 {
   std::vector< TriGlbIdx > ownedNodesIdcs;
@@ -536,16 +548,20 @@ struct Adjacency
   std::vector< std::vector< TriScalarInt > > values;
 };
 
+// Function to populate the adjacency struct defined directly above
+// Note that we really build I+A, as in the end we will essentially be doing graph traversal by multiplying by (I+A)^n, ((I+A)^T)^n
+// Inlcuding I gives you everything in the graph 'up to' n steps away, without it you only get things 'exactly' n steps away
 Adjacency buildAdjacency( MeshGraph const & owned,
                           MeshGraph const & present,
                           FindGeometricalType const & convert )
 {
   Adjacency adjacency;
 
+  // Num matrix indices owned by the current rank, and the num indices corresponding to the 'present' data
   std::size_t const numOwned = std::size( owned.n2pos ) + std::size( owned.e2n ) + std::size( owned.f2e ) + std::size( owned.c2f );
   std::size_t const numOther = std::size( present.n2pos ) + std::size( present.e2n ) + std::size( present.f2e );
 
-  // Aliases
+  // Aliases for the data in adjacency
   std::vector< TriGlbIdx > & ownedNodesIdcs = adjacency.ownedNodesIdcs;
   std::vector< TriGlbIdx > & ownedGlbIdcs = adjacency.ownedGlbIdcs;
   std::vector< TriGlbIdx > & otherGlbIdcs = adjacency.otherGlbIdcs;
@@ -553,6 +569,7 @@ Adjacency buildAdjacency( MeshGraph const & owned,
   std::vector< std::vector< TriGlbIdx > > & indices = adjacency.indices;
   std::vector< std::vector< TriScalarInt > > & values = adjacency.values;
 
+  // we can go ahead and set sizes for the current rank matrix data
   ownedNodesIdcs.reserve( std::size( owned.n2pos ) );
   ownedGlbIdcs.reserve( numOwned );
   otherGlbIdcs.reserve( numOther );
[…]
