### 2024-06-25 bf89fce
Fix sorting inconsistency
