### 2024-06-25 28b1541
Comments through ghostingFootprint (Ryan's version)
