+++ src/coreComponents/mesh/generators/NewGhosting.cpp
@@ -359,13 +359,27 @@ TriCrsMatrix multiply( int commSize,
   Teuchos::RCP< TriMap const > ownedMap = upward.getRowMap();
   Teuchos::RCP< TriMap const > mpiMap = indicator.getRangeMap();
 
-  // Upward (n -> e -> f -> c)
+  // row j of indicator was just ones in columns corresponding to present entities on rank j
+  // row i of upward was constructed as having nonzeros in the columns of entities on which the entity corresponding to row i directly depended 
 
+  // Begin graph traversal by matrix matrix multiplying
+  // Trilinos matrix matrix multiply https://docs.trilinos.org/dev/packages/tpetra/doc/html/namespaceTpetra_1_1MatrixMatrix.html#a4c3132fa56ba6d5b071f053486529111
+
+  // Upward (n -> e -> f -> c) 
+
+  // u0_0 (num_entities x num_rank) = upward (num_entities x num_entities) * indicator^T (num_entities x num_rank)
+  // Using the `upward` interpretation, multiplying upward by each column of indicator^T would give 
+  // a column vector noting the all the faces in the global mesh which contain the edges owned by that rank, all the global cells which contain the owned faces, etc
+  // These are the colums of the result `result_u0_0`
   TriCrsMatrix result_u0_0( ownedMap, commSize );
   Tpetra::MatrixMatrix::Multiply( upward, false, indicator, true, result_u0_0, false );
   result_u0_0.fillComplete( mpiMap, ownedMap );
   Tpetra::MatrixMarket::Writer< TriCrsMatrix >::writeSparseFile( "/tmp/matrices/result-0.mat", result_u0_0 );
 
+  // We repeat this process to make sure we catch all the dependencies, but im not convinced once isnt enough
+  // In theory, you iterate until the matrix stops changing
+  // TODO: test this hypothesis
+  // TODO: you might also be able to just do one multiplication with upward + upward^T to do everything at once, but perhaps decoding becomes an issue
   TriCrsMatrix result_u0_1( ownedMap, commSize );
   Tpetra::MatrixMatrix::Multiply( upward, false, result_u0_0, false, result_u0_1, false );
   result_u0_1.fillComplete( mpiMap, ownedMap );
@@ -376,8 +390,14 @@ TriCrsMatrix multiply( int commSize,
   result_u0_2.fillComplete( mpiMap, ownedMap );
   Tpetra::MatrixMarket::Writer< TriCrsMatrix >::writeSparseFile( "/tmp/matrices/result-2.mat", result_u0_2 );
 
-  // Downward (c -> f -> e -> n)
+  // Now we know, for each rank, all the geometrical quantities in the global mesh which include something from that rank
 
+  // Downward (c -> f -> e -> n)
+  // Note that we still need to do a traversal in the "opposite direction"
+  // Consider starting with some edge, and applying the process above. 
+  // You will recover the faces containing it, and the cells which contain those
+  // What you wont get, for example, is any other (faces, edges, nodes) in that cell
+  // Thus we repeat the process with downward = upward^T
   TriCrsMatrix result_d0_0( ownedMap, commSize );
   Tpetra::MatrixMatrix::Multiply( upward, true, result_u0_2, false, result_d0_0, false );
   result_d0_0.fillComplete( mpiMap, ownedMap );
@@ -394,6 +414,10 @@ TriCrsMatrix multiply( int commSize,
   Tpetra::MatrixMarket::Writer< TriCrsMatrix >::writeSparseFile( "/tmp/matrices/result-6.mat", result_d0_2 );
 
   MpiWrapper::barrier();
+
+  // Now we know, for each rank, the set of all global geometric entities which are included by any geometric entity that includes something from that rank
+  // (what a mouthful)
[…]
