### 2024-06-26 5aa2b77
add doc and python script for validation
