+++ src/docs/sphinx/advancedExamples/validationStudies/wellboreProblems/nonLinearThermalDiffusion_TemperatureDependentSinglePhaseThermalConductivity/plot.py
@@ -1,0 +1,198 @@
+import os
+import sys
+
+import numpy as np
+import matplotlib.pyplot as plt
+import pandas as pd
+import scipy.linalg
+from scipy import special
+
+from scipy.sparse import diags
+from scipy.sparse.linalg import spsolve
+
+from xml.etree import ElementTree
+
+# Analytical results for linear thermal behavior
+def steadyState(Tin, Tout, Rin, Rout, radialCoordinate):
+	return Tin + (Tout - Tin) * (np.log(radialCoordinate) - np.log(Rin)) / (np.log(Rout) - np.log(Rin))
+
+def diffusionFunction(radialCoordinate, Rin, diffusionCoefficient, diffusionTime):
+	return special.erfc(  (radialCoordinate - Rin) / 2.0 / np.sqrt( diffusionCoefficient * diffusionTime ) )
+
+def computeTransientTemperature(Tin, Tout, Rin, radialCoordinate, thermalDiffusionCoefficient, diffusionTime):
+	# Ref. Wang and Papamichos (1994), https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/94WR01774
+	return Tout + (Tin-Tout) * np.sqrt(Rin/radialCoordinate) * diffusionFunction(radialCoordinate, Rin, thermalDiffusionCoefficient, diffusionTime)
+	 
+def computeThermalDiffusionCoefficient(thermalConductivity, volumetricHeatCapacity):
+	return thermalConductivity / volumetricHeatCapacity
+	
+# Finite difference results for non-linear thermal behavior
+def temperatureDependentThermalConductivity(lambda0, lambda_gradient, T, Treference):
+    return lambda0 + lambda_gradient*(T-Treference)
+
+def temperatureDependentVolumetricHeat(c0, c_gradient, T, Treference):
+    return c0 + c_gradient*(T-Treference)
+
+def coefficientMatrix(thermalConductivity, volumetricHeatCapacity, r, dt, N):
+    # Coefficients for the matrix
+    A = np.zeros((N+1, N+1))
+    
+    for i in range(1,N):
+        dr = r[i] - r[i-1] 		
+        r_i = r[i]
+
+        diffusivity_i = thermalConductivity[i]/volumetricHeatCapacity[i]
+        A[i, i-1] = - diffusivity_i * (dt/(dr**2) - dt/(2 * r_i * dr)) \
+                    + (thermalConductivity[i+1] - thermalConductivity[i-1])/volumetricHeatCapacity[i]*dt/4/(dr**2)
+        A[i, i] = 1.0 + 2.0 * diffusivity_i * dt / (dr**2)
+        A[i, i+1] = - diffusivity_i * (dt/(dr**2) + dt/(2 * r_i * dr)) \
+                    - (thermalConductivity[i+1] - thermalConductivity[i-1])/volumetricHeatCapacity[i]*dt/4/(dr**2)
+
+    # Boundary conditions
+    # No-flux at r=0 approximated by setting the flux between the first two cells to zero
+    A[0, 0] = 1.0 
+    A[N, N] = 1.0
+    return A
+
+def solve_radial_diffusion(r, tmax, dt, Tin, Tout, lambda0, lambda_gradient, c0, c_gradient, Treference):
+    N = len(r)-1
+    # Time setup
+    n_steps = int(tmax / dt)
+    
+    # Time-stepping
+    T = np.zeros(N+1) + Tout  # initial condition u(r, 0)
+    T[0] = Tin
+    for step in range(n_steps):
+        thermalConductivity = temperatureDependentThermalConductivity(lambda0, lambda_gradient, T, Treference)
+        
[…]
