+++ src/docs/sphinx/advancedExamples/validationStudies/wellboreProblems/nonLinearThermalDiffusion_TemperatureDependentSinglePhaseThermalConductivity/Example.rst
@@ -1,0 +1,62 @@
+.. _AdvancedExampleWellboreNonLinearThermalDiffusionTemperatureDependentSinglePhaseThermalConductivity:
+
+
+#####################################################################################################################
+Non-Linear Thermal Diffusion Around a Wellbore: The Case with Temperature Dependent Single Phase Thermal Conductivity
+#####################################################################################################################
+
+------------------------------------------------------------------
+Problem description
+------------------------------------------------------------------
+
+This example is an extension of the linear thermal diffusion problem presented in :ref:`AdvancedExamplePureThermalDiffusionWellbore`. It uses the thermal single-phase flow solver to model a non-linear thermal diffusion problem around a wellbore where the single phase thermal conductivity of the porous rock depends linearly on the temperature.
+
+
+**Input file**
+
+This benchmark example uses no external input file and everything required is
+contained within two GEOS xml files that are located at:
+
+.. code-block:: console
+
+  inputFiles/???_base.xml
+
+and
+
+.. code-block:: console
+
+  inputFiles/???_benchmark.xml
+
+
+In this example, we focus on the ``Constitutive`` tag.
+
+-----------------------------------------------------------
+Constitutive
+-----------------------------------------------------------
+
+The reference value of the single phase thermal conductivity of the porous medium around the wellbore and its derivative with respect to temperature are defined in the ``SolidInternalEnergy`` XML block:  
+
+.. literalinclude:: ../../../../../../../inputFiles/???_base.xml
+  :language: xml
+  :start-after: <!--  -->
+  :end-before: <!--  -->
+
+
+---------------------------------
+Results and benchmark
+---------------------------------
+
+A good agreement between the results obtained using GEOS and the analytical results is shown in the figure below:
+
+
+.. plot:: docs/sphinx/advancedExamples/validationStudies/wellboreProblems/nonLinearThermalDiffusion_TemperatureDependentSinglePhaseThermalConductivity/plot.py
+
+
+------------------------------------------------------------------
+To go further
+------------------------------------------------------------------
+
+**Feedback on this example**
+
+This concludes the example of a non-linear thermal diffusion problem around a wellbore due to temperature dependent single phase thermal conductivity of porous rock.
+For any feedback on this example, please submit a `GitHub issue on the project's GitHub page <https://github.com/GEOS-DEV/GEOS/issues>`_.
