### 2024-06-26 26b7ac4
update thermal unit tests
