### 2024-06-26 cf70f7d
update doc and Python script
