+++ src/docs/sphinx/advancedExamples/validationStudies/wellboreProblems/nonLinearThermalDiffusion_TemperatureDependentVolumetricHeatCapacity/plot.py
@@ -101,8 +101,14 @@ def getLoadingFromXML(xmlFilePath):
 				Tin = float(fsParam.get('scale'))
 			if fsParam.get('setNames') == "{ rpos }":
 				Tout = float(fsParam.get('scale'))
-            
-	thermalConductivity = float( extractDataFromXMLList( tree.find('Constitutive/SinglePhaseConstantThermalConductivity').get('thermalConductivityComponents') )[0] )
+        
+	
+	tree_SinglePhaseThermalConductivities = tree.findall('Constitutive/SinglePhaseThermalConductivity')
+
+	for tree_SinglePhaseThermalConductivity in tree_SinglePhaseThermalConductivities:
+		if tree_SinglePhaseThermalConductivity.get('name') == "thermalCond_linear":
+			thermalConductivity = float( extractDataFromXMLList( tree_SinglePhaseThermalConductivity.get('defaultThermalConductivityComponents') )[0] )
+
 
 	tree_SolidInternalEnergies = tree.findall('Constitutive/SolidInternalEnergy')
 
