+++ src/coreComponents/unitTests/fluidFlowTests/testThermalCompMultiphaseFlow.cpp
@@ -95,15 +95,14 @@ char const * xmlInput =
     <ElementRegions>
       <CellElementRegion name="region"
                          cellBlocks="{ cb }"
-                         materialList="{ fluid, rock, relperm }" />
+                         materialList="{ fluid, rock, relperm, thermalCond }" />
     </ElementRegions>
     <Constitutive>
       <CompressibleSolidConstantPermeability name="rock"
                                              solidModelName="nullSolid"
                                              porosityModelName="rockPorosity"
                                              permeabilityModelName="rockPerm"
-                                             solidInternalEnergyModelName="rockInternalEnergy"
-                                             singlePhaseThermalConductivityModelName="thermalCond" />
+                                             solidInternalEnergyModelName="rockInternalEnergy"/>
       <NullModel name="nullSolid" />
       <PressurePorosity name="rockPorosity"
                         defaultReferencePorosity="0.2"
