### 2024-06-26 fd3a389
fix set apply default vector
