+++ src/coreComponents/constitutive/fluid/multifluid/reactive/ReactiveBrineFluid.hpp
@@ -162,9 +162,8 @@ private:
 
   /**
    * @brief Create a PVT Model and output them
-   * @param isClone If we are in the case of a clone of a constitutive mode, never the output
    */
-  void createPVTModels( bool isClone );
+  void createPVTModels();
 
   /// Names of the files defining the viscosity and density models
   path_array m_phasePVTParaFiles;
