### 2024-06-26 4ec9dae
Add warning about GPU use
