### 2024-06-26 d5fe6c6
Add EOS parameters to flash
