### 2024-06-26 e560eab
Update .github/workflows/ci_tests.yml
Co-authored-by: Randolph Settgast <settgast1@llnl.gov>
