### 2024-06-26 9b803e6
try reducing nproc
