+++ host-configs/LLNL/old/quartz-clang-14.cmake
@@ -1,10 +1,0 @@
-include(${CMAKE_CURRENT_LIST_DIR}../../../src/coreComponents/LvArray/host-configs/LLNL/old/quartz-clang-14.cmake)
-
-# Fortran
-set(CMAKE_Fortran_COMPILER /usr/tce/packages/gcc/gcc-12.1.1-magic/bin/gfortran CACHE PATH "")
-set(CMAKE_Fortran_FLAGS_RELEASE "-O3 -DNDEBUG -march=native -mtune=native" CACHE STRING "")
-
-# MPI
-set(MPI_HOME /usr/tce/packages/mvapich2/mvapich2-2.3.6-clang-14.0.6-magic CACHE PATH "")
-
-include(${CMAKE_CURRENT_LIST_DIR}/quartz-base.cmake)
