+++ src/coreComponents/mesh/generators/NewGhosting.cpp
@@ -891,19 +891,26 @@ std::tuple< MeshGraph, GhostRecv, GhostSend > performGhosting( MeshGraph const &
   // The `ghostExchange` matrix is rectangular,
   // the number of columns being the number of nodes in the mesh graph, ie the total number of mesh geometric quantities
   // the number of rows being the number of MPI ranks.
+  // ghostExchange (num_ranks x num_entities) = ghostingFootprint^T (num_ranks x num_entities) * ownership (num_entities x num_entities, diagonal)
+  // recall that ghostExchange has all the nonzeros set to 1
+  // diagonal matrix just multiplies each col of ghostingFootprint^T (row of ghostingFootprint) by the corresponding diagonal entry
+  // So each column of ghostExchange will have nonzeros which are all the same value, with the value being the rank owning that entity
+  // Thus the nonzeros in row i tell us the ranks which own the entities "adjacent" to the entities on rank i
   //
   // As the result of the multiplication between the `ghostingFootprint` matrix and the `ownership` matrix,
   // for each row owned (ie at the current MPI rank index),
-  // the value of the `ghostExchange` matrix term will provide the actual owning rank for all the .
+  // the value of the `ghostExchange` matrix term will provide the actual owning rank for all the needed geometric entities.
   //
   // From `ghostExchange` we can extract which other rank will send to the current rank any graph node.
-  TriCrsMatrix ghostExchange( mpiMap, 10000 ); // TODO having `0` there should be working!
+  TriCrsMatrix ghostExchange( mpiMap, 10000 ); // TODO having `0` there should be working! We need to get a better estimate of number of entries
   Tpetra::MatrixMatrix::Multiply( ghostingFootprint, true, ownership, false, ghostExchange, false );
   ghostExchange.fillComplete( ownedMap, mpiMap );
 
   // TODO Do I have to work with `ghostingFootprint` if I already have `ghostExchange` which may convey more information?
+  // As we shall see, below we go back to working with ghostingFootprint
   // TODO Maybe because of the ownership of the ranks: one is also the "scaled" transposed of the other.
 
+  // debug output
   if( curRank == 0_mpi )
   {
     GEOS_LOG_RANK( "ghostExchange->NumGlobalCols() = " << ghostExchange.getGlobalNumCols() );
@@ -911,20 +918,32 @@ std::tuple< MeshGraph, GhostRecv, GhostSend > performGhosting( MeshGraph const &
   }
   Tpetra::MatrixMarket::Writer< TriCrsMatrix >::writeSparseFile( "/tmp/matrices/ghostInfo.mat", ghostExchange );
 
+  // Now, for each of the entities owned by this current rank, we collect all the other ranks to which this entity is sent
+
+  // GhostSend defined in BuildPods.hpp, 
+  // struct of maps for each (node, edge, face, cell) giving the set of ranks each entity (given by node, edge, face, cell global index) must be sent to
+  GhostSend send;
+  
+  // temp data structures used in process
   std::size_t extracted = 0;
   Teuchos::Array< TriScalarInt > extractedValues;
   Teuchos::Array< TriGlbIdx > extractedIndices;
 
-  GhostSend send;
+  // Loop over the owned indiced in global matrix (owned geometric entities)
   for( TriGlbIdx const & index: ownedGlbIdcs )
   {
+    // get the number of ranks which needed this entity and resize our temp data storage
     std::size_t const length = ghostingFootprint.getNumEntriesInGlobalRow( index );
-    extractedValues.resize( length );
+    extractedValues.resize( length ); // note we never actually use the values, just needed the indices where the nonzeros were
     extractedIndices.resize( length );
[…]
