### 2024-06-26 450b5ed
Added particle location to warning that particle was no repartitioned, added weighting to surface normals in contact calculations
