### 2024-06-26 db1f07a
Added particle location to warning that particle was no repartitioned, added weighting to surface normals in contact calculations
