+++ src/coreComponents/physicsSolvers/solidMechanics/SolidMechanicsMPM.cpp
@@ -136,7 +136,11 @@ SolidMechanicsMPM::SolidMechanicsMPM( const string & name,
   m_stressControlLastError(),
   m_stressControlITerm(),
   m_boxAverageHistory( 0 ),
+  m_boxAverageWriteInterval( 0.0 ),
   m_reactionHistory( 0 ),
+  m_writeParticleData( 0 ).
+  m_reactionWriteInterval( 0.0 ),
+  m_particleDataWriteInterval( 0.0 ),
   m_explicitSurfaceNormalInfluence( 0.0 ),
   m_computeSurfaceNormalsOnlyOnInitialization( 0 ),
   m_referenceCohesiveZone( 0 ),
@@ -188,6 +192,7 @@ SolidMechanicsMPM::SolidMechanicsMPM( const string & name,
   m_damageFieldPartitioning( 0 ),
   m_useSurfacePositionForContact( 0 ),
   m_contactNormalType( ContactNormalTypeOption::MassWeighted ),
+  m_contactNormalExponent( 1.0 ),
   m_contactGapCorrection( ContactGapCorrectionOption::Simple ),
   m_resetDefGradForFullyDamagedParticles( 0 ),
   m_plotUnscaledParticles( 0 ),
@@ -283,7 +288,7 @@ SolidMechanicsMPM::SolidMechanicsMPM( const string & name,
 
   registerWrapper( "reactionWriteInterval", &m_reactionWriteInterval ).
     setInputFlag( InputFlags::OPTIONAL ).
-    setApplyDefaultValue( 0 ).
+    setApplyDefaultValue( m_reactionWriteInterval ).
     setRestartFlags( RestartFlags::NO_WRITE ).
     setDescription( "Interval between writing reactions to files" );
 
@@ -292,6 +297,24 @@ SolidMechanicsMPM::SolidMechanicsMPM( const string & name,
     setApplyDefaultValue( 0 ).
     setRestartFlags( RestartFlags::WRITE_AND_READ ).
     setDescription( "Next time to write reactions" );
+
+  registerWrapper( "writeParticleData", &m_writeParticleData ).
+    setInputFlag( InputFlags::OPTIONAL ).
+    setApplyDefaultValue( m_writeParticleData ).
+    setRestartFlags( RestartFlags::NO_WRITE ).
+    setDescription( "Flag to enable writing particle data to file" );
+
+  registerWrapper( "particleDataWriteInterval", &m_particleDataWriteInterval ).
+    setInputFlag( InputFlags::OPTIONAL ).
+    setApplyDefaultValue( m_particleDataWriteInterval ).
+    setRestartFlags( RestartFlags::NO_WRITE ).
+    setDescription( "Interval to write particle data to file" );
+
+  registerWrapper( "nextParticleDataWriteTime", &m_nextParticleDataWriteTime ).
+    setInputFlag( InputFlags::FALSE ).
+    setApplyDefaultValue( 0.0 ).
+    setRestartFlags( RestartFlags::NO_WRITE ).
+    setDescription( "Next time to write particle data" );
 
   registerWrapper( "explicitSurfaceNormalInfluence", &m_explicitSurfaceNormalInfluence ).
     setInputFlag( InputFlags::OPTIONAL ).
@@ -737,6 +760,12 @@ SolidMechanicsMPM::SolidMechanicsMPM( const string & name,
     setInputFlag( InputFlags::OPTIONAL ).
     setRestartFlags( RestartFlags::NO_WRITE ).
     setDescription( "Flag for contact normal type" );
+
+  registerWrapper( "contactNormalExponent", &m_contactNormalExponent ).
+    setApplyDefaultValue( m_contactNormalExponent ).
[…]
