### 2024-06-26 59a812d
Added particle location to warning that particle was no repartitioned, added weighting to surface normals in contact calculations
