### 2024-06-26 3a8acc2
Implement stability test (#3136)
