+++ src/coreComponents/constitutive/unitTests/testStabilityTest9Comp.cpp
@@ -1,0 +1,631 @@
+/*
+ * ------------------------------------------------------------------------------------------------------------
+ * SPDX-License-Identifier: LGPL-2.1-only
+ *
+ * Copyright (c) 2018-2020 Lawrence Livermore National Security LLC
+ * Copyright (c) 2018-2020 The Board of Trustees of the Leland Stanford Junior University
+ * Copyright (c) 2018-2020 TotalEnergies
+ * Copyright (c) 2019-     GEOSX Contributors
+ * All rights reserved
+ *
+ * See top level LICENSE, COPYRIGHT, CONTRIBUTORS, NOTICE, and ACKNOWLEDGEMENTS files for details.
+ * ------------------------------------------------------------------------------------------------------------
+ */
+
+// Source includes
+#include "codingUtilities/UnitTestUtilities.hpp"
+#include "constitutive/fluid/multifluid/MultiFluidConstants.hpp"
+#include "constitutive/fluid/multifluid/compositional/functions/StabilityTest.hpp"
+#include "constitutive/fluid/multifluid/compositional/functions/NegativeTwoPhaseFlash.hpp"
+#include "constitutive/fluid/multifluid/compositional/functions/CubicEOSPhaseModel.hpp"
+#include "TestFluid.hpp"
+#include "TestFluidUtilities.hpp"
+
+using namespace geos::constitutive;
+using namespace geos::constitutive::compositional;
+
+namespace geos
+{
+namespace testing
+{
+
+static constexpr integer numComps = 9;
+
+using FlashData = std::tuple<
+  real64 const,             // pressure
+  real64 const,             // temperature
+  Feed< numComps > const,   // total composition
+  real64 const              // expected tangent plane distance
+  >;
+
+template< typename EOS_TYPE >
+class StabilityTestTest9CompFixture :  public ::testing::TestWithParam< FlashData >
+{
+  static constexpr real64 relTol = 1.0e-5;
+  static constexpr real64 absTol = 1.0e-7;
+  static constexpr real64 tpdTol = MultiFluidConstants::fugacityTolerance;
+public:
+  StabilityTestTest9CompFixture()
+    : m_fluid( createFluid() )
+  {}
+
+  ~StabilityTestTest9CompFixture() = default;
+
+  void testStability( FlashData const & data )
+  {
+    auto componentProperties = this->m_fluid->createKernelWrapper();
+
+    real64 const pressure = std::get< 0 >( data );
+    real64 const temperature = std::get< 1 >( data );
+    stackArray1d< real64, numComps > composition;
+    TestFluid< numComps >::createArray( composition, std::get< 2 >( data ));
+
+    real64 const expectedTangentPlaneDistance = std::get< 3 >( data );
+
+    real64 tangentPlaneDistance = LvArray::NumericLimits< real64 >::max;
+    stackArray1d< real64, numComps > kValues( numComps );
+
[…]
