### 2024-06-26 62713f5
Implement stability test (#3136)
