+++ src/coreComponents/constitutive/fluid/multifluid/compositional/functions/StabilityTest.hpp
@@ -1,0 +1,233 @@
+/*
+ * ------------------------------------------------------------------------------------------------------------
+ * SPDX-License-Identifier: LGPL-2.1-only
+ *
+ * Copyright (c) 2018-2020 Lawrence Livermore National Security LLC
+ * Copyright (c) 2018-2020 The Board of Trustees of the Leland Stanford Junior University
+ * Copyright (c) 2018-2020 TotalEnergies
+ * Copyright (c) 2019-     GEOSX Contributors
+ * All rights reserved
+ *
+ * See top level LICENSE, COPYRIGHT, CONTRIBUTORS, NOTICE, and ACKNOWLEDGEMENTS files for details.
+ * ------------------------------------------------------------------------------------------------------------
+ */
+
+/**
+ * @file StabilityTest.hpp
+ */
+
+#ifndef GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_FUNCTIONS_STABILITYTEST_HPP_
+#define GEOS_CONSTITUTIVE_FLUID_MULTIFLUID_COMPOSITIONAL_FUNCTIONS_STABILITYTEST_HPP_
+
+#include "KValueInitialization.hpp"
+#include "constitutive/fluid/multifluid/Layouts.hpp"
+#include "constitutive/fluid/multifluid/MultiFluidConstants.hpp"
+#include "constitutive/fluid/multifluid/compositional/models/ComponentProperties.hpp"
+
+namespace geos
+{
+
+namespace constitutive
+{
+
+namespace compositional
+{
+
+struct StabilityTest
+{
+private:
+  static constexpr integer maxNumComps = MultiFluidConstants::MAX_NUM_COMPONENTS;
+public:
+  /**
+   * @brief Perform a two-phase stability test
+   * @param[in] numComps number of components
+   * @param[in] pressure pressure
+   * @param[in] temperature temperature
+   * @param[in] composition composition of the mixture
+   * @param[in] componentProperties The compositional component properties
+   * @param[out] tangentPlaneDistance the minimum tangent plane distance (TPD)
+   * @param[out] kValues the k-values estimated from the stationary points
+   * @return a flag indicating that 2 stationary points have been found
+   */
+  template< typename EOS_TYPE, integer USD1 >
+  GEOS_HOST_DEVICE
+  static bool compute( integer const numComps,
+                       real64 const pressure,
+                       real64 const temperature,
+                       arraySlice1d< real64 const, USD1 > const & composition,
+                       ComponentProperties::KernelWrapper const & componentProperties,
+                       real64 & tangentPlaneDistance,
+                       arraySlice1d< real64 > const & kValues )
+  {
+    constexpr integer numTrials = 2;    // Trial compositions
+    stackArray1d< real64, maxNumComps > logFugacity( numComps );
+    stackArray1d< real64, maxNumComps > normalizedComposition( numComps );
+    stackArray2d< real64, numTrials *maxNumComps > trialComposition( numTrials, numComps );
+    stackArray1d< real64, maxNumComps > logTrialComposition( numComps );
+    stackArray1d< real64, maxNumComps > hyperplane( numComps );     // h-parameter
+    stackArray1d< integer, maxNumComps > availableComponents( numComps );
+
+    calculatePresentComponents( numComps, composition, availableComponents );
+    auto const presentComponents = availableComponents.toSliceConst();
+
+    // Calculate the hyperplane parameter
+    // h_i = log( z_i ) + log( phi_i )
+    for( integer ic = 0; ic < numComps; ++ic )
+    {
+      hyperplane[ic] = 0.0;
+    }
+    EOS_TYPE::computeLogFugacityCoefficients( numComps,
+                                              pressure,
+                                              temperature,
+                                              composition,
[…]
