### 2024-06-26 716a082
complete comments for performGhosting (Ryan's version)
