+++ src/coreComponents/mesh/generators/NewGhosting.cpp
@@ -1041,7 +1041,8 @@ std::tuple< MeshGraph, GhostRecv, GhostSend > performGhosting( MeshGraph const &
     // matrix index of required entity
     TriGlbIdx const & index = extractedIndices[i];
 
-    // TODO: should this be notPresentIndices??
+    // Note this is not notPresentIndices, as information which is defined on geometric entities 
+    // which happen to already be present will still need to be communicated in the future
     if( receivedIndices.find( index ) == std::cend( receivedIndices ) )  // TODO make a map `receivedIndices -> mpi rank`
     {
       continue;
@@ -1096,10 +1097,13 @@ std::tuple< MeshGraph, GhostRecv, GhostSend > performGhosting( MeshGraph const &
   //   knowing the index of the nodes is not enough.
   //
   // In order to do that, we build the `missingIndicator` matrix, which is rectangular:
-  // - The number of columns is the number of graph nodes in the mesh graph.
+  // - The number of columns is the number of graph nodes in the mesh graph, ie the total number of geometrical entities in the mesh.
   // - The number of rows is the total number of graph nodes that are missing on ranks.
+  //   For me this was confusing, as its a little different from what we've done before
+  //   Each rank is missing some quantity of mesh entities, and we are going to MpiAllreduce to get the total across all ranks, and this is the number of rows in the global matrix
+  //   Its going to be incredible sparse, in fact only one nonzero per row marking the index of the entity in the graph
   //   Note that the same quantity can be missing on multiple ranks, and that's handled.
-  // - The non-zero terms equal `1`, meaning that a given quantity (column) is missing on a given rank (row).
+  // - The non-zero terms equal `1`, meaning that a given quantity (column) is missing on a given rank (row, each rank owns a chunk of rows).
   //
   // Combining `missingIndicator` with the adjacency matrix which conveys a lot of connections information,
   // we'll be able to create the final `missingIndices` matrix,
@@ -1120,21 +1124,31 @@ std::tuple< MeshGraph, GhostRecv, GhostSend > performGhosting( MeshGraph const &
   //   that actually are physical nodes in the mesh.
   // - The number of rows is the total number of graph nodes (that actually are physical nodes in the mesh) that are missing on ranks.
   // - The non-zero terms equal `1`, meaning that a given quantity (column) is missing on a given rank (row).
+
+  // Group together the number of missing mesh entities and the number of missing mesh nodes in particular for this rank
   std::size_t const numLocMissingCompound[2] = { std::size( notPresentIndices ), std::size( notPresentNodes ) };
+  // Now we can communicate to get the sum over missing entities and missing mesh nodes over all the ranks with simple reduction
   std::size_t numGlbMissingCompound[2] = { 0, 0 };
   MpiWrapper::allReduce( numLocMissingCompound, numGlbMissingCompound, 2, MPI_SUM );
+
+  // break the above back into individual variables now that parallel reduction is done
   std::size_t const & numLocMissingIndices = numLocMissingCompound[0];
   std::size_t const & numLocMissingNodes = numLocMissingCompound[1];
   Tpetra::global_size_t const numGlbMissingIndices = numGlbMissingCompound[0];
   Tpetra::global_size_t const numGlbMissingNodes = numGlbMissingCompound[1];
 
-  std::size_t const numGlbNodes = gis.nodes.get() + 1;  // TODO Use strongly typed integers
[…]
