### 2024-06-27 03d59f2
Fix compositional density
