### 2024-06-27 0546d6b
Remove EOS parameter from stability test
