### 2024-07-01 #6829 @lucydot
@MakisH @dglaeser @timokoch  - hello 👋 . Are you available to review this submission to JOSS?
