### 2024-07-01 #6829 @MakisH
@lucydot I can review this submission (FYI: this will be my first review for JOSS). What is the expected timeline?
