### 2024-07-08 #6829 @lucydot
Hi @MakisH  - excellent that this will be your first review, always good to bring someone into the fold! I'll be here to answer any questions you have, and the online docs are useful: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

You don't have to do anything right now. I will find another one or two reviewers, and then a new thread will start (which will give clear review instructions). Once that starts we ask that reviews are completed in ~4 weeks. The process is iterative, a conversation back and forth between reviewers/authors/editors, rather than a "one-shot" review, so we ask that the reviews start early in the timeframe. If you have leave planned, let us know, so we have expectations set. I know holidays for many people (myself included) are coming up.
