### 2024-07-08 #6829 @lucydot
@dglaeser @timokoch @justinsgray @jamiebramwell - hello 👋 . Are you available to review this submission to JOSS?
