### 2024-07-08 #6829 @timokoch
@lucydot I can do a review. However, I'll be away for much of July. It looks like a relatively big complex HPC-focused package. It could be good to have an additional reviewer with good GPU experience.
