### 2024-07-10 #6829 @lucydot
Hi @timokoch  - great that you can review. And being away in July is fine - it is a time of year when many of us will be taking leave - thanks for the heads up. Also thanks for your suggestion re: GPU expertise 👍
