### 2024-07-10 #6829 @lucydot
Reading the [GEOS docs](https://geosx-geosx.readthedocs-hosted.com/en/latest/index.html) it appears that GEOS can run on (and be reviewed using) a number of system types (laptop, HPC, GPU). The pre-requisites for building are [here]( https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/buildGuide/Prerequisites.html#list-of-prerequisites).

@timokoch and @MakisH  - a couple of questions to ensure that the reviewer team (I will find one more) can assess the main parallel computing modes enabled in GEOS:

- do you have the capability to build/run with MPI?
- do you have the capability to build/run with OpenMP?
- do you have the capability to build/run with GPU-acceleration (via CUDA)?
