### 2024-07-10 #6829 @lucydot
Hi @@berenger-eu 👋  - are you available to review this submission to JOSS? We have two reviewers who have expertise in multi-physics simulations. Looking for a third with expertise in parallel computing (MPI/OpenMP/GPU), as the software runs using a number of different parallel computing schemes. There is more information in the [GEOS docs](https://geosx-geosx.readthedocs-hosted.com/en/latest/index.html).
