### 2024-07-10 #6829 @MakisH
> @timokoch and @MakisH - a couple of questions to ensure that the reviewer team (I will find one more) can assess the main parallel computing models enabled in GEOS:
>
>     * do you have the capability to build/run with MPI?
>
>     * do you have the capability to build/run with OpenMP?
>
>     * do you have the capability to build/run with GPU-acceleration (via CUDA)?

I will run on a laptop with MPI and OpenMP, but not with CUDA.
