### 2024-07-10 #6829 @dglaeser
> @dglaeser @timokoch @JustinSGray @jamiebramwell - hello 👋 . Are you available to review this submission to JOSS?

@lucydot sorry for the late reply! Unfortunately I am too busy at the moment to review this...
