### 2024-07-10 #6829 @rrsettgast
> Reading the [GEOS docs](https://geosx-geosx.readthedocs-hosted.com/en/latest/index.html) it appears that GEOS can run on (and be reviewed using) a number of system types (laptop, HPC, GPU). The pre-requisites for building are [here](https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/buildGuide/Prerequisites.html#list-of-prerequisites).
>
> @timokoch and @MakisH - a couple of questions to ensure that the reviewer team (I will find one more) can assess the main parallel computing models enabled in GEOS:
>
> * do you have the capability to build/run with MPI?
> * do you have the capability to build/run with OpenMP?
> * do you have the capability to build/run with GPU-acceleration (via CUDA)?

Hello all
A point of clarification is that GEOS uses a portability layer, and some of the results included in the paper were generated on NVIDIA V100 present on LLNL/Lassen, and the AMD MI250X GPU present on the ORNL/Frontier exascale machine. Regular non-fp64 capable NVIDIA GPU's may be problematic to run GEOS on (for NVIDIA, we haven't tried anything aside from a V100, A100, H100). The AMD GPU's are not readily available outside the US DOE complex, so it may be hard to find someone with access.

Is the review process testing the building, or about running the code examples? There exist publicly accessible docker images which we use as part of our CI process (building and testing) that reviewers can download. These contain the third-party-libraries for GEOS, and as they are the largest challenge to the build process it might be a good place to start. I can provide paths to the docker images from dockerhub if any reviewers would like to go this route.

Thanks!!!
Randy S
