### 2024-07-10 #6829 @lucydot
Thanks @rrsettgast for that information 👍

> Is the review process testing the building, or about running the code examples?

Usually both. The review needs to assess the major claims of the software, to ensure that stated functionality is available to other researchers. Part of this process usually involves installation/build. As GEOS is pitched at HPC (amongst others), my first thought is that a review will need to include a build and run on a multi-node machine (or at least with MPI). My understanding is that Docker images are limited in this regard. Availability through Docker images only would not necessarily be an acceptance-blocker but this would need to be made clear in the paper/docs.

If Docker images are the suggested way to use GEOS (? this wasn't obvious to me from [the docs](https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/QuickStart.html)), then by all means it might make sense to use them during the review process.
