### 2024-07-10 #6973 @editorialbot
[REVIEW]: GEOS: A performance portable multi-physics simulation framework for subsurface applications
**Submitting author:** @rrsettgast (<a href="http://orcid.org/0000-0002-2536-7867">Randolph Settgast</a>)
**Repository:** https://github.com/GEOS-DEV/GEOS
**Branch with paper.md** (empty if default branch): docs/JOSS
**Version:** v1.1.0
