### 2024-07-24 #6973 @lucydot
@rrsettgast @MakisH @timokoch @berenger-eu

A note here that from this Monday 29th I will be on leave (⛺) for just over two weeks, returning on the 14th of August when I'll check in on progress ASAP. If you have any queries, it may be best to ask before I go.
