### 2024-07-31 GEOS-DEV/GEOS#3256 @berenger-eu
JOSS Review - Cannot compile
Hello there,
I would like to compile GEOS on an HPC platform.
I was able to checkout the two repo, and then I loaded the following modules `module load compiler/cuda/12.3 compiler/gcc/11.2.0 mpi/openmpi/5.0.1 build/cmake/3.27.0 linalg/mkl/2020_update4`.

I added my hostfile, `nano ../GEOS/host-configs/plafrim/a100.cmake`

```
set(CMAKE_CXX_COMPILER "g++" CACHE PATH "" FORCE)
set(CMAKE_C_COMPILER "gcc" CACHE PATH "" FORCE)
set(CMAKE_Fortran_COMPILER "gfortran" CACHE PATH "" FORCE)

set(ENABLE_FORTRAN OFF CACHE BOOL "")

set(ENABLE_MPI ON CACHE BOOL "")
set(MPI_HOME /cm/shared/modules/generic/mpi/openmpi/5.0.1/ CACHE PATH "")
set(MPI_C_COMPILER       ${MPI_HOME}/bin/mpicc   CACHE PATH "")
set(MPI_CXX_COMPILER     ${MPI_HOME}/bin/mpicxx  CACHE PATH "")
set(MPI_Fortran_COMPILER ${MPI_HOME}/bin/mpifort CACHE PATH "")

set(MPIEXEC              /usr/bin/srun CACHE PATH "")
set(MPIEXEC_NUMPROC_FLAG "-n" CACHE STRING "")

set(ENABLE_OPENMP ON CACHE BOOL "")

set(ENABLE_SPHINX_EXECUTABLE OFF CACHE BOOL "")
set(ENABLE_UNCRUSTIFY OFF CACHE BOOL "")
set(ENABLE_DOXYGEN OFF CACHE BOOL "")

set(ENABLE_GTEST_DEATH_TESTS OFF CACHE BOOL "")

set(ENABLE_PVTPackage OFF CACHE BOOL "")
set(ENABLE_CALIPER OFF CACHE BOOL "")
set(ENABLE_ADIAK OFF CACHE BOOL "")
set(ENABLE_PAPI OFF CACHE BOOL "")
set(ENABLE_PETSC OFF CACHE BOOL "" FORCE )
set(ENABLE_TRILINOS OFF CACHE BOOL "" FORCE )
set(ENABLE_HYPRE OFF CACHE BOOL "" FORCE )
set(ENABLE_TOTALVIEW_OUTPUT OFF CACHE BOOL "Enables Totalview custom view" FORCE)

set(ENABLE_MKL ON CACHE BOOL "")
set(MKL_ROOT /cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl/)
set(MKL_INCLUDE_DIRS ${MKL_ROOT}/include CACHE STRING "")
set(MKL_LIBRARIES ${MKL_ROOT}/lib/intel64/libmkl_intel_lp64.so
                  ${MKL_ROOT}/lib/intel64/libmkl_gnu_thread.so
                  ${MKL_ROOT}/lib/intel64/libmkl_core.so
                  CACHE STRING "")

# Now let's add what's dedicated to GPU.
set(ENABLE_CUDA ON CACHE PATH "" FORCE)
set(CUDA_TOOLKIT_ROOT_DIR /usr/local/cuda-12.3/ CACHE PATH "")
set(CMAKE_CUDA_HOST_COMPILER ${CMAKE_CXX_COMPILER} CACHE STRING "")
set(CMAKE_CUDA_COMPILER ${CUDA_TOOLKIT_ROOT_DIR}/bin/nvcc CACHE STRING "")

set(CUDA_ARCH sm_75 CACHE STRING "")
set(CMAKE_CUDA_FLAGS "-restrict -arch ${CUDA_ARCH} --expt-relaxed-constexpr --expt-extended-lambda -Werror cross-execution-space-call,reorder,deprecated-declarations" CACHE STRING "")
set(CMAKE_CUDA_FLAGS_RELEASE "-O3 -DNDEBUG -Xcompiler -DNDEBUG -Xcompiler -O3" CACHE STRING "")
set(CMAKE_CUDA_FLAGS_RELWITHDEBINFO "-g -lineinfo ${CMAKE_CUDA_FLAGS_RELEASE}" CACHE STRING "")
set(CMAKE_CUDA_FLAGS_DEBUG "-g -G -O0 -Xcompiler -O0" CACHE STRING "")

# Current version of hypre does not build with GPU support.
# Most recent version does build. Let's wait for an upgrade on our side.
set(ENABLE_HYPRE_DEVICE "CPU" CACHE STRING "" FORCE)
```

And then when I execute the python command `python scripts/config-build.py -hc ../GEOS/host-configs/plafrim/a100.cmake -bt Release`, I get:
```
thirdPartyLibs$ python scripts/config-build.py -hc ../GEOS/host-configs/plafrim/a100.cmake -bt Release
[config-build.py]:[INFO]: Using host config file: '/.projets_7/schnaps/GEOS/GEOS/host-configs/plafrim/a100.cmake'.
[config-build.py]:[INFO]: Build path is: build-a100-release
[config-build.py]:[INFO]: Build directory '/.projets_7/schnaps/GEOS/thirdPartyLibs/build-a100-release' already exists. Deleting...
[config-build.py]:[INFO]: Creating build directory '/.projets_7/schnaps/GEOS/thirdPartyLibs/build-a100-release'...
[config-build.py]:[INFO]: Install directory '/.projets_7/schnaps/GEOS/thirdPartyLibs/install-a100-release' already exists. Deleting...
[config-build.py]:[INFO]: Creating install path '/.projets_7/schnaps/GEOS/thirdPartyLibs/install-a100-release'...
[config-build.py]:[INFO]: Extracting cmake entry from host config file /.projets_7/schnaps/GEOS/GEOS/host-configs/plafrim/a100.cmake
[config-build.py]:[INFO]: Could not find a cmake entry in host config file. Using ${PATH}.
[config-build.py]:[INFO]: Changing to build directory '/.projets_7/schnaps/GEOS/thirdPartyLibs/build-a100-release'
[config-build.py]:[INFO]: Executing cmake line: 'cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=/.projets_7/schnaps/GEOS/thirdPartyLibs/install-a100-release -C/.projets_7/schnaps/GEOS/GEOS/host-configs/plafrim/a100.cmake /.projets_7/schnaps/GEOS/thirdPartyLibs'
loading initial cache file /.projets_7/schnaps/GEOS/GEOS/host-configs/plafrim/a100.cmake
-- The C compiler identification is GNU 11.2.0
-- The CXX compiler identification is GNU 11.2.0
-- The Fortran compiler identification is GNU 11.2.0
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working C compiler: /cm/shared/modules/amd/rome/compiler/gcc/11.2.0/bin/gcc - skipped
-- Detecting C compile features
-- Detecting C compile features - done
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /cm/shared/modules/amd/rome/compiler/gcc/11.2.0/bin/g++ - skipped
-- Detecting CXX compile features
-- Detecting Fortran compiler ABI info
-- Detecting Fortran compiler ABI info - done
-- Check for working Fortran compiler: /cm/shared/modules/amd/rome/compiler/gcc/11.2.0/bin/gfortran - skipped
-- BLT Version: 0.6.2
-- CMake Version: 3.27.0
-- CMake Build Type: Release
-- CMake Executable: /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/bin/cmake
-- Setting CMAKE_CXX_EXTENSIONS to OFF
-- Standard C++17 selected
-- Standard C++17 selected for CUDA
-- BLT MPI support is ON
-- BLT OpenMP support is ON
-- BLT CUDA support is ON
-- BLT HIP support is OFF
-- Creating BLT MPI targets...
-- FindMPI Enabled  (ENABLE_FIND_MPI == ON)
-- Could NOT find MPI_C (missing: MPI_C_WORKS)
-- Could NOT find MPI_CXX (missing: MPI_CXX_WORKS)
-- Could NOT find MPI_Fortran (missing: MPI_Fortran_WORKS)
CMake Error at /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/FindPackageHandleStandardArgs.cmake:230 (message):
  Could NOT find MPI (missing: MPI_C_FOUND MPI_CXX_FOUND MPI_Fortran_FOUND)
Call Stack (most recent call first):
  /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/FindPackageHandleStandardArgs.cmake:600 (_FPHSA_FAILURE_MESSAGE)
  /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/FindMPI.cmake:1837 (find_package_handle_standard_args)
  cmake/blt/cmake/thirdparty/BLTSetupMPI.cmake:43 (find_package)
  cmake/blt/cmake/BLTSetupTargets.cmake:75 (include)
  cmake/blt/cmake/SetupThirdParty.cmake:6 (include)
  cmake/blt/SetupBLT.cmake:129 (include)
  CMakeLists.txt:38 (include)

-- Configuring incomplete, errors occurred!
```

Any help will be welcome.
