### 2024-07-31 #6973 @berenger-eu
I would like to use GEOS on a platform with MPI+CUDA, I opened an issue because I have difficulties to do so https://github.com/GEOS-DEV/GEOS/issues/3256
