### 2024-07-31 GEOS-DEV/GEOS#3256 @victorapm
@berenger-eu thank you for your review!

Could you double check the following?

```
ls /cm/shared/modules/generic/mpi/openmpi/5.0.1/bin/mpicc
ls /cm/shared/modules/generic/mpi/openmpi/5.0.1/bin/mpicxx
ls /cm/shared/modules/generic/mpi/openmpi/5.0.1/bin/mpifort
```

In addition, could you attach the files `CMakeOutput.log` and `CMakeError.log`?

Thank you!
