### 2024-07-31 GEOS-DEV/GEOS#3256 @berenger-eu
That seems to exist:
```
sirocco22:thirdPartyLibs$ ls /cm/shared/modules/generic/mpi/openmpi/5.0.1/bin/mpicc
/cm/shared/modules/generic/mpi/openmpi/5.0.1/bin/mpicc
sirocco22:thirdPartyLibs$ ls /cm/shared/modules/generic/mpi/openmpi/5.0.1/bin/mpicxx
/cm/shared/modules/generic/mpi/openmpi/5.0.1/bin/mpicxx
sirocco22:thirdPartyLibs$ ls /cm/shared/modules/generic/mpi/openmpi/5.0.1/bin/mpifort
/cm/shared/modules/generic/mpi/openmpi/5.0.1/bin/mpifort
```
I did not find a CMakeError.log, but I have the file attached to this message.
[CMakeConfigureLog.txt](https://github.com/user-attachments/files/16442809/CMakeConfigureLog.txt)
I feel like the error is not on your side but on the MPI libs...
