### 2024-07-31 GEOS-DEV/GEOS#3256 @victorapm
Thank you for providing the log file.

Yes, this looks like an MPI installation issue and I would recommend checking with your HPC system admin:

```
        /usr/bin/ld: warning: libpsm2.so.2, needed by /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi_usempif08.so, not found (try using -rpath or -rpath-link)
        /usr/bin/ld: warning: libfabric.so.1, needed by /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi_usempif08.so, not found (try using -rpath or -rpath-link)

        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_ep_open@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « fi_freeinfo@FABRIC_1.1 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_init@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_isend2@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_error_get_string@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_init@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libopen-pal.so.80 : référence indéfinie vers « fi_version@FABRIC_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_get_capability_mask@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « fi_dupinfo@FABRIC_1.1 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_finalize@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_ep_connect@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_irecv2@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_ep_open_opts_get_defaults@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_send2@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « fi_strerror@FABRIC_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « fi_fabric@FABRIC_1.1 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_iprobe2@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « fi_getinfo@FABRIC_1.1 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libopen-pal.so.80 : référence indéfinie vers « fi_tostr@FABRIC_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_get_stats@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_error_defer@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_finalize@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_test2@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_improbe2@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_cancel@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_ipeek2@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_error_register_handler@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_test@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_mq_imrecv@PSM2_1.0 »
        /cm/shared/modules/generic/mpi/openmpi/5.0.1/lib/libmpi.so : référence indéfinie vers « psm2_ep_close@PSM2_1.0 »
        collect2: erreur: ld a retourné le statut de sortie 1
```
