### 2024-07-31 GEOS-DEV/GEOS#3256 @rrsettgast
Hello @berenger-eu ,
This reads as a failure of cmake/blt to find your MPI distribution. We have seen this in the past on cray systems...is this a cray system?

- What is the linux distro/version? `cat /etc/os-release`
- is there a default set of compilers and mpi installation?
