### 2024-07-31 GEOS-DEV/GEOS#3256 @timokoch
Hi, not sure if it's good to add this here but the error is similar and I'm also a JOSS reviewer. Same error but on a different system (macOS arm with OpenMPI 5.0.3). `CMake Error at /opt/homebrew/Cellar/cmake/3.26.0/share/cmake/Modules/FindPackageHandleStandardArgs.cmake:230 (message):
  Could NOT find MPI (missing: MPI_CXX_FOUND MPI_Fortran_FOUND) (found
  version "3.1")`

Note: The error seemed to have only appeared after installing openblas through homebrew. I usually use openblas on macOS through framework.Accelerate. Before `brew install openblas` I got past the error above but got an error from Kokkos which wasn't able to map framework.Accelerate on arm (different issue I guess). I'm not sure how this is related to not finding MPI now.

You find attached my host-config file and the CMakeOutput from `python3 scripts/config-build.py -hc ../host-configs/myhost.cmake -bt Release`. Since CMake 3.26 CMakeOutput.log and CMakeError.log are no longer emitted.

```
$ /opt/homebrew/Cellar/open-mpi/5.0.3_1/bin/mpicc --version
Apple clang version 14.0.3 (clang-1403.0.22.14.1)
Target: arm64-apple-darwin22.3.0
Thread model: posix
InstalledDir: /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin
```
```
$ /opt/homebrew/Cellar/open-mpi/5.0.3_1/bin/mpicc --version
Apple clang version 14.0.3 (clang-1403.0.22.14.1)
Target: arm64-apple-darwin22.3.0
Thread model: posix
InstalledDir: /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin
```
```
$ /opt/homebrew/Cellar/open-mpi/5.0.3_1/bin/mpifort --version
GNU Fortran (Homebrew GCC 14.1.0_2) 14.1.0
Copyright (C) 2024 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
```

[myhost.cmake.txt](https://github.com/user-attachments/files/16443006/myhost.cmake.txt) (added .txt to be able to upload to github)
[output.log](https://github.com/user-attachments/files/16443351/output.log)
[CMakeConfigureLog.txt](https://github.com/user-attachments/files/16443355/CMakeConfigureLog.txt)

I guess I would have to remove `-ld_classic` but not sure where and why.
Help is appreciated :). BTW, I use the same MPI installation in other projects without problem.
