### 2024-07-31 GEOS-DEV/GEOS#3256 @victorapm
@timokoch maybe this is relevant to you: https://discourse.cmake.org/t/appleclang-15-new-linker-general-issues/9106
