### 2024-07-31 GEOS-DEV/GEOS#3256 @rrsettgast
Hello @timokoch,
Sorry about the trouble. I was looking at the install on my local Mac, for which I use the unaltered `/host-configs/apple/macOS_arm.cmake` file and my home-brew installation. There are only a few things I notice:

- I am on a newer apple clang version 15, but that should not matter
- you have specified the full path to mpi executables rather than the top level bin dir
```
set(ENABLE_MPI ON CACHE PATH "")
set(MPI_C_COMPILER "${HOMEBREW_DIR}/bin/mpicc" CACHE PATH "")
set(MPI_CXX_COMPILER "${HOMEBREW_DIR}/bin/mpicxx" CACHE PATH "")
set(MPIEXEC "${HOMEBREW_DIR}/bin/mpirun" CACHE PATH "")
```
vs
```
set(MPI_C_COMPILER "${HOMEBREW_DIR}/Cellar/open-mpi/5.0.3_1/bin/mpicc" CACHE PATH "")
set(MPI_CXX_COMPILER "${HOMEBREW_DIR}Cellar/open-mpi/5.0.3_1/bin/mpicxx" CACHE PATH "")
set(MPI_FORTRAN_COMPILER "${HOMEBREW_DIR}Cellar/open-mpi/5.0.3_1/bin/mpifort" CACHE PATH "")
set(MPIEXEC "${HOMEBREW_DIR}Cellar/open-mpi/5.0.3_1/bin/mpirun" CACHE PATH "")
```
was this required for you?
- I don't understand why you had to add the creation of `${GEOS_TPL_DIR}`.
   What is the command that you are executing to configure the `thirdPartyLibs` build? it should be something like:
   `python3 scripts/config-build.py -hc ../GEOS/host-configs/apple/macOS_arm.cmake -bt Release`
   ah. The hostconfig assume a release build....that is why you needed to create`${GEOS_TPL_DIR}`.
- The blas/lapack issue is strange. I have both installed from home-brew, and I ended up just using the `lapack` path. Is `lapack` installed via home-brew? I think the directions are missing that.

I have to configure a travel laptop (MacBook) from scratch this afternoon. I will confirm the steps that work for that system.
