### 2024-07-31 GEOS-DEV/GEOS#3256 @timokoch
* I get the same error with the provided host config, i.e. just running `python3 scripts/config-build.py -hc ../GEOS/host-configs/apple/macOS_arm.cmake -bt Release` in the `thirdPartyLibs` repo as you suggested but then additionally I get

```
CONFIG_NAME = macOS_arm
in tpls.cmake GEOS_TPL_DIR=/Users/timokoch/tmp/geos/thirdPartyLibs/../../thirdPartyLibs/install-macOS_arm-release
CMake Warning at /Users/timokoch/tmp/geos/GEOS/host-configs/tpls.cmake:10 (message):
  'GEOS_TPL_DIR' does not exist.

Call Stack (most recent call first):
  /Users/timokoch/tmp/geos/GEOS/host-configs/apple/macOS_base.cmake:44 (include)
  /Users/timokoch/tmp/geos/GEOS/host-configs/apple/macOS_arm.cmake:4 (include)
```

* For blas/lapack, `brew install openblas` installs both lapack and blas libraries into /opt/homebrew/opt/openblas/ for me. I don't have the folder /opt/homebrew/opt/lapack/

* the apple clang 14 vs 15 might matter since apparently 14 doesn't understand -ld_classic while it seems to be recommended for 15, see @victorapm's link above. But I don't understand why/where the flag is added when using apple clang 14. I guess it could be some openmpi version incompatiblity with clang 14. However, I also don't understand why it tries to compile some fortran code if the host config explicitly disables fortran.
