### 2024-07-31 GEOS-DEV/GEOS#3256 @timokoch
My issue seems to be fixed in a recent open-mpi version (using brew install open-mpi --HEAD which gives 5.1.0a1). Possibly https://github.com/open-mpi/ompi/pull/12664 which fixes something related from 5.0.4. Stable homebrew version is 5.0.3 though which is unfortunate and the version that seems to be incompatible with apple clang 14. I'm pretty sure there is no issue with apple clang 15 as well from what I read.
