### 2024-07-31 GEOS-DEV/GEOS#3256 @rrsettgast
@timokoch I have openmpi 5.0.3 working on my primary laptop. I will do a vanilla install on a fresh system this afternoon and try to reproduce the errors.
