### 2024-07-31 GEOS-DEV/GEOS#3256 @rrsettgast
Hello @berenger-eu , @timokoch
I have created a branch/PR to address the problems you are having (#3258 ). I have also invited you to join the GEOS org so you can perform reviews on the PR. Hopefully we can figure this out without too much difficulty. Thanks!
