### 2024-08-01 GEOS-DEV/GEOS#3256 @rrsettgast
@berenger-eu what time zone are you in? Perhaps we can do a zoom call to get to a solution faster?
