### 2024-08-02 #6973 @timokoch
@rrsettgast The statement of need section is currently on the "shallow" side. CO2 injection scenarios (cited as the main goal but also "geothermal energy, hydrogen storage, and related subsurface applications") is of course a topic for many, with a large number of open source tools out there. You probably know of the [11th SPE CSP](https://www.spe.org/en/csp/) which has a large number of open source codes (and closed source codes) participating. The same goes for HPC multi-physics tools or coupling libraries. So I think it would be good to (a) mention other tools (at least the ones with a clear (partial) overlap of topics and capabilities, definitely tools like [OPM](https://opm-project.org/), [OpenGeoSys](https://www.opengeosys.org/), just to name two FOSS with a very clear overlap) and (b) highlight the specific strengths of GEOS in comparison to those tools (I guess you have both great HPC capabilities and some complex THCM models, I also like the extensive benchmark/verification suite, ...).
