### 2024-08-02 GEOS-DEV/GEOS#3256 @berenger-eu
So I was able to configure with CMake, but it seems that getting the dependencies does not work:
```
$ make -j 8
[  0%] Creating directories for 'raja'
[  0%] Creating directories for 'parmetis'
[  1%] Creating directories for 'scotch'
[  3%] Creating directories for 'uncrustify'
[  3%] Creating directories for 'hdf5'
[  5%] Creating directories for 'metis'
[  7%] Creating directories for 'mathpresso'
[  7%] Performing download step (verify and extract) for 'raja'
[  7%] Creating directories for 'pugixml'
[  8%] Performing download step (verify and extract) for 'parmetis'
CMake Warning at raja-stamp/verify-raja.cmake:15 (message):
  File will not be verified since no URL_HASH specified

[  9%] Performing download step (verify and extract) for 'hdf5'
[ 10%] Performing download step (verify and extract) for 'scotch'
CMake Warning at parmetis-stamp/verify-parmetis.cmake:15 (message):
  File will not be verified since no URL_HASH specified

[ 10%] Performing download step (verify and extract) for 'uncrustify'
-- extracting...
     src='/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/RAJA-v2024.02.2.tar.gz'
     dst='/.projets_7/schnaps/GEOS/thirdPartyLibs/build-a100-release/raja/src/raja'
[ 10%] Performing download step (verify and extract) for 'metis'
-- extracting... [tar xfz]
[ 11%] Performing download step (verify and extract) for 'pugixml'
CMake Warning at hdf5-stamp/verify-hdf5.cmake:15 (message):
  File will not be verified since no URL_HASH specified

[ 11%] Performing download step (verify and extract) for 'mathpresso'
CMake Warning at scotch-stamp/verify-scotch.cmake:15 (message):
  File will not be verified since no URL_HASH specified

-- extracting...
     src='/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/parmetis-4.0.3-IDXTYPEWIDTH_64.tar.gz'
     dst='/.projets_7/schnaps/GEOS/thirdPartyLibs/build-a100-release/parmetis/src/parmetis'

CMake Warning at uncrustify-stamp/verify-uncrustify.cmake:15 (message):
  File will not be verified since no URL_HASH specified

CMake Warning at metis-stamp/verify-metis.cmake:15 (message):
  File will not be verified since no URL_HASH specified

-- extracting... [tar xfz]
CMake Warning at pugixml-stamp/verify-pugixml.cmake:15 (message):
  File will not be verified since no URL_HASH specified

CMake Warning at mathpresso-stamp/verify-mathpresso.cmake:15 (message):
  File will not be verified since no URL_HASH specified

-- extracting...
     src='/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/hdf5-1.12.1.tar.gz'
     dst='/.projets_7/schnaps/GEOS/thirdPartyLibs/build-a100-release/hdf5/src/hdf5'
-- extracting...
     src='/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/scotch-v7.0.3.tar.gz'
     dst='/.projets_7/schnaps/GEOS/thirdPartyLibs/build-a100-release/scotch/src/scotch'
-- extracting...
     src='/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/uncrustify-401a409.zip'
     dst='/.projets_7/schnaps/GEOS/thirdPartyLibs/build-a100-release/uncrustify/src/uncrustify'
-- extracting... [tar xfz]
-- extracting...
     src='/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/parmetis-4.0.3-IDXTYPEWIDTH_64.tar.gz'
     dst='/.projets_7/schnaps/GEOS/thirdPartyLibs/build-a100-release/metis/src/metis'
-- extracting... [tar xfz]
-- extracting... [tar xfz]
-- extracting...
     src='/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/pugixml-1.13.tar.gz'
     dst='/.projets_7/schnaps/GEOS/thirdPartyLibs/build-a100-release/pugixml/src/pugixml'
-- extracting... [tar xfz]
-- extracting...
     src='/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/mathpresso-2015-12-15.tar.gz'
     dst='/.projets_7/schnaps/GEOS/thirdPartyLibs/build-a100-release/mathpresso/src/mathpresso'
-- extracting... [tar xfz]
-- extracting... [tar xfz]
CMake Error: Problem with archive_read_open_file(): Unrecognized archive format
CMake Error: Problem extracting tar: /.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/parmetis-4.0.3-IDXTYPEWIDTH_64.tar.gz
CMake Error: Problem with archive_read_open_file(): Unrecognized archive format
CMake Error: Problem extracting tar: /.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/parmetis-4.0.3-IDXTYPEWIDTH_64.tar.gz
-- extracting... [error clean up]
-- extracting... [error clean up]
CMake Error: Problem with archive_read_open_file(): Unrecognized archive format
CMake Error: Problem extracting tar: /.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/RAJA-v2024.02.2.tar.gz
-- extracting... [error clean up]
CMake Error at metis-stamp/extract-metis.cmake:40 (message):
  Extract of
  '/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/parmetis-4.0.3-IDXTYPEWIDTH_64.tar.gz'
  failed

CMake Error at parmetis-stamp/extract-parmetis.cmake:40 (message):
  Extract of
  '/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/parmetis-4.0.3-IDXTYPEWIDTH_64.tar.gz'
  failed

CMake Error: Problem with archive_read_open_file(): Unrecognized archive format
CMake Error: Problem extracting tar: /.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/hdf5-1.12.1.tar.gz
-- extracting... [error clean up]
make[2]: *** [parmetis/src/parmetis-stamp/parmetis-download] Erreur 1
make[2]: *** [metis/src/metis-stamp/metis-download] Erreur 1
CMake Error at raja-stamp/extract-raja.cmake:40 (message):
  Extract of
  '/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/RAJA-v2024.02.2.tar.gz'
  failed

make[1]: *** [CMakeFiles/parmetis.dir/all] Erreur 2
make[1]: *** Attente des tâches non terminées....
make[1]: *** [CMakeFiles/metis.dir/all] Erreur 2
make[2]: *** [raja/src/raja-stamp/raja-download] Erreur 1
make[1]: *** [CMakeFiles/raja.dir/all] Erreur 2
CMake Error at hdf5-stamp/extract-hdf5.cmake:40 (message):
  Extract of
  '/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/hdf5-1.12.1.tar.gz'
  failed

make[2]: *** [hdf5/src/hdf5-stamp/hdf5-download] Erreur 1
make[1]: *** [CMakeFiles/hdf5.dir/all] Erreur 2
CMake Error: Problem with archive_read_open_file(): Unrecognized archive format
CMake Error: Problem extracting tar: /.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/uncrustify-401a409.zip
-- extracting... [error clean up]
CMake Error: Problem with archive_read_open_file(): Unrecognized archive format
CMake Error: Problem extracting tar: /.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/scotch-v7.0.3.tar.gz
-- extracting... [error clean up]
CMake Error at uncrustify-stamp/extract-uncrustify.cmake:40 (message):
  Extract of
  '/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/uncrustify-401a409.zip'
  failed

CMake Error: Problem with archive_read_open_file(): Unrecognized archive format
CMake Error: Problem extracting tar: /.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/pugixml-1.13.tar.gz
make[2]: *** [uncrustify/src/uncrustify-stamp/uncrustify-download] Erreur 1
CMake Error at scotch-stamp/extract-scotch.cmake:40 (message):
  Extract of
  '/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/scotch-v7.0.3.tar.gz'
  failed

-- extracting... [error clean up]

make[1]: *** [CMakeFiles/uncrustify.dir/all] Erreur 2
make[2]: *** [scotch/src/scotch-stamp/scotch-download] Erreur 1
make[1]: *** [CMakeFiles/scotch.dir/all] Erreur 2
CMake Error at pugixml-stamp/extract-pugixml.cmake:40 (message):
  Extract of
  '/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/pugixml-1.13.tar.gz'
  failed

CMake Error: Problem with archive_read_open_file(): Unrecognized archive format
CMake Error: Problem extracting tar: /.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/mathpresso-2015-12-15.tar.gz
make[2]: *** [pugixml/src/pugixml-stamp/pugixml-download] Erreur 1
-- extracting... [error clean up]
make[1]: *** [CMakeFiles/pugixml.dir/all] Erreur 2
CMake Error at mathpresso-stamp/extract-mathpresso.cmake:40 (message):
  Extract of
  '/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/mathpresso-2015-12-15.tar.gz'
  failed

make[2]: *** [mathpresso/src/mathpresso-stamp/mathpresso-download] Erreur 1
make[1]: *** [CMakeFiles/mathpresso.dir/all] Erreur 2
make: *** [all] Erreur 2
```

The files seems to exist:
```
$ ls /.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/RAJA-v2024.02.2.tar.gz
/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/RAJA-v2024.02.2.tar.gz
sirocco25:build-a100-release$ ls /.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/mathpresso-2015-12-15.tar.gz
/.projets_7/schnaps/GEOS/thirdPartyLibs/tplMirror/mathpresso-2015-12-15.tar.gz
```
