### 2024-08-02 GEOS-DEV/GEOS#3256 @victorapm
@berenger-eu Could you run the following and try again? Note you need `git-lfs` in your system.
```
cd ${HOME}/.projets_7/schnaps/GEOS/thirdPartyLibs
git lfs install
git pull
git submodule init
git submodule update
```
