### 2024-08-05 GEOS-DEV/GEOS#3256 @berenger-eu
Right, lfs seems not to be present.
```
$ git lfs install
git: 'lfs' is not a git command. See 'git --help'.

Did you mean this?
	log
```
