### 2024-08-05 GEOS-DEV/GEOS#3256 @timokoch
@berenger-eu You actually have to install git-lfs (https://git-lfs.com/) which is a separate plugin to make this command work.
