### 2024-08-05 GEOS-DEV/GEOS#3256 @berenger-eu
Almost work, git lfs now works, but there are other issues.
```
$ git pull
remote: Enumerating objects: 44, done.
remote: Total 44 (delta 26), reused 39 (delta 26), pack-reused 0
Unpacking objects: 100% (44/44), done.
From https://github.com/GEOS-DEV/thirdPartyLibs
   0fd15c3..696a3ea  master     -> origin/master
 * [new branch]      feature/bubusch/test-omp -> origin/feature/bubusch/test-omp
 + b5432be...8db5145 feature/han12/docker_spack -> origin/feature/han12/docker_spack  (forced update)
First, rewinding head to replay your work on top of it...
Downloading tplMirror/hypre-v2.31.0-12-g06da35b1a.tar.gz (4.4 MB)
Error downloading object: tplMirror/hypre-v2.31.0-12-g06da35b1a.tar.gz (121ea0e): Smudge error: Error downloading tplMirror/hypre-v2.31.0-12-g06da35b1a.tar.gz (121ea0eebb170fc524de45979b92316627630e096ee878a6ab1b987849dd1d52): LFS: Get "https://github-cloud.githubusercontent.com/alambic/media/176071333/12/1e/121ea0eebb170fc524de45979b92316627630e096ee878a6ab1b987849dd1d52?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=AKIA5BA2674WPWWEFGQ5%2F20240805%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Date=20240805T100902Z&X-Amz-Expires=3600&X-Amz-Signature=7251469d06e8ed6194365a813f5603e704e99361bc9f5da25d319c591c0f11f8&X-Amz-SignedHeaders=host&actor_id=0&key_id=0&repo_id=132315668&token=1": Forbidden

Errors logged to '/.projets_7/schnaps/GEOS/thirdPartyLibs/.git/lfs/logs/20240805T120902.510837443.log'.
Use `git lfs logs last` to view the log.
error: external filter git-lfs smudge -- %f failed 2
error: external filter git-lfs smudge -- %f failed
fatal: tplMirror/hypre-v2.31.0-12-g06da35b1a.tar.gz: smudge filter lfs failed
could not detach HEAD
```

But for some dependencies it worked:
```
sirocco21:thirdPartyLibs$ git submodule init
sirocco21:thirdPartyLibs$ git submodule update
sirocco21:thirdPartyLibs$ cd build-a100-release/
sirocco21:build-a100-release$ make -j 3
-- BLT Version: 0.6.2
-- CMake Version: 3.27.0
-- CMake Build Type: Release
-- CMake Executable: /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/bin/cmake
-- Setting CMAKE_CXX_EXTENSIONS to OFF
-- Standard C++17 selected
-- Standard C++17 selected for CUDA
-- BLT MPI support is ON
-- BLT OpenMP support is ON
-- BLT CUDA support is ON
-- BLT HIP support is OFF
-- Creating BLT MPI targets...
-- FindMPI Enabled  (ENABLE_FIND_MPI == ON)
-- BLT MPI Compile Flags:  $<$<NOT:$<COMPILE_LANGUAGE:CUDA>>:-pthread>;$<$<COMPILE_LANGUAGE:CUDA>:-Xcompiler=-pthread>
-- BLT MPI Include Paths:  /cm/shared/modules/generic/mpi/openmpi/4.0.1/include;/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib
-- BLT MPI Libraries:      /cm/shared/modules/generic/mpi/openmpi/4.0.1/lib/libmpi.so;/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib/libmpi_usempif08.so;/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib/libmpi_usempi_ignore_tkr.so;/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib/libmpi_mpifh.so
-- BLT MPI Link Flags:     SHELL:-Wl,-rpath -Wl,/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib -Wl,--enable-new-dtags -pthread
-- MPI Executable:       /usr/bin/srun
-- MPI Num Proc Flag:    -n
-- MPI Command Append:
-- Using MPI Fortran header: mpif.h
-- Creating BLT OpenMP targets...
-- BLT OpenMP Compile Flags: $<$<NOT:$<COMPILE_LANGUAGE:CUDA>>:-fopenmp>;$<$<COMPILE_LANGUAGE:CUDA>:-Xcompiler=-fopenmp>
-- BLT OpenMP Link Flags:    -fopenmp
-- Creating BLT CUDA targets...
CMake Warning (dev) at cmake/blt/cmake/thirdparty/BLTSetupCUDA.cmake:101 (find_package):
  Policy CMP0146 is not set: The FindCUDA module is removed.  Run "cmake
  --help-policy CMP0146" for policy details.  Use the cmake_policy command to
  set the policy and suppress this warning.

Call Stack (most recent call first):
  cmake/blt/cmake/BLTSetupTargets.cmake:97 (include)
  cmake/blt/cmake/SetupThirdParty.cmake:6 (include)
  cmake/blt/SetupBLT.cmake:129 (include)
  CMakeLists.txt:38 (include)
This warning is for project developers.  Use -Wno-dev to suppress it.

-- CUDA Version:                   12.3
-- CUDA Toolkit Root Dir:          /usr/local/cuda-12.3
-- CUDA Compiler:                  /usr/local/cuda-12.3/bin/nvcc
-- CUDA Host Compiler:             g++
-- CUDA Include Path:              /usr/local/cuda-12.3/include
-- CUDA Libraries:                 /usr/local/cuda-12.3/lib64/libcudart_static.a;Threads::Threads;dl;/usr/lib64/librt.so
-- CUDA Compile Flags:             -restrict -arch sm_75 --expt-relaxed-constexpr --expt-extended-lambda -Werror cross-execution-space-call,reorder,deprecated-declarations
-- CUDA Link Flags:                -L/usr/local/cuda-12.3/lib64
-- BLT CUDA Separable Compilation: OFF
-- CUDA Separable Compilation:
-- CUDA Link with NVCC:
-- CUDA Implicit Link Libraries:   stdc++;m;gcc_s;gcc;c;gcc_s;gcc
-- CUDA Implicit Link Directories: /usr/local/cuda-12.3/targets/x86_64-linux/lib/stubs;/usr/local/cuda-12.3/targets/x86_64-linux/lib;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib/gcc/x86_64-pc-linux-gnu/11.2.0;/lib64;/usr/lib64;/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/lib/intel64;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl/lib/intel64;/usr/local/cuda-12.3/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib
-- Git Support is ON
-- Git Executable: /usr/bin/git
-- Git Version: 1.8.3.1
-- Sphinx support is ON
-- Failed to locate Sphinx executable (missing: SPHINX_EXECUTABLE)
-- Valgrind support is ON
-- AStyle support is ON
-- Failed to locate AStyle executable (missing: ASTYLE_EXECUTABLE)
-- ClangFormat support is ON
-- Uncrustify support is OFF
-- Yapf support is ON
-- Failed to locate Yapf executable (missing: YAPF_EXECUTABLE)
-- CMakeFormat support is ON
-- Failed to locate CMakeFormat executable (missing: CMAKEFORMAT_EXECUTABLE)
-- Cppcheck support is ON
-- Failed to locate Cppcheck executable (missing: CPPCHECK_EXECUTABLE)
-- ClangQuery support is ON
-- Failed to locate ClangQuery executable (missing: CLANGQUERY_EXECUTABLE)
-- ClangTidy support is ON
-- Failed to locate ClangTidy executable (missing: CLANGTIDY_EXECUTABLE)
-- ClangApplyReplacements support is ON
-- Failed to locate ClangApplyReplacements executable (missing: CLANGAPPLYREPLACEMENTS_EXECUTABLE)
-- C Compiler family is GNU
-- Fortran Compiler family is GNU
-- Adding optional BLT definitions and compiler flags
-- Enabling all compiler warnings on all targets.
-- Fortran support disabled.
-- CMAKE_C_FLAGS flags are:   -Wall -Wextra
-- CMAKE_CXX_FLAGS flags are:   -Wall -Wextra
-- CMAKE_EXE_LINKER_FLAGS flags are:
-- CMake's C implicit link directories: /cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib/gcc/x86_64-pc-linux-gnu/11.2.0;/lib64;/usr/lib64;/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/lib/intel64;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl/lib/intel64;/usr/local/cuda-12.3/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib
-- CMake's CXX implicit link directories: /cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib/gcc/x86_64-pc-linux-gnu/11.2.0;/lib64;/usr/lib64;/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/lib/intel64;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl/lib/intel64;/usr/local/cuda-12.3/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib
-- CMake's CUDA implicit link directories: /usr/local/cuda-12.3/targets/x86_64-linux/lib/stubs;/usr/local/cuda-12.3/targets/x86_64-linux/lib;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib/gcc/x86_64-pc-linux-gnu/11.2.0;/lib64;/usr/lib64;/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/lib/intel64;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl/lib/intel64;/usr/local/cuda-12.3/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib
-- CMake's C implicit link libraries: gcc;gcc_s;c;gcc;gcc_s
-- CMake's CXX implicit link libraries: stdc++;m;gcc_s;gcc;c;gcc_s;gcc
-- CMake's CUDA implicit link libraries: stdc++;m;gcc_s;gcc;c;gcc_s;gcc
-- Google Test Support is ON
-- Google Mock Support is OFF
Running with NUM_PROC = 64
-- Build command = make;-j;64
-- MKL found
MKL_INCLUDE_DIRS=/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//include
MKL_LIBRARIES=/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_intel_lp64.so;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_gnu_thread.so;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_core.so
-- BLAS_LIBRARIES = /cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_intel_lp64.so;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_gnu_thread.so;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_core.so
-- BLAS_LINK_LINE = -Wl,-rpath,/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl/lib/intel64 /cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_intel_lp64.so /cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_gnu_thread.so /cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_core.so
-- LAPACK_LIBRARIES = /cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_intel_lp64.so;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_gnu_thread.so;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_core.so
-- LAPACK_LINK_LINE = -Wl,-rpath,/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl/lib/intel64 /cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_intel_lp64.so /cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_gnu_thread.so /cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_core.so
-- Building uncrustify found at /projets/schnaps/GEOS/thirdPartyLibs/tplMirror/uncrustify-401a409.zip
CMake Warning (dev) at /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:3136 (message):
  The DOWNLOAD_EXTRACT_TIMESTAMP option was not given and policy CMP0135 is
  not set.  The policy's OLD behavior will be used.  When using a URL
  download, the timestamps of extracted files should preferably be that of
  the time of extraction, otherwise code that depends on the extracted
  contents might not be rebuilt if the URL changes.  The OLD behavior
  preserves the timestamps from the archive instead, but this is usually not
  what you want.  Update your project to the NEW behavior or specify the
  DOWNLOAD_EXTRACT_TIMESTAMP option with a value of true to avoid this
  robustness issue.
Call Stack (most recent call first):
  /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:4345 (_ep_add_download_command)
  CMakeLists.txt:134 (ExternalProject_Add)
This warning is for project developers.  Use -Wno-dev to suppress it.

-- Building HDF5 found at /projets/schnaps/GEOS/thirdPartyLibs/tplMirror/hdf5-1.12.1.tar.gz
CMake Warning (dev) at /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:3136 (message):
  The DOWNLOAD_EXTRACT_TIMESTAMP option was not given and policy CMP0135 is
  not set.  The policy's OLD behavior will be used.  When using a URL
  download, the timestamps of extracted files should preferably be that of
  the time of extraction, otherwise code that depends on the extracted
  contents might not be rebuilt if the URL changes.  The OLD behavior
  preserves the timestamps from the archive instead, but this is usually not
  what you want.  Update your project to the NEW behavior or specify the
  DOWNLOAD_EXTRACT_TIMESTAMP option with a value of true to avoid this
  robustness issue.
Call Stack (most recent call first):
  /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:4345 (_ep_add_download_command)
  CMakeLists.txt:203 (ExternalProject_Add)
This warning is for project developers.  Use -Wno-dev to suppress it.

-- Building Conduit found at /projets/schnaps/GEOS/thirdPartyLibs/tplMirror/conduit-0.9.2.tar.gz
CMake Warning (dev) at /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:3136 (message):
  The DOWNLOAD_EXTRACT_TIMESTAMP option was not given and policy CMP0135 is
  not set.  The policy's OLD behavior will be used.  When using a URL
  download, the timestamps of extracted files should preferably be that of
  the time of extraction, otherwise code that depends on the extracted
  contents might not be rebuilt if the URL changes.  The OLD behavior
  preserves the timestamps from the archive instead, but this is usually not
  what you want.  Update your project to the NEW behavior or specify the
  DOWNLOAD_EXTRACT_TIMESTAMP option with a value of true to avoid this
  robustness issue.
Call Stack (most recent call first):
  /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:4345 (_ep_add_download_command)
  CMakeLists.txt:247 (ExternalProject_Add)
This warning is for project developers.  Use -Wno-dev to suppress it.

-- Building SILO found at /projets/schnaps/GEOS/thirdPartyLibs/tplMirror/silo-4.11-bsd.tgz
CMake Warning (dev) at /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:3136 (message):
  The DOWNLOAD_EXTRACT_TIMESTAMP option was not given and policy CMP0135 is
  not set.  The policy's OLD behavior will be used.  When using a URL
  download, the timestamps of extracted files should preferably be that of
  the time of extraction, otherwise code that depends on the extracted
  contents might not be rebuilt if the URL changes.  The OLD behavior
  preserves the timestamps from the archive instead, but this is usually not
  what you want.  Update your project to the NEW behavior or specify the
  DOWNLOAD_EXTRACT_TIMESTAMP option with a value of true to avoid this
  robustness issue.
Call Stack (most recent call first):
  /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:4345 (_ep_add_download_command)
  CMakeLists.txt:316 (ExternalProject_Add)
This warning is for project developers.  Use -Wno-dev to suppress it.

-- Building RAJA found at /projets/schnaps/GEOS/thirdPartyLibs/tplMirror/RAJA-v2024.02.2.tar.gz
CMake Warning (dev) at /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:3136 (message):
  The DOWNLOAD_EXTRACT_TIMESTAMP option was not given and policy CMP0135 is
  not set.  The policy's OLD behavior will be used.  When using a URL
  download, the timestamps of extracted files should preferably be that of
  the time of extraction, otherwise code that depends on the extracted
  contents might not be rebuilt if the URL changes.  The OLD behavior
  preserves the timestamps from the archive instead, but this is usually not
  what you want.  Update your project to the NEW behavior or specify the
  DOWNLOAD_EXTRACT_TIMESTAMP option with a value of true to avoid this
  robustness issue.
Call Stack (most recent call first):
  /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:4345 (_ep_add_download_command)
  CMakeLists.txt:340 (ExternalProject_Add)
This warning is for project developers.  Use -Wno-dev to suppress it.

-- Building CHAI found at /projets/schnaps/GEOS/thirdPartyLibs/tplMirror/chai-2024.02.2.tar.gz
CMake Warning (dev) at /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:3136 (message):
  The DOWNLOAD_EXTRACT_TIMESTAMP option was not given and policy CMP0135 is
  not set.  The policy's OLD behavior will be used.  When using a URL
  download, the timestamps of extracted files should preferably be that of
  the time of extraction, otherwise code that depends on the extracted
  contents might not be rebuilt if the URL changes.  The OLD behavior
  preserves the timestamps from the archive instead, but this is usually not
  what you want.  Update your project to the NEW behavior or specify the
  DOWNLOAD_EXTRACT_TIMESTAMP option with a value of true to avoid this
  robustness issue.
Call Stack (most recent call first):
  /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:4345 (_ep_add_download_command)
  CMakeLists.txt:376 (ExternalProject_Add)
This warning is for project developers.  Use -Wno-dev to suppress it.

-- Building MATHPRESSO found at /projets/schnaps/GEOS/thirdPartyLibs/tplMirror/mathpresso-2015-12-15.tar.gz
CMake Warning (dev) at /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:3136 (message):
  The DOWNLOAD_EXTRACT_TIMESTAMP option was not given and policy CMP0135 is
  not set.  The policy's OLD behavior will be used.  When using a URL
  download, the timestamps of extracted files should preferably be that of
  the time of extraction, otherwise code that depends on the extracted
  contents might not be rebuilt if the URL changes.  The OLD behavior
  preserves the timestamps from the archive instead, but this is usually not
  what you want.  Update your project to the NEW behavior or specify the
  DOWNLOAD_EXTRACT_TIMESTAMP option with a value of true to avoid this
  robustness issue.
Call Stack (most recent call first):
  /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:4345 (_ep_add_download_command)
  CMakeLists.txt:503 (ExternalProject_Add)
This warning is for project developers.  Use -Wno-dev to suppress it.

-- Building PUGIXML found at /projets/schnaps/GEOS/thirdPartyLibs/tplMirror/pugixml-1.13.tar.gz
CMake Warning (dev) at /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:3136 (message):
  The DOWNLOAD_EXTRACT_TIMESTAMP option was not given and policy CMP0135 is
  not set.  The policy's OLD behavior will be used.  When using a URL
  download, the timestamps of extracted files should preferably be that of
  the time of extraction, otherwise code that depends on the extracted
  contents might not be rebuilt if the URL changes.  The OLD behavior
  preserves the timestamps from the archive instead, but this is usually not
  what you want.  Update your project to the NEW behavior or specify the
  DOWNLOAD_EXTRACT_TIMESTAMP option with a value of true to avoid this
  robustness issue.
Call Stack (most recent call first):
  /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:4345 (_ep_add_download_command)
  CMakeLists.txt:541 (ExternalProject_Add)
This warning is for project developers.  Use -Wno-dev to suppress it.

-- Building PARMETIS found at /projets/schnaps/GEOS/thirdPartyLibs/tplMirror/parmetis-4.0.3-IDXTYPEWIDTH_64.tar.gz
CMake Warning (dev) at /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/share/cmake-3.27/Modules/ExternalProject.cmake:3136 (message):
  The DOWNLOAD_EXTRACT_TIMESTAMP option was not given and policy CMP0135 is
  not set.  The policy's OLD behavior will be used.  When using a URL
  download, the timestamps of extracted files should preferably be that of
  the time of extraction, otherwise code that depends on the extracted
  contents might not be rebuilt if the URL changes.  The OLD behavior
  preserves the timestamps from the archive instead, but this is usually not
  what you want.  Update your project to the NEW behavior or specify the
  DOWNLOAD_EXTRACT_TIMESTAMP option with a value of true to avoid this
[…]
