### 2024-08-05 GEOS-DEV/GEOS#3256 @victorapm
@berenger-eu can you retry the download step?

```
git lfs fetch --all
git lfs checkout
```

If it doesn't work, can you attach the error log? `git lfs logs last`
