### 2024-08-05 GEOS-DEV/GEOS#3256 @berenger-eu
Yes it seems the archives are incomplete:
```
$ ls -al
total 13144
drwxr-xr-x  2 bramas camus     4096  5 août  12:08 .
drwxr-xr-x 11 bramas camus     4096  5 août  12:08 ..
-rw-r--r--  1 bramas camus      132 31 juil. 09:12 adiak-0.4.0.tar.gz
-rw-r--r--  1 bramas camus      131 31 juil. 09:12 asmjit-2016-07-20.tar.gz
-rw-r--r--  1 bramas camus      132 31 juil. 09:12 Caliper-2.11.0.tar.gz
-rw-r--r--  1 bramas camus      133 31 juil. 09:12 chai-2024.02.2.tar.gz
-rw-r--r--  1 bramas camus      133 31 juil. 09:12 conduit-0.9.2.tar.gz
-rw-r--r--  1 bramas camus      132 31 juil. 09:12 doxygen-1.8.20.tar.gz
-rw-r--r--  1 bramas camus      131 31 juil. 09:12 fmt-11.0.1.tar.gz
-rw-r--r--  1 bramas camus      133 31 juil. 09:12 hdf5-1.12.1.tar.gz
-rw-r--r--  1 bramas camus      130 31 juil. 09:12 mathpresso-2015-12-15.tar.gz
-rw-r--r--  1 bramas camus      132 31 juil. 09:12 parmetis-4.0.3-IDXTYPEWIDTH_64.tar.gz
-rw-r--r--  1 bramas camus      133 31 juil. 09:12 petsc-3.19.4.tar.gz
-rw-r--r--  1 bramas camus      131 31 juil. 09:12 pugixml-1.13.tar.gz
-rw-r--r--  1 bramas camus      133 31 juil. 09:12 RAJA-v2024.02.2.tar.gz
-rw-r--r--  1 bramas camus      132 31 juil. 09:12 scotch-v7.0.3.tar.gz
-rw-r--r--  1 bramas camus 13298515 31 juil. 09:12 silo-4.11-bsd.tgz
-rw-r--r--  1 bramas camus     1868 31 juil. 09:12 silo-4.11.patch
-rw-r--r--  1 bramas camus      133 31 juil. 09:12 SuiteSparse-5.10.1.tar.gz
-rw-r--r--  1 bramas camus      132 31 juil. 09:12 superlu_dist-0f6efc3.tar.gz
-rw-r--r--  1 bramas camus      132 31 juil. 09:12 superlu_dist-9.0.0.tar.gz
-rw-r--r--  1 bramas camus      134 31 juil. 09:12 trilinos-release-15-1-1.tar.gz
-rw-r--r--  1 bramas camus      464 31 juil. 09:12 umpire_fcompare-debug-second.patch
-rw-r--r--  1 bramas camus      132 31 juil. 09:12 uncrustify-401a409.zip
-rw-r--r--  1 bramas camus      133 31 juil. 09:12 VTK-9.2.6.tar.gz
-rw-r--r--  1 bramas camus      630 31 juil. 09:12 vtkXMLReader-fpe.patch
```
git lfs seems to freeze.
```
$ git lfs fetch --all
fetch: 135 objects found, done.
fetch: Fetching all references...
Downloading LFS objects:   0% (0/135), 0 B | 0 B/s
```
Where are those files hosted ? It could be that it is forbidden on the supercomputer I am using.
Maybe there are direct links somewhere ?
