### 2024-08-05 GEOS-DEV/GEOS#3256 @victorapm
They are hosted on github cloud on dedicated LFS servers. You can find direct links to each package here: https://github.com/GEOS-DEV/thirdPartyLibs/blob/master/tplMirror
