### 2024-08-05 GEOS-DEV/GEOS#3256 @rrsettgast
Hello @berenger-eu.
There is ample storage and bandwidth allocations remaining on our GitHub LFS account so that isn't the problem. I have seen this issue before with older versions of git/git-lfs. Can you try the following things and report back the results?:
```
git --version
git-lfs --version
git lfs pull
```

The use of "git lfs pull" is documented, but perhaps more emphasis should be placed on it:
https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/QuickStart.html#download
```
Additional Notes:

#. git-lfs may not function properly (or may be very slow) if your version of git and git-lfs are not current. If you are using an older version, you may need to add git lfs pull after git pull in the above procedures.
```
