### 2024-08-12 GEOS-DEV/GEOS#3256 @rrsettgast
Hello. There should be an install directory under the tpl directory. For instance, on my Mac, the install is at:

```
/Users/settgast1/Codes/geos/thirdPartyLibs/install-macOS_arm-release
```
