### 2024-08-13 GEOS-DEV/GEOS#3256 @berenger-eu
Everything is installed correctly for third party:
```
sirocco22:GEOS$ ls thirdPartyLibs
build-a100-release  cmake  CMakeLists.txt  docker  host-configs  install-a100-release  scripts  tpl.cpp  tplMirror
sirocco22:GEOS$ ls thirdPartyLibs/build-a100-release/
asmjit-master  chai            CMakeFiles             conduit                fmt   Makefile    parmetis  scotch               suitesparse   tests
bin            CMakeCache.txt  cmake_install.cmake    CTestTestfile.cmake    hdf5  mathpresso  pugixml   scotch_makefile.inc  superlu_dist  uncrustify
blt            cmake_cmd       compile_commands.json  DartConfiguration.tcl  lib   metis       raja      silo                 Testing       vtk
```
It is the build of GEOS that fails.
