### 2024-08-13 GEOS-DEV/GEOS#3256 @victorapm
@berenger-eu, can you try building GEOS with the following command?

```
python scripts/config-build.py -hc host-configs/your-platform.cmake -bt Release -DGEOS_TPL_DIR=/projets/schnaps/GEOS/thirdPartyLibs/install-a100-release
```
