### 2024-08-13 GEOS-DEV/GEOS#3256 @berenger-eu
Seems not to work.
```
sirocco22:GEOS$ rm -r ./build-a100-release/
sirocco22:GEOS$ python scripts/config-build.py -hc host-configs/plafrim/a100.cmake -bt Release -DGEOS_TPL_DIR=/projets/schnaps/GEOS/thirdPartyLibs/install-a100-release
[config-build.py]:[INFO]: Passing the following unknown arguments directly to cmake: ['-DGEOS_TPL_DIR=/projets/schnaps/GEOS/thirdPartyLibs/install-a100-release']
[config-build.py]:[INFO]: Using host config file: '/.projets_7/schnaps/GEOS/GEOS/host-configs/plafrim/a100.cmake'.
[config-build.py]:[INFO]: Build path is: build-a100-release
[config-build.py]:[INFO]: Creating build directory '/.projets_7/schnaps/GEOS/GEOS/build-a100-release'...
[config-build.py]:[INFO]: Extracting cmake entry from host config file /.projets_7/schnaps/GEOS/GEOS/host-configs/plafrim/a100.cmake
[config-build.py]:[INFO]: Could not find a cmake entry in host config file. Using ${PATH}.
[config-build.py]:[INFO]: Changing to build directory '/.projets_7/schnaps/GEOS/GEOS/build-a100-release'
[config-build.py]:[INFO]: Executing cmake line: 'cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=/.projets_7/schnaps/GEOS/GEOS/install-a100-release -DGEOS_TPL_DIR=/projets/schnaps/GEOS/thirdPartyLibs/install-a100-release -C/.projets_7/schnaps/GEOS/GEOS/host-configs/plafrim/a100.cmake /.projets_7/schnaps/GEOS/GEOS/src'
loading initial cache file /.projets_7/schnaps/GEOS/GEOS/host-configs/plafrim/a100.cmake
CMake Deprecation Warning at CMakeLists.txt:6 (cmake_policy):
  The OLD behavior for policy CMP0104 will be removed from a future version
  of CMake.

  The cmake-policies(7) manual explains that the OLD behaviors of all
  policies are deprecated and that a policy should be set to OLD only under
  specific short-term circumstances.  Projects should be ported to the NEW
  behavior and not rely on setting a policy to OLD.

-- The C compiler identification is GNU 11.2.0
-- The CXX compiler identification is GNU 11.2.0
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working C compiler: /cm/shared/modules/amd/rome/compiler/gcc/11.2.0/bin/gcc - skipped
-- Detecting C compile features
-- Detecting C compile features - done
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /cm/shared/modules/amd/rome/compiler/gcc/11.2.0/bin/g++ - skipped
-- Detecting CXX compile features
-- BLT Version: 0.6.2
-- CMake Version: 3.27.0
-- CMake Build Type: Release
-- CMake Executable: /cm/shared/dev/modules/generic/apps/build/cmake/3.27.0/bin/cmake
-- Setting CMAKE_CXX_EXTENSIONS to OFF
-- Standard C++17 selected
-- Standard C++17 selected for CUDA
-- BLT MPI support is ON
-- BLT OpenMP support is ON
-- BLT CUDA support is ON
-- BLT HIP support is OFF
-- Creating BLT MPI targets...
-- FindMPI Enabled  (ENABLE_FIND_MPI == ON)
-- Found MPI_C: /cm/shared/modules/generic/mpi/openmpi/4.0.1/lib/libmpi.so (found version "3.1")
-- Found MPI_CXX: /cm/shared/modules/generic/mpi/openmpi/4.0.1/lib/libmpi.so (found version "3.1")
-- Found MPI: TRUE (found version "3.1")
-- BLT MPI Compile Flags:  $<$<NOT:$<COMPILE_LANGUAGE:CUDA>>:-pthread>;$<$<COMPILE_LANGUAGE:CUDA>:-Xcompiler=-pthread>
-- BLT MPI Include Paths:  /cm/shared/modules/generic/mpi/openmpi/4.0.1/include
-- BLT MPI Libraries:      /cm/shared/modules/generic/mpi/openmpi/4.0.1/lib/libmpi.so
-- BLT MPI Link Flags:     SHELL:-Wl,-rpath -Wl,/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib -Wl,--enable-new-dtags -pthread
-- MPI Executable:       /usr/bin/srun
-- MPI Num Proc Flag:    -n
-- MPI Command Append:
-- Creating BLT OpenMP targets...
-- Found OpenMP_C: -fopenmp (found version "4.5")
-- Found OpenMP_CXX: -fopenmp (found version "4.5")
-- Found OpenMP: TRUE (found version "4.5")
-- BLT OpenMP Compile Flags: $<$<NOT:$<COMPILE_LANGUAGE:CUDA>>:-fopenmp>;$<$<COMPILE_LANGUAGE:CUDA>:-Xcompiler=-fopenmp>
-- BLT OpenMP Link Flags:    -fopenmp
-- Creating BLT CUDA targets...
-- The CUDA compiler identification is NVIDIA 12.3.107
-- Detecting CUDA compiler ABI info
-- Detecting CUDA compiler ABI info - done
-- Check for working CUDA compiler: /usr/local/cuda-12.3/bin/nvcc - skipped
-- Detecting CUDA compile features
-- Detecting CUDA compile features - done
CMake Warning (dev) at cmake/blt/cmake/thirdparty/BLTSetupCUDA.cmake:101 (find_package):
  Policy CMP0146 is not set: The FindCUDA module is removed.  Run "cmake
  --help-policy CMP0146" for policy details.  Use the cmake_policy command to
  set the policy and suppress this warning.

Call Stack (most recent call first):
  cmake/blt/cmake/BLTSetupTargets.cmake:97 (include)
  cmake/blt/cmake/SetupThirdParty.cmake:6 (include)
  cmake/blt/SetupBLT.cmake:129 (include)
  CMakeLists.txt:32 (include)
This warning is for project developers.  Use -Wno-dev to suppress it.

-- Performing Test CMAKE_HAVE_LIBC_PTHREAD
-- Performing Test CMAKE_HAVE_LIBC_PTHREAD - Failed
-- Looking for pthread_create in pthreads
-- Looking for pthread_create in pthreads - not found
-- Looking for pthread_create in pthread
-- Looking for pthread_create in pthread - found
-- Found Threads: TRUE
-- Found CUDA: /usr/local/cuda-12.3 (found version "12.3")
-- CUDA Version:                   12.3
-- CUDA Toolkit Root Dir:          /usr/local/cuda-12.3
-- CUDA Compiler:                  /usr/local/cuda-12.3/bin/nvcc
-- CUDA Host Compiler:             g++
-- CUDA Include Path:              /usr/local/cuda-12.3/include
-- CUDA Libraries:                 /usr/local/cuda-12.3/lib64/libcudart_static.a;Threads::Threads;dl;/usr/lib64/librt.so
-- CUDA Compile Flags:             -restrict -arch sm_75 --expt-relaxed-constexpr --expt-extended-lambda -Werror cross-execution-space-call,reorder,deprecated-declarations
-- CUDA Link Flags:                -L/usr/local/cuda-12.3/lib64
-- BLT CUDA Separable Compilation: OFF
-- CUDA Separable Compilation:
-- CUDA Link with NVCC:
-- CUDA Implicit Link Libraries:   stdc++;m;gcc_s;gcc;c;gcc_s;gcc
-- CUDA Implicit Link Directories: /usr/local/cuda-12.3/targets/x86_64-linux/lib/stubs;/usr/local/cuda-12.3/targets/x86_64-linux/lib;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib/gcc/x86_64-pc-linux-gnu/11.2.0;/lib64;/usr/lib64;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/lib/intel64;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl/lib/intel64;/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib;/usr/local/cuda-12.3/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib
-- Found Git: /usr/bin/git (found version "1.8.3.1")
-- Git Support is ON
-- Git Executable: /usr/bin/git
-- Git Version: 1.8.3.1
-- Sphinx support is ON
-- Failed to locate Sphinx executable (missing: SPHINX_EXECUTABLE)
-- Valgrind support is ON
-- Found Valgrind: /usr/bin/valgrind
-- AStyle support is ON
-- Failed to locate AStyle executable (missing: ASTYLE_EXECUTABLE)
-- ClangFormat support is ON
-- Found ClangFormat: /usr/bin/clang-format
-- Uncrustify support is OFF
-- Yapf support is ON
-- Failed to locate Yapf executable (missing: YAPF_EXECUTABLE)
-- CMakeFormat support is ON
-- Failed to locate CMakeFormat executable (missing: CMAKEFORMAT_EXECUTABLE)
-- Cppcheck support is ON
-- Failed to locate Cppcheck executable (missing: CPPCHECK_EXECUTABLE)
-- ClangQuery support is ON
-- Failed to locate ClangQuery executable (missing: CLANGQUERY_EXECUTABLE)
-- ClangTidy support is ON
-- Failed to locate ClangTidy executable (missing: CLANGTIDY_EXECUTABLE)
-- ClangApplyReplacements support is ON
-- Failed to locate ClangApplyReplacements executable (missing: CLANGAPPLYREPLACEMENTS_EXECUTABLE)
-- C Compiler family is GNU
-- Adding optional BLT definitions and compiler flags
-- Enabling all compiler warnings on all targets.
-- Enabling treatment of warnings as errors on all targets.
-- Fortran support disabled.
-- CMAKE_C_FLAGS flags are:   -Wall -Wextra    -Werror
-- CMAKE_CXX_FLAGS flags are:   -Wall -Wextra       -Werror
-- CMAKE_EXE_LINKER_FLAGS flags are:
-- CMake's C implicit link directories: /cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib/gcc/x86_64-pc-linux-gnu/11.2.0;/lib64;/usr/lib64;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/lib/intel64;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl/lib/intel64;/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib;/usr/local/cuda-12.3/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib
-- CMake's CXX implicit link directories: /cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib/gcc/x86_64-pc-linux-gnu/11.2.0;/lib64;/usr/lib64;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/lib/intel64;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl/lib/intel64;/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib;/usr/local/cuda-12.3/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib
-- CMake's CUDA implicit link directories: /usr/local/cuda-12.3/targets/x86_64-linux/lib/stubs;/usr/local/cuda-12.3/targets/x86_64-linux/lib;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib/gcc/x86_64-pc-linux-gnu/11.2.0;/lib64;/usr/lib64;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/lib/intel64;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl/lib/intel64;/cm/shared/modules/generic/mpi/openmpi/4.0.1/lib;/usr/local/cuda-12.3/lib64;/cm/shared/modules/amd/rome/compiler/gcc/11.2.0/lib
-- CMake's C implicit link libraries: gcc;gcc_s;c;gcc;gcc_s
-- CMake's CXX implicit link libraries: stdc++;m;gcc_s;gcc;c;gcc_s;gcc
-- CMake's CUDA implicit link libraries: stdc++;m;gcc_s;gcc;c;gcc_s;gcc
-- Google Test Support is ON
-- Google Mock Support is OFF
-- Google Benchmark Support is ON
-- Failed to find LLVM FileCheck
-- git version: v0.6.2 normalized to 0.6.2
-- Version: 0.6.2
-- Looking for shm_open in rt
-- Looking for shm_open in rt - found
-- Performing Test HAVE_CXX_FLAG_WALL
-- Performing Test HAVE_CXX_FLAG_WALL - Success
-- Performing Test HAVE_CXX_FLAG_WEXTRA
-- Performing Test HAVE_CXX_FLAG_WEXTRA - Success
-- Performing Test HAVE_CXX_FLAG_WSHADOW
-- Performing Test HAVE_CXX_FLAG_WSHADOW - Success
-- Performing Test HAVE_CXX_FLAG_WFLOAT_EQUAL
-- Performing Test HAVE_CXX_FLAG_WFLOAT_EQUAL - Success
-- Performing Test HAVE_CXX_FLAG_WERROR
-- Performing Test HAVE_CXX_FLAG_WERROR - Success
-- Performing Test HAVE_CXX_FLAG_WSUGGEST_OVERRIDE
-- Performing Test HAVE_CXX_FLAG_WSUGGEST_OVERRIDE - Success
-- Performing Test HAVE_CXX_FLAG_PEDANTIC
-- Performing Test HAVE_CXX_FLAG_PEDANTIC - Success
-- Performing Test HAVE_CXX_FLAG_PEDANTIC_ERRORS
-- Performing Test HAVE_CXX_FLAG_PEDANTIC_ERRORS - Success
-- Performing Test HAVE_CXX_FLAG_WSHORTEN_64_TO_32
-- Performing Test HAVE_CXX_FLAG_WSHORTEN_64_TO_32 - Failed
-- Performing Test HAVE_CXX_FLAG_FSTRICT_ALIASING
-- Performing Test HAVE_CXX_FLAG_FSTRICT_ALIASING - Success
-- Performing Test HAVE_CXX_FLAG_WNO_DEPRECATED_DECLARATIONS
-- Performing Test HAVE_CXX_FLAG_WNO_DEPRECATED_DECLARATIONS - Success
-- Performing Test HAVE_CXX_FLAG_WNO_DEPRECATED
-- Performing Test HAVE_CXX_FLAG_WNO_DEPRECATED - Success
-- Performing Test HAVE_CXX_FLAG_WSTRICT_ALIASING
-- Performing Test HAVE_CXX_FLAG_WSTRICT_ALIASING - Success
-- Performing Test HAVE_CXX_FLAG_WD654
-- Performing Test HAVE_CXX_FLAG_WD654 - Failed
-- Performing Test HAVE_CXX_FLAG_WTHREAD_SAFETY
-- Performing Test HAVE_CXX_FLAG_WTHREAD_SAFETY - Failed
-- Performing Test HAVE_CXX_FLAG_COVERAGE
-- Performing Test HAVE_CXX_FLAG_COVERAGE - Success
-- Compiling and running to test HAVE_STD_REGEX
-- Performing Test HAVE_STD_REGEX -- success
-- Compiling and running to test HAVE_GNU_POSIX_REGEX
-- Performing Test HAVE_GNU_POSIX_REGEX -- failed to compile
-- Compiling and running to test HAVE_POSIX_REGEX
-- Performing Test HAVE_POSIX_REGEX -- success
-- Compiling and running to test HAVE_STEADY_CLOCK
-- Performing Test HAVE_STEADY_CLOCK -- success
-- Check if compiler accepts -pthread
-- Check if compiler accepts -pthread - yes
-- Compiling and running to test HAVE_PTHREAD_AFFINITY
-- Performing Test HAVE_PTHREAD_AFFINITY -- failed to compile

Processing GeosxOptions.cmake
CMAKE_HOST_SYSTEM_NAME = Linux
CMAKE_SYSTEM_NAME = Linux
CMAKE_HOST_APPLE =
-- GEOS_LA_INTERFACE = Hypre
CMAKE_CXX_COMPILER_ID = GNU
-- Performing Test CXX_UNUSED_BUT_SET_VAR
-- Performing Test CXX_UNUSED_BUT_SET_VAR - Success
localIndex is an alias for int
globalIndex is an alias for long long int
GEOS_LOCALINDEX_TYPE_FLAG = 0
GEOS_GLOBALINDEX_TYPE_FLAG = 2
CMAKE_CXX_FLAGS =  -Wall -Wextra       -Werror  -fopenmp  -Wpedantic -pedantic-errors -Wshadow -Wfloat-equal -Wcast-align -Wcast-qual
GEOS_LINK_PREPEND_FLAG=-Wl,--whole-archive
GEOS_LINK_POSTPEND_FLAG=-Wl,--no-whole-archive
Leaving GeosxOptions.cmake

-- MKL found
-- BLAS_LIBRARIES = /cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_intel_lp64.so;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_gnu_thread.so;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_core.so
-- LAPACK_LIBRARIES = /cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_intel_lp64.so;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_gnu_thread.so;/cm/shared/modules/amd/rome/compiler/intel/2020_update4/mkl//lib/intel64/libmkl_core.so
-- Using Intel MKL
CMake Error at cmake/thirdparty/SetupGeosxThirdParty.cmake:118 (message):
  GEOSX requires Conduit, either :

    - Verify that you provided a valid TPL installation directory (GEOS_TPL_DIR = "/projets/schnaps/GEOS/thirdPartyLibs/install-a100-release"),
    - Or set CONDUIT_DIR to the Conduit installation directory (CONDUIT_DIR = "").

Call Stack (most recent call first):
  cmake/thirdparty/SetupGeosxThirdParty.cmake:210 (mandatory_tpl_doesnt_exist)
  cmake/CMakeBasics.cmake:49 (include)
  CMakeLists.txt:72 (include)

-- Configuring incomplete, errors occurred!
```
