### 2024-08-13 GEOS-DEV/GEOS#3256 @victorapm
Could you please verify the output of `ls /projets/schnaps/GEOS/thirdPartyLibs/install-a100-release`?

Also, could you please verify the lines 31-33 of `GEOS/host-configs/tpls.cmake`. Do they look like the following?

``` cmake
if(EXISTS ${GEOS_TPL_DIR}/conduit)
  set(CONDUIT_DIR ${GEOS_TPL_DIR}/conduit CACHE PATH "" FORCE)
endif()
```
