### 2024-08-13 GEOS-DEV/GEOS#3256 @timokoch
hi @berenger-eu is the file `GEOS/host-configs/tpls.cmake` included in your host config (`include(${CMAKE_CURRENT_LIST_DIR}/tpls.cmake)`)? I'm asking because it isn't there is the file you posted above, but I assume it probably has to be, @victorapm? It's easily missed in the docs (https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/QuickStart.html#compilation). Could be highlighted with a sentence in the docs in case this is essential.
