### 2024-08-13 GEOS-DEV/GEOS#3256 @victorapm
Good point, @timokoch, we will highlight that in the documentation. Thanks!

@berenger-eu, could you please add the include above to your `host-configs/your-platform.cmake`?

In addition:

- Make sure you turn on hypre with: `set(ENABLE_HYPRE ON CACHE BOOL "" FORCE )`
- Since you are compiling for A100s, you can use `set(CUDA_ARCH sm_80 CACHE STRING "")` instead of 75
- If you find any problems when compiling hypre for GPUs, feel free to open a GitHub issue [here](https://github.com/hypre-space/hypre/issues)
