### 2024-08-13 f3eb4dc
chore: Changing GEOSX to GEOS in external mesh import (#3278)
* Changing GEOSX to GEOS in external mesh import

* Update .integrated_tests.yaml

* Update BASELINE_NOTES.md
