### 2024-08-13 GEOS-DEV/GEOS#3256 @rrsettgast
@victorapm I think we need a single dedicated template hostconfig rather than having folks try to template off one of ours that are typically using some sort of file hierarchy.
