### 2024-08-13 GEOS-DEV/GEOS#3256 @berenger-eu
So `include(${CMAKE_CURRENT_LIST_DIR}/tpls.cmake` was missing. But I added it. I added a hard path to the file, otherwise third party libs was complaining the file does not exist (since it is in GEOS).
@victorapm the lines 31-33 of GEOS/host-configs/tpls.cmake are correct.

I am progressing.
I run `python scripts/config-build.py -hc host-configs/plafrim/a100.cmake -bt Release -DGEOS_TPL_DIR=/projets/schnaps/GEOS/thirdPartyLibs/build-a100-release/` and now I get:
```
-- CONDUIT_DIR = /projets/schnaps/GEOS/thirdPartyLibs/build-a100-release/conduit
CMake Error at cmake/thirdparty/SetupGeosxThirdParty.cmake:181 (find_package):
  Could not find a package configuration file provided by "Conduit" with any
  of the following names:

    ConduitConfig.cmake
    conduit-config.cmake

  Add the installation prefix of "Conduit" to CMAKE_PREFIX_PATH or set
  "Conduit_DIR" to a directory containing one of the above files.  If
  "Conduit" provides a separate development package or SDK, be sure it has
  been installed.
Call Stack (most recent call first):
  cmake/CMakeBasics.cmake:49 (include)
  CMakeLists.txt:72 (include)

-- Configuring incomplete, errors occurred!
```

And:
```
GEOS$ ls ../thirdPartyLibs/build-a100-release/conduit/
src  tmp
```
