### 2024-08-13 GEOS-DEV/GEOS#3256 @victorapm
@berenger-eu Could you try using `install-a100-release` instead of `build-a100-release`
