### 2024-08-14 GEOS-DEV/GEOS#3256 @berenger-eu
@victorapm thanks, it is currently compiling.

By the way, in the quickstart, there is this line `python scripts/config-build.py -hc host-configs/your-platform.cmake -bt Release -D GEOS_TPL_DIR=/full/path/to/thirdPartyLibs` I feel that it could be nice to specify that the GEOS_TPL_DIR is actually the install dir (so not the top level thirdPartyLibs dir, that I was expecting because I thought it will retrieve the right subdirectory based on my hostfile, and not the build dir neither).
Also, I never did make install in the thirdPartyLibs build dir, so I guess it installed it automatically (could be nice to say it after the
```
cd thirdPartyLibs
python scripts/config-build.py -hc ../GEOS/host-configs/your-platform.cmake -bt Release
cd build-your-platform-release
make
```
commands)
