### 2024-08-14 GEOS-DEV/GEOS#3256 @berenger-eu
The compilation did not finish:
```
sirocco21:build-a100-release$ make
[  0%] Generating version information
[  0%] Built target generate_version
[  0%] Built target hdf5_interface
[  0%] Built target fast_float
[  0%] Built target lvarray
[  1%] Built target events
[  3%] Built target fileIO
[  3%] Built target finiteElement
[  3%] Built target denseLinearAlgebra
[  8%] Built target mesh
[ 17%] Built target constitutive
[ 18%] Built target functions
[ 19%] Built target schema
[ 20%] Built target dataRepository
[ 20%] Built target codingUtilities
[ 21%] Built target common
[ 21%] Building CUDA object coreComponents/linearAlgebra/CMakeFiles/linearAlgebra.dir/interfaces/hypre/HypreMatrix.cpp.o
/.projets_7/schnaps/GEOS/GEOS/src/coreComponents/linearAlgebra/interfaces/hypre/HypreMatrix.cpp(1052): error: argument of type "HYPRE_BigInt *" is incompatible with parameter of type "HYPRE_Int *"
    do { auto const ierr = hypre_IJMatrixGetValuesParCSR( m_ij_mat, -1, &numEntries, &row, hypre::toHypreBigInt( colIndices ), values ); do { if( ierr != 0 ) { std::ostringstream __oss; __oss << "***** ERROR\n"; __oss << "***** LOCATION: " "/.projets_7/schnaps/GEOS/GEOS/src/coreComponents/linearAlgebra/interfaces/hypre/HypreMatrix.cpp" ":" "1052" "\n"; __oss << "***** Controlling expression (should be false): " "ierr != 0" "\n"; __oss << "***** Rank " << ::geos::logger::internal::rankString << ": " << "Error in call to " << "hypre_IJMatrixGetValuesParCSR( m_ij_mat, -1, &numEntries, &row, hypre::toHypreBigInt( colIndices ), values )" << "\n" << "Expected " << "ierr" << " " << "==" << " " << "0" << "\n" << "  " << "ierr" << " = " << ierr << "\n" << "  " << "0" << " = " << 0 << "\n" << "\n"; __oss << LvArray::system::stackTrace( true ); std::cout << __oss.str() << std::endl; LvArray::system::callErrorHandler(); } } while( false ); } while( false )
                                                                                           ^

/.projets_7/schnaps/GEOS/GEOS/src/coreComponents/linearAlgebra/interfaces/hypre/HypreMatrix.cpp(1052): error: no suitable conversion function from "const LvArray::ArraySlice<geos::real64, 1, 0, geos::localIndex>" to "HYPRE_BigInt *" exists
    do { auto const ierr = hypre_IJMatrixGetValuesParCSR( m_ij_mat, -1, &numEntries, &row, hypre::toHypreBigInt( colIndices ), values ); do { if( ierr != 0 ) { std::ostringstream __oss; __oss << "***** ERROR\n"; __oss << "***** LOCATION: " "/.projets_7/schnaps/GEOS/GEOS/src/coreComponents/linearAlgebra/interfaces/hypre/HypreMatrix.cpp" ":" "1052" "\n"; __oss << "***** Controlling expression (should be false): " "ierr != 0" "\n"; __oss << "***** Rank " << ::geos::logger::internal::rankString << ": " << "Error in call to " << "hypre_IJMatrixGetValuesParCSR( m_ij_mat, -1, &numEntries, &row, hypre::toHypreBigInt( colIndices ), values )" << "\n" << "Expected " << "ierr" << " " << "==" << " " << "0" << "\n" << "  " << "ierr" << " = " << ierr << "\n" << "  " << "0" << " = " << 0 << "\n" << "\n"; __oss << LvArray::system::stackTrace( true ); std::cout << __oss.str() << std::endl; LvArray::system::callErrorHandler(); } } while( false ); } while( false )
                                                                                                                               ^

/.projets_7/schnaps/GEOS/GEOS/src/coreComponents/linearAlgebra/interfaces/hypre/HypreMatrix.cpp(1052): error: too few arguments in function call
    do { auto const ierr = hypre_IJMatrixGetValuesParCSR( m_ij_mat, -1, &numEntries, &row, hypre::toHypreBigInt( colIndices ), values ); do { if( ierr != 0 ) { std::ostringstream __oss; __oss << "***** ERROR\n"; __oss << "***** LOCATION: " "/.projets_7/schnaps/GEOS/GEOS/src/coreComponents/linearAlgebra/interfaces/hypre/HypreMatrix.cpp" ":" "1052" "\n"; __oss << "***** Controlling expression (should be false): " "ierr != 0" "\n"; __oss << "***** Rank " << ::geos::logger::internal::rankString << ": " << "Error in call to " << "hypre_IJMatrixGetValuesParCSR( m_ij_mat, -1, &numEntries, &row, hypre::toHypreBigInt( colIndices ), values )" << "\n" << "Expected " << "ierr" << " " << "==" << " " << "0" << "\n" << "  " << "ierr" << " = " << ierr << "\n" << "  " << "0" << " = " << 0 << "\n" << "\n"; __oss << LvArray::system::stackTrace( true ); std::cout << __oss.str() << std::endl; LvArray::system::callErrorHandler(); } } while( false ); } while( false )
                                                                                                                                      ^

3 errors detected in the compilation of "/.projets_7/schnaps/GEOS/GEOS/src/coreComponents/linearAlgebra/interfaces/hypre/HypreMatrix.cpp".
make[2]: *** [coreComponents/linearAlgebra/CMakeFiles/linearAlgebra.dir/interfaces/hypre/HypreMatrix.cpp.o] Erreur 2
make[1]: *** [coreComponents/linearAlgebra/CMakeFiles/linearAlgebra.dir/all] Erreur 2
make: *** [all] Erreur 2
sirocco21:build-a100-release$ module li
Currently Loaded Modulefiles:
  1) compiler/cuda/12.3        2) compiler/gcc/11.2.0       3) mpi/openmpi/4.0.1         4) build/cmake/3.27.0        5) linalg/mkl/2020_update4
  ```
