### 2024-08-14 GEOS-DEV/GEOS#3256 @victorapm
@berenger-eu, thank you for the recommendation. We will implement the changes in the documentation.

Regarding your compilation error, it seems you are using an old version of GEOS. This is how that call should look like:

https://github.com/GEOS-DEV/GEOS/blob/f3eb4dccc4f27988adaffb6025460e9fc4ae883e/src/coreComponents/linearAlgebra/interfaces/hypre/HypreMatrix.cpp#L1052-L1059

Could you please do the following:
```
cd /.projets_7/schnaps/GEOS/GEOS/build-a100-release
git pull
make -j geosx
```
