### 2024-08-16 GEOS-DEV/GEOS#3256 @berenger-eu
Unfortunately, it does not work. It fails to buid `ElasticFirstOrderWaveEquationSEM.cpp.o`
```
$ VERBOSE=1 make -j geosx
[... everything fine here]
[ 34%] Building CUDA object coreComponents/physicsSolvers/CMakeFiles/physicsSolvers.dir/wavePropagation/sem/elastic/firstOrderEqn/isotropic/ElasticFirstOrderWaveEquationSEM.cpp.o
cd /.projets_7/schnaps/GEOS/GEOS/build-a100-release/coreComponents/physicsSolvers && /usr/local/cuda-12.3/bin/nvcc -forward-unknown-to-host-compiler -ccbin=g++ -DCAMP_HAVE_CUDA -DDIY_NO_THREADS -DFMT_HEADER_ONLY=1 -DH5_BUILT_AS_DYNAMIC_LIB -DMPICH_SKIP_MPICXX -DMPI_NO_CPPBIND -DOMPI_SKIP_MPICXX -D_FILE_OFFSET_BITS=64 -D_GNU_SOURCE -D_LARGEFILE64_SOURCE -D_LARGEFILE_SOURCE -D_MPICC_H -D_POSIX_C_SOURCE=200809L -Dkiss_fft_scalar=double --options-file CMakeFiles/physicsSolvers.dir/includes_CUDA.rsp -restrict -arch sm_75 --expt-relaxed-constexpr --expt-extended-lambda -Werror cross-execution-space-call,reorder,deprecated-declarations -O3 -DNDEBUG -Xcompiler -DNDEBUG -Xcompiler -O3 -std=c++17 -Xcompiler=-fPIC -Xcompiler=-fopenmp -Xcompiler -pthread -Xcompiler=-pthread -MD -MT coreComponents/physicsSolvers/CMakeFiles/physicsSolvers.dir/wavePropagation/sem/elastic/firstOrderEqn/isotropic/ElasticFirstOrderWaveEquationSEM.cpp.o -MF CMakeFiles/physicsSolvers.dir/wavePropagation/sem/elastic/firstOrderEqn/isotropic/ElasticFirstOrderWaveEquationSEM.cpp.o.d -x cu -c /.projets_7/schnaps/GEOS/GEOS/src/coreComponents/physicsSolvers/wavePropagation/sem/elastic/firstOrderEqn/isotropic/ElasticFirstOrderWaveEquationSEM.cpp -o CMakeFiles/physicsSolvers.dir/wavePropagation/sem/elastic/firstOrderEqn/isotropic/ElasticFirstOrderWaveEquationSEM.cpp.o
nvcc error   : 'cicc' died due to signal 11 (Invalid memory reference)
make[3]: *** [coreComponents/physicsSolvers/CMakeFiles/physicsSolvers.dir/wavePropagation/sem/elastic/firstOrderEqn/isotropic/ElasticFirstOrderWaveEquationSEM.cpp.o] Erreur 11
make[3] : on quitte le répertoire « /.projets_7/schnaps/GEOS/GEOS/build-a100-release »
make[2]: *** [coreComponents/physicsSolvers/CMakeFiles/physicsSolvers.dir/all] Erreur 2
make[2] : on quitte le répertoire « /.projets_7/schnaps/GEOS/GEOS/build-a100-release »
make[1]: *** [CMakeFiles/geosx.dir/rule] Erreur 2
make[1] : on quitte le répertoire « /.projets_7/schnaps/GEOS/GEOS/build-a100-release »
make: *** [geosx] Erreur 2
sirocco25:build-a100-release$ nvcc --version
nvcc: NVIDIA (R) Cuda compiler driver
Copyright (c) 2005-2023 NVIDIA Corporation
Built on Wed_Nov_22_10:17:15_PST_2023
Cuda compilation tools, release 12.3, V12.3.107
Build cuda_12.3.r12.3/compiler.33567101_0
```
I feel like it is an internal compiler error.... Maybe I should try another nvcc. What version do you use?
