### 2024-08-16 GEOS-DEV/GEOS#3256 @rrsettgast
@berenger-eu This is an internal compiler error. We have seen this in our RockyLinux CI builds which are currently not enabled. I think we can put this up on the priority list. @CusiniM can you confirm that this is the same error we saw?

Alternatively, if you can try cuda 11.8, this error has not come up in our testing.
