### 2024-08-16 GEOS-DEV/GEOS#3256 @CusiniM
> @berenger-eu This is an internal compiler error. We have seen this in our RockyLinux CI builds which are currently not enabled. I think we can put this up on the priority list. @CusiniM can you confirm that this is the same error we saw?
>
> Alternatively, if you can try cuda 11.8, this error has not come up in our testing.

yes, it's the same exact error we see on the rockylinux builds with `cuda12.5`. I have not seen that error with `cuda12.2` nor with previous versions.
