### 2024-08-17 GEOS-DEV/GEOS#3256 @berenger-eu
Ok, I will recompile everything with NVCC 11.7, I hope the host compiler has nothing to do with that.
