### 2024-08-18 GEOS-DEV/GEOS#3256 @berenger-eu
Ok, I compiled it, and it worked. I was able to run some tutorials. So I will close this issue. As a last remark, it could be great if you were sharing an archive (like a .tar.gz) that would contain the snapshot of the two git repo and allow getting everything in one file, like a release (especially, users might not be interested in having all the history of the repo, etc.). I will end my review on JOSS. Thanks for the support. Good work.
