### 2024-08-18 #6973 @berenger-eu
@rrsettgast as it has been stated by another reviewer, it would be great to add some references to competitors (i.e., other software doing related things). If this is updated, please could you regenerate the pdf? And I will check the last tick for my review. Thanks.
