### 2024-08-19 GEOS-DEV/GEOS#3256 @rrsettgast
> Ok, I compiled it, and it worked. I was able to run some tutorials. So I will close this issue. As a last remark, it could be great if you were sharing an archive (like a .tar.gz) that would contain the snapshot of the two git repo and allow getting everything in one file, like a release (especially, users might not be interested in having all the history of the repo, etc.). I will end my review on JOSS. Thanks for the support. Good work.

Thanks for your patience and persistence. We are discussing a way to link the two repos. In reality we are supposed to be using Spack with a single repository, but that hasn't gone so smoothly, so we are sort of in limbo with the TPL solution that nobody was ever supposed to see! We will come up with a strategy for ensuring that the main GEOS repo and the TPL's are in sync when a build is attempted.
