### 2024-08-20 #6973 @MakisH
@rrsettgast Some comments on the text (I still need to build and run the code):

- The summary is short and to the point. The comment about the complete rewrite nicely addresses the requirement to not publish the same research in different journals.
	- Note: I have not checked whether this is indeed a complete rewrite.
- The statement of need seems complete to me in the latest version. The cited tools are specific to the application discussed.
	- The list does not cover general coupling tools that have been used also for similar applications (e.g., preCICE -which I am involved into- has been used to couple the already mentioned DuMux for porous media applications, or PFLOTRAN for geothermal applications). But I think these is no need to expand the list besides application-specific examples.
	- The comparison to other tools is rather short. I have not checked whether the cited tools do not fulfill the respective criteria (lines 47-49).
- The GEOS Components section is clear. I understand that this is a complete framework that provides physical models and the coupling of them, similarly to MOOSE, for example. For applying GEOS to a new kind of application, I would need to write the respective physics code in either a stand-alone component that is then coupled with other components (partitioned / split-operator method), or write the fully monolithic model.
- The Applications section provides data for the statements in lines 47-49 of the statement of need. However, I am wondering whether the emphasis on "exascale" in line 49 is supported enough. On lines 94-95, there is a statement on the number of MPI ranks and degrees-of-freedom, with "close to 1/2 of Frontier" next to it. I think this needs further explanation, at least pointing out to the memory and not to the CPU capabilities.

Some typos and minor comments:

- Line 33: GHG is undefined (inconsistent with line 31)
- Line 45: "OMP" should be "OPM"
- Line 54: "mesh based" should be "mesh-based"

I have also added some notes on my checklist (cc @lucydot).

I have opened one PR to the documentation (https://github.com/GEOS-DEV/GEOS/pull/3289), I might open more at the end.
