### 2024-08-20 #6973 @lucydot
Thanks for your thorough review so far @MakisH.
I've noted your notes!

@rrsettgast, in case you missed it -  there are three open issues to address listed under points in [@MakisH's checklist](https://github.com/openjournals/joss-reviews/issues/6973#issuecomment-2268259113).
