### 2024-08-20 GEOS-DEV/GEOS#3291 @MakisH
Suggestions for the Quick Start Guide (JOSS review)
Some comments for the [QuickStart guide](https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/QuickStart.html), triggered by the [JOSS paper review](https://github.com/openjournals/joss-reviews/issues/6973):

- [ ] It sounds strange that for a QuickStart I have to provide a custom host config. I guess that the first time someone would try out GEOS, they would try on their laptop and expect to just run `cmake <path>` with a reasonable minimum config. Providing such a minimal host config for a popular setup (e.g., an ubuntu laptop) would help. Similarly, if many dependencies are disabled in the Quickstart, maybe guide the user to not build them either.
- [ ] In the example host config, the `GEOS_TPL_DIR` example could point to the already defined structure of `codes/thirdPartyLibs` and the next step could suggest a specific build directory.
- [ ] Before everything, it would be helpful to describe that there will be automatically generated `build-platform-target` and `install-platform-targer` directories, so that I don't need to start thinking about an installation prefix, for example.
- [ ] In any case, you could also point to your CI workflow configs for an example.
- [ ] In the prerequisites page, you ask for CMake 3.17, but the thirdPartyLibs needs 3.23.1 or higher.
- [ ] After cloning the thirdPartyLibs repository, the `git pull` should not be necessary.
