### 2024-08-21 #6973 @lucydot
> I am happy to try anything specific.

@MakisH thanks for your clear updates. I encourage you to confirm the main functional claims of the software. This might correspond to the four core tutorials, I'd probably play around with a few of the basic/advanced examples also depending on your capacity. There is so much on the website (in a good way!) that its certainly not possible to run through it all.

A quick note that I am on leave for the following week and will be back at desk on 2nd Sept.
