### 2024-08-26 #6973 @rrsettgast
@MakisH
The Frontier results are indeed previously unpublished. We will be publishing results with more detail at a later date. However, that publication would focus more on the hypre solution strategy, and reference this paper ideally.

I have added more details about Frontier, and modified the discussion... but the runs themselves would be difficult to verify as Frontier is not accessible to the general public.

@lucydot Do you have guidance on whether some results can stay in this publication without an in-depth discussion? Perhaps a statement indicating that the results are presented simply to show execution on a large GPU system?
