### 2024-08-26 #6973 @rrsettgast
@MakisH
Regarding your comments about the lack of a C++ api. Does this mean that you don't consider the Doxygen documentation as a valid API documentation? Or was it just to hard to get to? For instance, the main data structure object `Group` is documented here:
https://geosx-geosx.readthedocs-hosted.com/en/latest/doxygen_output/html/classgeos_1_1data_repository_1_1_group.html

Which I would think of as API documentation for that class. But there is no top level description with targets entry points to the doxygen documentation in the develop guide. Perhaps This is what is missing for you?
