### 2024-08-26 #6973 @rrsettgast
> @rrsettgast
>
> > Do you have guidance on whether some results can stay in this publication without an in-depth discussion? Perhaps a statement indicating that the results are presented simply to show execution on a large GPU system?
>
> The important point here is `Performance: If there are any performance claims of the software, have they been confirmed?`. As you state, these cannot be confirmed as our reviewers do not have access to a suitable system. However it should be possible to include details on how to reproduce these results _if_ someone did have access to the system. Ideally there would be a reference to the pre-print or similar which dives into the performance claims, but it sounds like that is not possible at this point. As you mention, we want to avoid in-depth discussion in the paper itself. Is it possible to include details within a benchmarking/performance section of the online documentation?

Ah. Yes. I certainly can include some guidance to reproduce in the documentation. Thanks!
