### 2024-08-28 #6973 @MakisH
> @MakisH Regarding your comments about the lack of a C++ api. Does this mean that you don't consider the Doxygen documentation as a valid API documentation? Or was it just to hard to get to? For instance, the main data structure object `Group` is documented here: https://geosx-geosx.readthedocs-hosted.com/en/latest/doxygen_output/html/classgeos_1_1data_repository_1_1_group.html
>
> Which I would think of as API documentation for that class. But there is no top level description with targeted entry points to the doxygen documentation in the develop guide. Perhaps This is what is missing for you?

@rrsettgast the Doxygen pages are of course great as API documentation, but the webpage does not guide to the specific user-facing classes. As an external, first-time observer, I would need to guess that `geos::dataRepository::Group` is the main entry point. As a user, I primarily need that, and not any classes not exposed to the user.

Just adding such a summary on the webpage of where one should look at in the Doxygen pages would solve the issue.

P.S.: I will run a few more tutorials/examples asap.
