### 2024-08-29 #6973 @MakisH
I ran the following cases:

- [Tutorial 1](https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/tutorials/step01/Tutorial.html) (see [above](https://github.com/openjournals/joss-reviews/issues/6973#issuecomment-2299683199))
- [Tutorial 2](https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/tutorials/step02/Tutorial.html)
- [Tutorial 3](https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/tutorials/step03/Tutorial.html)
- [CO2 Injection Example](https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/basicExamples/co2Injection/Example.html)

With an external view (I am not an expert in this type of simulations), I can see that the software indeed simulates problems relevant to CO2 storage and generally injection of fluids into wells.

Tutorial 2 (with two meshes):

![tutorial2hex](https://github.com/user-attachments/assets/7d773f5b-6cc9-467d-9a37-a95764b4f9d2)

![tutorial2tet](https://github.com/user-attachments/assets/a5104e81-11f5-485a-9bd5-654bba3dd6b0)

Tutorial 3:

![tutorial3](https://github.com/user-attachments/assets/07a2c4e7-b590-420d-8230-0e04cba4e282)

CO2 injection:

![co2-injection](https://github.com/user-attachments/assets/e8cbf323-6867-461c-97ea-57ece1ad8277)

Related PR: https://github.com/GEOS-DEV/GEOS/pull/3315
