### 2024-09-04 #6973 @MakisH
From my side, all points in the checklist are covered. Thank you for all the improvements!

I can confirm that at least the software seems to address the applications claimed (based on the documentation and examples, I am not working in this application field), which are indeed multi-physics applications. I can confirm that one can build and run the software with the provided documentation, and that there are sufficient pointers to getting support and to contributing.

The performance claims seem to be sufficiently documented: if someone had access to Frontier or a similar system, it is now clear which example to run and how. I did not run this example. However, I also don't think that JOSS is the right place to discuss such results, but including the results is appropriate, since the software emphasizes its suitability for HPC.

___

Looking at the paper again, I would like to draw the attention to the title:

> GEOS-2024: A portable multi-physics simulation framework

I think that the "multi-physics simulation framework" is clear (it is a code framework in which one can set up multi-physics simulations). However, looking at lines 47-49:

> However, GEOS stands out in two key areas: the explicit fault modeling coupled
with flow and mechanical deformation, and the focus on performance portability on platforms
ranging from workstations to exascale supercomputers.

- The "portable" in the title is not immediately clear. You are focusing on performance-portability, but this could also mean portable execution to different platforms (macOS, Windows, ...). This might be the case, but it is not yet addressed explicitly in the paper.
- While in principle the software seems to be general-purpose (as every framework, provided that the user codes the respective models), in the paper there is a lot of focus on porous media applications, and the project is compared to other porous media frameworks. This aspect is missing from the title, and could be very helpful to someone first encountering this paper.
