### 2024-09-04 #6973 @rrsettgast
@MakisH Thank you again for spending your time on this. I understand your remaining points.

- I will make a modification to the paper regarding the 'portable'. I think we actually strive for both "portable" and "performance portable"...except for windows...we haven't spent much time on windows portability 😁

- It is true that the emphasis is currently on porous media, but we are currently the application space to things like batteries...but perhaps that should be addressed with a follow on publication. It seems appropriate to emphasize the current application space in the title, but I was trying to avoid a very long title. I think it would be something like:

`GEOS-2024: A performance portable multi-physics simulation framework for subsurface applications `

@lucydot Are we allowed to modify the title?
