### 2024-09-05 #6973 @timokoch
@rrsettgast Thanks for all the great updates. I fully support @MakisH comments/suggestions and I like your title suggestion since it more clearly states the current focus. I now completed the tutorials (it was fun, sorry that it took some time) and also successfully ran some examples. The tool is clearly useful, seems performant, is easy to configure, and covers a broad range of applications and models within the subsurface scope.

Some comments below
* all tests passed (macOS sonoma / arm / apple clang 15)
* mpirun -np 2 ./geosx --help does not call MPI to finalize (would be good to always call this)
* Tutorial 1: discretization: "To know the list of admissible values of an attribute, please see GEOS’s XML schema." where can I find this? I link here would be nice.
* https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/tutorials/step01/Tutorial.html#single-phase-solver
here it should be mentioned in which file I can find the configuration. Since the command uses the _smoke version one has to first find that that one includes the base version.
Also a link to the file on GitHub would be nice so it can be viewed in another browser window as an alternative to looking at it in the terminal/or a text editor.
* https://geosx-geosx.readthedocs-hosted.com/en/latest/docs/sphinx/tutorials/step01/Tutorial.html#numerical-methods
I didn't understand the numerical methods section. Which additional information is supplied here?
The text mentions "fieldName" and "coefficientName" attributes that don't appear in the xml file.
* Tutorials were fun and are well described and give many clues about what could be changed and how
* Tutorial 3: "GEOS handles permeability as a diagonal matrix, so the three values of the permeability tensor are set individually using the component field." What if my permeability is a full (symmetric) tensor?
* Running a quick test on a laptop showed nice scaling even for the relatively simple tutorial 3 with runtime almost halving by doubling processors (of course not a super accurate performance test but it indicates that there is not much communication overhead)
* The section with benchmarks and verification problems is impressive, comprehensive, and convincing and demonstrates the maturity of the tool. The tests provide Python scripts e.g. generate comparisons between analytical solutions and numerical data
* I computed, for example, the Mandel problem benchmark and could reproduce the exact figure which is very nice.
(The only problem with the Python script was that relative paths to the input files are hard coded. But when you download them from the doc pages it is not clear where to put them that the relative path is correct. Even better would be to make the user supply the path to the `GEOS/inputFiles` directory.)
* The documentation regarding the physics and tuning of solvers is also very comprehensive
* I couldn't test on various hardware. But given that the code is based on portable libraries (e.g. RAJA), I believe the claim.
