### 2024-09-05 #6973 @lucydot
@rrsettgast

> Are we allowed to modify the title?

Yes, that should be no problem. Once you update the title on your paper let me know and I will update the issue title.

It seems you have a few very minor points to address above from @timokoch. I can see that all points in checkboxes are ✅  apart from one regarding references (@timokoch ?).

Once this is done, we can proceed to final stages of the review process.

I want to give a big thank you to @MakisH, @timokoch, @berenger-eu for your time and expertise. JOSS could not exist without your generosity!
