### 2024-09-05 #6973 @rrsettgast
@timokoch I the following list of your remaining points in https://github.com/GEOS-DEV/GEOS/pull/3332
