### 2024-09-09 #6973 @lucydot
I can see some action in https://github.com/GEOS-DEV/GEOS/pull/3332 - can you confirm you are happy with changes made, and recommend acceptance @timokoch ?
