### 2024-09-09 #6973 @timokoch
> I can see some action in [GEOS-DEV/GEOS#3332](https://github.com/GEOS-DEV/GEOS/pull/3332) - can you confirm you are happy with changes made, and recommend acceptance @timokoch ?

@lucydot Yes everything has been thoroughly addressed and I can recommend acceptance :)
