### 2024-09-09 #6973 @rrsettgast
> Some small edits for the paper @rrsettgast :
>
> * line 46: `, Darts` ---> `and Darts`
> * line 71:  `(“CHAI,” 2023)` The formatting of the title doesn't look correct, remove quotes?
> * line 73: provide an inline link to the performance portable website if required, rather than link via a reference. As a general rule, we try to only give references when there is a unique and persistent identifier (such as DOI).

done
