### 2024-09-16 #6973 @lucydot
> Do you have guidance on what you would like me to do regarding this difference?

Reading our editors Slack channel, I can see that there are previous examples where author list on the Zenodo and JOSS paper differ, for the reason you give above. So this is fine 👍
