### 2024-09-16 #6973 @lucydot
@rrsettgast I see different licenses listed on the repository (GNU LPGL) and Zenodo (CC-BY). Is this intentional? We usually expect them to be the same.
