### 2024-09-16 #6973 @rrsettgast
> @rrsettgast I understand this may take some time given the length of the author list, but please could you include Orcid IDs where they are available.

Yes. I will get the ones that are available..but I don't think all the authors have an Orcid ID. I assume this is ok?
