### 2024-09-16 #6973 @lucydot
> but I don't think all the authors have an Orcid ID. I assume this is ok?

@rrsettgast, yes - that's fine
