### 2024-09-24 #6973 @rrsettgast
> @rrsettgast  - the title on the paper and the zenodo and the paper archive also need to match - could you update the title on Zenodo?

Hello. I'm not sure I understand. Do I need to change the title of the zenodo archive release to match the title of the joss paper?
