### 2024-09-25 #6973 @rrsettgast
> @lucydot OK. I changed the title on zenodo. Also...the paper draft says it is under Creative Commons again...let me fix it...wait...there is nowhere this is listed in the paper.md file. Can you change it somewhere on your end?

@lucydot My mistake. I read the CC4 license statement on the draft as the license for the code, but it is the license of the paper itself, which is the same on all JOSS publications. So I took care of the title of the zenodo release, and so we should good?

Thanks again.
