### 2024-09-30 #6973 @lucydot
> and so we should be good?

Yes, we are all good now. Thanks for making those final few changes. I'm going to recommend acceptance :)

I'd like to thank you again @MakisH @timokoch @berenger-e for your excellent reviewing. This was a relatively complex piece of software to review, but your expertise made the whole process smooth. Thank you!
