### 2024-10-01 #6973 @rrsettgast
Hello @lucydot @MakisH @timokoch @berenger-eu ,
Thank you for spending your time reviewing this submission. I was very impressed with the level of detail given by the review process. The review comments/suggestions have lead to many improvements and we are left with substantively better documentation as a result. The entire GEOS team thanks you for your efforts.
