### 2024-10-01 #6973 @rrsettgast
@lucydot Do I send the bot the accept message, or do you?
