### 2024-10-01 #6973 @rrsettgast
@lucydot Do I send the accept message to the bot, or do you?
