### 2024-10-02 #6973 @rrsettgast
> Hi @rrsettgast  - there is nothing for either of us to do at this stage. The Editor-in-Chief for this track has been alerted and will go through their process before final acceptance.

Hi @lucydot. Ah. That makes sense. Thanks again!!!
