### 2024-10-10 #6973 @arfon
AUTHOR – your paper is now accepted and published in JOSS :zap::rocket::boom:
